Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Re-exports

pub use js_sys;

pub use wasm_bindgen;

Modules

console

Structs

AnimationEvent

The AnimationEvent class.

Blob

The Blob class.

CharacterData

The CharacterData class.

ClipboardEvent

The ClipboardEvent class.

CompositionEvent

The CompositionEvent class.

DataTransfer

The DataTransfer class.

Document

The Document class.

DomRect

The DomRect class.

DomRectReadOnly

The DomRectReadOnly class.

DragEvent

The DragEvent class.

Element

The Element class.

Event

The Event class.

EventTarget

The EventTarget class.

File

The File class.

FileList

The FileList class.

FileReader

The FileReader class.

FocusEvent

The FocusEvent class.

HtmlElement

The HtmlElement class.

HtmlFormElement

The HtmlFormElement class.

HtmlInputElement

The HtmlInputElement class.

HtmlSelectElement

The HtmlSelectElement class.

HtmlTextAreaElement

The HtmlTextAreaElement class.

InputEvent

The InputEvent class.

KeyboardEvent

The KeyboardEvent class.

Location

The Location class.

MessageEvent

The MessageEvent class.

MouseEvent

The MouseEvent class.

Node

The Node class.

NodeList

The NodeList class.

PointerEvent

The PointerEvent class.

ScrollIntoViewOptions

The ScrollIntoViewOptions dictionary.

Text

The Text class.

Touch

The Touch class.

TouchEvent

The TouchEvent class.

TouchList

The TouchList class.

TransitionEvent

The TransitionEvent class.

UiEvent

The UiEvent class.

WebSocket

The WebSocket class.

WheelEvent

The WheelEvent class.

Window

The Window class.

Enums

ScrollBehavior

The ScrollBehavior enum.

ScrollLogicalPosition

The ScrollLogicalPosition enum.

Functions

window

Getter for the Window object