web_sys::css

Struct JsNamespaceCss

pub struct JsNamespaceCss { /* private fields */ }

Implementations

impl JsNamespaceCss

pub fn highlights() -> HighlightRegistry

Getter for the CSS.highlights field.

MDN Documentation

This API requires the following crate features to be activated: HighlightRegistry, css

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide