web_sys
Enum GpuFeatureName

#[non_exhaustive]

#[repr(u32)]
pub enum GpuFeatureName {

    DepthClipControl = 0,
    Depth32floatStencil8 = 1,
    TextureCompressionBc = 2,
    TextureCompressionBcSliced3d = 3,
    TextureCompressionEtc2 = 4,
    TextureCompressionAstc = 5,
    TextureCompressionAstcSliced3d = 6,
    TimestampQuery = 7,
    IndirectFirstInstance = 8,
    ShaderF16 = 9,
    Rg11b10ufloatRenderable = 10,
    Bgra8unormStorage = 11,
    Float32Filterable = 12,
    Float32Blendable = 13,
    ClipDistances = 14,
    DualSourceBlending = 15,

}

The GpuFeatureName enum.

This API requires the following crate features to be activated: GpuFeatureName

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

DepthClipControl = 0

Depth32floatStencil8 = 1

TextureCompressionBc = 2

TextureCompressionBcSliced3d = 3

TextureCompressionEtc2 = 4

TextureCompressionAstc = 5

TextureCompressionAstcSliced3d = 6

TimestampQuery = 7

IndirectFirstInstance = 8

ShaderF16 = 9

Rg11b10ufloatRenderable = 10

Bgra8unormStorage = 11

Float32Filterable = 12

Float32Blendable = 13

ClipDistances = 14

DualSourceBlending = 15

Implementations

impl GpuFeatureName

pub fn from_js_value(obj: &JsValue) -> Option<GpuFeatureName>