web_sys
Enum GpuVertexFormat

#[non_exhaustive]

#[repr(u32)]
pub enum GpuVertexFormat {

    Uint8 = 0,
    Uint8x2 = 1,
    Uint8x4 = 2,
    Sint8 = 3,
    Sint8x2 = 4,
    Sint8x4 = 5,
    Unorm8 = 6,
    Unorm8x2 = 7,
    Unorm8x4 = 8,
    Snorm8 = 9,
    Snorm8x2 = 10,
    Snorm8x4 = 11,
    Uint16 = 12,
    Uint16x2 = 13,
    Uint16x4 = 14,
    Sint16 = 15,
    Sint16x2 = 16,
    Sint16x4 = 17,
    Unorm16 = 18,
    Unorm16x2 = 19,
    Unorm16x4 = 20,
    Snorm16 = 21,
    Snorm16x2 = 22,
    Snorm16x4 = 23,
    Float16 = 24,
    Float16x2 = 25,
    Float16x4 = 26,
    Float32 = 27,
    Float32x2 = 28,
    Float32x3 = 29,
    Float32x4 = 30,
    Uint32 = 31,
    Uint32x2 = 32,
    Uint32x3 = 33,
    Uint32x4 = 34,
    Sint32 = 35,
    Sint32x2 = 36,
    Sint32x3 = 37,
    Sint32x4 = 38,
    Unorm1010102 = 39,
    Unorm8x4Bgra = 40,

}

The GpuVertexFormat enum.

This API requires the following crate features to be activated: GpuVertexFormat

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Uint8 = 0

Uint8x2 = 1

Uint8x4 = 2

Sint8 = 3

Sint8x2 = 4

Sint8x4 = 5

Unorm8 = 6

Unorm8x2 = 7

Unorm8x4 = 8

Snorm8 = 9

Snorm8x2 = 10

Snorm8x4 = 11

Uint16 = 12

Uint16x2 = 13

Uint16x4 = 14

Sint16 = 15

Sint16x2 = 16

Sint16x4 = 17

Unorm16 = 18

Unorm16x2 = 19

Unorm16x4 = 20

Snorm16 = 21

Snorm16x2 = 22

Snorm16x4 = 23

Float16 = 24

Float16x2 = 25

Float16x4 = 26

Float32 = 27

Float32x2 = 28

Float32x3 = 29

Float32x4 = 30

Uint32 = 31

Uint32x2 = 32

Uint32x3 = 33

Uint32x4 = 34

Sint32 = 35

Sint32x2 = 36

Sint32x3 = 37

Sint32x4 = 38

Unorm1010102 = 39

Unorm8x4Bgra = 40

Implementations

impl GpuVertexFormat

pub fn from_js_value(obj: &JsValue) -> Option<GpuVertexFormat>