web_sys
Enum XrHandJoint

#[non_exhaustive]

#[repr(u32)]
pub enum XrHandJoint {

    Wrist = 0,
    ThumbMetacarpal = 1,
    ThumbPhalanxProximal = 2,
    ThumbPhalanxDistal = 3,
    ThumbTip = 4,
    IndexFingerMetacarpal = 5,
    IndexFingerPhalanxProximal = 6,
    IndexFingerPhalanxIntermediate = 7,
    IndexFingerPhalanxDistal = 8,
    IndexFingerTip = 9,
    MiddleFingerMetacarpal = 10,
    MiddleFingerPhalanxProximal = 11,
    MiddleFingerPhalanxIntermediate = 12,
    MiddleFingerPhalanxDistal = 13,
    MiddleFingerTip = 14,
    RingFingerMetacarpal = 15,
    RingFingerPhalanxProximal = 16,
    RingFingerPhalanxIntermediate = 17,
    RingFingerPhalanxDistal = 18,
    RingFingerTip = 19,
    PinkyFingerMetacarpal = 20,
    PinkyFingerPhalanxProximal = 21,
    PinkyFingerPhalanxIntermediate = 22,
    PinkyFingerPhalanxDistal = 23,
    PinkyFingerTip = 24,

}

The XrHandJoint enum.

This API requires the following crate features to be activated: XrHandJoint

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Wrist = 0

ThumbMetacarpal = 1

ThumbPhalanxProximal = 2

ThumbPhalanxDistal = 3

ThumbTip = 4

IndexFingerMetacarpal = 5

IndexFingerPhalanxProximal = 6

IndexFingerPhalanxIntermediate = 7

IndexFingerPhalanxDistal = 8

IndexFingerTip = 9

MiddleFingerMetacarpal = 10

MiddleFingerPhalanxProximal = 11

MiddleFingerPhalanxIntermediate = 12

MiddleFingerPhalanxDistal = 13

MiddleFingerTip = 14

RingFingerMetacarpal = 15

RingFingerPhalanxProximal = 16

RingFingerPhalanxIntermediate = 17

RingFingerPhalanxDistal = 18

RingFingerTip = 19

PinkyFingerMetacarpal = 20

PinkyFingerPhalanxProximal = 21

PinkyFingerPhalanxIntermediate = 22

PinkyFingerPhalanxDistal = 23

PinkyFingerTip = 24

Implementations

impl XrHandJoint

pub fn from_js_value(obj: &JsValue) -> Option<XrHandJoint>