Struct web_sys::Animation

#[repr(transparent)]

pub struct Animation { /* fields omitted */ }

The Animation object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl Animation

pub fn new() -> Result<Animation, JsValue>

This function requires the following crate features to be activated: Animation

impl Animation

pub fn new_with_effect(
effect: Option<&AnimationEffect>
) -> Result<Animation, JsValue>

This function requires the following crate features to be activated: Animation, AnimationEffect

impl Animation

pub fn new_with_effect_and_timeline(
effect: Option<&AnimationEffect>,
timeline: Option<&AnimationTimeline>
) -> Result<Animation, JsValue>

This function requires the following crate features to be activated: Animation, AnimationEffect, AnimationTimeline

impl Animation

pub fn cancel(&self)

The cancel() method

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn finish(&self) -> Result<(), JsValue>

The finish() method

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn pause(&self) -> Result<(), JsValue>

The pause() method

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn play(&self) -> Result<(), JsValue>

The play() method

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn reverse(&self) -> Result<(), JsValue>

The reverse() method

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn update_playback_rate(&self, playback_rate: f64)

The updatePlaybackRate() method

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn id(&self) -> String

The id getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn set_id(&self, id: &str)

The id setter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn effect(&self) -> Option<AnimationEffect>

The effect getter

MDN Documentation

This function requires the following crate features to be activated: Animation, AnimationEffect

impl Animation

pub fn set_effect(&self, effect: Option<&AnimationEffect>)

The effect setter

MDN Documentation

This function requires the following crate features to be activated: Animation, AnimationEffect

impl Animation

pub fn timeline(&self) -> Option<AnimationTimeline>

The timeline getter

MDN Documentation

This function requires the following crate features to be activated: Animation, AnimationTimeline

impl Animation

pub fn set_timeline(&self, timeline: Option<&AnimationTimeline>)

The timeline setter

MDN Documentation

This function requires the following crate features to be activated: Animation, AnimationTimeline

impl Animation

pub fn start_time(&self) -> Option<f64>

The startTime getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn set_start_time(&self, start_time: Option<f64>)

The startTime setter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn current_time(&self) -> Option<f64>

The currentTime getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn set_current_time(&self, current_time: Option<f64>)

The currentTime setter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn playback_rate(&self) -> f64

The playbackRate getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn set_playback_rate(&self, playback_rate: f64)

The playbackRate setter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn play_state(&self) -> AnimationPlayState

The playState getter

MDN Documentation

This function requires the following crate features to be activated: Animation, AnimationPlayState

impl Animation

pub fn pending(&self) -> bool

The pending getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn ready(&self) -> Result<Promise, JsValue>

The ready getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn finished(&self) -> Result<Promise, JsValue>

The finished getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn onfinish(&self) -> Option<Function>

The onfinish getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn set_onfinish(&self, onfinish: Option<&Function>)

The onfinish setter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn oncancel(&self) -> Option<Function>

The oncancel getter

MDN Documentation

This function requires the following crate features to be activated: Animation

impl Animation

pub fn set_oncancel(&self, oncancel: Option<&Function>)

The oncancel setter

MDN Documentation

This function requires the following crate features to be activated: Animation