Struct web_sys::BaseAudioContext

#[repr(transparent)]

pub struct BaseAudioContext { /* fields omitted */ }

The BaseAudioContext object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl BaseAudioContext

pub fn create_analyser(&self) -> Result<AnalyserNode, JsValue>

The createAnalyser() method

MDN Documentation

This function requires the following crate features to be activated: AnalyserNode, BaseAudioContext

impl BaseAudioContext

pub fn create_biquad_filter(&self) -> Result<BiquadFilterNode, JsValue>

The createBiquadFilter() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, BiquadFilterNode

impl BaseAudioContext

pub fn create_buffer(
&self,
number_of_channels: u32,
length: u32,
sample_rate: f32
) -> Result<AudioBuffer, JsValue>

The createBuffer() method

MDN Documentation

This function requires the following crate features to be activated: AudioBuffer, BaseAudioContext

impl BaseAudioContext

pub fn create_buffer_source(&self) -> Result<AudioBufferSourceNode, JsValue>

The createBufferSource() method

MDN Documentation

This function requires the following crate features to be activated: AudioBufferSourceNode, BaseAudioContext

impl BaseAudioContext

pub fn create_channel_merger(&self) -> Result<ChannelMergerNode, JsValue>

The createChannelMerger() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ChannelMergerNode

impl BaseAudioContext

pub fn create_channel_merger_with_number_of_inputs(
&self,
number_of_inputs: u32
) -> Result<ChannelMergerNode, JsValue>

The createChannelMerger() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ChannelMergerNode

impl BaseAudioContext

pub fn create_channel_splitter(&self) -> Result<ChannelSplitterNode, JsValue>

The createChannelSplitter() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ChannelSplitterNode

impl BaseAudioContext

pub fn create_channel_splitter_with_number_of_outputs(
&self,
number_of_outputs: u32
) -> Result<ChannelSplitterNode, JsValue>

The createChannelSplitter() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ChannelSplitterNode

impl BaseAudioContext

pub fn create_constant_source(&self) -> Result<ConstantSourceNode, JsValue>

The createConstantSource() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ConstantSourceNode

impl BaseAudioContext

pub fn create_convolver(&self) -> Result<ConvolverNode, JsValue>

The createConvolver() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ConvolverNode

impl BaseAudioContext

pub fn create_delay(&self) -> Result<DelayNode, JsValue>

The createDelay() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, DelayNode

impl BaseAudioContext

pub fn create_delay_with_max_delay_time(
&self,
max_delay_time: f64
) -> Result<DelayNode, JsValue>

The createDelay() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, DelayNode

impl BaseAudioContext

pub fn create_dynamics_compressor(
&self
) -> Result<DynamicsCompressorNode, JsValue>

The createDynamicsCompressor() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, DynamicsCompressorNode

impl BaseAudioContext

pub fn create_gain(&self) -> Result<GainNode, JsValue>

The createGain() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, GainNode

impl BaseAudioContext

pub fn create_oscillator(&self) -> Result<OscillatorNode, JsValue>

The createOscillator() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, OscillatorNode

impl BaseAudioContext

pub fn create_panner(&self) -> Result<PannerNode, JsValue>

The createPanner() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, PannerNode

impl BaseAudioContext

pub fn create_periodic_wave(
&self,
real: &[f32],
imag: &[f32]
) -> Result<PeriodicWave, JsValue>

The createPeriodicWave() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, PeriodicWave

impl BaseAudioContext

pub fn create_periodic_wave_with_constraints(
&self,
real: &[f32],
imag: &[f32],
constraints: &PeriodicWaveConstraints
) -> Result<PeriodicWave, JsValue>

The createPeriodicWave() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, PeriodicWave, PeriodicWaveConstraints

impl BaseAudioContext

pub fn create_script_processor(&self) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ScriptProcessorNode

impl BaseAudioContext

pub fn create_script_processor_with_buffer_size(
&self,
buffer_size: u32
) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ScriptProcessorNode

impl BaseAudioContext

pub fn create_script_processor_with_buffer_size_and_number_of_input_channels(
&self,
buffer_size: u32,
number_of_input_channels: u32
) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ScriptProcessorNode

impl BaseAudioContext

pub fn create_script_processor_with_buffer_size_and_number_of_input_channels_and_number_of_output_channels(
&self,
buffer_size: u32,
number_of_input_channels: u32,
number_of_output_channels: u32
) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, ScriptProcessorNode

impl BaseAudioContext

pub fn create_stereo_panner(&self) -> Result<StereoPannerNode, JsValue>

The createStereoPanner() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, StereoPannerNode

impl BaseAudioContext

pub fn create_wave_shaper(&self) -> Result<WaveShaperNode, JsValue>

The createWaveShaper() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext, WaveShaperNode

impl BaseAudioContext

pub fn decode_audio_data(
&self,
audio_data: &ArrayBuffer
) -> Result<Promise, JsValue>

The decodeAudioData() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext

impl BaseAudioContext

pub fn decode_audio_data_with_success_callback(
&self,
audio_data: &ArrayBuffer,
success_callback: &Function
) -> Result<Promise, JsValue>

The decodeAudioData() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext

impl BaseAudioContext

pub fn decode_audio_data_with_success_callback_and_error_callback(
&self,
audio_data: &ArrayBuffer,
success_callback: &Function,
error_callback: &Function
) -> Result<Promise, JsValue>

The decodeAudioData() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext

impl BaseAudioContext

pub fn resume(&self) -> Result<Promise, JsValue>

The resume() method

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext

impl BaseAudioContext

pub fn destination(&self) -> AudioDestinationNode

The destination getter

MDN Documentation

This function requires the following crate features to be activated: AudioDestinationNode, BaseAudioContext

impl BaseAudioContext

pub fn sample_rate(&self) -> f32

The sampleRate getter

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext

impl BaseAudioContext

pub fn current_time(&self) -> f64

The currentTime getter

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext

impl BaseAudioContext

pub fn listener(&self) -> AudioListener

The listener getter

MDN Documentation

This function requires the following crate features to be activated: AudioListener, BaseAudioContext

impl BaseAudioContext

pub fn state(&self) -> AudioContextState

The state getter

MDN Documentation

This function requires the following crate features to be activated: AudioContextState, BaseAudioContext

impl BaseAudioContext

pub fn audio_worklet(&self) -> Result<AudioWorklet, JsValue>

The audioWorklet getter

MDN Documentation

This function requires the following crate features to be activated: AudioWorklet, BaseAudioContext

impl BaseAudioContext

pub fn onstatechange(&self) -> Option<Function>

The onstatechange getter

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext

impl BaseAudioContext

pub fn set_onstatechange(&self, onstatechange: Option<&Function>)

The onstatechange setter

MDN Documentation

This function requires the following crate features to be activated: BaseAudioContext