Struct web_sys::CanvasRenderingContext2d

pub struct CanvasRenderingContext2d { /* private fields */ }

The CanvasRenderingContext2d class.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

Implementations

impl CanvasRenderingContext2d

pub fn canvas(&self) -> Option<HtmlCanvasElement>

Getter for the canvas field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn global_alpha(&self) -> f64

Getter for the globalAlpha field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_global_alpha(&self, value: f64)

Setter for the globalAlpha field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn global_composite_operation(&self) -> Result<String, JsValue>

Getter for the globalCompositeOperation field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_global_composite_operation(&self, value: &str) -> Result<(), JsValue>

Setter for the globalCompositeOperation field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_style(&self) -> JsValue

Getter for the strokeStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_stroke_style(&self, value: &JsValue)

Setter for the strokeStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_style(&self) -> JsValue

Getter for the fillStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_fill_style(&self, value: &JsValue)

Setter for the fillStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn filter(&self) -> String

Getter for the filter field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_filter(&self, value: &str)

Setter for the filter field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn image_smoothing_enabled(&self) -> bool

Getter for the imageSmoothingEnabled field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_image_smoothing_enabled(&self, value: bool)

Setter for the imageSmoothingEnabled field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_width(&self) -> f64

Getter for the lineWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_width(&self, value: f64)

Setter for the lineWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_cap(&self) -> String

Getter for the lineCap field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_cap(&self, value: &str)

Setter for the lineCap field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_join(&self) -> String

Getter for the lineJoin field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_join(&self, value: &str)

Setter for the lineJoin field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn miter_limit(&self) -> f64

Getter for the miterLimit field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_miter_limit(&self, value: f64)

Setter for the miterLimit field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_dash_offset(&self) -> f64

Getter for the lineDashOffset field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_dash_offset(&self, value: f64)

Setter for the lineDashOffset field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_offset_x(&self) -> f64

Getter for the shadowOffsetX field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_offset_x(&self, value: f64)

Setter for the shadowOffsetX field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_offset_y(&self) -> f64

Getter for the shadowOffsetY field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_offset_y(&self, value: f64)

Setter for the shadowOffsetY field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_blur(&self) -> f64

Getter for the shadowBlur field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_blur(&self, value: f64)

Setter for the shadowBlur field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_color(&self) -> String

Getter for the shadowColor field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_color(&self, value: &str)

Setter for the shadowColor field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn font(&self) -> String

Getter for the font field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_font(&self, value: &str)

Setter for the font field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn text_align(&self) -> String

Getter for the textAlign field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_text_align(&self, value: &str)

Setter for the textAlign field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn text_baseline(&self) -> String

Getter for the textBaseline field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_text_baseline(&self, value: &str)

Setter for the textBaseline field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn draw_window( &self, window: &Window, x: f64, y: f64, w: f64, h: f64, bg_color: &str ) -> Result<(), JsValue>

The drawWindow() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Window

impl CanvasRenderingContext2d

pub fn draw_window_with_flags( &self, window: &Window, x: f64, y: f64, w: f64, h: f64, bg_color: &str, flags: u32 ) -> Result<(), JsValue>

The drawWindow() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Window

impl CanvasRenderingContext2d

pub fn draw_image_with_html_canvas_element( &self, image: &HtmlCanvasElement, dx: f64, dy: f64 ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_canvas_element_and_dw_and_dh( &self, image: &HtmlCanvasElement, dx: f64, dy: f64, dw: f64, dh: f64 ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_canvas_element_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &HtmlCanvasElement, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64 ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn begin_path(&self)

The beginPath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn clip(&self)

The clip() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill(&self)

The fill() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn is_point_in_path_with_f64(&self, x: f64, y: f64) -> bool

The isPointInPath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn is_point_in_stroke_with_x_and_y(&self, x: f64, y: f64) -> bool

The isPointInStroke() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke(&self)

The stroke() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn add_hit_region(&self) -> Result<(), JsValue>

The addHitRegion() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn clear_hit_regions(&self)

The clearHitRegions() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn remove_hit_region(&self, id: &str)

The removeHitRegion() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn arc( &self, x: f64, y: f64, radius: f64, start_angle: f64, end_angle: f64 ) -> Result<(), JsValue>

The arc() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn arc_with_anticlockwise( &self, x: f64, y: f64, radius: f64, start_angle: f64, end_angle: f64, anticlockwise: bool ) -> Result<(), JsValue>

The arc() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn arc_to( &self, x1: f64, y1: f64, x2: f64, y2: f64, radius: f64 ) -> Result<(), JsValue>

The arcTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn bezier_curve_to( &self, cp1x: f64, cp1y: f64, cp2x: f64, cp2y: f64, x: f64, y: f64 )

The bezierCurveTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn close_path(&self)

The closePath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn ellipse( &self, x: f64, y: f64, radius_x: f64, radius_y: f64, rotation: f64, start_angle: f64, end_angle: f64 ) -> Result<(), JsValue>

The ellipse() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn ellipse_with_anticlockwise( &self, x: f64, y: f64, radius_x: f64, radius_y: f64, rotation: f64, start_angle: f64, end_angle: f64, anticlockwise: bool ) -> Result<(), JsValue>

The ellipse() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_to(&self, x: f64, y: f64)

The lineTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn move_to(&self, x: f64, y: f64)

The moveTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn quadratic_curve_to(&self, cpx: f64, cpy: f64, x: f64, y: f64)

The quadraticCurveTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn rect(&self, x: f64, y: f64, w: f64, h: f64)

The rect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect(&self, x: f64, y: f64, w: f64, h: f64) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect_with_f64( &self, x: f64, y: f64, w: f64, h: f64, radii: f64 ) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect_with_f64_sequence( &self, x: f64, y: f64, w: f64, h: f64, radii: &JsValue ) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect_with_dom_point_init_sequence( &self, x: f64, y: f64, w: f64, h: f64, radii: &JsValue ) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn get_line_dash(&self) -> Array

The getLineDash() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_dash(&self, segments: &JsValue) -> Result<(), JsValue>

The setLineDash() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn clear_rect(&self, x: f64, y: f64, w: f64, h: f64)

The clearRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_rect(&self, x: f64, y: f64, w: f64, h: f64)

The fillRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_rect(&self, x: f64, y: f64, w: f64, h: f64)

The strokeRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn reset(&self)

The reset() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn restore(&self)

The restore() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn save(&self)

The save() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_text(&self, text: &str, x: f64, y: f64) -> Result<(), JsValue>

The fillText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_text_with_max_width( &self, text: &str, x: f64, y: f64, max_width: f64 ) -> Result<(), JsValue>

The fillText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_text(&self, text: &str, x: f64, y: f64) -> Result<(), JsValue>

The strokeText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_text_with_max_width( &self, text: &str, x: f64, y: f64, max_width: f64 ) -> Result<(), JsValue>

The strokeText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn reset_transform(&self) -> Result<(), JsValue>

The resetTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn rotate(&self, angle: f64) -> Result<(), JsValue>

The rotate() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn scale(&self, x: f64, y: f64) -> Result<(), JsValue>

The scale() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_transform( &self, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64 ) -> Result<(), JsValue>

The setTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_transform_with_default_dom_matrix_2d_init( &self ) -> Result<(), JsValue>

The setTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn transform( &self, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64 ) -> Result<(), JsValue>

The transform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn translate(&self, x: f64, y: f64) -> Result<(), JsValue>

The translate() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn draw_custom_focus_ring(&self, element: &Element) -> bool

The drawCustomFocusRing() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Element

impl CanvasRenderingContext2d

pub fn draw_focus_if_needed(&self, element: &Element) -> Result<(), JsValue>

The drawFocusIfNeeded() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Element

impl CanvasRenderingContext2d

pub const DRAWWINDOW_DRAW_CARET: u32 = 1u32

The CanvasRenderingContext2D.DRAWWINDOW_DRAW_CARET const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_DO_NOT_FLUSH: u32 = 2u32

The CanvasRenderingContext2D.DRAWWINDOW_DO_NOT_FLUSH const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_DRAW_VIEW: u32 = 4u32

The CanvasRenderingContext2D.DRAWWINDOW_DRAW_VIEW const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_USE_WIDGET_LAYERS: u32 = 8u32

The CanvasRenderingContext2D.DRAWWINDOW_USE_WIDGET_LAYERS const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_ASYNC_DECODE_IMAGES: u32 = 16u32

The CanvasRenderingContext2D.DRAWWINDOW_ASYNC_DECODE_IMAGES const.

This API requires the following crate features to be activated: CanvasRenderingContext2d