Struct web_sys::CssCounterStyleRule

#[repr(transparent)]

pub struct CssCounterStyleRule { /* fields omitted */ }

The CSSCounterStyleRule object

MDN Documentation

This function requires the following crate features to be activated: CssRule

Methods

impl CssCounterStyleRule

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_name(&self, name: &str)

The name setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn system(&self) -> String

The system getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_system(&self, system: &str)

The system setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn symbols(&self) -> String

The symbols getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_symbols(&self, symbols: &str)

The symbols setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn additive_symbols(&self) -> String

The additiveSymbols getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_additive_symbols(&self, additive_symbols: &str)

The additiveSymbols setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn negative(&self) -> String

The negative getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_negative(&self, negative: &str)

The negative setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn prefix(&self) -> String

The prefix getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_prefix(&self, prefix: &str)

The prefix setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn suffix(&self) -> String

The suffix getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_suffix(&self, suffix: &str)

The suffix setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn range(&self) -> String

The range getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_range(&self, range: &str)

The range setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn pad(&self) -> String

The pad getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_pad(&self, pad: &str)

The pad setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn speak_as(&self) -> String

The speakAs getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_speak_as(&self, speak_as: &str)

The speakAs setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn fallback(&self) -> String

The fallback getter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule

impl CssCounterStyleRule

pub fn set_fallback(&self, fallback: &str)

The fallback setter

MDN Documentation

This function requires the following crate features to be activated: CssCounterStyleRule