Struct web_sys::DocumentType

#[repr(transparent)]

pub struct DocumentType { /* fields omitted */ }

The DocumentType object

MDN Documentation

This function requires the following crate features to be activated: EventTarget, Node

Methods

impl DocumentType

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn public_id(&self) -> String

The publicId getter

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn system_id(&self) -> String

The systemId getter

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_node_0(&self) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn after_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn after_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn after_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn after_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn after_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn after_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn after_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_0(&self) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn after_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The after() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_node_0(&self) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn before_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn before_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn before_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn before_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn before_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn before_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn before_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_0(&self) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn before_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The before() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn remove(&self)

The remove() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_node_0(&self) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn replace_with_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn replace_with_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn replace_with_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn replace_with_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn replace_with_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn replace_with_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType, Node

impl DocumentType

pub fn replace_with_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_0(&self) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType

impl DocumentType

pub fn replace_with_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The replaceWith() method

MDN Documentation

This function requires the following crate features to be activated: DocumentType