Struct web_sys::DomMatrixReadOnly

#[repr(transparent)]pub struct DomMatrixReadOnly { /* fields omitted */ }

The DomMatrixReadOnly class.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

Implementations

impl DomMatrixReadOnly

pub fn a(&self) -> f64

Getter for the a field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn b(&self) -> f64

Getter for the b field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn c(&self) -> f64

Getter for the c field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn d(&self) -> f64

Getter for the d field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn e(&self) -> f64

Getter for the e field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn f(&self) -> f64

Getter for the f field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m11(&self) -> f64

Getter for the m11 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m12(&self) -> f64

Getter for the m12 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m13(&self) -> f64

Getter for the m13 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m14(&self) -> f64

Getter for the m14 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m21(&self) -> f64

Getter for the m21 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m22(&self) -> f64

Getter for the m22 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m23(&self) -> f64

Getter for the m23 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m24(&self) -> f64

Getter for the m24 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m31(&self) -> f64

Getter for the m31 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m32(&self) -> f64

Getter for the m32 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m33(&self) -> f64

Getter for the m33 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m34(&self) -> f64

Getter for the m34 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m41(&self) -> f64

Getter for the m41 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m42(&self) -> f64

Getter for the m42 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m43(&self) -> f64

Getter for the m43 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn m44(&self) -> f64

Getter for the m44 field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn is_2d(&self) -> bool

Getter for the is2D field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn is_identity(&self) -> bool

Getter for the isIdentity field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn new() -> Result<DomMatrixReadOnly, JsValue>

The new DomMatrixReadOnly(..) constructor, creating a new instance of DomMatrixReadOnly.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn new_with_str(init: &str) -> Result<DomMatrixReadOnly, JsValue>

The new DomMatrixReadOnly(..) constructor, creating a new instance of DomMatrixReadOnly.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn new_with_f64_sequence(
init: &JsValue
) -> Result<DomMatrixReadOnly, JsValue>

The new DomMatrixReadOnly(..) constructor, creating a new instance of DomMatrixReadOnly.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn flip_x(&self) -> DomMatrix

The flipX() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn flip_y(&self) -> DomMatrix

The flipY() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn inverse(&self) -> DomMatrix

The inverse() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn multiply(&self, other: &DomMatrix) -> DomMatrix

The multiply() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn rotate(&self, angle: f64) -> DomMatrix

The rotate() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn rotate_with_origin_x(&self, angle: f64, origin_x: f64) -> DomMatrix

The rotate() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn rotate_with_origin_x_and_origin_y(
&self,
angle: f64,
origin_x: f64,
origin_y: f64
) -> DomMatrix

The rotate() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn rotate_axis_angle(&self, x: f64, y: f64, z: f64, angle: f64) -> DomMatrix

The rotateAxisAngle() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn rotate_from_vector(&self, x: f64, y: f64) -> DomMatrix

The rotateFromVector() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale(&self, scale: f64) -> DomMatrix

The scale() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_with_origin_x(&self, scale: f64, origin_x: f64) -> DomMatrix

The scale() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_with_origin_x_and_origin_y(
&self,
scale: f64,
origin_x: f64,
origin_y: f64
) -> DomMatrix

The scale() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale3d(&self, scale: f64) -> DomMatrix

The scale3d() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale3d_with_origin_x(&self, scale: f64, origin_x: f64) -> DomMatrix

The scale3d() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale3d_with_origin_x_and_origin_y(
&self,
scale: f64,
origin_x: f64,
origin_y: f64
) -> DomMatrix

The scale3d() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale3d_with_origin_x_and_origin_y_and_origin_z(
&self,
scale: f64,
origin_x: f64,
origin_y: f64,
origin_z: f64
) -> DomMatrix

The scale3d() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_non_uniform(&self, scale_x: f64) -> DomMatrix

The scaleNonUniform() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_non_uniform_with_scale_y(
&self,
scale_x: f64,
scale_y: f64
) -> DomMatrix

The scaleNonUniform() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_non_uniform_with_scale_y_and_scale_z(
&self,
scale_x: f64,
scale_y: f64,
scale_z: f64
) -> DomMatrix

The scaleNonUniform() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_non_uniform_with_scale_y_and_scale_z_and_origin_x(
&self,
scale_x: f64,
scale_y: f64,
scale_z: f64,
origin_x: f64
) -> DomMatrix

The scaleNonUniform() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_non_uniform_with_scale_y_and_scale_z_and_origin_x_and_origin_y(
&self,
scale_x: f64,
scale_y: f64,
scale_z: f64,
origin_x: f64,
origin_y: f64
) -> DomMatrix

The scaleNonUniform() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn scale_non_uniform_with_scale_y_and_scale_z_and_origin_x_and_origin_y_and_origin_z(
&self,
scale_x: f64,
scale_y: f64,
scale_z: f64,
origin_x: f64,
origin_y: f64,
origin_z: f64
) -> DomMatrix

The scaleNonUniform() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn skew_x(&self, sx: f64) -> DomMatrix

The skewX() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn skew_y(&self, sy: f64) -> DomMatrix

The skewY() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn to_float32_array(&self) -> Result<Vec<f32>, JsValue>

The toFloat32Array() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn to_float64_array(&self) -> Result<Vec<f64>, JsValue>

The toFloat64Array() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn to_json(&self) -> Object

The toJSON() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn transform_point(&self) -> DomPoint

The transformPoint() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly, DomPoint

impl DomMatrixReadOnly

pub fn transform_point_with_point(&self, point: &DomPointInit) -> DomPoint

The transformPoint() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrixReadOnly, DomPoint, DomPointInit

impl DomMatrixReadOnly

pub fn translate(&self, tx: f64, ty: f64) -> DomMatrix

The translate() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly

impl DomMatrixReadOnly

pub fn translate_with_tz(&self, tx: f64, ty: f64, tz: f64) -> DomMatrix

The translate() method.

MDN Documentation

This API requires the following crate features to be activated: DomMatrix, DomMatrixReadOnly