Struct web_sys::EventTarget

#[repr(transparent)]
pub struct EventTarget { /* private fields */ }

The EventTarget class.

MDN Documentation

This API requires the following crate features to be activated: EventTarget

Implementations

impl EventTarget

pub fn new() -> Result<EventTarget, JsValue>

The new EventTarget(..) constructor, creating a new instance of EventTarget.

MDN Documentation

This API requires the following crate features to be activated: EventTarget

impl EventTarget

pub fn add_event_listener_with_callback( &self, type_: &str, listener: &Function ) -> Result<(), JsValue>

The addEventListener() method.

MDN Documentation

This API requires the following crate features to be activated: EventTarget

impl EventTarget

pub fn add_event_listener_with_callback_and_add_event_listener_options( &self, type_: &str, listener: &Function, options: &AddEventListenerOptions ) -> Result<(), JsValue>

The addEventListener() method.

MDN Documentation

This API requires the following crate features to be activated: AddEventListenerOptions, EventTarget

impl EventTarget

pub fn add_event_listener_with_callback_and_bool( &self, type_: &str, listener: &Function, options: bool ) -> Result<(), JsValue>

The addEventListener() method.

MDN Documentation

This API requires the following crate features to be activated: EventTarget

impl EventTarget

pub fn add_event_listener_with_callback_and_add_event_listener_options_and_wants_untrusted( &self, type_: &str, listener: &Function, options: &AddEventListenerOptions, wants_untrusted: Option<bool> ) -> Result<(), JsValue>

The addEventListener() method.

MDN Documentation

This API requires the following crate features to be activated: AddEventListenerOptions, EventTarget

impl EventTarget

pub fn add_event_listener_with_callback_and_bool_and_wants_untrusted( &self, type_: &str, listener: &Function, options: bool, wants_untrusted: Option<bool> ) -> Result<(), JsValue>

The addEventListener() method.

MDN Documentation

This API requires the following crate features to be activated: EventTarget

impl EventTarget

pub fn dispatch_event(&self, event: &Event) -> Result<bool, JsValue>

The dispatchEvent() method.

MDN Documentation

This API requires the following crate features to be activated: Event, EventTarget

impl EventTarget

pub fn remove_event_listener_with_callback( &self, type_: &str, listener: &Function ) -> Result<(), JsValue>

The removeEventListener() method.

MDN Documentation

This API requires the following crate features to be activated: EventTarget

impl EventTarget

pub fn remove_event_listener_with_callback_and_bool( &self, type_: &str, listener: &Function, options: bool ) -> Result<(), JsValue>

The removeEventListener() method.

MDN Documentation

This API requires the following crate features to be activated: EventTarget