Struct web_sys::FileReader

#[repr(transparent)]

pub struct FileReader { /* fields omitted */ }

The FileReader object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl FileReader

pub fn new() -> Result<FileReader, JsValue>

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn abort(&self)

The abort() method

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn read_as_array_buffer(&self, blob: &Blob) -> Result<(), JsValue>

The readAsArrayBuffer() method

MDN Documentation

This function requires the following crate features to be activated: Blob, FileReader

impl FileReader

pub fn read_as_binary_string(&self, filedata: &Blob) -> Result<(), JsValue>

The readAsBinaryString() method

MDN Documentation

This function requires the following crate features to be activated: Blob, FileReader

impl FileReader

pub fn read_as_data_url(&self, blob: &Blob) -> Result<(), JsValue>

The readAsDataURL() method

MDN Documentation

This function requires the following crate features to be activated: Blob, FileReader

impl FileReader

pub fn read_as_text(&self, blob: &Blob) -> Result<(), JsValue>

The readAsText() method

MDN Documentation

This function requires the following crate features to be activated: Blob, FileReader

impl FileReader

pub fn read_as_text_with_label(
&self,
blob: &Blob,
label: &str
) -> Result<(), JsValue>

The readAsText() method

MDN Documentation

This function requires the following crate features to be activated: Blob, FileReader

impl FileReader

pub fn ready_state(&self) -> u16

The readyState getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn result(&self) -> Result<JsValue, JsValue>

The result getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn error(&self) -> Option<DomException>

The error getter

MDN Documentation

This function requires the following crate features to be activated: DomException, FileReader

impl FileReader

pub fn onloadstart(&self) -> Option<Function>

The onloadstart getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn set_onloadstart(&self, onloadstart: Option<&Function>)

The onloadstart setter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn onprogress(&self) -> Option<Function>

The onprogress getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn set_onprogress(&self, onprogress: Option<&Function>)

The onprogress setter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn onload(&self) -> Option<Function>

The onload getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn set_onload(&self, onload: Option<&Function>)

The onload setter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn onabort(&self) -> Option<Function>

The onabort getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn set_onabort(&self, onabort: Option<&Function>)

The onabort setter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn onloadend(&self) -> Option<Function>

The onloadend getter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub fn set_onloadend(&self, onloadend: Option<&Function>)

The onloadend setter

MDN Documentation

This function requires the following crate features to be activated: FileReader

impl FileReader

pub const EMPTY: u16

impl FileReader

pub const LOADING: u16

impl FileReader

pub const DONE: u16