Struct web_sys::FileSystemEntry

#[repr(transparent)]

pub struct FileSystemEntry { /* fields omitted */ }

The FileSystemEntry object

MDN Documentation

Methods

impl FileSystemEntry

pub fn get_parent(&self)

The getParent() method

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry

impl FileSystemEntry

pub fn get_parent_with_callback(&self, success_callback: &Function)

The getParent() method

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry

impl FileSystemEntry

pub fn get_parent_with_file_system_entry_callback(
&self,
success_callback: &FileSystemEntryCallback
)

The getParent() method

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry, FileSystemEntryCallback

impl FileSystemEntry

pub fn get_parent_with_callback_and_callback(
&self,
success_callback: &Function,
error_callback: &Function
)

The getParent() method

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry

impl FileSystemEntry

pub fn get_parent_with_file_system_entry_callback_and_callback(
&self,
success_callback: &FileSystemEntryCallback,
error_callback: &Function
)

The getParent() method

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry, FileSystemEntryCallback

impl FileSystemEntry

pub fn get_parent_with_callback_and_error_callback(
&self,
success_callback: &Function,
error_callback: &ErrorCallback
)

The getParent() method

MDN Documentation

This function requires the following crate features to be activated: ErrorCallback, FileSystemEntry

impl FileSystemEntry

pub fn get_parent_with_file_system_entry_callback_and_error_callback(
&self,
success_callback: &FileSystemEntryCallback,
error_callback: &ErrorCallback
)

The getParent() method

MDN Documentation

This function requires the following crate features to be activated: ErrorCallback, FileSystemEntry, FileSystemEntryCallback

impl FileSystemEntry

pub fn is_file(&self) -> bool

The isFile getter

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry

impl FileSystemEntry

pub fn is_directory(&self) -> bool

The isDirectory getter

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry

impl FileSystemEntry

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry

impl FileSystemEntry

pub fn full_path(&self) -> String

The fullPath getter

MDN Documentation

This function requires the following crate features to be activated: FileSystemEntry

impl FileSystemEntry

pub fn filesystem(&self) -> FileSystem

The filesystem getter

MDN Documentation

This function requires the following crate features to be activated: FileSystem, FileSystemEntry