Struct web_sys::FontFace

#[repr(transparent)]

pub struct FontFace { /* fields omitted */ }

The FontFace object

MDN Documentation

Methods

impl FontFace

pub fn new_with_str(family: &str, source: &str) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn new_with_array_buffer(
family: &str,
source: &ArrayBuffer
) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn new_with_array_buffer_view(
family: &str,
source: &Object
) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn new_with_u8_array(
family: &str,
source: &mut [u8]
) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn new_with_str_and_descriptors(
family: &str,
source: &str,
descriptors: &FontFaceDescriptors
) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace, FontFaceDescriptors

impl FontFace

pub fn new_with_array_buffer_and_descriptors(
family: &str,
source: &ArrayBuffer,
descriptors: &FontFaceDescriptors
) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace, FontFaceDescriptors

impl FontFace

pub fn new_with_array_buffer_view_and_descriptors(
family: &str,
source: &Object,
descriptors: &FontFaceDescriptors
) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace, FontFaceDescriptors

impl FontFace

pub fn new_with_u8_array_and_descriptors(
family: &str,
source: &mut [u8],
descriptors: &FontFaceDescriptors
) -> Result<FontFace, JsValue>

This function requires the following crate features to be activated: FontFace, FontFaceDescriptors

impl FontFace

pub fn load(&self) -> Result<Promise, JsValue>

The load() method

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn family(&self) -> String

The family getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_family(&self, family: &str)

The family setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn style(&self) -> String

The style getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_style(&self, style: &str)

The style setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn weight(&self) -> String

The weight getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_weight(&self, weight: &str)

The weight setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn stretch(&self) -> String

The stretch getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_stretch(&self, stretch: &str)

The stretch setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn unicode_range(&self) -> String

The unicodeRange getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_unicode_range(&self, unicode_range: &str)

The unicodeRange setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn variant(&self) -> String

The variant getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_variant(&self, variant: &str)

The variant setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn feature_settings(&self) -> String

The featureSettings getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_feature_settings(&self, feature_settings: &str)

The featureSettings setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn variation_settings(&self) -> String

The variationSettings getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_variation_settings(&self, variation_settings: &str)

The variationSettings setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn display(&self) -> String

The display getter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn set_display(&self, display: &str)

The display setter

MDN Documentation

This function requires the following crate features to be activated: FontFace

impl FontFace

pub fn status(&self) -> FontFaceLoadStatus

The status getter

MDN Documentation

This function requires the following crate features to be activated: FontFace, FontFaceLoadStatus

impl FontFace

pub fn loaded(&self) -> Result<Promise, JsValue>

The loaded getter

MDN Documentation

This function requires the following crate features to be activated: FontFace