Struct web_sys::HtmlCanvasElement

#[repr(transparent)]
pub struct HtmlCanvasElement { /* fields omitted */ }

The HtmlCanvasElement class.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

Implementations

impl HtmlCanvasElement

pub fn width(&self) -> u32

Getter for the width field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn set_width(&self, value: u32)

Setter for the width field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn height(&self) -> u32

Getter for the height field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn set_height(&self, value: u32)

Setter for the height field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn get_context(&self, context_id: &str) -> Result<Option<Object>, JsValue>

The getContext() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn get_context_with_context_options(
&self,
context_id: &str,
context_options: &JsValue
) -> Result<Option<Object>, JsValue>

The getContext() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn to_blob(&self, callback: &Function) -> Result<(), JsValue>

The toBlob() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn to_blob_with_type(
&self,
callback: &Function,
type_: &str
) -> Result<(), JsValue>

The toBlob() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn to_blob_with_type_and_encoder_options(
&self,
callback: &Function,
type_: &str,
encoder_options: &JsValue
) -> Result<(), JsValue>

The toBlob() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn to_data_url(&self) -> Result<String, JsValue>

The toDataURL() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn to_data_url_with_type(&self, type_: &str) -> Result<String, JsValue>

The toDataURL() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement

impl HtmlCanvasElement

pub fn to_data_url_with_type_and_encoder_options(
&self,
type_: &str,
encoder_options: &JsValue
) -> Result<String, JsValue>

The toDataURL() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlCanvasElement