Struct web_sys::HtmlIFrameElement

#[repr(transparent)]

pub struct HtmlIFrameElement { /* fields omitted */ }

The HTMLIFrameElement object

MDN Documentation

This function requires the following crate features to be activated: Element, EventTarget, HtmlElement, Node

Methods

impl HtmlIFrameElement

pub fn get_svg_document(&self) -> Option<Document>

The getSVGDocument() method

MDN Documentation

This function requires the following crate features to be activated: Document, HtmlIFrameElement

impl HtmlIFrameElement

pub fn src(&self) -> String

The src getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_src(&self, src: &str)

The src setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn srcdoc(&self) -> String

The srcdoc getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_srcdoc(&self, srcdoc: &str)

The srcdoc setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_name(&self, name: &str)

The name setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn sandbox(&self) -> DomTokenList

The sandbox getter

MDN Documentation

This function requires the following crate features to be activated: DomTokenList, HtmlIFrameElement

impl HtmlIFrameElement

pub fn allow_fullscreen(&self) -> bool

The allowFullscreen getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_allow_fullscreen(&self, allow_fullscreen: bool)

The allowFullscreen setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn allow_payment_request(&self) -> bool

The allowPaymentRequest getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_allow_payment_request(&self, allow_payment_request: bool)

The allowPaymentRequest setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn width(&self) -> String

The width getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_width(&self, width: &str)

The width setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn height(&self) -> String

The height getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_height(&self, height: &str)

The height setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn referrer_policy(&self) -> String

The referrerPolicy getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_referrer_policy(&self, referrer_policy: &str)

The referrerPolicy setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn content_document(&self) -> Option<Document>

The contentDocument getter

MDN Documentation

This function requires the following crate features to be activated: Document, HtmlIFrameElement

impl HtmlIFrameElement

pub fn align(&self) -> String

The align getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_align(&self, align: &str)

The align setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn scrolling(&self) -> String

The scrolling getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_scrolling(&self, scrolling: &str)

The scrolling setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn frame_border(&self) -> String

The frameBorder getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_frame_border(&self, frame_border: &str)

The frameBorder setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn long_desc(&self) -> String

The longDesc getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_long_desc(&self, long_desc: &str)

The longDesc setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn margin_height(&self) -> String

The marginHeight getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_margin_height(&self, margin_height: &str)

The marginHeight setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn margin_width(&self) -> String

The marginWidth getter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement

impl HtmlIFrameElement

pub fn set_margin_width(&self, margin_width: &str)

The marginWidth setter

MDN Documentation

This function requires the following crate features to be activated: HtmlIFrameElement