Struct web_sys::HtmlInputElement

#[repr(transparent)]
pub struct HtmlInputElement { /* private fields */ }

The HtmlInputElement class.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

Implementations

impl HtmlInputElement

pub fn accept(&self) -> String

Getter for the accept field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_accept(&self, value: &str)

Setter for the accept field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn alt(&self) -> String

Getter for the alt field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_alt(&self, value: &str)

Setter for the alt field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn autocomplete(&self) -> String

Getter for the autocomplete field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_autocomplete(&self, value: &str)

Setter for the autocomplete field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn autofocus(&self) -> bool

Getter for the autofocus field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_autofocus(&self, value: bool)

Setter for the autofocus field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn default_checked(&self) -> bool

Getter for the defaultChecked field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_default_checked(&self, value: bool)

Setter for the defaultChecked field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn checked(&self) -> bool

Getter for the checked field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_checked(&self, value: bool)

Setter for the checked field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn disabled(&self) -> bool

Getter for the disabled field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_disabled(&self, value: bool)

Setter for the disabled field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn files(&self) -> Option<FileList>

Getter for the files field of this object.

MDN Documentation

This API requires the following crate features to be activated: FileList, HtmlInputElement

impl HtmlInputElement

pub fn set_files(&self, value: Option<&FileList>)

Setter for the files field of this object.

MDN Documentation

This API requires the following crate features to be activated: FileList, HtmlInputElement

impl HtmlInputElement

pub fn form_action(&self) -> String

Getter for the formAction field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_action(&self, value: &str)

Setter for the formAction field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_enctype(&self) -> String

Getter for the formEnctype field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_enctype(&self, value: &str)

Setter for the formEnctype field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_method(&self) -> String

Getter for the formMethod field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_method(&self, value: &str)

Setter for the formMethod field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_no_validate(&self) -> bool

Getter for the formNoValidate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_no_validate(&self, value: bool)

Setter for the formNoValidate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn form_target(&self) -> String

Getter for the formTarget field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_form_target(&self, value: &str)

Setter for the formTarget field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn height(&self) -> u32

Getter for the height field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_height(&self, value: u32)

Setter for the height field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn indeterminate(&self) -> bool

Getter for the indeterminate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_indeterminate(&self, value: bool)

Setter for the indeterminate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn input_mode(&self) -> String

Getter for the inputMode field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_input_mode(&self, value: &str)

Setter for the inputMode field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn list(&self) -> Option<HtmlElement>

Getter for the list field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn max(&self) -> String

Getter for the max field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_max(&self, value: &str)

Setter for the max field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn max_length(&self) -> i32

Getter for the maxLength field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_max_length(&self, value: i32)

Setter for the maxLength field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn min(&self) -> String

Getter for the min field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_min(&self, value: &str)

Setter for the min field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn min_length(&self) -> i32

Getter for the minLength field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_min_length(&self, value: i32)

Setter for the minLength field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn multiple(&self) -> bool

Getter for the multiple field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_multiple(&self, value: bool)

Setter for the multiple field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn name(&self) -> String

Getter for the name field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_name(&self, value: &str)

Setter for the name field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn pattern(&self) -> String

Getter for the pattern field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_pattern(&self, value: &str)

Setter for the pattern field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn placeholder(&self) -> String

Getter for the placeholder field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_placeholder(&self, value: &str)

Setter for the placeholder field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn read_only(&self) -> bool

Getter for the readOnly field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_read_only(&self, value: bool)

Setter for the readOnly field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn required(&self) -> bool

Getter for the required field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_required(&self, value: bool)

Setter for the required field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn size(&self) -> u32

Getter for the size field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_size(&self, value: u32)

Setter for the size field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn src(&self) -> String

Getter for the src field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_src(&self, value: &str)

Setter for the src field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn step(&self) -> String

Getter for the step field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_step(&self, value: &str)

Setter for the step field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn type_(&self) -> String

Getter for the type field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_type(&self, value: &str)

Setter for the type field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn default_value(&self) -> String

Getter for the defaultValue field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_default_value(&self, value: &str)

Setter for the defaultValue field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn value(&self) -> String

Getter for the value field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_value(&self, value: &str)

Setter for the value field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn value_as_number(&self) -> f64

Getter for the valueAsNumber field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_value_as_number(&self, value: f64)

Setter for the valueAsNumber field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn width(&self) -> u32

Getter for the width field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_width(&self, value: u32)

Setter for the width field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn will_validate(&self) -> bool

Getter for the willValidate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn validation_message(&self) -> Result<String, JsValue>

Getter for the validationMessage field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn labels(&self) -> Option<NodeList>

Getter for the labels field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement, NodeList

impl HtmlInputElement

pub fn selection_start(&self) -> Result<Option<u32>, JsValue>

Getter for the selectionStart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_start(&self, value: Option<u32>) -> Result<(), JsValue>

Setter for the selectionStart field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn selection_end(&self) -> Result<Option<u32>, JsValue>

Getter for the selectionEnd field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_end(&self, value: Option<u32>) -> Result<(), JsValue>

Setter for the selectionEnd field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn selection_direction(&self) -> Result<Option<String>, JsValue>

Getter for the selectionDirection field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_direction(&self, value: Option<&str>) -> Result<(), JsValue>

Setter for the selectionDirection field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn align(&self) -> String

Getter for the align field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_align(&self, value: &str)

Setter for the align field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn use_map(&self) -> String

Getter for the useMap field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_use_map(&self, value: &str)

Setter for the useMap field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn webkit_entries(&self) -> Array

Getter for the webkitEntries field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn webkitdirectory(&self) -> bool

Getter for the webkitdirectory field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_webkitdirectory(&self, value: bool)

Setter for the webkitdirectory field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn check_validity(&self) -> bool

The checkValidity() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn report_validity(&self) -> bool

The reportValidity() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn select(&self)

The select() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_custom_validity(&self, error: &str)

The setCustomValidity() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_range_text(&self, replacement: &str) -> Result<(), JsValue>

The setRangeText() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_range_text_with_start_and_end(
&self,
replacement: &str,
start: u32,
end: u32
) -> Result<(), JsValue>

The setRangeText() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_range(&self, start: u32, end: u32) -> Result<(), JsValue>

The setSelectionRange() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement

impl HtmlInputElement

pub fn set_selection_range_with_direction(
&self,
start: u32,
end: u32,
direction: &str
) -> Result<(), JsValue>

The setSelectionRange() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlInputElement