Struct web_sys::HtmlObjectElement

#[repr(transparent)]

pub struct HtmlObjectElement { /* fields omitted */ }

The HTMLObjectElement object

MDN Documentation

This function requires the following crate features to be activated: Element, EventTarget, HtmlElement, Node

Methods

impl HtmlObjectElement

pub fn check_validity(&self) -> bool

The checkValidity() method

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn get_svg_document(&self) -> Option<Document>

The getSVGDocument() method

MDN Documentation

This function requires the following crate features to be activated: Document, HtmlObjectElement

impl HtmlObjectElement

pub fn report_validity(&self) -> bool

The reportValidity() method

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_custom_validity(&self, error: &str)

The setCustomValidity() method

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn data(&self) -> String

The data getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_data(&self, data: &str)

The data setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn type_(&self) -> String

The type getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_type(&self, type_: &str)

The type setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn type_must_match(&self) -> bool

The typeMustMatch getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_type_must_match(&self, type_must_match: bool)

The typeMustMatch setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_name(&self, name: &str)

The name setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn use_map(&self) -> String

The useMap getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_use_map(&self, use_map: &str)

The useMap setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn form(&self) -> Option<HtmlFormElement>

The form getter

MDN Documentation

This function requires the following crate features to be activated: HtmlFormElement, HtmlObjectElement

impl HtmlObjectElement

pub fn width(&self) -> String

The width getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_width(&self, width: &str)

The width setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn height(&self) -> String

The height getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_height(&self, height: &str)

The height setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn content_document(&self) -> Option<Document>

The contentDocument getter

MDN Documentation

This function requires the following crate features to be activated: Document, HtmlObjectElement

impl HtmlObjectElement

pub fn will_validate(&self) -> bool

The willValidate getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn validity(&self) -> ValidityState

The validity getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement, ValidityState

impl HtmlObjectElement

pub fn validation_message(&self) -> Result<String, JsValue>

The validationMessage getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn align(&self) -> String

The align getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_align(&self, align: &str)

The align setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn archive(&self) -> String

The archive getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_archive(&self, archive: &str)

The archive setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn code(&self) -> String

The code getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_code(&self, code: &str)

The code setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn declare(&self) -> bool

The declare getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_declare(&self, declare: bool)

The declare setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn hspace(&self) -> u32

The hspace getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_hspace(&self, hspace: u32)

The hspace setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn standby(&self) -> String

The standby getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_standby(&self, standby: &str)

The standby setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn vspace(&self) -> u32

The vspace getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_vspace(&self, vspace: u32)

The vspace setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn code_base(&self) -> String

The codeBase getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_code_base(&self, code_base: &str)

The codeBase setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn code_type(&self) -> String

The codeType getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_code_type(&self, code_type: &str)

The codeType setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn border(&self) -> String

The border getter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement

impl HtmlObjectElement

pub fn set_border(&self, border: &str)

The border setter

MDN Documentation

This function requires the following crate features to be activated: HtmlObjectElement