Struct web_sys::HtmlOptionElement

#[repr(transparent)]

pub struct HtmlOptionElement { /* fields omitted */ }

The HTMLOptionElement object

MDN Documentation

This function requires the following crate features to be activated: Element, EventTarget, HtmlElement, Node

Methods

impl HtmlOptionElement

pub fn new() -> Result<HtmlOptionElement, JsValue>

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn new_with_text(text: &str) -> Result<HtmlOptionElement, JsValue>

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn new_with_text_and_value(
text: &str,
value: &str
) -> Result<HtmlOptionElement, JsValue>

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn new_with_text_and_value_and_default_selected(
text: &str,
value: &str,
default_selected: bool
) -> Result<HtmlOptionElement, JsValue>

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn new_with_text_and_value_and_default_selected_and_selected(
text: &str,
value: &str,
default_selected: bool,
selected: bool
) -> Result<HtmlOptionElement, JsValue>

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn disabled(&self) -> bool

The disabled getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn set_disabled(&self, disabled: bool)

The disabled setter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn form(&self) -> Option<HtmlFormElement>

The form getter

MDN Documentation

This function requires the following crate features to be activated: HtmlFormElement, HtmlOptionElement

impl HtmlOptionElement

pub fn label(&self) -> String

The label getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn set_label(&self, label: &str)

The label setter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn default_selected(&self) -> bool

The defaultSelected getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn set_default_selected(&self, default_selected: bool)

The defaultSelected setter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn selected(&self) -> bool

The selected getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn set_selected(&self, selected: bool)

The selected setter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn value(&self) -> String

The value getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn set_value(&self, value: &str)

The value setter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn text(&self) -> String

The text getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn set_text(&self, text: &str)

The text setter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement

impl HtmlOptionElement

pub fn index(&self) -> i32

The index getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement