Struct web_sys::HtmlOutputElement

#[repr(transparent)]
pub struct HtmlOutputElement { /* fields omitted */ }

The HtmlOutputElement class.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

Implementations

impl HtmlOutputElement

pub fn html_for(&self) -> DomTokenList

Getter for the htmlFor field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomTokenList, HtmlOutputElement

impl HtmlOutputElement

pub fn form(&self) -> Option<HtmlFormElement>

Getter for the form field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlFormElement, HtmlOutputElement

impl HtmlOutputElement

pub fn name(&self) -> String

Getter for the name field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn set_name(&self, value: &str)

Setter for the name field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn type_(&self) -> String

Getter for the type field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn default_value(&self) -> String

Getter for the defaultValue field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn set_default_value(&self, value: &str)

Setter for the defaultValue field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn value(&self) -> String

Getter for the value field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn set_value(&self, value: &str)

Setter for the value field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn will_validate(&self) -> bool

Getter for the willValidate field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn validity(&self) -> ValidityState

Getter for the validity field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement, ValidityState

impl HtmlOutputElement

pub fn validation_message(&self) -> Result<String, JsValue>

Getter for the validationMessage field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn labels(&self) -> NodeList

Getter for the labels field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement, NodeList

impl HtmlOutputElement

pub fn check_validity(&self) -> bool

The checkValidity() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn report_validity(&self) -> bool

The reportValidity() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement

impl HtmlOutputElement

pub fn set_custom_validity(&self, error: &str)

The setCustomValidity() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlOutputElement