Struct web_sys::HtmlSelectElement

#[repr(transparent)]

pub struct HtmlSelectElement { /* fields omitted */ }

The HTMLSelectElement object

MDN Documentation

This function requires the following crate features to be activated: Element, EventTarget, HtmlElement, Node

Methods

impl HtmlSelectElement

pub fn add_with_html_option_element(
&self,
element: &HtmlOptionElement
) -> Result<(), JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement, HtmlSelectElement

impl HtmlSelectElement

pub fn add_with_html_opt_group_element(
&self,
element: &HtmlOptGroupElement
) -> Result<(), JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: HtmlOptGroupElement, HtmlSelectElement

impl HtmlSelectElement

pub fn add_with_html_option_element_and_opt_html_element(
&self,
element: &HtmlOptionElement,
before: Option<&HtmlElement>
) -> Result<(), JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: HtmlElement, HtmlOptionElement, HtmlSelectElement

impl HtmlSelectElement

pub fn add_with_html_opt_group_element_and_opt_html_element(
&self,
element: &HtmlOptGroupElement,
before: Option<&HtmlElement>
) -> Result<(), JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: HtmlElement, HtmlOptGroupElement, HtmlSelectElement

impl HtmlSelectElement

pub fn add_with_html_option_element_and_opt_i32(
&self,
element: &HtmlOptionElement,
before: Option<i32>
) -> Result<(), JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement, HtmlSelectElement

impl HtmlSelectElement

pub fn add_with_html_opt_group_element_and_opt_i32(
&self,
element: &HtmlOptGroupElement,
before: Option<i32>
) -> Result<(), JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: HtmlOptGroupElement, HtmlSelectElement

impl HtmlSelectElement

pub fn check_validity(&self) -> bool

The checkValidity() method

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn item(&self, index: u32) -> Option<Element>

The item() method

MDN Documentation

This function requires the following crate features to be activated: Element, HtmlSelectElement

impl HtmlSelectElement

pub fn named_item(&self, name: &str) -> Option<HtmlOptionElement>

The namedItem() method

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionElement, HtmlSelectElement

impl HtmlSelectElement

pub fn remove(&self)

The remove() method

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn report_validity(&self) -> bool

The reportValidity() method

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_custom_validity(&self, error: &str)

The setCustomValidity() method

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn get(&self, index: u32) -> Option<Element>

The indexing getter

This function requires the following crate features to be activated: Element, HtmlSelectElement

impl HtmlSelectElement

pub fn set(
&self,
index: u32,
option: Option<&HtmlOptionElement>
) -> Result<(), JsValue>

The indexing setter

This function requires the following crate features to be activated: HtmlOptionElement, HtmlSelectElement

impl HtmlSelectElement

pub fn autofocus(&self) -> bool

The autofocus getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_autofocus(&self, autofocus: bool)

The autofocus setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn autocomplete(&self) -> String

The autocomplete getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_autocomplete(&self, autocomplete: &str)

The autocomplete setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn disabled(&self) -> bool

The disabled getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_disabled(&self, disabled: bool)

The disabled setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn form(&self) -> Option<HtmlFormElement>

The form getter

MDN Documentation

This function requires the following crate features to be activated: HtmlFormElement, HtmlSelectElement

impl HtmlSelectElement

pub fn multiple(&self) -> bool

The multiple getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_multiple(&self, multiple: bool)

The multiple setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_name(&self, name: &str)

The name setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn required(&self) -> bool

The required getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_required(&self, required: bool)

The required setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn size(&self) -> u32

The size getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_size(&self, size: u32)

The size setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn type_(&self) -> String

The type getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn options(&self) -> HtmlOptionsCollection

The options getter

MDN Documentation

This function requires the following crate features to be activated: HtmlOptionsCollection, HtmlSelectElement

impl HtmlSelectElement

pub fn length(&self) -> u32

The length getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_length(&self, length: u32)

The length setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn selected_options(&self) -> HtmlCollection

The selectedOptions getter

MDN Documentation

This function requires the following crate features to be activated: HtmlCollection, HtmlSelectElement

impl HtmlSelectElement

pub fn selected_index(&self) -> i32

The selectedIndex getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_selected_index(&self, selected_index: i32)

The selectedIndex setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn value(&self) -> String

The value getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn set_value(&self, value: &str)

The value setter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn will_validate(&self) -> bool

The willValidate getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn validity(&self) -> ValidityState

The validity getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement, ValidityState

impl HtmlSelectElement

pub fn validation_message(&self) -> Result<String, JsValue>

The validationMessage getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement

impl HtmlSelectElement

pub fn labels(&self) -> NodeList

The labels getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSelectElement, NodeList