Struct web_sys::HtmlVideoElement

#[repr(transparent)]pub struct HtmlVideoElement { /* fields omitted */ }

The HtmlVideoElement class.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

Implementations

impl HtmlVideoElement

pub fn width(&self) -> u32

Getter for the width field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn set_width(&self, value: u32)

Setter for the width field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn height(&self) -> u32

Getter for the height field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn set_height(&self, value: u32)

Setter for the height field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn video_width(&self) -> u32

Getter for the videoWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn video_height(&self) -> u32

Getter for the videoHeight field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn poster(&self) -> String

Getter for the poster field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn set_poster(&self, value: &str)

Setter for the poster field of this object.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement

impl HtmlVideoElement

pub fn get_video_playback_quality(&self) -> VideoPlaybackQuality

The getVideoPlaybackQuality() method.

MDN Documentation

This API requires the following crate features to be activated: HtmlVideoElement, VideoPlaybackQuality