Struct web_sys::IdbDatabase

#[repr(transparent)]

pub struct IdbDatabase { /* fields omitted */ }

The IDBDatabase object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl IdbDatabase

pub fn close(&self)

The close() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn create_mutable_file(&self, name: &str) -> Result<IdbRequest, JsValue>

The createMutableFile() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbRequest

impl IdbDatabase

pub fn create_mutable_file_with_type(
&self,
name: &str,
type_: &str
) -> Result<IdbRequest, JsValue>

The createMutableFile() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbRequest

impl IdbDatabase

pub fn create_object_store(&self, name: &str) -> Result<IdbObjectStore, JsValue>

The createObjectStore() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbObjectStore

impl IdbDatabase

pub fn create_object_store_with_optional_parameters(
&self,
name: &str,
optional_parameters: &IdbObjectStoreParameters
) -> Result<IdbObjectStore, JsValue>

The createObjectStore() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbObjectStore, IdbObjectStoreParameters

impl IdbDatabase

pub fn delete_object_store(&self, name: &str) -> Result<(), JsValue>

The deleteObjectStore() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn moz_create_file_handle(&self, name: &str) -> Result<IdbRequest, JsValue>

The mozCreateFileHandle() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbRequest

impl IdbDatabase

pub fn moz_create_file_handle_with_type(
&self,
name: &str,
type_: &str
) -> Result<IdbRequest, JsValue>

The mozCreateFileHandle() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbRequest

impl IdbDatabase

pub fn transaction_with_str(
&self,
store_names: &str
) -> Result<IdbTransaction, JsValue>

The transaction() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbTransaction

impl IdbDatabase

pub fn transaction_with_str_and_mode(
&self,
store_names: &str,
mode: IdbTransactionMode
) -> Result<IdbTransaction, JsValue>

The transaction() method

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, IdbTransaction, IdbTransactionMode

impl IdbDatabase

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn object_store_names(&self) -> DomStringList

The objectStoreNames getter

MDN Documentation

This function requires the following crate features to be activated: DomStringList, IdbDatabase

impl IdbDatabase

pub fn onabort(&self) -> Option<Function>

The onabort getter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn set_onabort(&self, onabort: Option<&Function>)

The onabort setter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn onclose(&self) -> Option<Function>

The onclose getter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn set_onclose(&self, onclose: Option<&Function>)

The onclose setter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn onversionchange(&self) -> Option<Function>

The onversionchange getter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn set_onversionchange(&self, onversionchange: Option<&Function>)

The onversionchange setter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase

impl IdbDatabase

pub fn storage(&self) -> StorageType

The storage getter

MDN Documentation

This function requires the following crate features to be activated: IdbDatabase, StorageType