Struct web_sys::IdbObjectStore

#[repr(transparent)]

pub struct IdbObjectStore { /* fields omitted */ }

The IDBObjectStore object

MDN Documentation

Methods

impl IdbObjectStore

pub fn add(&self, value: &JsValue) -> Result<IdbRequest, JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn add_with_key(
&self,
value: &JsValue,
key: &JsValue
) -> Result<IdbRequest, JsValue>

The add() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn clear(&self) -> Result<IdbRequest, JsValue>

The clear() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn count(&self) -> Result<IdbRequest, JsValue>

The count() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn count_with_key(&self, key: &JsValue) -> Result<IdbRequest, JsValue>

The count() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn create_index_with_str(
&self,
name: &str,
key_path: &str
) -> Result<IdbIndex, JsValue>

The createIndex() method

MDN Documentation

This function requires the following crate features to be activated: IdbIndex, IdbObjectStore

impl IdbObjectStore

pub fn create_index_with_str_and_optional_parameters(
&self,
name: &str,
key_path: &str,
optional_parameters: &IdbIndexParameters
) -> Result<IdbIndex, JsValue>

The createIndex() method

MDN Documentation

This function requires the following crate features to be activated: IdbIndex, IdbIndexParameters, IdbObjectStore

impl IdbObjectStore

pub fn delete(&self, key: &JsValue) -> Result<IdbRequest, JsValue>

The delete() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn delete_index(&self, index_name: &str) -> Result<(), JsValue>

The deleteIndex() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore

impl IdbObjectStore

pub fn get(&self, key: &JsValue) -> Result<IdbRequest, JsValue>

The get() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn get_all(&self) -> Result<IdbRequest, JsValue>

The getAll() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn get_all_with_key(&self, key: &JsValue) -> Result<IdbRequest, JsValue>

The getAll() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn get_all_with_key_and_limit(
&self,
key: &JsValue,
limit: u32
) -> Result<IdbRequest, JsValue>

The getAll() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn get_all_keys(&self) -> Result<IdbRequest, JsValue>

The getAllKeys() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn get_all_keys_with_key(
&self,
key: &JsValue
) -> Result<IdbRequest, JsValue>

The getAllKeys() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn get_all_keys_with_key_and_limit(
&self,
key: &JsValue,
limit: u32
) -> Result<IdbRequest, JsValue>

The getAllKeys() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn get_key(&self, key: &JsValue) -> Result<IdbRequest, JsValue>

The getKey() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn index(&self, name: &str) -> Result<IdbIndex, JsValue>

The index() method

MDN Documentation

This function requires the following crate features to be activated: IdbIndex, IdbObjectStore

impl IdbObjectStore

pub fn moz_get_all(&self) -> Result<IdbRequest, JsValue>

The mozGetAll() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn moz_get_all_with_key(&self, key: &JsValue) -> Result<IdbRequest, JsValue>

The mozGetAll() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn moz_get_all_with_key_and_limit(
&self,
key: &JsValue,
limit: u32
) -> Result<IdbRequest, JsValue>

The mozGetAll() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn open_cursor(&self) -> Result<IdbRequest, JsValue>

The openCursor() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn open_cursor_with_range(
&self,
range: &JsValue
) -> Result<IdbRequest, JsValue>

The openCursor() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn open_cursor_with_range_and_direction(
&self,
range: &JsValue,
direction: IdbCursorDirection
) -> Result<IdbRequest, JsValue>

The openCursor() method

MDN Documentation

This function requires the following crate features to be activated: IdbCursorDirection, IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn open_key_cursor(&self) -> Result<IdbRequest, JsValue>

The openKeyCursor() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn open_key_cursor_with_range(
&self,
range: &JsValue
) -> Result<IdbRequest, JsValue>

The openKeyCursor() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn open_key_cursor_with_range_and_direction(
&self,
range: &JsValue,
direction: IdbCursorDirection
) -> Result<IdbRequest, JsValue>

The openKeyCursor() method

MDN Documentation

This function requires the following crate features to be activated: IdbCursorDirection, IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn put(&self, value: &JsValue) -> Result<IdbRequest, JsValue>

The put() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn put_with_key(
&self,
value: &JsValue,
key: &JsValue
) -> Result<IdbRequest, JsValue>

The put() method

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbRequest

impl IdbObjectStore

pub fn name(&self) -> String

The name getter

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore

impl IdbObjectStore

pub fn set_name(&self, name: &str)

The name setter

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore

impl IdbObjectStore

pub fn key_path(&self) -> Result<JsValue, JsValue>

The keyPath getter

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore

impl IdbObjectStore

pub fn index_names(&self) -> DomStringList

The indexNames getter

MDN Documentation

This function requires the following crate features to be activated: DomStringList, IdbObjectStore

impl IdbObjectStore

pub fn transaction(&self) -> IdbTransaction

The transaction getter

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore, IdbTransaction

impl IdbObjectStore

pub fn auto_increment(&self) -> bool

The autoIncrement getter

MDN Documentation

This function requires the following crate features to be activated: IdbObjectStore