Struct web_sys::KeyEvent

#[repr(transparent)]

pub struct KeyEvent { /* fields omitted */ }

The KeyEvent object

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

Methods

impl KeyEvent

pub fn init_key_event(&self, type_: &str)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble(&self, type_: &str, can_bubble: bool)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_key_code(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
key_code: u32
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_key_code_and_char_code(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
key_code: u32,
char_code: u32
)

The initKeyEvent() method

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub const DOM_VK_CANCEL: u32

impl KeyEvent

pub const DOM_VK_HELP: u32

impl KeyEvent

pub const DOM_VK_BACK_SPACE: u32

impl KeyEvent

pub const DOM_VK_TAB: u32

impl KeyEvent

pub const DOM_VK_CLEAR: u32

impl KeyEvent

pub const DOM_VK_RETURN: u32

impl KeyEvent

pub const DOM_VK_SHIFT: u32

impl KeyEvent

pub const DOM_VK_CONTROL: u32

impl KeyEvent

pub const DOM_VK_ALT: u32

impl KeyEvent

pub const DOM_VK_PAUSE: u32

impl KeyEvent

pub const DOM_VK_CAPS_LOCK: u32

impl KeyEvent

pub const DOM_VK_KANA: u32

impl KeyEvent

pub const DOM_VK_HANGUL: u32

impl KeyEvent

pub const DOM_VK_EISU: u32

impl KeyEvent

pub const DOM_VK_JUNJA: u32

impl KeyEvent

pub const DOM_VK_FINAL: u32

impl KeyEvent

pub const DOM_VK_HANJA: u32

impl KeyEvent

pub const DOM_VK_KANJI: u32

impl KeyEvent

pub const DOM_VK_ESCAPE: u32

impl KeyEvent

pub const DOM_VK_CONVERT: u32

impl KeyEvent

pub const DOM_VK_NONCONVERT: u32

impl KeyEvent

pub const DOM_VK_ACCEPT: u32

impl KeyEvent

pub const DOM_VK_MODECHANGE: u32

impl KeyEvent

pub const DOM_VK_SPACE: u32

impl KeyEvent

pub const DOM_VK_PAGE_UP: u32

impl KeyEvent

pub const DOM_VK_PAGE_DOWN: u32

impl KeyEvent

pub const DOM_VK_END: u32

impl KeyEvent

pub const DOM_VK_HOME: u32

impl KeyEvent

pub const DOM_VK_LEFT: u32

impl KeyEvent

pub const DOM_VK_UP: u32

impl KeyEvent

pub const DOM_VK_RIGHT: u32

impl KeyEvent

pub const DOM_VK_DOWN: u32

impl KeyEvent

pub const DOM_VK_SELECT: u32

impl KeyEvent

pub const DOM_VK_PRINT: u32

impl KeyEvent

pub const DOM_VK_EXECUTE: u32

impl KeyEvent

pub const DOM_VK_PRINTSCREEN: u32

impl KeyEvent

pub const DOM_VK_INSERT: u32

impl KeyEvent

pub const DOM_VK_DELETE: u32

impl KeyEvent

pub const DOM_VK_0: u32

impl KeyEvent

pub const DOM_VK_1: u32

impl KeyEvent

pub const DOM_VK_2: u32

impl KeyEvent

pub const DOM_VK_3: u32

impl KeyEvent

pub const DOM_VK_4: u32

impl KeyEvent

pub const DOM_VK_5: u32

impl KeyEvent

pub const DOM_VK_6: u32

impl KeyEvent

pub const DOM_VK_7: u32

impl KeyEvent

pub const DOM_VK_8: u32

impl KeyEvent

pub const DOM_VK_9: u32

impl KeyEvent

pub const DOM_VK_COLON: u32

impl KeyEvent

pub const DOM_VK_SEMICOLON: u32

impl KeyEvent

pub const DOM_VK_LESS_THAN: u32

impl KeyEvent

pub const DOM_VK_EQUALS: u32

impl KeyEvent

pub const DOM_VK_GREATER_THAN: u32

impl KeyEvent

pub const DOM_VK_QUESTION_MARK: u32

impl KeyEvent

pub const DOM_VK_AT: u32

impl KeyEvent

pub const DOM_VK_A: u32

impl KeyEvent

pub const DOM_VK_B: u32

impl KeyEvent

pub const DOM_VK_C: u32

impl KeyEvent

pub const DOM_VK_D: u32

impl KeyEvent

pub const DOM_VK_E: u32

impl KeyEvent

pub const DOM_VK_F: u32

impl KeyEvent

pub const DOM_VK_G: u32

impl KeyEvent

pub const DOM_VK_H: u32

impl KeyEvent

pub const DOM_VK_I: u32

impl KeyEvent

pub const DOM_VK_J: u32

impl KeyEvent

pub const DOM_VK_K: u32

impl KeyEvent

pub const DOM_VK_L: u32

impl KeyEvent

pub const DOM_VK_M: u32

impl KeyEvent

pub const DOM_VK_N: u32

impl KeyEvent

pub const DOM_VK_O: u32

impl KeyEvent

pub const DOM_VK_P: u32

impl KeyEvent

pub const DOM_VK_Q: u32

impl KeyEvent

pub const DOM_VK_R: u32

impl KeyEvent

pub const DOM_VK_S: u32

impl KeyEvent

pub const DOM_VK_T: u32

impl KeyEvent

pub const DOM_VK_U: u32

impl KeyEvent

pub const DOM_VK_V: u32

impl KeyEvent

pub const DOM_VK_W: u32

impl KeyEvent

pub const DOM_VK_X: u32

impl KeyEvent

pub const DOM_VK_Y: u32

impl KeyEvent

pub const DOM_VK_Z: u32

impl KeyEvent

pub const DOM_VK_WIN: u32

impl KeyEvent

pub const DOM_VK_CONTEXT_MENU: u32

impl KeyEvent

pub const DOM_VK_SLEEP: u32

impl KeyEvent

pub const DOM_VK_NUMPAD0: u32

impl KeyEvent

pub const DOM_VK_NUMPAD1: u32

impl KeyEvent

pub const DOM_VK_NUMPAD2: u32

impl KeyEvent

pub const DOM_VK_NUMPAD3: u32

impl KeyEvent

pub const DOM_VK_NUMPAD4: u32

impl KeyEvent

pub const DOM_VK_NUMPAD5: u32

impl KeyEvent

pub const DOM_VK_NUMPAD6: u32

impl KeyEvent

pub const DOM_VK_NUMPAD7: u32

impl KeyEvent

pub const DOM_VK_NUMPAD8: u32

impl KeyEvent

pub const DOM_VK_NUMPAD9: u32

impl KeyEvent

pub const DOM_VK_MULTIPLY: u32

impl KeyEvent

pub const DOM_VK_ADD: u32

impl KeyEvent

pub const DOM_VK_SEPARATOR: u32

impl KeyEvent

pub const DOM_VK_SUBTRACT: u32

impl KeyEvent

pub const DOM_VK_DECIMAL: u32

impl KeyEvent

pub const DOM_VK_DIVIDE: u32

impl KeyEvent

pub const DOM_VK_F1: u32

impl KeyEvent

pub const DOM_VK_F2: u32

impl KeyEvent

pub const DOM_VK_F3: u32

impl KeyEvent

pub const DOM_VK_F4: u32

impl KeyEvent

pub const DOM_VK_F5: u32

impl KeyEvent

pub const DOM_VK_F6: u32

impl KeyEvent

pub const DOM_VK_F7: u32

impl KeyEvent

pub const DOM_VK_F8: u32

impl KeyEvent

pub const DOM_VK_F9: u32

impl KeyEvent

pub const DOM_VK_F10: u32

impl KeyEvent

pub const DOM_VK_F11: u32

impl KeyEvent

pub const DOM_VK_F12: u32

impl KeyEvent

pub const DOM_VK_F13: u32

impl KeyEvent

pub const DOM_VK_F14: u32

impl KeyEvent

pub const DOM_VK_F15: u32

impl KeyEvent

pub const DOM_VK_F16: u32

impl KeyEvent

pub const DOM_VK_F17: u32

impl KeyEvent

pub const DOM_VK_F18: u32

impl KeyEvent

pub const DOM_VK_F19: u32

impl KeyEvent

pub const DOM_VK_F20: u32

impl KeyEvent

pub const DOM_VK_F21: u32

impl KeyEvent

pub const DOM_VK_F22: u32

impl KeyEvent

pub const DOM_VK_F23: u32

impl KeyEvent

pub const DOM_VK_F24: u32

impl KeyEvent

pub const DOM_VK_NUM_LOCK: u32

impl KeyEvent

pub const DOM_VK_SCROLL_LOCK: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_FJ_JISHO: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_FJ_MASSHOU: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_FJ_TOUROKU: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_FJ_LOYA: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_FJ_ROYA: u32

impl KeyEvent

pub const DOM_VK_CIRCUMFLEX: u32

impl KeyEvent

pub const DOM_VK_EXCLAMATION: u32

impl KeyEvent

pub const DOM_VK_DOUBLE_QUOTE: u32

impl KeyEvent

pub const DOM_VK_HASH: u32

impl KeyEvent

pub const DOM_VK_DOLLAR: u32

impl KeyEvent

pub const DOM_VK_PERCENT: u32

impl KeyEvent

pub const DOM_VK_AMPERSAND: u32

impl KeyEvent

pub const DOM_VK_UNDERSCORE: u32

impl KeyEvent

pub const DOM_VK_OPEN_PAREN: u32

impl KeyEvent

pub const DOM_VK_CLOSE_PAREN: u32

impl KeyEvent

pub const DOM_VK_ASTERISK: u32

impl KeyEvent

pub const DOM_VK_PLUS: u32

impl KeyEvent

pub const DOM_VK_PIPE: u32

impl KeyEvent

pub const DOM_VK_HYPHEN_MINUS: u32

impl KeyEvent

pub const DOM_VK_OPEN_CURLY_BRACKET: u32

impl KeyEvent

pub const DOM_VK_CLOSE_CURLY_BRACKET: u32

impl KeyEvent

pub const DOM_VK_TILDE: u32

impl KeyEvent

pub const DOM_VK_VOLUME_MUTE: u32

impl KeyEvent

pub const DOM_VK_VOLUME_DOWN: u32

impl KeyEvent

pub const DOM_VK_VOLUME_UP: u32

impl KeyEvent

pub const DOM_VK_COMMA: u32

impl KeyEvent

pub const DOM_VK_PERIOD: u32

impl KeyEvent

pub const DOM_VK_SLASH: u32

impl KeyEvent

pub const DOM_VK_BACK_QUOTE: u32

impl KeyEvent

pub const DOM_VK_OPEN_BRACKET: u32

impl KeyEvent

pub const DOM_VK_BACK_SLASH: u32

impl KeyEvent

pub const DOM_VK_CLOSE_BRACKET: u32

impl KeyEvent

pub const DOM_VK_QUOTE: u32

impl KeyEvent

pub const DOM_VK_META: u32

impl KeyEvent

pub const DOM_VK_ALTGR: u32

impl KeyEvent

pub const DOM_VK_WIN_ICO_HELP: u32

impl KeyEvent

pub const DOM_VK_WIN_ICO_00: u32

impl KeyEvent

pub const DOM_VK_PROCESSKEY: u32

impl KeyEvent

pub const DOM_VK_WIN_ICO_CLEAR: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_RESET: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_JUMP: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_PA1: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_PA2: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_PA3: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_WSCTRL: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_CUSEL: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_ATTN: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_FINISH: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_COPY: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_AUTO: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_ENLW: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_BACKTAB: u32

impl KeyEvent

pub const DOM_VK_ATTN: u32

impl KeyEvent

pub const DOM_VK_CRSEL: u32

impl KeyEvent

pub const DOM_VK_EXSEL: u32

impl KeyEvent

pub const DOM_VK_EREOF: u32

impl KeyEvent

pub const DOM_VK_PLAY: u32

impl KeyEvent

pub const DOM_VK_ZOOM: u32

impl KeyEvent

pub const DOM_VK_PA1: u32

impl KeyEvent

pub const DOM_VK_WIN_OEM_CLEAR: u32