Struct web_sys::MediaRecorder

#[repr(transparent)]

pub struct MediaRecorder { /* fields omitted */ }

The MediaRecorder object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl MediaRecorder

pub fn new_with_media_stream(
stream: &MediaStream
) -> Result<MediaRecorder, JsValue>

This function requires the following crate features to be activated: MediaRecorder, MediaStream

impl MediaRecorder

pub fn new_with_media_stream_and_media_recorder_options(
stream: &MediaStream,
options: &MediaRecorderOptions
) -> Result<MediaRecorder, JsValue>

This function requires the following crate features to be activated: MediaRecorder, MediaRecorderOptions, MediaStream

impl MediaRecorder

pub fn new_with_audio_node(node: &AudioNode) -> Result<MediaRecorder, JsValue>

This function requires the following crate features to be activated: AudioNode, MediaRecorder

impl MediaRecorder

pub fn new_with_audio_node_and_u32(
node: &AudioNode,
output: u32
) -> Result<MediaRecorder, JsValue>

This function requires the following crate features to be activated: AudioNode, MediaRecorder

impl MediaRecorder

pub fn new_with_audio_node_and_u32_and_options(
node: &AudioNode,
output: u32,
options: &MediaRecorderOptions
) -> Result<MediaRecorder, JsValue>

This function requires the following crate features to be activated: AudioNode, MediaRecorder, MediaRecorderOptions

impl MediaRecorder

pub fn is_type_supported(type_: &str) -> bool

The isTypeSupported() method

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn pause(&self) -> Result<(), JsValue>

The pause() method

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn request_data(&self) -> Result<(), JsValue>

The requestData() method

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn resume(&self) -> Result<(), JsValue>

The resume() method

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn start(&self) -> Result<(), JsValue>

The start() method

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn start_with_time_slice(&self, time_slice: i32) -> Result<(), JsValue>

The start() method

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn stop(&self) -> Result<(), JsValue>

The stop() method

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn stream(&self) -> MediaStream

The stream getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder, MediaStream

impl MediaRecorder

pub fn state(&self) -> RecordingState

The state getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder, RecordingState

impl MediaRecorder

pub fn mime_type(&self) -> String

The mimeType getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn ondataavailable(&self) -> Option<Function>

The ondataavailable getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn set_ondataavailable(&self, ondataavailable: Option<&Function>)

The ondataavailable setter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn onstart(&self) -> Option<Function>

The onstart getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn set_onstart(&self, onstart: Option<&Function>)

The onstart setter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn onstop(&self) -> Option<Function>

The onstop getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn set_onstop(&self, onstop: Option<&Function>)

The onstop setter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn onwarning(&self) -> Option<Function>

The onwarning getter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder

impl MediaRecorder

pub fn set_onwarning(&self, onwarning: Option<&Function>)

The onwarning setter

MDN Documentation

This function requires the following crate features to be activated: MediaRecorder