Struct web_sys::MediaSource

#[repr(transparent)]

pub struct MediaSource { /* fields omitted */ }

The MediaSource object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl MediaSource

pub fn new() -> Result<MediaSource, JsValue>

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn add_source_buffer(&self, type_: &str) -> Result<SourceBuffer, JsValue>

The addSourceBuffer() method

MDN Documentation

This function requires the following crate features to be activated: MediaSource, SourceBuffer

impl MediaSource

pub fn clear_live_seekable_range(&self) -> Result<(), JsValue>

The clearLiveSeekableRange() method

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn end_of_stream(&self) -> Result<(), JsValue>

The endOfStream() method

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn end_of_stream_with_error(
&self,
error: MediaSourceEndOfStreamError
) -> Result<(), JsValue>

The endOfStream() method

MDN Documentation

This function requires the following crate features to be activated: MediaSource, MediaSourceEndOfStreamError

impl MediaSource

pub fn is_type_supported(type_: &str) -> bool

The isTypeSupported() method

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn remove_source_buffer(
&self,
source_buffer: &SourceBuffer
) -> Result<(), JsValue>

The removeSourceBuffer() method

MDN Documentation

This function requires the following crate features to be activated: MediaSource, SourceBuffer

impl MediaSource

pub fn set_live_seekable_range(
&self,
start: f64,
end: f64
) -> Result<(), JsValue>

The setLiveSeekableRange() method

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn source_buffers(&self) -> SourceBufferList

The sourceBuffers getter

MDN Documentation

This function requires the following crate features to be activated: MediaSource, SourceBufferList

impl MediaSource

pub fn active_source_buffers(&self) -> SourceBufferList

The activeSourceBuffers getter

MDN Documentation

This function requires the following crate features to be activated: MediaSource, SourceBufferList

impl MediaSource

pub fn ready_state(&self) -> MediaSourceReadyState

The readyState getter

MDN Documentation

This function requires the following crate features to be activated: MediaSource, MediaSourceReadyState

impl MediaSource

pub fn duration(&self) -> f64

The duration getter

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn set_duration(&self, duration: f64)

The duration setter

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn onsourceopen(&self) -> Option<Function>

The onsourceopen getter

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn set_onsourceopen(&self, onsourceopen: Option<&Function>)

The onsourceopen setter

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn onsourceended(&self) -> Option<Function>

The onsourceended getter

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn set_onsourceended(&self, onsourceended: Option<&Function>)

The onsourceended setter

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn onsourceclosed(&self) -> Option<Function>

The onsourceclosed getter

MDN Documentation

This function requires the following crate features to be activated: MediaSource

impl MediaSource

pub fn set_onsourceclosed(&self, onsourceclosed: Option<&Function>)

The onsourceclosed setter

MDN Documentation

This function requires the following crate features to be activated: MediaSource