Struct web_sys::MediaStreamTrack

#[repr(transparent)]

pub struct MediaStreamTrack { /* fields omitted */ }

The MediaStreamTrack object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl MediaStreamTrack

pub fn apply_constraints(&self) -> Result<Promise, JsValue>

The applyConstraints() method

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn apply_constraints_with_constraints(
&self,
constraints: &MediaTrackConstraints
) -> Result<Promise, JsValue>

The applyConstraints() method

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack, MediaTrackConstraints

impl MediaStreamTrack

pub fn clone(&self) -> MediaStreamTrack

The clone() method

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn get_constraints(&self) -> MediaTrackConstraints

The getConstraints() method

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack, MediaTrackConstraints

impl MediaStreamTrack

pub fn get_settings(&self) -> MediaTrackSettings

The getSettings() method

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack, MediaTrackSettings

impl MediaStreamTrack

pub fn stop(&self)

The stop() method

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn kind(&self) -> String

The kind getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn id(&self) -> String

The id getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn label(&self) -> String

The label getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn enabled(&self) -> bool

The enabled getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn set_enabled(&self, enabled: bool)

The enabled setter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn muted(&self) -> bool

The muted getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn onmute(&self) -> Option<Function>

The onmute getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn set_onmute(&self, onmute: Option<&Function>)

The onmute setter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn onunmute(&self) -> Option<Function>

The onunmute getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn set_onunmute(&self, onunmute: Option<&Function>)

The onunmute setter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn ready_state(&self) -> MediaStreamTrackState

The readyState getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack, MediaStreamTrackState

impl MediaStreamTrack

pub fn onended(&self) -> Option<Function>

The onended getter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack

impl MediaStreamTrack

pub fn set_onended(&self, onended: Option<&Function>)

The onended setter

MDN Documentation

This function requires the following crate features to be activated: MediaStreamTrack