Struct web_sys::MouseScrollEvent

#[repr(transparent)]

pub struct MouseScrollEvent { /* fields omitted */ }

The MouseScrollEvent object

MDN Documentation

This function requires the following crate features to be activated: Event, MouseEvent, UiEvent

Methods

impl MouseScrollEvent

pub fn init_mouse_scroll_event(&self, type_: &str)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble(
&self,
type_: &str,
can_bubble: bool
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y_and_ctrl_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32,
ctrl_key: bool
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y_and_ctrl_key_and_alt_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32,
ctrl_key: bool,
alt_key: bool
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y_and_ctrl_key_and_alt_key_and_shift_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_button(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
button: i16
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_button_and_related_target(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
button: i16,
related_target: Option<&EventTarget>
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: EventTarget, MouseScrollEvent, Window

impl MouseScrollEvent

pub fn init_mouse_scroll_event_with_can_bubble_and_cancelable_and_view_and_detail_and_screen_x_and_screen_y_and_client_x_and_client_y_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_button_and_related_target_and_axis(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
screen_x: i32,
screen_y: i32,
client_x: i32,
client_y: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
button: i16,
related_target: Option<&EventTarget>,
axis: i32
)

The initMouseScrollEvent() method

MDN Documentation

This function requires the following crate features to be activated: EventTarget, MouseScrollEvent, Window

impl MouseScrollEvent

pub fn axis(&self) -> i32

The axis getter

MDN Documentation

This function requires the following crate features to be activated: MouseScrollEvent

impl MouseScrollEvent

pub const HORIZONTAL_AXIS: i32

impl MouseScrollEvent

pub const VERTICAL_AXIS: i32