Struct web_sys::Navigator

#[repr(transparent)]
pub struct Navigator { /* fields omitted */ }

The Navigator class.

MDN Documentation

This API requires the following crate features to be activated: Navigator

Implementations

impl Navigator

pub fn permissions(&self) -> Result<Permissions, JsValue>

Getter for the permissions field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, Permissions

impl Navigator

pub fn mime_types(&self) -> Result<MimeTypeArray, JsValue>

Getter for the mimeTypes field of this object.

MDN Documentation

This API requires the following crate features to be activated: MimeTypeArray, Navigator

impl Navigator

pub fn plugins(&self) -> Result<PluginArray, JsValue>

Getter for the plugins field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, PluginArray

impl Navigator

pub fn do_not_track(&self) -> String

Getter for the doNotTrack field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn max_touch_points(&self) -> i32

Getter for the maxTouchPoints field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn media_capabilities(&self) -> MediaCapabilities

Getter for the mediaCapabilities field of this object.

MDN Documentation

This API requires the following crate features to be activated: MediaCapabilities, Navigator

impl Navigator

pub fn connection(&self) -> Result<NetworkInformation, JsValue>

Getter for the connection field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, NetworkInformation

impl Navigator

pub fn active_vr_displays(&self) -> Array

Getter for the activeVRDisplays field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn media_devices(&self) -> Result<MediaDevices, JsValue>

Getter for the mediaDevices field of this object.

MDN Documentation

This API requires the following crate features to be activated: MediaDevices, Navigator

impl Navigator

pub fn service_worker(&self) -> ServiceWorkerContainer

Getter for the serviceWorker field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, ServiceWorkerContainer

impl Navigator

pub fn presentation(&self) -> Result<Option<Presentation>, JsValue>

Getter for the presentation field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, Presentation

impl Navigator

pub fn credentials(&self) -> CredentialsContainer

Getter for the credentials field of this object.

MDN Documentation

This API requires the following crate features to be activated: CredentialsContainer, Navigator

impl Navigator

pub fn bluetooth(&self) -> Option<Bluetooth>

Getter for the bluetooth field of this object.

MDN Documentation

This API requires the following crate features to be activated: Bluetooth, Navigator

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Navigator

pub fn clipboard(&self) -> Option<Clipboard>

Getter for the clipboard field of this object.

MDN Documentation

This API requires the following crate features to be activated: Clipboard, Navigator

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Navigator

pub fn wake_lock(&self) -> WakeLock

Getter for the wakeLock field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, WakeLock

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Navigator

pub fn hid(&self) -> Hid

Getter for the hid field of this object.

MDN Documentation

This API requires the following crate features to be activated: Hid, Navigator

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Navigator

pub fn usb(&self) -> Usb

Getter for the usb field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, Usb

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Navigator

pub fn xr(&self) -> Xr

Getter for the xr field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, Xr

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Navigator

pub fn hardware_concurrency(&self) -> f64

Getter for the hardwareConcurrency field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn gpu(&self) -> Gpu

Getter for the gpu field of this object.

MDN Documentation

This API requires the following crate features to be activated: Gpu, Navigator

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Navigator

pub fn geolocation(&self) -> Result<Geolocation, JsValue>

Getter for the geolocation field of this object.

MDN Documentation

This API requires the following crate features to be activated: Geolocation, Navigator

impl Navigator

pub fn app_code_name(&self) -> Result<String, JsValue>

Getter for the appCodeName field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn app_name(&self) -> String

Getter for the appName field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn app_version(&self) -> Result<String, JsValue>

Getter for the appVersion field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn platform(&self) -> Result<String, JsValue>

Getter for the platform field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn user_agent(&self) -> Result<String, JsValue>

Getter for the userAgent field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn product(&self) -> String

Getter for the product field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn language(&self) -> Option<String>

Getter for the language field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn languages(&self) -> Array

Getter for the languages field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn on_line(&self) -> bool

Getter for the onLine field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn storage(&self) -> StorageManager

Getter for the storage field of this object.

MDN Documentation

This API requires the following crate features to be activated: Navigator, StorageManager

impl Navigator

pub fn get_gamepads(&self) -> Result<Array, JsValue>

The getGamepads() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn get_vr_displays(&self) -> Result<Promise, JsValue>

The getVRDisplays() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn request_gamepad_service_test(&self) -> GamepadServiceTest

The requestGamepadServiceTest() method.

MDN Documentation

This API requires the following crate features to be activated: GamepadServiceTest, Navigator

impl Navigator

pub fn request_midi_access(&self) -> Result<Promise, JsValue>

The requestMIDIAccess() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn request_midi_access_with_options(
&self,
options: &MidiOptions
) -> Result<Promise, JsValue>

The requestMIDIAccess() method.

MDN Documentation

This API requires the following crate features to be activated: MidiOptions, Navigator

impl Navigator

pub fn request_media_key_system_access(
&self,
key_system: &str,
supported_configurations: &JsValue
) -> Promise

The requestMediaKeySystemAccess() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn request_vr_service_test(&self) -> VrServiceTest

The requestVRServiceTest() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator, VrServiceTest

impl Navigator

pub fn send_beacon(&self, url: &str) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn send_beacon_with_opt_blob(
&self,
url: &str,
data: Option<&Blob>
) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: Blob, Navigator

impl Navigator

pub fn send_beacon_with_opt_buffer_source(
&self,
url: &str,
data: Option<&Object>
) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn send_beacon_with_opt_u8_array(
&self,
url: &str,
data: Option<&mut [u8]>
) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn send_beacon_with_opt_form_data(
&self,
url: &str,
data: Option<&FormData>
) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: FormData, Navigator

impl Navigator

pub fn send_beacon_with_opt_url_search_params(
&self,
url: &str,
data: Option<&UrlSearchParams>
) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator, UrlSearchParams

impl Navigator

pub fn send_beacon_with_opt_str(
&self,
url: &str,
data: Option<&str>
) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn send_beacon_with_opt_readable_stream(
&self,
url: &str,
data: Option<&ReadableStream>
) -> Result<bool, JsValue>

The sendBeacon() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator, ReadableStream

impl Navigator

pub fn vibrate_with_duration(&self, duration: u32) -> bool

The vibrate() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn vibrate_with_pattern(&self, pattern: &JsValue) -> bool

The vibrate() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn register_content_handler(
&self,
mime_type: &str,
url: &str,
title: &str
) -> Result<(), JsValue>

The registerContentHandler() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn register_protocol_handler(
&self,
scheme: &str,
url: &str,
title: &str
) -> Result<(), JsValue>

The registerProtocolHandler() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator

impl Navigator

pub fn taint_enabled(&self) -> bool

The taintEnabled() method.

MDN Documentation

This API requires the following crate features to be activated: Navigator