Struct web_sys::NetworkCommandOptions

#[repr(transparent)]
pub struct NetworkCommandOptions { /* private fields */ }

The NetworkCommandOptions dictionary.

This API requires the following crate features to be activated: NetworkCommandOptions

Implementations

impl NetworkCommandOptions

pub fn new() -> Self

Construct a new NetworkCommandOptions.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn cmd(&mut self, val: &str) -> &mut Self

Change the cmd field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn cur_external_ifname(&mut self, val: &str) -> &mut Self

Change the curExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn cur_internal_ifname(&mut self, val: &str) -> &mut Self

Change the curInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn dns1(&mut self, val: &str) -> &mut Self

Change the dns1 field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn dns1_long(&mut self, val: i32) -> &mut Self

Change the dns1_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn dns2(&mut self, val: &str) -> &mut Self

Change the dns2 field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn dns2_long(&mut self, val: i32) -> &mut Self

Change the dns2_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn dnses(&mut self, val: &JsValue) -> &mut Self

Change the dnses field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn domain(&mut self, val: &str) -> &mut Self

Change the domain field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn enable(&mut self, val: bool) -> &mut Self

Change the enable field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn enabled(&mut self, val: bool) -> &mut Self

Change the enabled field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn end_ip(&mut self, val: &str) -> &mut Self

Change the endIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn external_ifname(&mut self, val: &str) -> &mut Self

Change the externalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn gateway(&mut self, val: &str) -> &mut Self

Change the gateway field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn gateway_long(&mut self, val: i32) -> &mut Self

Change the gateway_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn gateways(&mut self, val: &JsValue) -> &mut Self

Change the gateways field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn id(&mut self, val: i32) -> &mut Self

Change the id field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn ifname(&mut self, val: &str) -> &mut Self

Change the ifname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn interface_list(&mut self, val: &JsValue) -> &mut Self

Change the interfaceList field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn internal_ifname(&mut self, val: &str) -> &mut Self

Change the internalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn ip(&mut self, val: &str) -> &mut Self

Change the ip field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn ipaddr(&mut self, val: i32) -> &mut Self

Change the ipaddr field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn key(&mut self, val: &str) -> &mut Self

Change the key field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn link(&mut self, val: &str) -> &mut Self

Change the link field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn mask(&mut self, val: i32) -> &mut Self

Change the mask field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn mask_length(&mut self, val: &str) -> &mut Self

Change the maskLength field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn mode(&mut self, val: &str) -> &mut Self

Change the mode field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn mtu(&mut self, val: i32) -> &mut Self

Change the mtu field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn pre_external_ifname(&mut self, val: &str) -> &mut Self

Change the preExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn pre_internal_ifname(&mut self, val: &str) -> &mut Self

Change the preInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn prefix(&mut self, val: &str) -> &mut Self

Change the prefix field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn prefix_length(&mut self, val: u32) -> &mut Self

Change the prefixLength field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn report(&mut self, val: bool) -> &mut Self

Change the report field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn security(&mut self, val: &str) -> &mut Self

Change the security field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn server_ip(&mut self, val: &str) -> &mut Self

Change the serverIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn ssid(&mut self, val: &str) -> &mut Self

Change the ssid field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn start_ip(&mut self, val: &str) -> &mut Self

Change the startIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn threshold(&mut self, val: f64) -> &mut Self

Change the threshold field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn usb_end_ip(&mut self, val: &str) -> &mut Self

Change the usbEndIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn usb_start_ip(&mut self, val: &str) -> &mut Self

Change the usbStartIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn wifi_end_ip(&mut self, val: &str) -> &mut Self

Change the wifiEndIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn wifi_start_ip(&mut self, val: &str) -> &mut Self

Change the wifiStartIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn wifictrlinterfacename(&mut self, val: &str) -> &mut Self

Change the wifictrlinterfacename field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions