web_sys
Struct NetworkResultOptions

pub struct NetworkResultOptions { /* private fields */ }

The NetworkResultOptions dictionary.

This API requires the following crate features to be activated: NetworkResultOptions

Implementations

impl NetworkResultOptions

pub fn get_broadcast(&self) -> Option<bool>

Get the broadcast field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_broadcast(&self, val: bool)

Change the broadcast field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_cur_external_ifname(&self) -> Option<String>

Get the curExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_cur_external_ifname(&self, val: &str)

Change the curExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_cur_internal_ifname(&self) -> Option<String>

Get the curInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_cur_internal_ifname(&self, val: &str)

Change the curInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_dns1(&self) -> Option<i32>

Get the dns1 field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_dns1(&self, val: i32)

Change the dns1 field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_dns1_str(&self) -> Option<String>

Get the dns1_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_dns1_str(&self, val: &str)

Change the dns1_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_dns2(&self) -> Option<i32>

Get the dns2 field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_dns2(&self, val: i32)

Change the dns2 field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_dns2_str(&self) -> Option<String>

Get the dns2_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_dns2_str(&self, val: &str)

Change the dns2_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_enable(&self) -> Option<bool>

Get the enable field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_enable(&self, val: bool)

Change the enable field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_error(&self) -> Option<bool>

Get the error field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_error(&self, val: bool)

Change the error field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_flag(&self) -> Option<String>

Get the flag field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_flag(&self, val: &str)

Change the flag field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_gateway(&self) -> Option<i32>

Get the gateway field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_gateway(&self, val: i32)

Change the gateway field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_gateway_str(&self) -> Option<String>

Get the gateway_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_gateway_str(&self, val: &str)

Change the gateway_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_id(&self) -> Option<i32>

Get the id field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_id(&self, val: i32)

Change the id field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_interface_list(&self) -> Option<Array>

Get the interfaceList field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_interface_list(&self, val: &JsValue)

Change the interfaceList field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_ip_addr(&self) -> Option<String>

Get the ipAddr field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_ip_addr(&self, val: &str)

Change the ipAddr field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_ipaddr(&self) -> Option<i32>

Get the ipaddr field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_ipaddr(&self, val: i32)

Change the ipaddr field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_ipaddr_str(&self) -> Option<String>

Get the ipaddr_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_ipaddr_str(&self, val: &str)

Change the ipaddr_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_lease(&self) -> Option<i32>

Get the lease field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_lease(&self, val: i32)

Change the lease field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_mac_addr(&self) -> Option<String>

Get the macAddr field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_mac_addr(&self, val: &str)

Change the macAddr field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_mask(&self) -> Option<i32>

Get the mask field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_mask(&self, val: i32)

Change the mask field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_mask_str(&self) -> Option<String>

Get the mask_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_mask_str(&self, val: &str)

Change the mask_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_net_id(&self) -> Option<String>

Get the netId field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_net_id(&self, val: &str)

Change the netId field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_prefix_length(&self) -> Option<i32>

Get the prefixLength field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_prefix_length(&self, val: i32)

Change the prefixLength field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_reason(&self) -> Option<String>

Get the reason field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_reason(&self, val: &str)

Change the reason field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_reply(&self) -> Option<String>

Get the reply field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_reply(&self, val: &str)

Change the reply field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_result(&self) -> Option<bool>

Get the result field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_result(&self, val: bool)

Change the result field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_result_code(&self) -> Option<i32>

Get the resultCode field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_result_code(&self, val: i32)

Change the resultCode field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_result_reason(&self) -> Option<String>

Get the resultReason field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_result_reason(&self, val: &str)

Change the resultReason field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_ret(&self) -> Option<bool>

Get the ret field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_ret(&self, val: bool)

Change the ret field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_route(&self) -> Option<String>

Get the route field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_route(&self, val: &str)

Change the route field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_server(&self) -> Option<i32>

Get the server field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_server(&self, val: i32)

Change the server field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_server_str(&self) -> Option<String>

Get the server_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_server_str(&self, val: &str)

Change the server_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_success(&self) -> Option<bool>

Get the success field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_success(&self, val: bool)

Change the success field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_topic(&self) -> Option<String>

Get the topic field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_topic(&self, val: &str)

Change the topic field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn get_vendor_str(&self) -> Option<String>

Get the vendor_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn set_vendor_str(&self, val: &str)

Change the vendor_str field of this object.

This API requires the following crate features to be activated: NetworkResultOptions

impl NetworkResultOptions

pub fn new() -> Self

Construct a new NetworkResultOptions.

This API requires the following crate features to be activated: NetworkResultOptions

pub fn broadcast(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_broadcast() instead.

pub fn cur_external_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_cur_external_ifname() instead.

pub fn cur_internal_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_cur_internal_ifname() instead.

pub fn dns1(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_dns1() instead.

pub fn dns1_str(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_dns1_str() instead.

pub fn dns2(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_dns2() instead.

pub fn dns2_str(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_dns2_str() instead.

pub fn enable(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_enable() instead.

pub fn error(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_error() instead.

pub fn flag(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_flag() instead.

pub fn gateway(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_gateway() instead.

pub fn gateway_str(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_gateway_str() instead.

pub fn id(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_id() instead.

pub fn interface_list(&mut self, val: &JsValue) -> &mut Self

👎Deprecated: Use set_interface_list() instead.

pub fn ip_addr(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_ip_addr() instead.

pub fn ipaddr(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_ipaddr() instead.

pub fn ipaddr_str(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_ipaddr_str() instead.

pub fn lease(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_lease() instead.

pub fn mac_addr(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_mac_addr() instead.

pub fn mask(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_mask() instead.

pub fn mask_str(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_mask_str() instead.

pub fn net_id(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_net_id() instead.

pub fn prefix_length(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_prefix_length() instead.

pub fn reason(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_reason() instead.

pub fn reply(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_reply() instead.

pub fn result(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_result() instead.

pub fn result_code(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_result_code() instead.

pub fn result_reason(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_result_reason() instead.

pub fn ret(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_ret() instead.

pub fn route(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_route() instead.

pub fn server(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_server() instead.

pub fn server_str(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_server_str() instead.

pub fn success(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_success() instead.

pub fn topic(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_topic() instead.

pub fn vendor_str(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_vendor_str() instead.