Struct web_sys::OfflineAudioContext

#[repr(transparent)]
pub struct OfflineAudioContext { /* private fields */ }

The OfflineAudioContext class.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

Implementations

impl OfflineAudioContext

pub fn length(&self) -> u32

Getter for the length field of this object.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn oncomplete(&self) -> Option<Function>

Getter for the oncomplete field of this object.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn set_oncomplete(&self, value: Option<&Function>)

Setter for the oncomplete field of this object.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn destination(&self) -> AudioDestinationNode

Getter for the destination field of this object.

MDN Documentation

This API requires the following crate features to be activated: AudioDestinationNode, OfflineAudioContext

impl OfflineAudioContext

pub fn sample_rate(&self) -> f32

Getter for the sampleRate field of this object.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn current_time(&self) -> f64

Getter for the currentTime field of this object.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn listener(&self) -> AudioListener

Getter for the listener field of this object.

MDN Documentation

This API requires the following crate features to be activated: AudioListener, OfflineAudioContext

impl OfflineAudioContext

pub fn state(&self) -> AudioContextState

Getter for the state field of this object.

MDN Documentation

This API requires the following crate features to be activated: AudioContextState, OfflineAudioContext

impl OfflineAudioContext

pub fn audio_worklet(&self) -> Result<AudioWorklet, JsValue>

Getter for the audioWorklet field of this object.

MDN Documentation

This API requires the following crate features to be activated: AudioWorklet, OfflineAudioContext

impl OfflineAudioContext

pub fn onstatechange(&self) -> Option<Function>

Getter for the onstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn set_onstatechange(&self, value: Option<&Function>)

Setter for the onstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn new_with_context_options( context_options: &OfflineAudioContextOptions ) -> Result<OfflineAudioContext, JsValue>

The new OfflineAudioContext(..) constructor, creating a new instance of OfflineAudioContext.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, OfflineAudioContextOptions

impl OfflineAudioContext

pub fn new_with_number_of_channels_and_length_and_sample_rate( number_of_channels: u32, length: u32, sample_rate: f32 ) -> Result<OfflineAudioContext, JsValue>

The new OfflineAudioContext(..) constructor, creating a new instance of OfflineAudioContext.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn start_rendering(&self) -> Result<Promise, JsValue>

The startRendering() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn create_analyser(&self) -> Result<AnalyserNode, JsValue>

The createAnalyser() method.

MDN Documentation

This API requires the following crate features to be activated: AnalyserNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_biquad_filter(&self) -> Result<BiquadFilterNode, JsValue>

The createBiquadFilter() method.

MDN Documentation

This API requires the following crate features to be activated: BiquadFilterNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_buffer( &self, number_of_channels: u32, length: u32, sample_rate: f32 ) -> Result<AudioBuffer, JsValue>

The createBuffer() method.

MDN Documentation

This API requires the following crate features to be activated: AudioBuffer, OfflineAudioContext

impl OfflineAudioContext

pub fn create_buffer_source(&self) -> Result<AudioBufferSourceNode, JsValue>

The createBufferSource() method.

MDN Documentation

This API requires the following crate features to be activated: AudioBufferSourceNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_channel_merger(&self) -> Result<ChannelMergerNode, JsValue>

The createChannelMerger() method.

MDN Documentation

This API requires the following crate features to be activated: ChannelMergerNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_channel_merger_with_number_of_inputs( &self, number_of_inputs: u32 ) -> Result<ChannelMergerNode, JsValue>

The createChannelMerger() method.

MDN Documentation

This API requires the following crate features to be activated: ChannelMergerNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_channel_splitter(&self) -> Result<ChannelSplitterNode, JsValue>

The createChannelSplitter() method.

MDN Documentation

This API requires the following crate features to be activated: ChannelSplitterNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_channel_splitter_with_number_of_outputs( &self, number_of_outputs: u32 ) -> Result<ChannelSplitterNode, JsValue>

The createChannelSplitter() method.

MDN Documentation

This API requires the following crate features to be activated: ChannelSplitterNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_constant_source(&self) -> Result<ConstantSourceNode, JsValue>

The createConstantSource() method.

MDN Documentation

This API requires the following crate features to be activated: ConstantSourceNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_convolver(&self) -> Result<ConvolverNode, JsValue>

The createConvolver() method.

MDN Documentation

This API requires the following crate features to be activated: ConvolverNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_delay(&self) -> Result<DelayNode, JsValue>

The createDelay() method.

MDN Documentation

This API requires the following crate features to be activated: DelayNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_delay_with_max_delay_time( &self, max_delay_time: f64 ) -> Result<DelayNode, JsValue>

The createDelay() method.

MDN Documentation

This API requires the following crate features to be activated: DelayNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_dynamics_compressor( &self ) -> Result<DynamicsCompressorNode, JsValue>

The createDynamicsCompressor() method.

MDN Documentation

This API requires the following crate features to be activated: DynamicsCompressorNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_gain(&self) -> Result<GainNode, JsValue>

The createGain() method.

MDN Documentation

This API requires the following crate features to be activated: GainNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_iir_filter( &self, feedforward: &JsValue, feedback: &JsValue ) -> Result<IirFilterNode, JsValue>

The createIIRFilter() method.

MDN Documentation

This API requires the following crate features to be activated: IirFilterNode, OfflineAudioContext

impl OfflineAudioContext

pub fn create_oscillator(&self) -> Result<OscillatorNode, JsValue>

The createOscillator() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, OscillatorNode

impl OfflineAudioContext

pub fn create_panner(&self) -> Result<PannerNode, JsValue>

The createPanner() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, PannerNode

impl OfflineAudioContext

pub fn create_periodic_wave( &self, real: &mut [f32], imag: &mut [f32] ) -> Result<PeriodicWave, JsValue>

The createPeriodicWave() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, PeriodicWave

impl OfflineAudioContext

pub fn create_periodic_wave_with_constraints( &self, real: &mut [f32], imag: &mut [f32], constraints: &PeriodicWaveConstraints ) -> Result<PeriodicWave, JsValue>

The createPeriodicWave() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, PeriodicWave, PeriodicWaveConstraints

impl OfflineAudioContext

pub fn create_script_processor(&self) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, ScriptProcessorNode

impl OfflineAudioContext

pub fn create_script_processor_with_buffer_size( &self, buffer_size: u32 ) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, ScriptProcessorNode

impl OfflineAudioContext

pub fn create_script_processor_with_buffer_size_and_number_of_input_channels( &self, buffer_size: u32, number_of_input_channels: u32 ) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, ScriptProcessorNode

impl OfflineAudioContext

pub fn create_script_processor_with_buffer_size_and_number_of_input_channels_and_number_of_output_channels( &self, buffer_size: u32, number_of_input_channels: u32, number_of_output_channels: u32 ) -> Result<ScriptProcessorNode, JsValue>

The createScriptProcessor() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, ScriptProcessorNode

impl OfflineAudioContext

pub fn create_stereo_panner(&self) -> Result<StereoPannerNode, JsValue>

The createStereoPanner() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, StereoPannerNode

impl OfflineAudioContext

pub fn create_wave_shaper(&self) -> Result<WaveShaperNode, JsValue>

The createWaveShaper() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext, WaveShaperNode

impl OfflineAudioContext

pub fn decode_audio_data( &self, audio_data: &ArrayBuffer ) -> Result<Promise, JsValue>

The decodeAudioData() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn decode_audio_data_with_success_callback( &self, audio_data: &ArrayBuffer, success_callback: &Function ) -> Result<Promise, JsValue>

The decodeAudioData() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn decode_audio_data_with_success_callback_and_error_callback( &self, audio_data: &ArrayBuffer, success_callback: &Function, error_callback: &Function ) -> Result<Promise, JsValue>

The decodeAudioData() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext

impl OfflineAudioContext

pub fn resume(&self) -> Result<Promise, JsValue>

The resume() method.

MDN Documentation

This API requires the following crate features to be activated: OfflineAudioContext