Struct web_sys::OfflineResourceList

#[repr(transparent)]

pub struct OfflineResourceList { /* fields omitted */ }

The OfflineResourceList object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl OfflineResourceList

pub fn swap_cache(&self) -> Result<(), JsValue>

The swapCache() method

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn update(&self) -> Result<(), JsValue>

The update() method

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn status(&self) -> Result<u16, JsValue>

The status getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn onchecking(&self) -> Option<Function>

The onchecking getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_onchecking(&self, onchecking: Option<&Function>)

The onchecking setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn onnoupdate(&self) -> Option<Function>

The onnoupdate getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_onnoupdate(&self, onnoupdate: Option<&Function>)

The onnoupdate setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn ondownloading(&self) -> Option<Function>

The ondownloading getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_ondownloading(&self, ondownloading: Option<&Function>)

The ondownloading setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn onprogress(&self) -> Option<Function>

The onprogress getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_onprogress(&self, onprogress: Option<&Function>)

The onprogress setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn onupdateready(&self) -> Option<Function>

The onupdateready getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_onupdateready(&self, onupdateready: Option<&Function>)

The onupdateready setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn oncached(&self) -> Option<Function>

The oncached getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_oncached(&self, oncached: Option<&Function>)

The oncached setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn onobsolete(&self) -> Option<Function>

The onobsolete getter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub fn set_onobsolete(&self, onobsolete: Option<&Function>)

The onobsolete setter

MDN Documentation

This function requires the following crate features to be activated: OfflineResourceList

impl OfflineResourceList

pub const UNCACHED: u16

impl OfflineResourceList

pub const IDLE: u16

impl OfflineResourceList

pub const CHECKING: u16

impl OfflineResourceList

pub const DOWNLOADING: u16

impl OfflineResourceList

pub const UPDATEREADY: u16

impl OfflineResourceList

pub const OBSOLETE: u16