Struct web_sys::Performance

#[repr(transparent)]

pub struct Performance { /* fields omitted */ }

The Performance object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl Performance

pub fn clear_marks(&self)

The clearMarks() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn clear_marks_with_mark_name(&self, mark_name: &str)

The clearMarks() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn clear_measures(&self)

The clearMeasures() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn clear_measures_with_measure_name(&self, measure_name: &str)

The clearMeasures() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn clear_resource_timings(&self)

The clearResourceTimings() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn mark(&self, mark_name: &str) -> Result<(), JsValue>

The mark() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn measure(&self, measure_name: &str) -> Result<(), JsValue>

The measure() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn measure_with_start_mark(
&self,
measure_name: &str,
start_mark: &str
) -> Result<(), JsValue>

The measure() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn measure_with_start_mark_and_end_mark(
&self,
measure_name: &str,
start_mark: &str,
end_mark: &str
) -> Result<(), JsValue>

The measure() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn now(&self) -> f64

The now() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn set_resource_timing_buffer_size(&self, max_size: u32)

The setResourceTimingBufferSize() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn to_json(&self) -> Object

The toJSON() method

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn time_origin(&self) -> f64

The timeOrigin getter

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn timing(&self) -> PerformanceTiming

The timing getter

MDN Documentation

This function requires the following crate features to be activated: Performance, PerformanceTiming

impl Performance

pub fn navigation(&self) -> PerformanceNavigation

The navigation getter

MDN Documentation

This function requires the following crate features to be activated: Performance, PerformanceNavigation

impl Performance

pub fn onresourcetimingbufferfull(&self) -> Option<Function>

The onresourcetimingbufferfull getter

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn set_onresourcetimingbufferfull(
&self,
onresourcetimingbufferfull: Option<&Function>
)

The onresourcetimingbufferfull setter

MDN Documentation

This function requires the following crate features to be activated: Performance

impl Performance

pub fn moz_memory(&self) -> Object

The mozMemory getter

MDN Documentation

This function requires the following crate features to be activated: Performance