Struct web_sys::PointerEvent

#[repr(transparent)]

pub struct PointerEvent { /* fields omitted */ }

The PointerEvent object

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

Methods

impl PointerEvent

pub fn new(type_: &str) -> Result<PointerEvent, JsValue>

The new PointerEvent(..) constructor, creating a new instance of PointerEvent

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn new_with_event_init_dict(
type_: &str,
event_init_dict: &PointerEventInit
) -> Result<PointerEvent, JsValue>

The new PointerEvent(..) constructor, creating a new instance of PointerEvent

MDN Documentation

This API requires the following crate features to be activated: PointerEvent, PointerEventInit

impl PointerEvent

pub fn get_coalesced_events(&self) -> Array

The getCoalescedEvents() method

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn pointer_id(&self) -> i32

The pointerId getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn width(&self) -> i32

The width getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn height(&self) -> i32

The height getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn pressure(&self) -> f32

The pressure getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn tangential_pressure(&self) -> f32

The tangentialPressure getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn tilt_x(&self) -> i32

The tiltX getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn tilt_y(&self) -> i32

The tiltY getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn twist(&self) -> i32

The twist getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn pointer_type(&self) -> String

The pointerType getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent

impl PointerEvent

pub fn is_primary(&self) -> bool

The isPrimary getter

MDN Documentation

This API requires the following crate features to be activated: PointerEvent