Struct web_sys::ScrollAreaEvent

#[repr(transparent)]

pub struct ScrollAreaEvent { /* fields omitted */ }

The ScrollAreaEvent object

MDN Documentation

This function requires the following crate features to be activated: Event, UiEvent

Methods

impl ScrollAreaEvent

pub fn init_scroll_area_event(&self, type_: &str)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble(
&self,
type_: &str,
can_bubble: bool
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble_and_cancelable(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble_and_cancelable_and_view(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent, Window

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble_and_cancelable_and_view_and_detail(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent, Window

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble_and_cancelable_and_view_and_detail_and_x(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
x: f32
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent, Window

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble_and_cancelable_and_view_and_detail_and_x_and_y(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
x: f32,
y: f32
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent, Window

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble_and_cancelable_and_view_and_detail_and_x_and_y_and_width(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
x: f32,
y: f32,
width: f32
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent, Window

impl ScrollAreaEvent

pub fn init_scroll_area_event_with_can_bubble_and_cancelable_and_view_and_detail_and_x_and_y_and_width_and_height(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
x: f32,
y: f32,
width: f32,
height: f32
)

The initScrollAreaEvent() method

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent, Window

impl ScrollAreaEvent

pub fn x(&self) -> f32

The x getter

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent

impl ScrollAreaEvent

pub fn y(&self) -> f32

The y getter

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent

impl ScrollAreaEvent

pub fn width(&self) -> f32

The width getter

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent

impl ScrollAreaEvent

pub fn height(&self) -> f32

The height getter

MDN Documentation

This function requires the following crate features to be activated: ScrollAreaEvent