Struct web_sys::SecurityPolicyViolationEvent

#[repr(transparent)]

pub struct SecurityPolicyViolationEvent { /* fields omitted */ }

The SecurityPolicyViolationEvent object

MDN Documentation

This function requires the following crate features to be activated: Event

Methods

impl SecurityPolicyViolationEvent

pub fn new(type_: &str) -> Result<SecurityPolicyViolationEvent, JsValue>

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn new_with_event_init_dict(
type_: &str,
event_init_dict: &SecurityPolicyViolationEventInit
) -> Result<SecurityPolicyViolationEvent, JsValue>

This function requires the following crate features to be activated: SecurityPolicyViolationEvent, SecurityPolicyViolationEventInit

impl SecurityPolicyViolationEvent

pub fn document_uri(&self) -> String

The documentURI getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn referrer(&self) -> String

The referrer getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn blocked_uri(&self) -> String

The blockedURI getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn violated_directive(&self) -> String

The violatedDirective getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn effective_directive(&self) -> String

The effectiveDirective getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn original_policy(&self) -> String

The originalPolicy getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn source_file(&self) -> String

The sourceFile getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn sample(&self) -> String

The sample getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn disposition(&self) -> SecurityPolicyViolationEventDisposition

The disposition getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent, SecurityPolicyViolationEventDisposition

impl SecurityPolicyViolationEvent

pub fn status_code(&self) -> u16

The statusCode getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn line_number(&self) -> i32

The lineNumber getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent

impl SecurityPolicyViolationEvent

pub fn column_number(&self) -> i32

The columnNumber getter

MDN Documentation

This function requires the following crate features to be activated: SecurityPolicyViolationEvent