Struct web_sys::Selection

#[repr(transparent)]

pub struct Selection { /* fields omitted */ }

The Selection object

MDN Documentation

Methods

impl Selection

pub fn add_range(&self, range: &Range) -> Result<(), JsValue>

The addRange() method

MDN Documentation

This function requires the following crate features to be activated: Range, Selection

impl Selection

pub fn collapse(&self, node: Option<&Node>) -> Result<(), JsValue>

The collapse() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn collapse_with_offset(
&self,
node: Option<&Node>,
offset: u32
) -> Result<(), JsValue>

The collapse() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn collapse_to_end(&self) -> Result<(), JsValue>

The collapseToEnd() method

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn collapse_to_start(&self) -> Result<(), JsValue>

The collapseToStart() method

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn contains_node(&self, node: &Node) -> Result<bool, JsValue>

The containsNode() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn contains_node_with_allow_partial_containment(
&self,
node: &Node,
allow_partial_containment: bool
) -> Result<bool, JsValue>

The containsNode() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn delete_from_document(&self) -> Result<(), JsValue>

The deleteFromDocument() method

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn empty(&self) -> Result<(), JsValue>

The empty() method

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn extend(&self, node: &Node) -> Result<(), JsValue>

The extend() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn extend_with_offset(
&self,
node: &Node,
offset: u32
) -> Result<(), JsValue>

The extend() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn get_range_at(&self, index: u32) -> Result<Range, JsValue>

The getRangeAt() method

MDN Documentation

This function requires the following crate features to be activated: Range, Selection

impl Selection

pub fn modify(
&self,
alter: &str,
direction: &str,
granularity: &str
) -> Result<(), JsValue>

The modify() method

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn remove_all_ranges(&self) -> Result<(), JsValue>

The removeAllRanges() method

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn remove_range(&self, range: &Range) -> Result<(), JsValue>

The removeRange() method

MDN Documentation

This function requires the following crate features to be activated: Range, Selection

impl Selection

pub fn select_all_children(&self, node: &Node) -> Result<(), JsValue>

The selectAllChildren() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn set_base_and_extent(
&self,
anchor_node: &Node,
anchor_offset: u32,
focus_node: &Node,
focus_offset: u32
) -> Result<(), JsValue>

The setBaseAndExtent() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn set_position(&self, node: Option<&Node>) -> Result<(), JsValue>

The setPosition() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn set_position_with_offset(
&self,
node: Option<&Node>,
offset: u32
) -> Result<(), JsValue>

The setPosition() method

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn anchor_node(&self) -> Option<Node>

The anchorNode getter

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn anchor_offset(&self) -> u32

The anchorOffset getter

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn focus_node(&self) -> Option<Node>

The focusNode getter

MDN Documentation

This function requires the following crate features to be activated: Node, Selection

impl Selection

pub fn focus_offset(&self) -> u32

The focusOffset getter

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn is_collapsed(&self) -> bool

The isCollapsed getter

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn range_count(&self) -> u32

The rangeCount getter

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn type_(&self) -> String

The type getter

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn caret_bidi_level(&self) -> Result<Option<i16>, JsValue>

The caretBidiLevel getter

MDN Documentation

This function requires the following crate features to be activated: Selection

impl Selection

pub fn set_caret_bidi_level(
&self,
caret_bidi_level: Option<i16>
) -> Result<(), JsValue>

The caretBidiLevel setter

MDN Documentation

This function requires the following crate features to be activated: Selection