Struct web_sys::SourceBuffer

#[repr(transparent)]

pub struct SourceBuffer { /* fields omitted */ }

The SourceBuffer object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl SourceBuffer

pub fn abort(&self) -> Result<(), JsValue>

The abort() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_buffer_with_array_buffer(
&self,
data: &ArrayBuffer
) -> Result<(), JsValue>

The appendBuffer() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_buffer_with_array_buffer_view(
&self,
data: &Object
) -> Result<(), JsValue>

The appendBuffer() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_buffer_with_u8_array(
&self,
data: &mut [u8]
) -> Result<(), JsValue>

The appendBuffer() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_buffer_async_with_array_buffer(
&self,
data: &ArrayBuffer
) -> Result<Promise, JsValue>

The appendBufferAsync() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_buffer_async_with_array_buffer_view(
&self,
data: &Object
) -> Result<Promise, JsValue>

The appendBufferAsync() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_buffer_async_with_u8_array(
&self,
data: &mut [u8]
) -> Result<Promise, JsValue>

The appendBufferAsync() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn change_type(&self, type_: &str) -> Result<(), JsValue>

The changeType() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn remove(&self, start: f64, end: f64) -> Result<(), JsValue>

The remove() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn remove_async(&self, start: f64, end: f64) -> Result<Promise, JsValue>

The removeAsync() method

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn mode(&self) -> SourceBufferAppendMode

The mode getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer, SourceBufferAppendMode

impl SourceBuffer

pub fn set_mode(&self, mode: SourceBufferAppendMode)

The mode setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer, SourceBufferAppendMode

impl SourceBuffer

pub fn updating(&self) -> bool

The updating getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn buffered(&self) -> Result<TimeRanges, JsValue>

The buffered getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer, TimeRanges

impl SourceBuffer

pub fn timestamp_offset(&self) -> f64

The timestampOffset getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_timestamp_offset(&self, timestamp_offset: f64)

The timestampOffset setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_window_start(&self) -> f64

The appendWindowStart getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_append_window_start(&self, append_window_start: f64)

The appendWindowStart setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn append_window_end(&self) -> f64

The appendWindowEnd getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_append_window_end(&self, append_window_end: f64)

The appendWindowEnd setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn onupdatestart(&self) -> Option<Function>

The onupdatestart getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_onupdatestart(&self, onupdatestart: Option<&Function>)

The onupdatestart setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn onupdate(&self) -> Option<Function>

The onupdate getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_onupdate(&self, onupdate: Option<&Function>)

The onupdate setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn onupdateend(&self) -> Option<Function>

The onupdateend getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_onupdateend(&self, onupdateend: Option<&Function>)

The onupdateend setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn onabort(&self) -> Option<Function>

The onabort getter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer

impl SourceBuffer

pub fn set_onabort(&self, onabort: Option<&Function>)

The onabort setter

MDN Documentation

This function requires the following crate features to be activated: SourceBuffer