Struct web_sys::SpeechRecognition

#[repr(transparent)]

pub struct SpeechRecognition { /* fields omitted */ }

The SpeechRecognition object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl SpeechRecognition

pub fn new() -> Result<SpeechRecognition, JsValue>

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn abort(&self)

The abort() method

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn start(&self) -> Result<(), JsValue>

The start() method

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn start_with_stream(&self, stream: &MediaStream) -> Result<(), JsValue>

The start() method

MDN Documentation

This function requires the following crate features to be activated: MediaStream, SpeechRecognition

impl SpeechRecognition

pub fn stop(&self)

The stop() method

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn grammars(&self) -> SpeechGrammarList

The grammars getter

MDN Documentation

This function requires the following crate features to be activated: SpeechGrammarList, SpeechRecognition

impl SpeechRecognition

pub fn set_grammars(&self, grammars: &SpeechGrammarList)

The grammars setter

MDN Documentation

This function requires the following crate features to be activated: SpeechGrammarList, SpeechRecognition

impl SpeechRecognition

pub fn lang(&self) -> String

The lang getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_lang(&self, lang: &str)

The lang setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn continuous(&self) -> Result<bool, JsValue>

The continuous getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_continuous(&self, continuous: bool) -> Result<(), JsValue>

The continuous setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn interim_results(&self) -> bool

The interimResults getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_interim_results(&self, interim_results: bool)

The interimResults setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn max_alternatives(&self) -> u32

The maxAlternatives getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_max_alternatives(&self, max_alternatives: u32)

The maxAlternatives setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn service_uri(&self) -> Result<String, JsValue>

The serviceURI getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_service_uri(&self, service_uri: &str) -> Result<(), JsValue>

The serviceURI setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onaudiostart(&self) -> Option<Function>

The onaudiostart getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onaudiostart(&self, onaudiostart: Option<&Function>)

The onaudiostart setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onsoundstart(&self) -> Option<Function>

The onsoundstart getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onsoundstart(&self, onsoundstart: Option<&Function>)

The onsoundstart setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onspeechstart(&self) -> Option<Function>

The onspeechstart getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onspeechstart(&self, onspeechstart: Option<&Function>)

The onspeechstart setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onspeechend(&self) -> Option<Function>

The onspeechend getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onspeechend(&self, onspeechend: Option<&Function>)

The onspeechend setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onsoundend(&self) -> Option<Function>

The onsoundend getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onsoundend(&self, onsoundend: Option<&Function>)

The onsoundend setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onaudioend(&self) -> Option<Function>

The onaudioend getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onaudioend(&self, onaudioend: Option<&Function>)

The onaudioend setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onresult(&self) -> Option<Function>

The onresult getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onresult(&self, onresult: Option<&Function>)

The onresult setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onnomatch(&self) -> Option<Function>

The onnomatch getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onnomatch(&self, onnomatch: Option<&Function>)

The onnomatch setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onstart(&self) -> Option<Function>

The onstart getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onstart(&self, onstart: Option<&Function>)

The onstart setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn onend(&self) -> Option<Function>

The onend getter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition

impl SpeechRecognition

pub fn set_onend(&self, onend: Option<&Function>)

The onend setter

MDN Documentation

This function requires the following crate features to be activated: SpeechRecognition