Struct web_sys::SpeechSynthesisUtterance

#[repr(transparent)]

pub struct SpeechSynthesisUtterance { /* fields omitted */ }

The SpeechSynthesisUtterance object

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

Methods

impl SpeechSynthesisUtterance

pub fn new() -> Result<SpeechSynthesisUtterance, JsValue>

The new SpeechSynthesisUtterance(..) constructor, creating a new instance of SpeechSynthesisUtterance

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn new_with_text(text: &str) -> Result<SpeechSynthesisUtterance, JsValue>

The new SpeechSynthesisUtterance(..) constructor, creating a new instance of SpeechSynthesisUtterance

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn text(&self) -> String

The text getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_text(&self, text: &str)

The text setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn lang(&self) -> String

The lang getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_lang(&self, lang: &str)

The lang setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn voice(&self) -> Option<SpeechSynthesisVoice>

The voice getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance, SpeechSynthesisVoice

impl SpeechSynthesisUtterance

pub fn set_voice(&self, voice: Option<&SpeechSynthesisVoice>)

The voice setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance, SpeechSynthesisVoice

impl SpeechSynthesisUtterance

pub fn volume(&self) -> f32

The volume getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_volume(&self, volume: f32)

The volume setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn rate(&self) -> f32

The rate getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_rate(&self, rate: f32)

The rate setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn pitch(&self) -> f32

The pitch getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_pitch(&self, pitch: f32)

The pitch setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn onstart(&self) -> Option<Function>

The onstart getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_onstart(&self, onstart: Option<&Function>)

The onstart setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn onend(&self) -> Option<Function>

The onend getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_onend(&self, onend: Option<&Function>)

The onend setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn onpause(&self) -> Option<Function>

The onpause getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_onpause(&self, onpause: Option<&Function>)

The onpause setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn onresume(&self) -> Option<Function>

The onresume getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_onresume(&self, onresume: Option<&Function>)

The onresume setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn onmark(&self) -> Option<Function>

The onmark getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_onmark(&self, onmark: Option<&Function>)

The onmark setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn onboundary(&self) -> Option<Function>

The onboundary getter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance

impl SpeechSynthesisUtterance

pub fn set_onboundary(&self, onboundary: Option<&Function>)

The onboundary setter

MDN Documentation

This API requires the following crate features to be activated: SpeechSynthesisUtterance