Struct web_sys::SvgGraphicsElement

#[repr(transparent)]

pub struct SvgGraphicsElement { /* fields omitted */ }

The SVGGraphicsElement object

MDN Documentation

This function requires the following crate features to be activated: Element, EventTarget, Node, SvgElement

Methods

impl SvgGraphicsElement

pub fn get_b_box(&self) -> Result<SvgRect, JsValue>

The getBBox() method

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement, SvgRect

impl SvgGraphicsElement

pub fn get_b_box_with_a_options(
&self,
a_options: &SvgBoundingBoxOptions
) -> Result<SvgRect, JsValue>

The getBBox() method

MDN Documentation

This function requires the following crate features to be activated: SvgBoundingBoxOptions, SvgGraphicsElement, SvgRect

impl SvgGraphicsElement

pub fn get_ctm(&self) -> Option<SvgMatrix>

The getCTM() method

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement, SvgMatrix

impl SvgGraphicsElement

pub fn get_screen_ctm(&self) -> Option<SvgMatrix>

The getScreenCTM() method

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement, SvgMatrix

impl SvgGraphicsElement

pub fn get_transform_to_element(
&self,
element: &SvgGraphicsElement
) -> Result<SvgMatrix, JsValue>

The getTransformToElement() method

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement, SvgMatrix

impl SvgGraphicsElement

pub fn transform(&self) -> SvgAnimatedTransformList

The transform getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedTransformList, SvgGraphicsElement

impl SvgGraphicsElement

pub fn nearest_viewport_element(&self) -> Option<SvgElement>

The nearestViewportElement getter

MDN Documentation

This function requires the following crate features to be activated: SvgElement, SvgGraphicsElement

impl SvgGraphicsElement

pub fn farthest_viewport_element(&self) -> Option<SvgElement>

The farthestViewportElement getter

MDN Documentation

This function requires the following crate features to be activated: SvgElement, SvgGraphicsElement

impl SvgGraphicsElement

pub fn has_extension(&self, extension: &str) -> bool

The hasExtension() method

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement

impl SvgGraphicsElement

pub fn required_features(&self) -> SvgStringList

The requiredFeatures getter

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement, SvgStringList

impl SvgGraphicsElement

pub fn required_extensions(&self) -> SvgStringList

The requiredExtensions getter

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement, SvgStringList

impl SvgGraphicsElement

pub fn system_language(&self) -> SvgStringList

The systemLanguage getter

MDN Documentation

This function requires the following crate features to be activated: SvgGraphicsElement, SvgStringList