Struct web_sys::SvgLineElement

#[repr(transparent)]
pub struct SvgLineElement { /* fields omitted */ }

The SvgLineElement class.

MDN Documentation

This API requires the following crate features to be activated: SvgLineElement

Implementations

impl SvgLineElement

pub fn x1(&self) -> SvgAnimatedLength

Getter for the x1 field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgLineElement

impl SvgLineElement

pub fn y1(&self) -> SvgAnimatedLength

Getter for the y1 field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgLineElement

impl SvgLineElement

pub fn x2(&self) -> SvgAnimatedLength

Getter for the x2 field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgLineElement

impl SvgLineElement

pub fn y2(&self) -> SvgAnimatedLength

Getter for the y2 field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgLineElement