web_sys
Struct SvgMarkerElement

pub struct SvgMarkerElement { /* private fields */ }

The SvgMarkerElement class.

MDN Documentation

This API requires the following crate features to be activated: SvgMarkerElement

Implementations

impl SvgMarkerElement

pub fn ref_x(&self) -> SvgAnimatedLength

Getter for the refX field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgMarkerElement

impl SvgMarkerElement

pub fn ref_y(&self) -> SvgAnimatedLength

Getter for the refY field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgMarkerElement

impl SvgMarkerElement

pub fn marker_units(&self) -> SvgAnimatedEnumeration

Getter for the markerUnits field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedEnumeration, SvgMarkerElement

impl SvgMarkerElement

pub fn marker_width(&self) -> SvgAnimatedLength

Getter for the markerWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgMarkerElement

impl SvgMarkerElement

pub fn marker_height(&self) -> SvgAnimatedLength

Getter for the markerHeight field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgMarkerElement

impl SvgMarkerElement

pub fn orient_type(&self) -> SvgAnimatedEnumeration

Getter for the orientType field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedEnumeration, SvgMarkerElement

impl SvgMarkerElement

pub fn orient_angle(&self) -> SvgAnimatedAngle

Getter for the orientAngle field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedAngle, SvgMarkerElement

impl SvgMarkerElement

pub fn view_box(&self) -> SvgAnimatedRect

Getter for the viewBox field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedRect, SvgMarkerElement

impl SvgMarkerElement

pub fn preserve_aspect_ratio(&self) -> SvgAnimatedPreserveAspectRatio

Getter for the preserveAspectRatio field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedPreserveAspectRatio, SvgMarkerElement

impl SvgMarkerElement

pub fn set_orient_to_angle(&self, angle: &SvgAngle) -> Result<(), JsValue>

The setOrientToAngle() method.

MDN Documentation

This API requires the following crate features to be activated: SvgAngle, SvgMarkerElement

impl SvgMarkerElement

pub fn set_orient_to_auto(&self)

The setOrientToAuto() method.

MDN Documentation

This API requires the following crate features to be activated: SvgMarkerElement

impl SvgMarkerElement

pub const SVG_MARKERUNITS_UNKNOWN: u16 = 0u16

The SVGMarkerElement.SVG_MARKERUNITS_UNKNOWN const.

This API requires the following crate features to be activated: SvgMarkerElement

pub const SVG_MARKERUNITS_USERSPACEONUSE: u16 = 1u16

The SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE const.

This API requires the following crate features to be activated: SvgMarkerElement

pub const SVG_MARKERUNITS_STROKEWIDTH: u16 = 2u16

The SVGMarkerElement.SVG_MARKERUNITS_STROKEWIDTH const.

This API requires the following crate features to be activated: SvgMarkerElement

pub const SVG_MARKER_ORIENT_UNKNOWN: u16 = 0u16

The SVGMarkerElement.SVG_MARKER_ORIENT_UNKNOWN const.

This API requires the following crate features to be activated: SvgMarkerElement

pub const SVG_MARKER_ORIENT_AUTO: u16 = 1u16

The SVGMarkerElement.SVG_MARKER_ORIENT_AUTO const.

This API requires the following crate features to be activated: SvgMarkerElement

pub const SVG_MARKER_ORIENT_ANGLE: u16 = 2u16

The SVGMarkerElement.SVG_MARKER_ORIENT_ANGLE const.

This API requires the following crate features to be activated: SvgMarkerElement