Struct web_sys::SvgTextContentElement

#[repr(transparent)]
pub struct SvgTextContentElement { /* private fields */ }

The SvgTextContentElement class.

MDN Documentation

This API requires the following crate features to be activated: SvgTextContentElement

Implementations

impl SvgTextContentElement

pub fn text_length(&self) -> SvgAnimatedLength

Getter for the textLength field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgTextContentElement

impl SvgTextContentElement

pub fn length_adjust(&self) -> SvgAnimatedEnumeration

Getter for the lengthAdjust field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedEnumeration, SvgTextContentElement

impl SvgTextContentElement

pub fn get_char_num_at_position(&self, point: &SvgPoint) -> i32

The getCharNumAtPosition() method.

MDN Documentation

This API requires the following crate features to be activated: SvgPoint, SvgTextContentElement

impl SvgTextContentElement

pub fn get_computed_text_length(&self) -> f32

The getComputedTextLength() method.

MDN Documentation

This API requires the following crate features to be activated: SvgTextContentElement

impl SvgTextContentElement

pub fn get_end_position_of_char( &self, charnum: u32 ) -> Result<SvgPoint, JsValue>

The getEndPositionOfChar() method.

MDN Documentation

This API requires the following crate features to be activated: SvgPoint, SvgTextContentElement

impl SvgTextContentElement

pub fn get_extent_of_char(&self, charnum: u32) -> Result<SvgRect, JsValue>

The getExtentOfChar() method.

MDN Documentation

This API requires the following crate features to be activated: SvgRect, SvgTextContentElement

impl SvgTextContentElement

pub fn get_number_of_chars(&self) -> i32

The getNumberOfChars() method.

MDN Documentation

This API requires the following crate features to be activated: SvgTextContentElement

impl SvgTextContentElement

pub fn get_rotation_of_char(&self, charnum: u32) -> Result<f32, JsValue>

The getRotationOfChar() method.

MDN Documentation

This API requires the following crate features to be activated: SvgTextContentElement

impl SvgTextContentElement

pub fn get_start_position_of_char( &self, charnum: u32 ) -> Result<SvgPoint, JsValue>

The getStartPositionOfChar() method.

MDN Documentation

This API requires the following crate features to be activated: SvgPoint, SvgTextContentElement

impl SvgTextContentElement

pub fn get_sub_string_length( &self, charnum: u32, nchars: u32 ) -> Result<f32, JsValue>

The getSubStringLength() method.

MDN Documentation

This API requires the following crate features to be activated: SvgTextContentElement

impl SvgTextContentElement

pub fn select_sub_string( &self, charnum: u32, nchars: u32 ) -> Result<(), JsValue>

The selectSubString() method.

MDN Documentation

This API requires the following crate features to be activated: SvgTextContentElement

impl SvgTextContentElement

pub const LENGTHADJUST_UNKNOWN: u16 = 0u16

The SVGTextContentElement.LENGTHADJUST_UNKNOWN const.

This API requires the following crate features to be activated: SvgTextContentElement

pub const LENGTHADJUST_SPACING: u16 = 1u16

The SVGTextContentElement.LENGTHADJUST_SPACING const.

This API requires the following crate features to be activated: SvgTextContentElement

pub const LENGTHADJUST_SPACINGANDGLYPHS: u16 = 2u16

The SVGTextContentElement.LENGTHADJUST_SPACINGANDGLYPHS const.

This API requires the following crate features to be activated: SvgTextContentElement