Struct web_sys::SvgfeConvolveMatrixElement

#[repr(transparent)]

pub struct SvgfeConvolveMatrixElement { /* fields omitted */ }

The SVGFEConvolveMatrixElement object

MDN Documentation

This function requires the following crate features to be activated: Element, EventTarget, Node, SvgElement

Methods

impl SvgfeConvolveMatrixElement

pub fn in1(&self) -> SvgAnimatedString

The in1 getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedString, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn order_x(&self) -> SvgAnimatedInteger

The orderX getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedInteger, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn order_y(&self) -> SvgAnimatedInteger

The orderY getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedInteger, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn kernel_matrix(&self) -> SvgAnimatedNumberList

The kernelMatrix getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedNumberList, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn divisor(&self) -> SvgAnimatedNumber

The divisor getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedNumber, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn bias(&self) -> SvgAnimatedNumber

The bias getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedNumber, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn target_x(&self) -> SvgAnimatedInteger

The targetX getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedInteger, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn target_y(&self) -> SvgAnimatedInteger

The targetY getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedInteger, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn edge_mode(&self) -> SvgAnimatedEnumeration

The edgeMode getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedEnumeration, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn kernel_unit_length_x(&self) -> SvgAnimatedNumber

The kernelUnitLengthX getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedNumber, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn kernel_unit_length_y(&self) -> SvgAnimatedNumber

The kernelUnitLengthY getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedNumber, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn preserve_alpha(&self) -> SvgAnimatedBoolean

The preserveAlpha getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedBoolean, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn x(&self) -> SvgAnimatedLength

The x getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn y(&self) -> SvgAnimatedLength

The y getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn width(&self) -> SvgAnimatedLength

The width getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn height(&self) -> SvgAnimatedLength

The height getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub fn result(&self) -> SvgAnimatedString

The result getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedString, SvgfeConvolveMatrixElement

impl SvgfeConvolveMatrixElement

pub const SVG_EDGEMODE_UNKNOWN: u16

impl SvgfeConvolveMatrixElement

pub const SVG_EDGEMODE_DUPLICATE: u16

impl SvgfeConvolveMatrixElement

pub const SVG_EDGEMODE_WRAP: u16

impl SvgfeConvolveMatrixElement

pub const SVG_EDGEMODE_NONE: u16