Struct web_sys::SvgsvgElement

#[repr(transparent)]

pub struct SvgsvgElement { /* fields omitted */ }

The SVGSVGElement object

MDN Documentation

This function requires the following crate features to be activated: Element, EventTarget, Node, SvgElement, SvgGraphicsElement

Methods

impl SvgsvgElement

pub fn animations_paused(&self) -> bool

The animationsPaused() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn create_svg_angle(&self) -> SvgAngle

The createSVGAngle() method

MDN Documentation

This function requires the following crate features to be activated: SvgAngle, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_length(&self) -> SvgLength

The createSVGLength() method

MDN Documentation

This function requires the following crate features to be activated: SvgLength, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_matrix(&self) -> SvgMatrix

The createSVGMatrix() method

MDN Documentation

This function requires the following crate features to be activated: SvgMatrix, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_number(&self) -> SvgNumber

The createSVGNumber() method

MDN Documentation

This function requires the following crate features to be activated: SvgNumber, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_point(&self) -> SvgPoint

The createSVGPoint() method

MDN Documentation

This function requires the following crate features to be activated: SvgPoint, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_rect(&self) -> SvgRect

The createSVGRect() method

MDN Documentation

This function requires the following crate features to be activated: SvgRect, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_transform(&self) -> SvgTransform

The createSVGTransform() method

MDN Documentation

This function requires the following crate features to be activated: SvgTransform, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_transform_from_matrix(
&self,
matrix: &SvgMatrix
) -> SvgTransform

The createSVGTransformFromMatrix() method

MDN Documentation

This function requires the following crate features to be activated: SvgMatrix, SvgTransform, SvgsvgElement

impl SvgsvgElement

pub fn deselect_all(&self)

The deselectAll() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn force_redraw(&self)

The forceRedraw() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn get_current_time(&self) -> f32

The getCurrentTime() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn get_element_by_id(&self, element_id: &str) -> Option<Element>

The getElementById() method

MDN Documentation

This function requires the following crate features to be activated: Element, SvgsvgElement

impl SvgsvgElement

pub fn pause_animations(&self)

The pauseAnimations() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn set_current_time(&self, seconds: f32)

The setCurrentTime() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn suspend_redraw(&self, max_wait_milliseconds: u32) -> u32

The suspendRedraw() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn unpause_animations(&self)

The unpauseAnimations() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn unsuspend_redraw(&self, suspend_handle_id: u32)

The unsuspendRedraw() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn unsuspend_redraw_all(&self)

The unsuspendRedrawAll() method

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn x(&self) -> SvgAnimatedLength

The x getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn y(&self) -> SvgAnimatedLength

The y getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn width(&self) -> SvgAnimatedLength

The width getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn height(&self) -> SvgAnimatedLength

The height getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn use_current_view(&self) -> bool

The useCurrentView getter

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn current_scale(&self) -> f32

The currentScale getter

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn set_current_scale(&self, current_scale: f32)

The currentScale setter

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn current_translate(&self) -> SvgPoint

The currentTranslate getter

MDN Documentation

This function requires the following crate features to be activated: SvgPoint, SvgsvgElement

impl SvgsvgElement

pub fn view_box(&self) -> SvgAnimatedRect

The viewBox getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedRect, SvgsvgElement

impl SvgsvgElement

pub fn preserve_aspect_ratio(&self) -> SvgAnimatedPreserveAspectRatio

The preserveAspectRatio getter

MDN Documentation

This function requires the following crate features to be activated: SvgAnimatedPreserveAspectRatio, SvgsvgElement

impl SvgsvgElement

pub fn zoom_and_pan(&self) -> u16

The zoomAndPan getter

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn set_zoom_and_pan(&self, zoom_and_pan: u16)

The zoomAndPan setter

MDN Documentation

This function requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub const SVG_ZOOMANDPAN_UNKNOWN: u16

impl SvgsvgElement

pub const SVG_ZOOMANDPAN_DISABLE: u16

impl SvgsvgElement

pub const SVG_ZOOMANDPAN_MAGNIFY: u16