Struct web_sys::Text

#[repr(transparent)]

pub struct Text { /* fields omitted */ }

The Text object

MDN Documentation

This function requires the following crate features to be activated: CharacterData, EventTarget, Node

Methods

impl Text

pub fn new() -> Result<Text, JsValue>

This function requires the following crate features to be activated: Text

impl Text

pub fn new_with_data(data: &str) -> Result<Text, JsValue>

This function requires the following crate features to be activated: Text

impl Text

pub fn split_text(&self, offset: u32) -> Result<Text, JsValue>

The splitText() method

MDN Documentation

This function requires the following crate features to be activated: Text

impl Text

pub fn whole_text(&self) -> Result<String, JsValue>

The wholeText getter

MDN Documentation

This function requires the following crate features to be activated: Text

impl Text

pub fn assigned_slot(&self) -> Option<HtmlSlotElement>

The assignedSlot getter

MDN Documentation

This function requires the following crate features to be activated: HtmlSlotElement, Text

impl Text

pub fn convert_point_from_node_with_text(
&self,
point: &DomPointInit,
from: &Text
) -> Result<DomPoint, JsValue>

The convertPointFromNode() method

MDN Documentation

This function requires the following crate features to be activated: DomPoint, DomPointInit, Text

impl Text

pub fn convert_point_from_node_with_element(
&self,
point: &DomPointInit,
from: &Element
) -> Result<DomPoint, JsValue>

The convertPointFromNode() method

MDN Documentation

This function requires the following crate features to be activated: DomPoint, DomPointInit, Element, Text

impl Text

pub fn convert_point_from_node_with_document(
&self,
point: &DomPointInit,
from: &Document
) -> Result<DomPoint, JsValue>

The convertPointFromNode() method

MDN Documentation

This function requires the following crate features to be activated: Document, DomPoint, DomPointInit, Text

impl Text

pub fn convert_point_from_node_with_text_and_options(
&self,
point: &DomPointInit,
from: &Text,
options: &ConvertCoordinateOptions
) -> Result<DomPoint, JsValue>

The convertPointFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, DomPoint, DomPointInit, Text

impl Text

pub fn convert_point_from_node_with_element_and_options(
&self,
point: &DomPointInit,
from: &Element,
options: &ConvertCoordinateOptions
) -> Result<DomPoint, JsValue>

The convertPointFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, DomPoint, DomPointInit, Element, Text

impl Text

pub fn convert_point_from_node_with_document_and_options(
&self,
point: &DomPointInit,
from: &Document,
options: &ConvertCoordinateOptions
) -> Result<DomPoint, JsValue>

The convertPointFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, Document, DomPoint, DomPointInit, Text

impl Text

pub fn convert_quad_from_node_with_text(
&self,
quad: &DomQuad,
from: &Text
) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method

MDN Documentation

This function requires the following crate features to be activated: DomQuad, Text

impl Text

pub fn convert_quad_from_node_with_element(
&self,
quad: &DomQuad,
from: &Element
) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method

MDN Documentation

This function requires the following crate features to be activated: DomQuad, Element, Text

impl Text

pub fn convert_quad_from_node_with_document(
&self,
quad: &DomQuad,
from: &Document
) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method

MDN Documentation

This function requires the following crate features to be activated: Document, DomQuad, Text

impl Text

pub fn convert_quad_from_node_with_text_and_options(
&self,
quad: &DomQuad,
from: &Text,
options: &ConvertCoordinateOptions
) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, Text

impl Text

pub fn convert_quad_from_node_with_element_and_options(
&self,
quad: &DomQuad,
from: &Element,
options: &ConvertCoordinateOptions
) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, Element, Text

impl Text

pub fn convert_quad_from_node_with_document_and_options(
&self,
quad: &DomQuad,
from: &Document,
options: &ConvertCoordinateOptions
) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, Document, DomQuad, Text

impl Text

pub fn convert_rect_from_node_with_text(
&self,
rect: &DomRectReadOnly,
from: &Text
) -> Result<DomQuad, JsValue>

The convertRectFromNode() method

MDN Documentation

This function requires the following crate features to be activated: DomQuad, DomRectReadOnly, Text

impl Text

pub fn convert_rect_from_node_with_element(
&self,
rect: &DomRectReadOnly,
from: &Element
) -> Result<DomQuad, JsValue>

The convertRectFromNode() method

MDN Documentation

This function requires the following crate features to be activated: DomQuad, DomRectReadOnly, Element, Text

impl Text

pub fn convert_rect_from_node_with_document(
&self,
rect: &DomRectReadOnly,
from: &Document
) -> Result<DomQuad, JsValue>

The convertRectFromNode() method

MDN Documentation

This function requires the following crate features to be activated: Document, DomQuad, DomRectReadOnly, Text

impl Text

pub fn convert_rect_from_node_with_text_and_options(
&self,
rect: &DomRectReadOnly,
from: &Text,
options: &ConvertCoordinateOptions
) -> Result<DomQuad, JsValue>

The convertRectFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, DomRectReadOnly, Text

impl Text

pub fn convert_rect_from_node_with_element_and_options(
&self,
rect: &DomRectReadOnly,
from: &Element,
options: &ConvertCoordinateOptions
) -> Result<DomQuad, JsValue>

The convertRectFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, DomRectReadOnly, Element, Text

impl Text

pub fn convert_rect_from_node_with_document_and_options(
&self,
rect: &DomRectReadOnly,
from: &Document,
options: &ConvertCoordinateOptions
) -> Result<DomQuad, JsValue>

The convertRectFromNode() method

MDN Documentation

This function requires the following crate features to be activated: ConvertCoordinateOptions, Document, DomQuad, DomRectReadOnly, Text