Struct web_sys::TouchEvent

#[repr(transparent)]

pub struct TouchEvent { /* fields omitted */ }

The TouchEvent object

MDN Documentation

This function requires the following crate features to be activated: Event, UiEvent

Methods

impl TouchEvent

pub fn new(type_: &str) -> Result<TouchEvent, JsValue>

This function requires the following crate features to be activated: TouchEvent

impl TouchEvent

pub fn new_with_event_init_dict(
type_: &str,
event_init_dict: &TouchEventInit
) -> Result<TouchEvent, JsValue>

This function requires the following crate features to be activated: TouchEvent, TouchEventInit

impl TouchEvent

pub fn init_touch_event(&self, type_: &str)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent

impl TouchEvent

pub fn init_touch_event_with_can_bubble(&self, type_: &str, can_bubble: bool)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail_and_ctrl_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
ctrl_key: bool
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail_and_ctrl_key_and_alt_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
ctrl_key: bool,
alt_key: bool
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail_and_ctrl_key_and_alt_key_and_shift_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_touches(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
touches: Option<&TouchList>
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, TouchList, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_touches_and_target_touches(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
touches: Option<&TouchList>,
target_touches: Option<&TouchList>
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, TouchList, Window

impl TouchEvent

pub fn init_touch_event_with_can_bubble_and_cancelable_and_view_and_detail_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_touches_and_target_touches_and_changed_touches(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
detail: i32,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
touches: Option<&TouchList>,
target_touches: Option<&TouchList>,
changed_touches: Option<&TouchList>
)

The initTouchEvent() method

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, TouchList, Window

impl TouchEvent

pub fn touches(&self) -> TouchList

The touches getter

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, TouchList

impl TouchEvent

pub fn target_touches(&self) -> TouchList

The targetTouches getter

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, TouchList

impl TouchEvent

pub fn changed_touches(&self) -> TouchList

The changedTouches getter

MDN Documentation

This function requires the following crate features to be activated: TouchEvent, TouchList

impl TouchEvent

pub fn alt_key(&self) -> bool

The altKey getter

MDN Documentation

This function requires the following crate features to be activated: TouchEvent

impl TouchEvent

pub fn meta_key(&self) -> bool

The metaKey getter

MDN Documentation

This function requires the following crate features to be activated: TouchEvent

impl TouchEvent

pub fn ctrl_key(&self) -> bool

The ctrlKey getter

MDN Documentation

This function requires the following crate features to be activated: TouchEvent

impl TouchEvent

pub fn shift_key(&self) -> bool

The shiftKey getter

MDN Documentation

This function requires the following crate features to be activated: TouchEvent