Struct web_sys::UsbDevice

#[repr(transparent)]
pub struct UsbDevice { /* private fields */ }

The UsbDevice class.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

Implementations

impl UsbDevice

pub fn usb_version_major(&self) -> u8

Getter for the usbVersionMajor field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn usb_version_minor(&self) -> u8

Getter for the usbVersionMinor field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn usb_version_subminor(&self) -> u8

Getter for the usbVersionSubminor field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn device_class(&self) -> u8

Getter for the deviceClass field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn device_subclass(&self) -> u8

Getter for the deviceSubclass field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn device_protocol(&self) -> u8

Getter for the deviceProtocol field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn vendor_id(&self) -> u16

Getter for the vendorId field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn product_id(&self) -> u16

Getter for the productId field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn device_version_major(&self) -> u8

Getter for the deviceVersionMajor field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn device_version_minor(&self) -> u8

Getter for the deviceVersionMinor field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn device_version_subminor(&self) -> u8

Getter for the deviceVersionSubminor field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn manufacturer_name(&self) -> Option<String>

Getter for the manufacturerName field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn product_name(&self) -> Option<String>

Getter for the productName field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn serial_number(&self) -> Option<String>

Getter for the serialNumber field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn configuration(&self) -> Option<UsbConfiguration>

Getter for the configuration field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbConfiguration, UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn configurations(&self) -> Array

Getter for the configurations field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn opened(&self) -> bool

Getter for the opened field of this object.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn claim_interface(&self, interface_number: u8) -> Promise

The claimInterface() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn clear_halt(&self, direction: UsbDirection, endpoint_number: u8) -> Promise

The clearHalt() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice, UsbDirection

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn close(&self) -> Promise

The close() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn control_transfer_in(
&self,
setup: &UsbControlTransferParameters,
length: u16
) -> Promise

The controlTransferIn() method.

MDN Documentation

This API requires the following crate features to be activated: UsbControlTransferParameters, UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn control_transfer_out(
&self,
setup: &UsbControlTransferParameters
) -> Promise

The controlTransferOut() method.

MDN Documentation

This API requires the following crate features to be activated: UsbControlTransferParameters, UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn control_transfer_out_with_buffer_source(
&self,
setup: &UsbControlTransferParameters,
data: &Object
) -> Promise

The controlTransferOut() method.

MDN Documentation

This API requires the following crate features to be activated: UsbControlTransferParameters, UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn control_transfer_out_with_u8_array(
&self,
setup: &UsbControlTransferParameters,
data: &mut [u8]
) -> Promise

The controlTransferOut() method.

MDN Documentation

This API requires the following crate features to be activated: UsbControlTransferParameters, UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn isochronous_transfer_in(
&self,
endpoint_number: u8,
packet_lengths: &JsValue
) -> Promise

The isochronousTransferIn() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn isochronous_transfer_out_with_buffer_source(
&self,
endpoint_number: u8,
data: &Object,
packet_lengths: &JsValue
) -> Promise

The isochronousTransferOut() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn isochronous_transfer_out_with_u8_array(
&self,
endpoint_number: u8,
data: &mut [u8],
packet_lengths: &JsValue
) -> Promise

The isochronousTransferOut() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn open(&self) -> Promise

The open() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn release_interface(&self, interface_number: u8) -> Promise

The releaseInterface() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn reset(&self) -> Promise

The reset() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn select_alternate_interface(
&self,
interface_number: u8,
alternate_setting: u8
) -> Promise

The selectAlternateInterface() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn select_configuration(&self, configuration_value: u8) -> Promise

The selectConfiguration() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn transfer_in(&self, endpoint_number: u8, length: u32) -> Promise

The transferIn() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn transfer_out_with_buffer_source(
&self,
endpoint_number: u8,
data: &Object
) -> Promise

The transferOut() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl UsbDevice

pub fn transfer_out_with_u8_array(
&self,
endpoint_number: u8,
data: &mut [u8]
) -> Promise

The transferOut() method.

MDN Documentation

This API requires the following crate features to be activated: UsbDevice

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide