Struct web_sys::VttCue

#[repr(transparent)]
pub struct VttCue { /* fields omitted */ }

The VTTCue object

MDN Documentation

This API requires the following crate features to be activated: VttCue

Methods

impl VttCue

pub fn new(
start_time: f64,
end_time: f64,
text: &str
) -> Result<VttCue, JsValue>

The new VTTCue(..) constructor, creating a new instance of VTTCue

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn get_cue_as_html(&self) -> DocumentFragment

The getCueAsHTML() method

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment, VttCue

impl VttCue

pub fn region(&self) -> Option<VttRegion>

The region getter

MDN Documentation

This API requires the following crate features to be activated: VttCue, VttRegion

impl VttCue

pub fn set_region(&self, region: Option<&VttRegion>)

The region setter

MDN Documentation

This API requires the following crate features to be activated: VttCue, VttRegion

impl VttCue

pub fn vertical(&self) -> DirectionSetting

The vertical getter

MDN Documentation

This API requires the following crate features to be activated: DirectionSetting, VttCue

impl VttCue

pub fn set_vertical(&self, vertical: DirectionSetting)

The vertical setter

MDN Documentation

This API requires the following crate features to be activated: DirectionSetting, VttCue

impl VttCue

pub fn snap_to_lines(&self) -> bool

The snapToLines getter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn set_snap_to_lines(&self, snap_to_lines: bool)

The snapToLines setter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn line(&self) -> JsValue

The line getter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn set_line(&self, line: &JsValue)

The line setter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn line_align(&self) -> LineAlignSetting

The lineAlign getter

MDN Documentation

This API requires the following crate features to be activated: LineAlignSetting, VttCue

impl VttCue

pub fn set_line_align(&self, line_align: LineAlignSetting)

The lineAlign setter

MDN Documentation

This API requires the following crate features to be activated: LineAlignSetting, VttCue

impl VttCue

pub fn position(&self) -> JsValue

The position getter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn set_position(&self, position: &JsValue)

The position setter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn position_align(&self) -> PositionAlignSetting

The positionAlign getter

MDN Documentation

This API requires the following crate features to be activated: PositionAlignSetting, VttCue

impl VttCue

pub fn set_position_align(&self, position_align: PositionAlignSetting)

The positionAlign setter

MDN Documentation

This API requires the following crate features to be activated: PositionAlignSetting, VttCue

impl VttCue

pub fn size(&self) -> f64

The size getter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn set_size(&self, size: f64)

The size setter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn align(&self) -> AlignSetting

The align getter

MDN Documentation

This API requires the following crate features to be activated: AlignSetting, VttCue

impl VttCue

pub fn set_align(&self, align: AlignSetting)

The align setter

MDN Documentation

This API requires the following crate features to be activated: AlignSetting, VttCue

impl VttCue

pub fn text(&self) -> String

The text getter

MDN Documentation

This API requires the following crate features to be activated: VttCue

impl VttCue

pub fn set_text(&self, text: &str)

The text setter

MDN Documentation

This API requires the following crate features to be activated: VttCue