Struct web_sys::WebGpuDevice

#[repr(transparent)]

pub struct WebGpuDevice { /* fields omitted */ }

The WebGPUDevice object

MDN Documentation

Methods

impl WebGpuDevice

pub fn create_attachment_state(&self) -> WebGpuAttachmentState

The createAttachmentState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuAttachmentState, WebGpuDevice

impl WebGpuDevice

pub fn create_attachment_state_with_descriptor(
&self,
descriptor: &WebGpuAttachmentStateDescriptor
) -> WebGpuAttachmentState

The createAttachmentState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuAttachmentState, WebGpuAttachmentStateDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_bind_group(&self) -> WebGpuBindGroup

The createBindGroup() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBindGroup, WebGpuDevice

impl WebGpuDevice

pub fn create_bind_group_with_descriptor(
&self,
descriptor: &WebGpuBindGroupDescriptor
) -> WebGpuBindGroup

The createBindGroup() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBindGroup, WebGpuBindGroupDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_bind_group_layout(&self) -> WebGpuBindGroupLayout

The createBindGroupLayout() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBindGroupLayout, WebGpuDevice

impl WebGpuDevice

pub fn create_bind_group_layout_with_descriptor(
&self,
descriptor: &WebGpuBindGroupLayoutDescriptor
) -> WebGpuBindGroupLayout

The createBindGroupLayout() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBindGroupLayout, WebGpuBindGroupLayoutDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_blend_state(&self) -> WebGpuBlendState

The createBlendState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBlendState, WebGpuDevice

impl WebGpuDevice

pub fn create_blend_state_with_descriptor(
&self,
descriptor: &WebGpuBlendStateDescriptor
) -> WebGpuBlendState

The createBlendState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBlendState, WebGpuBlendStateDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_buffer(&self) -> WebGpuBuffer

The createBuffer() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBuffer, WebGpuDevice

impl WebGpuDevice

pub fn create_buffer_with_descriptor(
&self,
descriptor: &WebGpuBufferDescriptor
) -> WebGpuBuffer

The createBuffer() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBuffer, WebGpuBufferDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_command_encoder(&self) -> WebGpuCommandEncoder

The createCommandEncoder() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuCommandEncoder, WebGpuDevice

impl WebGpuDevice

pub fn create_command_encoder_with_descriptor(
&self,
descriptor: &WebGpuCommandEncoderDescriptor
) -> WebGpuCommandEncoder

The createCommandEncoder() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuCommandEncoder, WebGpuCommandEncoderDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_compute_pipeline(
&self,
descriptor: &WebGpuComputePipelineDescriptor
) -> WebGpuComputePipeline

The createComputePipeline() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuComputePipeline, WebGpuComputePipelineDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_depth_stencil_state(&self) -> WebGpuDepthStencilState

The createDepthStencilState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDepthStencilState, WebGpuDevice

impl WebGpuDevice

pub fn create_depth_stencil_state_with_descriptor(
&self,
descriptor: &WebGpuDepthStencilStateDescriptor
) -> WebGpuDepthStencilState

The createDepthStencilState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDepthStencilState, WebGpuDepthStencilStateDescriptor, WebGpuDevice

impl WebGpuDevice

pub fn create_input_state(&self) -> WebGpuInputState

The createInputState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuInputState

impl WebGpuDevice

pub fn create_input_state_with_descriptor(
&self,
descriptor: &WebGpuInputStateDescriptor
) -> WebGpuInputState

The createInputState() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuInputState, WebGpuInputStateDescriptor

impl WebGpuDevice

pub fn create_pipeline_layout(&self) -> WebGpuPipelineLayout

The createPipelineLayout() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuPipelineLayout

impl WebGpuDevice

pub fn create_pipeline_layout_with_descriptor(
&self,
descriptor: &WebGpuPipelineLayoutDescriptor
) -> WebGpuPipelineLayout

The createPipelineLayout() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuPipelineLayout, WebGpuPipelineLayoutDescriptor

impl WebGpuDevice

pub fn create_render_pipeline(
&self,
descriptor: &WebGpuRenderPipelineDescriptor
) -> WebGpuRenderPipeline

The createRenderPipeline() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuRenderPipeline, WebGpuRenderPipelineDescriptor

impl WebGpuDevice

pub fn create_sampler(&self) -> WebGpuSampler

The createSampler() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuSampler

impl WebGpuDevice

pub fn create_sampler_with_descriptor(
&self,
descriptor: &WebGpuSamplerDescriptor
) -> WebGpuSampler

The createSampler() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuSampler, WebGpuSamplerDescriptor

impl WebGpuDevice

pub fn create_shader_module(
&self,
descriptor: &WebGpuShaderModuleDescriptor
) -> WebGpuShaderModule

The createShaderModule() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuShaderModule, WebGpuShaderModuleDescriptor

impl WebGpuDevice

pub fn create_texture(&self) -> WebGpuTexture

The createTexture() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuTexture

impl WebGpuDevice

pub fn create_texture_with_descriptor(
&self,
descriptor: &WebGpuTextureDescriptor
) -> WebGpuTexture

The createTexture() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuTexture, WebGpuTextureDescriptor

impl WebGpuDevice

pub fn extensions(&self) -> WebGpuExtensions

The extensions() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuExtensions

impl WebGpuDevice

pub fn get_object_status_with_web_gpu_buffer(
&self,
obj: &WebGpuBuffer
) -> Promise

The getObjectStatus() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuBuffer, WebGpuDevice

impl WebGpuDevice

pub fn get_object_status_with_web_gpu_texture(
&self,
obj: &WebGpuTexture
) -> Promise

The getObjectStatus() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuTexture

impl WebGpuDevice

pub fn get_queue(&self) -> WebGpuQueue

The getQueue() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuQueue

impl WebGpuDevice

pub fn limits(&self) -> WebGpuLimits

The limits() method

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice, WebGpuLimits

impl WebGpuDevice

pub fn adapter(&self) -> WebGpuAdapter

The adapter getter

MDN Documentation

This function requires the following crate features to be activated: WebGpuAdapter, WebGpuDevice

impl WebGpuDevice

pub fn on_log(&self) -> Function

The onLog getter

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice

impl WebGpuDevice

pub fn set_on_log(&self, on_log: &Function)

The onLog setter

MDN Documentation

This function requires the following crate features to be activated: WebGpuDevice