Struct web_sys::WebSocket

#[repr(transparent)]

pub struct WebSocket { /* fields omitted */ }

The WebSocket object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl WebSocket

pub fn new(url: &str) -> Result<WebSocket, JsValue>

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn new_with_str(url: &str, protocols: &str) -> Result<WebSocket, JsValue>

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn close(&self) -> Result<(), JsValue>

The close() method

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn close_with_code(&self, code: u16) -> Result<(), JsValue>

The close() method

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn close_with_code_and_reason(
&self,
code: u16,
reason: &str
) -> Result<(), JsValue>

The close() method

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_str(&self, data: &str) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_blob(&self, data: &Blob) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: Blob, WebSocket

impl WebSocket

pub fn send_with_array_buffer(&self, data: &ArrayBuffer) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_array_buffer_view(&self, data: &Object) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_u8_array(&self, data: &mut [u8]) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn url(&self) -> String

The url getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn ready_state(&self) -> u16

The readyState getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn buffered_amount(&self) -> u32

The bufferedAmount getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onopen(&self) -> Option<Function>

The onopen getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onopen(&self, onopen: Option<&Function>)

The onopen setter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onclose(&self) -> Option<Function>

The onclose getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onclose(&self, onclose: Option<&Function>)

The onclose setter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn extensions(&self) -> String

The extensions getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn protocol(&self) -> String

The protocol getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onmessage(&self) -> Option<Function>

The onmessage getter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onmessage(&self, onmessage: Option<&Function>)

The onmessage setter

MDN Documentation

This function requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn binary_type(&self) -> BinaryType

The binaryType getter

MDN Documentation

This function requires the following crate features to be activated: BinaryType, WebSocket

impl WebSocket

pub fn set_binary_type(&self, binary_type: BinaryType)

The binaryType setter

MDN Documentation

This function requires the following crate features to be activated: BinaryType, WebSocket

impl WebSocket

pub const CONNECTING: u16

impl WebSocket

pub const OPEN: u16

impl WebSocket

pub const CLOSING: u16

impl WebSocket

pub const CLOSED: u16