Struct web_sys::XPathResult

#[repr(transparent)]

pub struct XPathResult { /* fields omitted */ }

The XPathResult object

MDN Documentation

Methods

impl XPathResult

pub fn iterate_next(&self) -> Result<Option<Node>, JsValue>

The iterateNext() method

MDN Documentation

This function requires the following crate features to be activated: Node, XPathResult

impl XPathResult

pub fn snapshot_item(&self, index: u32) -> Result<Option<Node>, JsValue>

The snapshotItem() method

MDN Documentation

This function requires the following crate features to be activated: Node, XPathResult

impl XPathResult

pub fn result_type(&self) -> u16

The resultType getter

MDN Documentation

This function requires the following crate features to be activated: XPathResult

impl XPathResult

pub fn number_value(&self) -> Result<f64, JsValue>

The numberValue getter

MDN Documentation

This function requires the following crate features to be activated: XPathResult

impl XPathResult

pub fn string_value(&self) -> Result<String, JsValue>

The stringValue getter

MDN Documentation

This function requires the following crate features to be activated: XPathResult

impl XPathResult

pub fn boolean_value(&self) -> Result<bool, JsValue>

The booleanValue getter

MDN Documentation

This function requires the following crate features to be activated: XPathResult

impl XPathResult

pub fn single_node_value(&self) -> Result<Option<Node>, JsValue>

The singleNodeValue getter

MDN Documentation

This function requires the following crate features to be activated: Node, XPathResult

impl XPathResult

pub fn invalid_iterator_state(&self) -> bool

The invalidIteratorState getter

MDN Documentation

This function requires the following crate features to be activated: XPathResult

impl XPathResult

pub fn snapshot_length(&self) -> Result<u32, JsValue>

The snapshotLength getter

MDN Documentation

This function requires the following crate features to be activated: XPathResult

impl XPathResult

pub const ANY_TYPE: u16

impl XPathResult

pub const NUMBER_TYPE: u16

impl XPathResult

pub const STRING_TYPE: u16

impl XPathResult

pub const BOOLEAN_TYPE: u16

impl XPathResult

pub const UNORDERED_NODE_ITERATOR_TYPE: u16

impl XPathResult

pub const ORDERED_NODE_ITERATOR_TYPE: u16

impl XPathResult

pub const UNORDERED_NODE_SNAPSHOT_TYPE: u16

impl XPathResult

pub const ORDERED_NODE_SNAPSHOT_TYPE: u16

impl XPathResult

pub const ANY_UNORDERED_NODE_TYPE: u16

impl XPathResult

pub const FIRST_ORDERED_NODE_TYPE: u16