Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Re-exports

pub use js_sys;

pub use wasm_bindgen;

Modules

console

Structs

AbortSignal

The AbortSignal class.

Element

The Element class.

EventTarget

The EventTarget class.

Headers

The Headers class.

HtmlElement

The HtmlElement class.

HtmlTextAreaElement

The HtmlTextAreaElement class.

Node

The Node class.

ReadableStream

The ReadableStream class.

Request

The Request class.

RequestInit

The RequestInit dictionary.

Response

The Response class.

Window

The Window class.

WritableStream

The WritableStream class.

Enums

RequestMode

The RequestMode enum.

Functions

window

Getter for the Window object