Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Modules

css

Structs

AddEventListenerOptions

The AddEventListenerOptions dictionary.

AnimationEvent

The AnimationEvent class.

CharacterData

The CharacterData class.

Comment

The Comment class.

CssKeyframesRule

The CssKeyframesRule class.

CssRule

The CssRule class.

CssRuleList

The CssRuleList class.

CssStyleDeclaration

The CssStyleDeclaration class.

CssStyleRule

The CssStyleRule class.

CssStyleSheet

The CssStyleSheet class.

DataTransfer

The DataTransfer class.

Document

The Document class.

DocumentFragment

The DocumentFragment class.

DomRect

The DomRect class.

DomRectReadOnly

The DomRectReadOnly class.

DomTokenList

The DomTokenList class.

DragEvent

The DragEvent class.

Element

The Element class.

Event

The Event class.

EventTarget

The EventTarget class.

FocusEvent

The FocusEvent class.

Headers

The Headers class.

History

The History class.

HtmlAnchorElement

The HtmlAnchorElement class.

HtmlBodyElement

The HtmlBodyElement class.

HtmlButtonElement

The HtmlButtonElement class.

HtmlCanvasElement

The HtmlCanvasElement class.

HtmlDivElement

The HtmlDivElement class.

HtmlElement

The HtmlElement class.

HtmlHeadElement

The HtmlHeadElement class.

HtmlImageElement

The HtmlImageElement class.

HtmlInputElement

The HtmlInputElement class.

HtmlLabelElement

The HtmlLabelElement class.

HtmlLinkElement

The HtmlLinkElement class.

HtmlMediaElement

The HtmlMediaElement class.

HtmlParagraphElement

The HtmlParagraphElement class.

HtmlScriptElement

The HtmlScriptElement class.

HtmlStyleElement

The HtmlStyleElement class.

HtmlTemplateElement

The HtmlTemplateElement class.

HtmlTextAreaElement

The HtmlTextAreaElement class.

HtmlVideoElement

The HtmlVideoElement class.

InputEvent

The InputEvent class.

KeyboardEvent

The KeyboardEvent class.

Location

The Location class.

MessageEvent

The MessageEvent class.

MouseEvent

The MouseEvent class.

Node

The Node class.

PointerEvent

The PointerEvent class.

Request

The Request class.

RequestInit

The RequestInit dictionary.

Response

The Response class.

ShadowRoot

The ShadowRoot class.

ShadowRootInit

The ShadowRootInit dictionary.

Storage

The Storage class.

StyleSheet

The StyleSheet class.

SvgElement

The SvgElement class.

SvgGraphicsElement

The SvgGraphicsElement class.

SvgsvgElement

The SvgsvgElement class.

Text

The Text class.

Touch

The Touch class.

TouchEvent

The TouchEvent class.

TouchList

The TouchList class.

UiEvent

The UiEvent class.

WheelEvent

The WheelEvent class.

Window

The Window class.

Worker

The Worker class.

Enums

ShadowRootMode

The ShadowRootMode enum.

Functions

window

Getter for the Window object