Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Modules

console

Structs

DragEvent

The DragEvent class.

Event

The Event class.

EventTarget

The EventTarget class.

FocusEvent

The FocusEvent class.

KeyboardEvent

The KeyboardEvent class.

MouseEvent

The MouseEvent class.

UiEvent

The UiEvent class.