Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Re-exports

pub use js_sys;

pub use wasm_bindgen;

Modules

console

Structs

Blob

The Blob class.

CloseEvent

The CloseEvent class.

DedicatedWorkerGlobalScope

The DedicatedWorkerGlobalScope class.

DomException

The DomException class.

Event

The Event class.

EventTarget

The EventTarget class.

History

The History class.

Location

The Location class.

MessageEvent

The MessageEvent class.

Navigator

The Navigator class.

Storage

The Storage class.

Url

The Url class.

WebSocket

The WebSocket class.

Window

The Window class.

Worker

The Worker class.

WorkerGlobalScope

The WorkerGlobalScope class.

WorkerNavigator

The WorkerNavigator class.

WorkerOptions

The WorkerOptions dictionary.

Enums

BinaryType

The BinaryType enum.

WorkerType

The WorkerType enum.

Functions

window

Getter for the Window object