Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browser provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Structs

Blob
The Blob class.

Document
The Document class.

Element
The Element class.

ErrorEvent
The ErrorEvent class.

Event
The Event class.

EventTarget
The EventTarget class.

FileReader
The FileReader class.

HtmlElement
The HtmlElement class.

MessageEvent
The MessageEvent class.

Node
The Node class.

ProgressEvent
The ProgressEvent class.

WebSocket
The WebSocket class.

Window
The Window class.

Enums

BinaryType
The BinaryType enum.

Functions

window
Getter for the Window object