Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Modules

console

Structs

AbortSignal

The AbortSignal class.

AddEventListenerOptions

The AddEventListenerOptions dictionary.

AnimationEvent

The AnimationEvent class.

BeforeUnloadEvent

The BeforeUnloadEvent class.

Blob

The Blob class.

CharacterData

The CharacterData class.

CloseEvent

The CloseEvent class.

CloseEventInit

The CloseEventInit dictionary.

Comment

The Comment class.

CompositionEvent

The CompositionEvent class.

CssStyleDeclaration

The CssStyleDeclaration class.

CustomEvent

The CustomEvent class.

DeviceMotionEvent

The DeviceMotionEvent class.

DeviceOrientationEvent

The DeviceOrientationEvent class.

Document

The Document class.

DocumentFragment

The DocumentFragment class.

DomTokenList

The DomTokenList class.

DragEvent

The DragEvent class.

Element

The Element class.

ErrorEvent

The ErrorEvent class.

Event

The Event class.

EventSource

The EventSource class.

EventTarget

The EventTarget class.

FileReader

The FileReader class.

FocusEvent

The FocusEvent class.

FormData

The FormData class.

GamepadEvent

The GamepadEvent class.

HashChangeEvent

The HashChangeEvent class.

Headers

The Headers class.

History

The History class.

HtmlAnchorElement

The HtmlAnchorElement class.

HtmlAreaElement

The HtmlAreaElement class.

HtmlAudioElement

The HtmlAudioElement class.

HtmlBaseElement

The HtmlBaseElement class.

HtmlBodyElement

The HtmlBodyElement class.

HtmlBrElement

The HtmlBrElement class.

HtmlButtonElement

The HtmlButtonElement class.

HtmlCanvasElement

The HtmlCanvasElement class.

HtmlCollection

The HtmlCollection class.

HtmlDListElement

The HtmlDListElement class.

HtmlDataElement

The HtmlDataElement class.

HtmlDataListElement

The HtmlDataListElement class.

HtmlDetailsElement

The HtmlDetailsElement class.

HtmlDialogElement

The HtmlDialogElement class.

HtmlDivElement

The HtmlDivElement class.

HtmlElement

The HtmlElement class.

HtmlEmbedElement

The HtmlEmbedElement class.

HtmlFieldSetElement

The HtmlFieldSetElement class.

HtmlFormElement

The HtmlFormElement class.

HtmlHeadElement

The HtmlHeadElement class.

HtmlHeadingElement

The HtmlHeadingElement class.

HtmlHrElement

The HtmlHrElement class.

HtmlHtmlElement

The HtmlHtmlElement class.

HtmlIFrameElement

The HtmlIFrameElement class.

HtmlImageElement

The HtmlImageElement class.

HtmlInputElement

The HtmlInputElement class.

HtmlLabelElement

The HtmlLabelElement class.

HtmlLegendElement

The HtmlLegendElement class.

HtmlLiElement

The HtmlLiElement class.

HtmlLinkElement

The HtmlLinkElement class.

HtmlMapElement

The HtmlMapElement class.

HtmlMediaElement

The HtmlMediaElement class.

HtmlMenuElement

The HtmlMenuElement class.

HtmlMetaElement

The HtmlMetaElement class.

HtmlMeterElement

The HtmlMeterElement class.

HtmlModElement

The HtmlModElement class.

HtmlOListElement

The HtmlOListElement class.

HtmlObjectElement

The HtmlObjectElement class.

HtmlOptGroupElement

The HtmlOptGroupElement class.

HtmlOptionElement

The HtmlOptionElement class.

HtmlOutputElement

The HtmlOutputElement class.

HtmlParagraphElement

The HtmlParagraphElement class.

HtmlParamElement

The HtmlParamElement class.

HtmlPictureElement

The HtmlPictureElement class.

HtmlPreElement

The HtmlPreElement class.

HtmlProgressElement

The HtmlProgressElement class.

HtmlQuoteElement

The HtmlQuoteElement class.

HtmlScriptElement

The HtmlScriptElement class.

HtmlSelectElement

The HtmlSelectElement class.

HtmlSlotElement

The HtmlSlotElement class.

HtmlSourceElement

The HtmlSourceElement class.

HtmlSpanElement

The HtmlSpanElement class.

HtmlStyleElement

The HtmlStyleElement class.

HtmlTableCaptionElement

The HtmlTableCaptionElement class.

HtmlTableCellElement

The HtmlTableCellElement class.

HtmlTableColElement

The HtmlTableColElement class.

HtmlTableElement

The HtmlTableElement class.

HtmlTableRowElement

The HtmlTableRowElement class.

HtmlTableSectionElement

The HtmlTableSectionElement class.

HtmlTemplateElement

The HtmlTemplateElement class.

HtmlTextAreaElement

The HtmlTextAreaElement class.

HtmlTimeElement

The HtmlTimeElement class.

HtmlTitleElement

The HtmlTitleElement class.

HtmlTrackElement

The HtmlTrackElement class.

HtmlUListElement

The HtmlUListElement class.

HtmlVideoElement

The HtmlVideoElement class.

InputEvent

The InputEvent class.

KeyboardEvent

The KeyboardEvent class.

Location

The Location class.

MessageEvent

The MessageEvent class.

MouseEvent

The MouseEvent class.

Node

The Node class.

NodeList

The NodeList class.

ObserverCallback

The ObserverCallback dictionary.

PageTransitionEvent

The PageTransitionEvent class.

PointerEvent

The PointerEvent class.

PopStateEvent

The PopStateEvent class.

ProgressEvent

The ProgressEvent class.

PromiseRejectionEvent

The PromiseRejectionEvent class.

Range

The Range class.

ReadableStream

The ReadableStream class.

Request

The Request class.

RequestInit

The RequestInit dictionary.

Response

The Response class.

SecurityPolicyViolationEvent

The SecurityPolicyViolationEvent class.

StorageEvent

The StorageEvent class.

SubmitEvent

The SubmitEvent class.

SvgElement

The SvgElement class.

Text

The Text class.

TouchEvent

The TouchEvent class.

TransitionEvent

The TransitionEvent class.

TreeWalker

The TreeWalker class.

UiEvent

The UiEvent class.

Url

The Url class.

UrlSearchParams

The UrlSearchParams class.

WebSocket

The WebSocket class.

WheelEvent

The WheelEvent class.

Window

The Window class.

WorkerGlobalScope

The WorkerGlobalScope class.

Enums

BinaryType

The BinaryType enum.

ReferrerPolicy

The ReferrerPolicy enum.

RequestCache

The RequestCache enum.

RequestCredentials

The RequestCredentials enum.

RequestMode

The RequestMode enum.

RequestRedirect

The RequestRedirect enum.

ResponseType

The ResponseType enum.

Functions

window

Getter for the Window object