Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Re-exports

pub use js_sys;

pub use wasm_bindgen;

Modules

console

Structs

AddEventListenerOptions

The AddEventListenerOptions dictionary.

CanvasRenderingContext2d

The CanvasRenderingContext2d class.

CssStyleDeclaration

The CssStyleDeclaration class.

Document

The Document class.

DomMatrix

The DomMatrix class.

DomMatrixReadOnly

The DomMatrixReadOnly class.

DomPoint

The DomPoint class.

DomPointInit

The DomPointInit dictionary.

DomPointReadOnly

The DomPointReadOnly class.

Element

The Element class.

Event

The Event class.

EventTarget

The EventTarget class.

HtmlCanvasElement

The HtmlCanvasElement class.

HtmlElement

The HtmlElement class.

HtmlInputElement

The HtmlInputElement class.

InputEvent

The InputEvent class.

MouseEvent

The MouseEvent class.

Node

The Node class.

Path2d

The Path2d class.

TextMetrics

The TextMetrics class.

Touch

The Touch class.

TouchEvent

The TouchEvent class.

TouchList

The TouchList class.

UiEvent

The UiEvent class.

Window

The Window class.

Functions

window

Getter for the Window object