Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Modules

console

Structs

AbortSignal

The AbortSignal class.

AddEventListenerOptions

The AddEventListenerOptions dictionary.

AnimationEvent

The AnimationEvent class.

Blob

The Blob class.

BlobPropertyBag

The BlobPropertyBag dictionary.

CharacterData

The CharacterData class.

CloseEvent

The CloseEvent class.

CloseEventInit

The CloseEventInit dictionary.

DedicatedWorkerGlobalScope

The DedicatedWorkerGlobalScope class.

Document

The Document class.

DocumentFragment

The DocumentFragment class.

DomException

The DomException class.

DragEvent

The DragEvent class.

Element

The Element class.

ErrorEvent

The ErrorEvent class.

Event

The Event class.

EventInit

The EventInit dictionary.

EventSource

The EventSource class.

EventTarget

The EventTarget class.

File

The File class.

FileList

The FileList class.

FilePropertyBag

The FilePropertyBag dictionary.

FileReader

The FileReader class.

FocusEvent

The FocusEvent class.

FormData

The FormData class.

Headers

The Headers class.

History

The History class.

HtmlCollection

The HtmlCollection class.

HtmlElement

The HtmlElement class.

HtmlHeadElement

The HtmlHeadElement class.

HtmlInputElement

The HtmlInputElement class.

HtmlScriptElement

The HtmlScriptElement class.

HtmlStyleElement

The HtmlStyleElement class.

HtmlTextAreaElement

The HtmlTextAreaElement class.

InputEvent

The InputEvent class.

InputEventInit

The InputEventInit dictionary.

KeyboardEvent

The KeyboardEvent class.

Location

The Location class.

MessageEvent

The MessageEvent class.

MouseEvent

The MouseEvent class.

Node

The Node class.

NodeList

The NodeList class.

ObserverCallback

The ObserverCallback dictionary.

PointerEvent

The PointerEvent class.

ProgressEvent

The ProgressEvent class.

ReadableStream

The ReadableStream class.

Request

The Request class.

RequestInit

The RequestInit dictionary.

Response

The Response class.

ShadowRoot

The ShadowRoot class.

Storage

The Storage class.

SubmitEvent

The SubmitEvent class.

Text

The Text class.

TouchEvent

The TouchEvent class.

TransitionEvent

The TransitionEvent class.

UiEvent

The UiEvent class.

Url

The Url class.

UrlSearchParams

The UrlSearchParams class.

WebSocket

The WebSocket class.

WheelEvent

The WheelEvent class.

Window

The Window class.

Worker

The Worker class.

WorkerGlobalScope

The WorkerGlobalScope class.

WorkerOptions

The WorkerOptions dictionary.

Enums

BinaryType

The BinaryType enum.

ReferrerPolicy

The ReferrerPolicy enum.

RequestCache

The RequestCache enum.

RequestCredentials

The RequestCredentials enum.

RequestMode

The RequestMode enum.

RequestRedirect

The RequestRedirect enum.

ResponseType

The ResponseType enum.

Functions

window

Getter for the Window object