Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browsers provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Modules

console

Structs

AbortController

The AbortController class.

AbortSignal

The AbortSignal class.

AddEventListenerOptions

The AddEventListenerOptions dictionary.

AnimationEvent

The AnimationEvent class.

Blob

The Blob class.

BlobPropertyBag

The BlobPropertyBag dictionary.

CanvasGradient

The CanvasGradient class.

CanvasRenderingContext2d

The CanvasRenderingContext2d class.

CharacterData

The CharacterData class.

CssStyleDeclaration

The CssStyleDeclaration class.

DedicatedWorkerGlobalScope

The DedicatedWorkerGlobalScope class.

Document

The Document class.

DomException

The DomException class.

DomTokenList

The DomTokenList class.

DragEvent

The DragEvent class.

Element

The Element class.

ErrorEvent

The ErrorEvent class.

Event

The Event class.

EventTarget

The EventTarget class.

File

The File class.

FileList

The FileList class.

FilePropertyBag

The FilePropertyBag dictionary.

FileReader

The FileReader class.

FocusEvent

The FocusEvent class.

Headers

The Headers class.

HtmlAnchorElement

The HtmlAnchorElement class.

HtmlButtonElement

The HtmlButtonElement class.

HtmlCanvasElement

The HtmlCanvasElement class.

HtmlElement

The HtmlElement class.

HtmlImageElement

The HtmlImageElement class.

HtmlInputElement

The HtmlInputElement class.

HtmlSelectElement

The HtmlSelectElement class.

HtmlTextAreaElement

The HtmlTextAreaElement class.

ImageBitmap

The ImageBitmap class.

InputEvent

The InputEvent class.

KeyboardEvent

The KeyboardEvent class.

Location

The Location class.

MessageEvent

The MessageEvent class.

MouseEvent

The MouseEvent class.

Node

The Node class.

ObserverCallback

The ObserverCallback dictionary.

Performance

The Performance class.

PointerEvent

The PointerEvent class.

ProgressEvent

The ProgressEvent class.

Request

The Request class.

RequestInit

The RequestInit dictionary.

Response

The Response class.

Storage

The Storage class.

Text

The Text class.

Touch

The Touch class.

TouchEvent

The TouchEvent class.

TouchList

The TouchList class.

TransitionEvent

The TransitionEvent class.

UiEvent

The UiEvent class.

Url

The Url class.

WebGlBuffer

The WebGlBuffer class.

WebGlProgram

The WebGlProgram class.

WebGlRenderingContext

The WebGlRenderingContext class.

WebGlShader

The WebGlShader class.

WebGlTexture

The WebGlTexture class.

WebGlUniformLocation

The WebGlUniformLocation class.

WebSocket

The WebSocket class.

WheelEvent

The WheelEvent class.

Window

The Window class.

Worker

The Worker class.

WorkerGlobalScope

The WorkerGlobalScope class.

WorkerOptions

The WorkerOptions dictionary.

Enums

BinaryType

The BinaryType enum.

ReferrerPolicy

The ReferrerPolicy enum.

RequestCache

The RequestCache enum.

RequestCredentials

The RequestCredentials enum.

RequestMode

The RequestMode enum.

RequestRedirect

The RequestRedirect enum.

Functions

window

Getter for the Window object