Crate web_sys

Raw API bindings for Web APIs

This is a procedurally generated crate from browser WebIDL which provides a binding to all APIs that browser provide on the web.

This crate by default contains very little when compiled as almost all of its exposed APIs are gated by Cargo features. The exhaustive list of features can be found in crates/web-sys/Cargo.toml, but the rule of thumb for web-sys is that each type has its own cargo feature (named after the type). Using an API requires enabling the features for all types used in the API, and APIs should mention in the documentation what features they require.

Modules

console

css

Structs

AbortController
The AbortController object

AbortSignal
The AbortSignal object

AddEventListenerOptions
The AddEventListenerOptions dictionary

AesCbcParams
The AesCbcParams dictionary

AesCtrParams
The AesCtrParams dictionary

AesDerivedKeyParams
The AesDerivedKeyParams dictionary

AesGcmParams
The AesGcmParams dictionary

AesKeyAlgorithm
The AesKeyAlgorithm dictionary

AesKeyGenParams
The AesKeyGenParams dictionary

Algorithm
The Algorithm dictionary

AnalyserNode
The AnalyserNode object

AnalyserOptions
The AnalyserOptions dictionary

AngleInstancedArrays
The ANGLE_instanced_arrays object

Animation
The Animation object

AnimationEffect
The AnimationEffect object

AnimationEvent
The AnimationEvent object

AnimationEventInit
The AnimationEventInit dictionary

AnimationPlaybackEvent
The AnimationPlaybackEvent object

AnimationPlaybackEventInit
The AnimationPlaybackEventInit dictionary

AnimationPropertyDetails
The AnimationPropertyDetails dictionary

AnimationPropertyValueDetails
The AnimationPropertyValueDetails dictionary

AnimationTimeline
The AnimationTimeline object

AssignedNodesOptions
The AssignedNodesOptions dictionary

Attr
The Attr object

AttributeNameValue
The AttributeNameValue dictionary

AudioBuffer
The AudioBuffer object

AudioBufferOptions
The AudioBufferOptions dictionary

AudioBufferSourceNode
The AudioBufferSourceNode object

AudioBufferSourceOptions
The AudioBufferSourceOptions dictionary

AudioConfiguration
The AudioConfiguration dictionary

AudioContext
The AudioContext object

AudioContextOptions
The AudioContextOptions dictionary

AudioDestinationNode
The AudioDestinationNode object

AudioListener
The AudioListener object

AudioNode
The AudioNode object

AudioNodeOptions
The AudioNodeOptions dictionary

AudioParam
The AudioParam object

AudioParamMap
The AudioParamMap object

AudioProcessingEvent
The AudioProcessingEvent object

AudioScheduledSourceNode	Deprecated
The AudioScheduledSourceNode object

AudioStreamTrack
The AudioStreamTrack object

AudioTrack
The AudioTrack object

AudioTrackList
The AudioTrackList object

AudioWorklet
The AudioWorklet object

AudioWorkletGlobalScope
The AudioWorkletGlobalScope object

AudioWorkletNode
The AudioWorkletNode object

AudioWorkletNodeOptions
The AudioWorkletNodeOptions dictionary

AudioWorkletProcessor
The AudioWorkletProcessor object

AuthenticationExtensionsClientInputs
The AuthenticationExtensionsClientInputs dictionary

AuthenticationExtensionsClientOutputs
The AuthenticationExtensionsClientOutputs dictionary

AuthenticatorAssertionResponse
The AuthenticatorAssertionResponse object

AuthenticatorAttestationResponse
The AuthenticatorAttestationResponse object

AuthenticatorResponse
The AuthenticatorResponse object

AuthenticatorSelectionCriteria
The AuthenticatorSelectionCriteria dictionary

AutocompleteInfo
The AutocompleteInfo dictionary

BarProp
The BarProp object

BaseAudioContext	Deprecated
The BaseAudioContext object

BaseComputedKeyframe
The BaseComputedKeyframe dictionary

BaseKeyframe
The BaseKeyframe dictionary

BasePropertyIndexedKeyframe
The BasePropertyIndexedKeyframe dictionary

BasicCardRequest
The BasicCardRequest dictionary

BasicCardResponse
The BasicCardResponse dictionary

BatteryManager
The BatteryManager object

BeforeUnloadEvent
The BeforeUnloadEvent object

BiquadFilterNode
The BiquadFilterNode object

BiquadFilterOptions
The BiquadFilterOptions dictionary

Blob
The Blob object

BlobEvent
The BlobEvent object

BlobEventInit
The BlobEventInit dictionary

BlobPropertyBag
The BlobPropertyBag dictionary

BlockParsingOptions
The BlockParsingOptions dictionary

BoxQuadOptions
The BoxQuadOptions dictionary

BroadcastChannel
The BroadcastChannel object

BrowserElementDownloadOptions
The BrowserElementDownloadOptions dictionary

BrowserElementExecuteScriptOptions
The BrowserElementExecuteScriptOptions dictionary

BrowserFeedWriter
The BrowserFeedWriter object

Cache
The Cache object

CacheBatchOperation
The CacheBatchOperation dictionary

CacheQueryOptions
The CacheQueryOptions dictionary

CacheStorage
The CacheStorage object

CanvasCaptureMediaStream
The CanvasCaptureMediaStream object

CanvasGradient
The CanvasGradient object

CanvasPattern
The CanvasPattern object

CanvasRenderingContext2d
The CanvasRenderingContext2D object

CaretPosition
The CaretPosition object

CaretStateChangedEventInit
The CaretStateChangedEventInit dictionary

CdataSection
The CDATASection object

ChannelMergerNode
The ChannelMergerNode object

ChannelMergerOptions
The ChannelMergerOptions dictionary

ChannelPixelLayout
The ChannelPixelLayout dictionary

ChannelSplitterNode
The ChannelSplitterNode object

ChannelSplitterOptions
The ChannelSplitterOptions dictionary

CharacterData
The CharacterData object

CheckerboardReport
The CheckerboardReport dictionary

CheckerboardReportService
The CheckerboardReportService object

ChromeFilePropertyBag
The ChromeFilePropertyBag dictionary

ChromeWorker
The ChromeWorker object

Client
The Client object

ClientQueryOptions
The ClientQueryOptions dictionary

ClientRectsAndTexts
The ClientRectsAndTexts dictionary

Clients
The Clients object

ClipboardEvent
The ClipboardEvent object

ClipboardEventInit
The ClipboardEventInit dictionary

CloseEvent
The CloseEvent object

CloseEventInit
The CloseEventInit dictionary

CollectedClientData
The CollectedClientData dictionary

Comment
The Comment object

CompositionEvent
The CompositionEvent object

CompositionEventInit
The CompositionEventInit dictionary

ComputedEffectTiming
The ComputedEffectTiming dictionary

ConnStatusDict
The ConnStatusDict dictionary

ConsoleCounter
The ConsoleCounter dictionary

ConsoleCounterError
The ConsoleCounterError dictionary

ConsoleEvent
The ConsoleEvent dictionary

ConsoleInstance
The ConsoleInstance object

ConsoleInstanceOptions
The ConsoleInstanceOptions dictionary

ConsoleProfileEvent
The ConsoleProfileEvent dictionary

ConsoleStackEntry
The ConsoleStackEntry dictionary

ConsoleTimerError
The ConsoleTimerError dictionary

ConsoleTimerLogOrEnd
The ConsoleTimerLogOrEnd dictionary

ConsoleTimerStart
The ConsoleTimerStart dictionary

ConstantSourceNode
The ConstantSourceNode object

ConstantSourceOptions
The ConstantSourceOptions dictionary

ConstrainBooleanParameters
The ConstrainBooleanParameters dictionary

ConstrainDomStringParameters
The ConstrainDOMStringParameters dictionary

ConstrainDoubleRange
The ConstrainDoubleRange dictionary

ConstrainLongRange
The ConstrainLongRange dictionary

ContextAttributes2d
The ContextAttributes2D dictionary

ConvertCoordinateOptions
The ConvertCoordinateOptions dictionary

ConvolverNode
The ConvolverNode object

ConvolverOptions
The ConvolverOptions dictionary

Coordinates
The Coordinates object

Credential
The Credential object

CredentialCreationOptions
The CredentialCreationOptions dictionary

CredentialRequestOptions
The CredentialRequestOptions dictionary

CredentialsContainer
The CredentialsContainer object

Crypto
The Crypto object

CryptoKey
The CryptoKey object

CryptoKeyPair
The CryptoKeyPair dictionary

Csp
The CSP dictionary

CspPolicies
The CSPPolicies dictionary

CspReport
The CSPReport dictionary

CspReportProperties
The CSPReportProperties dictionary

CssAnimation
The CSSAnimation object

CssConditionRule
The CSSConditionRule object

CssCounterStyleRule
The CSSCounterStyleRule object

CssFontFaceRule
The CSSFontFaceRule object

CssFontFeatureValuesRule
The CSSFontFeatureValuesRule object

CssGroupingRule
The CSSGroupingRule object

CssImportRule
The CSSImportRule object

CssKeyframeRule
The CSSKeyframeRule object

CssKeyframesRule
The CSSKeyframesRule object

CssMediaRule
The CSSMediaRule object

CssNamespaceRule
The CSSNamespaceRule object

CssPageRule
The CSSPageRule object

CssPseudoElement
The CSSPseudoElement object

CssRule
The CSSRule object

CssRuleList
The CSSRuleList object

CssStyleDeclaration
The CSSStyleDeclaration object

CssStyleRule
The CSSStyleRule object

CssStyleSheet
The CSSStyleSheet object

CssSupportsRule
The CSSSupportsRule object

CssTransition
The CSSTransition object

CustomElementRegistry
The CustomElementRegistry object

CustomEvent
The CustomEvent object

CustomEventInit
The CustomEventInit dictionary

DataTransfer
The DataTransfer object

DataTransferItem
The DataTransferItem object

DataTransferItemList
The DataTransferItemList object

DateTimeValue
The DateTimeValue dictionary

DecoderDoctorNotification
The DecoderDoctorNotification dictionary

DedicatedWorkerGlobalScope
The DedicatedWorkerGlobalScope object

DelayNode
The DelayNode object

DelayOptions
The DelayOptions dictionary

DeviceAcceleration
The DeviceAcceleration object

DeviceAccelerationInit
The DeviceAccelerationInit dictionary

DeviceLightEvent
The DeviceLightEvent object

DeviceLightEventInit
The DeviceLightEventInit dictionary

DeviceMotionEvent
The DeviceMotionEvent object

DeviceMotionEventInit
The DeviceMotionEventInit dictionary

DeviceOrientationEvent
The DeviceOrientationEvent object

DeviceOrientationEventInit
The DeviceOrientationEventInit dictionary

DeviceProximityEvent
The DeviceProximityEvent object

DeviceProximityEventInit
The DeviceProximityEventInit dictionary

DeviceRotationRate
The DeviceRotationRate object

DeviceRotationRateInit
The DeviceRotationRateInit dictionary

DhKeyDeriveParams
The DhKeyDeriveParams dictionary

Directory
The Directory object

DisplayNameOptions
The DisplayNameOptions dictionary

DisplayNameResult
The DisplayNameResult dictionary

DnsCacheDict
The DNSCacheDict dictionary

DnsCacheEntry
The DnsCacheEntry dictionary

DnsLookupDict
The DNSLookupDict dictionary

Document
The Document object

DocumentFragment
The DocumentFragment object

DocumentTimeline
The DocumentTimeline object

DocumentTimelineOptions
The DocumentTimelineOptions dictionary

DocumentType
The DocumentType object

DomError
The DOMError object

DomException
The DOMException object

DomImplementation
The DOMImplementation object

DomMatrix
The DOMMatrix object

DomMatrixReadOnly
The DOMMatrixReadOnly object

DomParser
The DOMParser object

DomPoint
The DOMPoint object

DomPointInit
The DOMPointInit dictionary

DomPointReadOnly
The DOMPointReadOnly object

DomQuad
The DOMQuad object

DomQuadInit
The DOMQuadInit dictionary

DomQuadJson
The DOMQuadJSON dictionary

DomRect
The DOMRect object

DomRectInit
The DOMRectInit dictionary

DomRectList
The DOMRectList object

DomRectReadOnly
The DOMRectReadOnly object

DomRequest
The DOMRequest object

DomStringList
The DOMStringList object

DomStringMap
The DOMStringMap object

DomTokenList
The DOMTokenList object

DomWindowResizeEventDetail
The DOMWindowResizeEventDetail dictionary

DragEvent
The DragEvent object

DragEventInit
The DragEventInit dictionary

DynamicsCompressorNode
The DynamicsCompressorNode object

DynamicsCompressorOptions
The DynamicsCompressorOptions dictionary

EcKeyAlgorithm
The EcKeyAlgorithm dictionary

EcKeyGenParams
The EcKeyGenParams dictionary

EcKeyImportParams
The EcKeyImportParams dictionary

EcdhKeyDeriveParams
The EcdhKeyDeriveParams dictionary

EcdsaParams
The EcdsaParams dictionary

EffectTiming
The EffectTiming dictionary

Element
The Element object

ElementCreationOptions
The ElementCreationOptions dictionary

ElementDefinitionOptions
The ElementDefinitionOptions dictionary

ErrorCallback

ErrorEvent
The ErrorEvent object

ErrorEventInit
The ErrorEventInit dictionary

Event
The Event object

EventInit
The EventInit dictionary

EventListener

EventListenerOptions
The EventListenerOptions dictionary

EventModifierInit
The EventModifierInit dictionary

EventSource
The EventSource object

EventSourceInit
The EventSourceInit dictionary

EventTarget
The EventTarget object

Exception
The Exception object

ExtBlendMinmax
The EXT_blend_minmax object

ExtColorBufferFloat
The EXT_color_buffer_float object

ExtColorBufferHalfFloat
The EXT_color_buffer_half_float object

ExtDisjointTimerQuery
The EXT_disjoint_timer_query object

ExtFragDepth
The EXT_frag_depth object

ExtSRgb
The EXT_sRGB object

ExtShaderTextureLod
The EXT_shader_texture_lod object

ExtTextureFilterAnisotropic
The EXT_texture_filter_anisotropic object

ExtendableEvent
The ExtendableEvent object

ExtendableEventInit
The ExtendableEventInit dictionary

ExtendableMessageEvent
The ExtendableMessageEvent object

ExtendableMessageEventInit
The ExtendableMessageEventInit dictionary

External
The External object

FakePluginMimeEntry
The FakePluginMimeEntry dictionary

FakePluginTagInit
The FakePluginTagInit dictionary

FetchEvent
The FetchEvent object

FetchEventInit
The FetchEventInit dictionary

FetchObserver
The FetchObserver object

FetchReadableStreamReadDataArray
The FetchReadableStreamReadDataArray dictionary

FetchReadableStreamReadDataDone
The FetchReadableStreamReadDataDone dictionary

File
The File object

FileCallback

FileList
The FileList object

FilePropertyBag
The FilePropertyBag dictionary

FileReader
The FileReader object

FileReaderSync
The FileReaderSync object

FileSystem
The FileSystem object

FileSystemDirectoryEntry
The FileSystemDirectoryEntry object

FileSystemDirectoryReader
The FileSystemDirectoryReader object

FileSystemEntriesCallback

FileSystemEntry
The FileSystemEntry object

FileSystemEntryCallback

FileSystemFileEntry
The FileSystemFileEntry object

FileSystemFlags
The FileSystemFlags dictionary

FocusEvent
The FocusEvent object

FocusEventInit
The FocusEventInit dictionary

FontFace
The FontFace object

FontFaceDescriptors
The FontFaceDescriptors dictionary

FontFaceSet
The FontFaceSet object

FontFaceSetIterator
The FontFaceSetIterator object

FontFaceSetIteratorResult
The FontFaceSetIteratorResult dictionary

FontFaceSetLoadEvent
The FontFaceSetLoadEvent object

FontFaceSetLoadEventInit
The FontFaceSetLoadEventInit dictionary

FormData
The FormData object

FuzzingFunctions
The FuzzingFunctions object

GainNode
The GainNode object

GainOptions
The GainOptions dictionary

Gamepad
The Gamepad object

GamepadAxisMoveEvent
The GamepadAxisMoveEvent object

GamepadAxisMoveEventInit
The GamepadAxisMoveEventInit dictionary

GamepadButton
The GamepadButton object

GamepadButtonEvent
The GamepadButtonEvent object

GamepadButtonEventInit
The GamepadButtonEventInit dictionary

GamepadEvent
The GamepadEvent object

GamepadEventInit
The GamepadEventInit dictionary

GamepadHapticActuator
The GamepadHapticActuator object

GamepadPose
The GamepadPose object

GamepadServiceTest
The GamepadServiceTest object

Geolocation
The Geolocation object

GetNotificationOptions
The GetNotificationOptions dictionary

GetRootNodeOptions
The GetRootNodeOptions dictionary

GetUserMediaRequest
The GetUserMediaRequest object

GroupedHistoryEventInit
The GroupedHistoryEventInit dictionary

HalfOpenInfoDict
The HalfOpenInfoDict dictionary

HashChangeEvent
The HashChangeEvent object

HashChangeEventInit
The HashChangeEventInit dictionary

Headers
The Headers object

HiddenPluginEventInit
The HiddenPluginEventInit dictionary

History
The History object

HitRegionOptions
The HitRegionOptions dictionary

HkdfParams
The HkdfParams dictionary

HmacDerivedKeyParams
The HmacDerivedKeyParams dictionary

HmacImportParams
The HmacImportParams dictionary

HmacKeyAlgorithm
The HmacKeyAlgorithm dictionary

HmacKeyGenParams
The HmacKeyGenParams dictionary

HtmlAllCollection
The HTMLAllCollection object

HtmlAnchorElement
The HTMLAnchorElement object

HtmlAreaElement
The HTMLAreaElement object

HtmlAudioElement
The HTMLAudioElement object

HtmlBaseElement
The HTMLBaseElement object

HtmlBodyElement
The HTMLBodyElement object

HtmlBrElement
The HTMLBRElement object

HtmlButtonElement
The HTMLButtonElement object

HtmlCanvasElement
The HTMLCanvasElement object

HtmlCollection
The HTMLCollection object

HtmlDListElement
The HTMLDListElement object

HtmlDataElement
The HTMLDataElement object

HtmlDataListElement
The HTMLDataListElement object

HtmlDetailsElement
The HTMLDetailsElement object

HtmlDialogElement
The HTMLDialogElement object

HtmlDirectoryElement
The HTMLDirectoryElement object

HtmlDivElement
The HTMLDivElement object

HtmlDocument
The HTMLDocument object

HtmlElement
The HTMLElement object

HtmlEmbedElement
The HTMLEmbedElement object

HtmlFieldSetElement
The HTMLFieldSetElement object

HtmlFontElement
The HTMLFontElement object

HtmlFormControlsCollection
The HTMLFormControlsCollection object

HtmlFormElement
The HTMLFormElement object

HtmlFrameElement
The HTMLFrameElement object

HtmlFrameSetElement
The HTMLFrameSetElement object

HtmlHeadElement
The HTMLHeadElement object

HtmlHeadingElement
The HTMLHeadingElement object

HtmlHrElement
The HTMLHRElement object

HtmlHtmlElement
The HTMLHtmlElement object

HtmlIFrameElement
The HTMLIFrameElement object

HtmlImageElement
The HTMLImageElement object

HtmlInputElement
The HTMLInputElement object

HtmlLabelElement
The HTMLLabelElement object

HtmlLegendElement
The HTMLLegendElement object

HtmlLiElement
The HTMLLIElement object

HtmlLinkElement
The HTMLLinkElement object

HtmlMapElement
The HTMLMapElement object

HtmlMediaElement
The HTMLMediaElement object

HtmlMenuElement
The HTMLMenuElement object

HtmlMenuItemElement
The HTMLMenuItemElement object

HtmlMetaElement
The HTMLMetaElement object

HtmlMeterElement
The HTMLMeterElement object

HtmlModElement
The HTMLModElement object

HtmlOListElement
The HTMLOListElement object

HtmlObjectElement
The HTMLObjectElement object

HtmlOptGroupElement
The HTMLOptGroupElement object

HtmlOptionElement
The HTMLOptionElement object

HtmlOptionsCollection
The HTMLOptionsCollection object

HtmlOutputElement
The HTMLOutputElement object

HtmlParagraphElement
The HTMLParagraphElement object

HtmlParamElement
The HTMLParamElement object

HtmlPictureElement
The HTMLPictureElement object

HtmlPreElement
The HTMLPreElement object

HtmlProgressElement
The HTMLProgressElement object

HtmlQuoteElement
The HTMLQuoteElement object

HtmlScriptElement
The HTMLScriptElement object

HtmlSelectElement
The HTMLSelectElement object

HtmlSlotElement
The HTMLSlotElement object

HtmlSourceElement
The HTMLSourceElement object

HtmlSpanElement
The HTMLSpanElement object

HtmlStyleElement
The HTMLStyleElement object

HtmlTableCaptionElement
The HTMLTableCaptionElement object

HtmlTableCellElement
The HTMLTableCellElement object

HtmlTableColElement
The HTMLTableColElement object

HtmlTableElement
The HTMLTableElement object

HtmlTableRowElement
The HTMLTableRowElement object

HtmlTableSectionElement
The HTMLTableSectionElement object

HtmlTemplateElement
The HTMLTemplateElement object

HtmlTextAreaElement
The HTMLTextAreaElement object

HtmlTimeElement
The HTMLTimeElement object

HtmlTitleElement
The HTMLTitleElement object

HtmlTrackElement
The HTMLTrackElement object

HtmlUListElement
The HTMLUListElement object

HtmlUnknownElement
The HTMLUnknownElement object

HtmlVideoElement
The HTMLVideoElement object

HttpConnDict
The HttpConnDict dictionary

HttpConnInfo
The HttpConnInfo dictionary

HttpConnectionElement
The HttpConnectionElement dictionary

IdbCursor
The IDBCursor object

IdbCursorWithValue
The IDBCursorWithValue object

IdbDatabase
The IDBDatabase object

IdbFactory
The IDBFactory object

IdbFileHandle
The IDBFileHandle object

IdbFileMetadataParameters
The IDBFileMetadataParameters dictionary

IdbFileRequest
The IDBFileRequest object

IdbIndex
The IDBIndex object

IdbIndexParameters
The IDBIndexParameters dictionary

IdbKeyRange
The IDBKeyRange object

IdbLocaleAwareKeyRange
The IDBLocaleAwareKeyRange object

IdbMutableFile
The IDBMutableFile object

IdbObjectStore
The IDBObjectStore object

IdbObjectStoreParameters
The IDBObjectStoreParameters dictionary

IdbOpenDbOptions
The IDBOpenDBOptions dictionary

IdbOpenDbRequest
The IDBOpenDBRequest object

IdbRequest
The IDBRequest object

IdbTransaction
The IDBTransaction object

IdbVersionChangeEvent
The IDBVersionChangeEvent object

IdbVersionChangeEventInit
The IDBVersionChangeEventInit dictionary

IdleDeadline
The IdleDeadline object

IdleRequestOptions
The IdleRequestOptions dictionary

IirFilterNode
The IIRFilterNode object

IirFilterOptions
The IIRFilterOptions dictionary

ImageBitmap
The ImageBitmap object

ImageBitmapRenderingContext
The ImageBitmapRenderingContext object

ImageCapture
The ImageCapture object

ImageCaptureError
The ImageCaptureError object

ImageCaptureErrorEvent
The ImageCaptureErrorEvent object

ImageCaptureErrorEventInit
The ImageCaptureErrorEventInit dictionary

ImageData
The ImageData object

InputEvent
The InputEvent object

InputEventInit
The InputEventInit dictionary

InstallTriggerData
The InstallTriggerData dictionary

IntersectionObserver
The IntersectionObserver object

IntersectionObserverEntry
The IntersectionObserverEntry object

IntersectionObserverEntryInit
The IntersectionObserverEntryInit dictionary

IntersectionObserverInit
The IntersectionObserverInit dictionary

IntlUtils
The IntlUtils object

IterableKeyAndValueResult
The IterableKeyAndValueResult dictionary

IterableKeyOrValueResult
The IterableKeyOrValueResult dictionary

JsonWebKey
The JsonWebKey dictionary

KeyAlgorithm
The KeyAlgorithm dictionary

KeyEvent
The KeyEvent object

KeyIdsInitData
The KeyIdsInitData dictionary

KeyboardEvent
The KeyboardEvent object

KeyboardEventInit
The KeyboardEventInit dictionary

KeyframeEffect
The KeyframeEffect object

KeyframeEffectOptions
The KeyframeEffectOptions dictionary

L10nElement
The L10nElement dictionary

L10nValue
The L10nValue dictionary

LifecycleCallbacks
The LifecycleCallbacks dictionary

ListBoxObject
The ListBoxObject object

LocalMediaStream
The LocalMediaStream object

LocaleInfo
The LocaleInfo dictionary

Location
The Location object

MediaCapabilities
The MediaCapabilities object

MediaCapabilitiesInfo
The MediaCapabilitiesInfo object

MediaConfiguration
The MediaConfiguration dictionary

MediaDecodingConfiguration
The MediaDecodingConfiguration dictionary

MediaDeviceInfo
The MediaDeviceInfo object

MediaDevices
The MediaDevices object

MediaElementAudioSourceNode
The MediaElementAudioSourceNode object

MediaElementAudioSourceOptions
The MediaElementAudioSourceOptions dictionary

MediaEncodingConfiguration
The MediaEncodingConfiguration dictionary

MediaEncryptedEvent
The MediaEncryptedEvent object

MediaError
The MediaError object

MediaKeyError
The MediaKeyError object

MediaKeyMessageEvent
The MediaKeyMessageEvent object

MediaKeyMessageEventInit
The MediaKeyMessageEventInit dictionary

MediaKeyNeededEventInit
The MediaKeyNeededEventInit dictionary

MediaKeySession
The MediaKeySession object

MediaKeyStatusMap
The MediaKeyStatusMap object

MediaKeySystemAccess
The MediaKeySystemAccess object

MediaKeySystemConfiguration
The MediaKeySystemConfiguration dictionary

MediaKeySystemMediaCapability
The MediaKeySystemMediaCapability dictionary

MediaKeys
The MediaKeys object

MediaKeysPolicy
The MediaKeysPolicy dictionary

MediaList
The MediaList object

MediaQueryList
The MediaQueryList object

MediaQueryListEvent
The MediaQueryListEvent object

MediaQueryListEventInit
The MediaQueryListEventInit dictionary

MediaRecorder
The MediaRecorder object

MediaRecorderErrorEvent
The MediaRecorderErrorEvent object

MediaRecorderErrorEventInit
The MediaRecorderErrorEventInit dictionary

MediaRecorderOptions
The MediaRecorderOptions dictionary

MediaSource
The MediaSource object

MediaStream
The MediaStream object

MediaStreamAudioDestinationNode
The MediaStreamAudioDestinationNode object

MediaStreamAudioSourceNode
The MediaStreamAudioSourceNode object

MediaStreamAudioSourceOptions
The MediaStreamAudioSourceOptions dictionary

MediaStreamConstraints
The MediaStreamConstraints dictionary

MediaStreamError
The MediaStreamError object

MediaStreamEvent
The MediaStreamEvent object

MediaStreamEventInit
The MediaStreamEventInit dictionary

MediaStreamTrack
The MediaStreamTrack object

MediaStreamTrackEvent
The MediaStreamTrackEvent object

MediaStreamTrackEventInit
The MediaStreamTrackEventInit dictionary

MediaTrackConstraintSet
The MediaTrackConstraintSet dictionary

MediaTrackConstraints
The MediaTrackConstraints dictionary

MediaTrackSettings
The MediaTrackSettings dictionary

MediaTrackSupportedConstraints
The MediaTrackSupportedConstraints dictionary

MessageChannel
The MessageChannel object

MessageEvent
The MessageEvent object

MessageEventInit
The MessageEventInit dictionary

MessagePort
The MessagePort object

MidiAccess
The MIDIAccess object

MidiConnectionEvent
The MIDIConnectionEvent object

MidiConnectionEventInit
The MIDIConnectionEventInit dictionary

MidiInput
The MIDIInput object

MidiInputMap
The MIDIInputMap object

MidiMessageEvent
The MIDIMessageEvent object

MidiMessageEventInit
The MIDIMessageEventInit dictionary

MidiOptions
The MIDIOptions dictionary

MidiOutput
The MIDIOutput object

MidiOutputMap
The MIDIOutputMap object

MidiPort
The MIDIPort object

MimeType
The MimeType object

MimeTypeArray
The MimeTypeArray object

MouseEvent
The MouseEvent object

MouseEventInit
The MouseEventInit dictionary

MouseScrollEvent
The MouseScrollEvent object

MozDebug
The MOZ_debug object

MutationEvent
The MutationEvent object

MutationObserver
The MutationObserver object

MutationObserverInit
The MutationObserverInit dictionary

MutationObservingInfo
The MutationObservingInfo dictionary

MutationRecord
The MutationRecord object

NamedNodeMap
The NamedNodeMap object

NativeOsFileReadOptions
The NativeOSFileReadOptions dictionary

NativeOsFileWriteAtomicOptions
The NativeOSFileWriteAtomicOptions dictionary

Navigator
The Navigator object

NavigatorAutomationInformation
The NavigatorAutomationInformation object

NetworkCommandOptions
The NetworkCommandOptions dictionary

NetworkInformation
The NetworkInformation object

NetworkResultOptions
The NetworkResultOptions dictionary

Node
The Node object

NodeFilter

NodeIterator
The NodeIterator object

NodeList
The NodeList object

Notification
The Notification object

NotificationBehavior
The NotificationBehavior dictionary

NotificationEvent
The NotificationEvent object

NotificationEventInit
The NotificationEventInit dictionary

NotificationOptions
The NotificationOptions dictionary

ObserverCallback

OesElementIndexUint
The OES_element_index_uint object

OesStandardDerivatives
The OES_standard_derivatives object

OesTextureFloat
The OES_texture_float object

OesTextureFloatLinear
The OES_texture_float_linear object

OesTextureHalfFloat
The OES_texture_half_float object

OesTextureHalfFloatLinear
The OES_texture_half_float_linear object

OesVertexArrayObject
The OES_vertex_array_object object

OfflineAudioCompletionEvent
The OfflineAudioCompletionEvent object

OfflineAudioCompletionEventInit
The OfflineAudioCompletionEventInit dictionary

OfflineAudioContext
The OfflineAudioContext object

OfflineAudioContextOptions
The OfflineAudioContextOptions dictionary

OfflineResourceList
The OfflineResourceList object

OffscreenCanvas
The OffscreenCanvas object

OpenWindowEventDetail
The OpenWindowEventDetail dictionary

OptionalEffectTiming
The OptionalEffectTiming dictionary

OscillatorNode
The OscillatorNode object

OscillatorOptions
The OscillatorOptions dictionary

PageTransitionEvent
The PageTransitionEvent object

PageTransitionEventInit
The PageTransitionEventInit dictionary

PaintRequest
The PaintRequest object

PaintRequestList
The PaintRequestList object

PaintWorkletGlobalScope
The PaintWorkletGlobalScope object

PannerNode
The PannerNode object

PannerOptions
The PannerOptions dictionary

Path2d
The Path2D object

PaymentAddress
The PaymentAddress object

PaymentMethodChangeEvent
The PaymentMethodChangeEvent object

PaymentMethodChangeEventInit
The PaymentMethodChangeEventInit dictionary

PaymentRequestUpdateEvent
The PaymentRequestUpdateEvent object

PaymentRequestUpdateEventInit
The PaymentRequestUpdateEventInit dictionary

PaymentResponse
The PaymentResponse object

Pbkdf2Params
The Pbkdf2Params dictionary

Performance
The Performance object

PerformanceEntry
The PerformanceEntry object

PerformanceEntryEventInit
The PerformanceEntryEventInit dictionary

PerformanceEntryFilterOptions
The PerformanceEntryFilterOptions dictionary

PerformanceMark
The PerformanceMark object

PerformanceMeasure
The PerformanceMeasure object

PerformanceNavigation
The PerformanceNavigation object

PerformanceNavigationTiming
The PerformanceNavigationTiming object

PerformanceObserver
The PerformanceObserver object

PerformanceObserverEntryList
The PerformanceObserverEntryList object

PerformanceObserverInit
The PerformanceObserverInit dictionary

PerformanceResourceTiming
The PerformanceResourceTiming object

PerformanceServerTiming
The PerformanceServerTiming object

PerformanceTiming
The PerformanceTiming object

PeriodicWave
The PeriodicWave object

PeriodicWaveConstraints
The PeriodicWaveConstraints dictionary

PeriodicWaveOptions
The PeriodicWaveOptions dictionary

PermissionDescriptor
The PermissionDescriptor dictionary

PermissionStatus
The PermissionStatus object

Permissions
The Permissions object

Plugin
The Plugin object

PluginArray
The PluginArray object

PluginCrashedEventInit
The PluginCrashedEventInit dictionary

PointerEvent
The PointerEvent object

PointerEventInit
The PointerEventInit dictionary

PopStateEvent
The PopStateEvent object

PopStateEventInit
The PopStateEventInit dictionary

PopupBlockedEvent
The PopupBlockedEvent object

PopupBlockedEventInit
The PopupBlockedEventInit dictionary

Position
The Position object

PositionError
The PositionError object

PositionOptions
The PositionOptions dictionary

Presentation
The Presentation object

PresentationAvailability
The PresentationAvailability object

PresentationConnection
The PresentationConnection object

PresentationConnectionAvailableEvent
The PresentationConnectionAvailableEvent object

PresentationConnectionAvailableEventInit
The PresentationConnectionAvailableEventInit dictionary

PresentationConnectionCloseEvent
The PresentationConnectionCloseEvent object

PresentationConnectionCloseEventInit
The PresentationConnectionCloseEventInit dictionary

PresentationConnectionList
The PresentationConnectionList object

PresentationReceiver
The PresentationReceiver object

PresentationRequest
The PresentationRequest object

ProcessingInstruction
The ProcessingInstruction object

ProfileTimelineLayerRect
The ProfileTimelineLayerRect dictionary

ProfileTimelineMarker
The ProfileTimelineMarker dictionary

ProfileTimelineStackFrame
The ProfileTimelineStackFrame dictionary

ProgressEvent
The ProgressEvent object

ProgressEventInit
The ProgressEventInit dictionary

PromiseNativeHandler
The PromiseNativeHandler object

PromiseRejectionEvent
The PromiseRejectionEvent object

PromiseRejectionEventInit
The PromiseRejectionEventInit dictionary

PublicKeyCredential
The PublicKeyCredential object

PublicKeyCredentialCreationOptions
The PublicKeyCredentialCreationOptions dictionary

PublicKeyCredentialDescriptor
The PublicKeyCredentialDescriptor dictionary

PublicKeyCredentialEntity
The PublicKeyCredentialEntity dictionary

PublicKeyCredentialParameters
The PublicKeyCredentialParameters dictionary

PublicKeyCredentialRequestOptions
The PublicKeyCredentialRequestOptions dictionary

PublicKeyCredentialRpEntity
The PublicKeyCredentialRpEntity dictionary

PublicKeyCredentialUserEntity
The PublicKeyCredentialUserEntity dictionary

PushEvent
The PushEvent object

PushEventInit
The PushEventInit dictionary

PushManager
The PushManager object

PushMessageData
The PushMessageData object

PushSubscription
The PushSubscription object

PushSubscriptionInit
The PushSubscriptionInit dictionary

PushSubscriptionJson
The PushSubscriptionJSON dictionary

PushSubscriptionKeys
The PushSubscriptionKeys dictionary

PushSubscriptionOptions
The PushSubscriptionOptions object

PushSubscriptionOptionsInit
The PushSubscriptionOptionsInit dictionary

RadioNodeList
The RadioNodeList object

Range
The Range object

RcwnPerfStats
The RcwnPerfStats dictionary

RcwnStatus
The RcwnStatus dictionary

ReadableStream
The ReadableStream object

RegisterRequest
The RegisterRequest dictionary

RegisterResponse
The RegisterResponse dictionary

RegisteredKey
The RegisteredKey dictionary

RegistrationOptions
The RegistrationOptions dictionary

Request
The Request object

RequestInit
The RequestInit dictionary

RequestMediaKeySystemAccessNotification
The RequestMediaKeySystemAccessNotification dictionary

Response
The Response object

ResponseInit
The ResponseInit dictionary

RsaHashedImportParams
The RsaHashedImportParams dictionary

RsaOaepParams
The RsaOaepParams dictionary

RsaOtherPrimesInfo
The RsaOtherPrimesInfo dictionary

RsaPssParams
The RsaPssParams dictionary

RtcAnswerOptions
The RTCAnswerOptions dictionary

RtcCertificate
The RTCCertificate object

RtcCertificateExpiration
The RTCCertificateExpiration dictionary

RtcCodecStats
The RTCCodecStats dictionary

RtcConfiguration
The RTCConfiguration dictionary

RtcDataChannel
The RTCDataChannel object

RtcDataChannelEvent
The RTCDataChannelEvent object

RtcDataChannelEventInit
The RTCDataChannelEventInit dictionary

RtcDataChannelInit
The RTCDataChannelInit dictionary

RtcFecParameters
The RTCFecParameters dictionary

RtcIceCandidate
The RTCIceCandidate object

RtcIceCandidateInit
The RTCIceCandidateInit dictionary

RtcIceCandidatePairStats
The RTCIceCandidatePairStats dictionary

RtcIceCandidateStats
The RTCIceCandidateStats dictionary

RtcIceComponentStats
The RTCIceComponentStats dictionary

RtcIceServer
The RTCIceServer dictionary

RtcIdentityAssertion
The RTCIdentityAssertion dictionary

RtcIdentityAssertionResult
The RTCIdentityAssertionResult dictionary

RtcIdentityProvider
The RTCIdentityProvider dictionary

RtcIdentityProviderDetails
The RTCIdentityProviderDetails dictionary

RtcIdentityProviderOptions
The RTCIdentityProviderOptions dictionary

RtcIdentityProviderRegistrar
The RTCIdentityProviderRegistrar object

RtcIdentityValidationResult
The RTCIdentityValidationResult dictionary

RtcInboundRtpStreamStats
The RTCInboundRTPStreamStats dictionary

RtcMediaStreamStats
The RTCMediaStreamStats dictionary

RtcMediaStreamTrackStats
The RTCMediaStreamTrackStats dictionary

RtcOfferAnswerOptions
The RTCOfferAnswerOptions dictionary

RtcOfferOptions
The RTCOfferOptions dictionary

RtcOutboundRtpStreamStats
The RTCOutboundRTPStreamStats dictionary

RtcPeerConnection
The RTCPeerConnection object

RtcPeerConnectionIceEvent
The RTCPeerConnectionIceEvent object

RtcPeerConnectionIceEventInit
The RTCPeerConnectionIceEventInit dictionary

RtcRtcpParameters
The RTCRtcpParameters dictionary

RtcRtpCodecParameters
The RTCRtpCodecParameters dictionary

RtcRtpContributingSource
The RTCRtpContributingSource dictionary

RtcRtpEncodingParameters
The RTCRtpEncodingParameters dictionary

RtcRtpHeaderExtensionParameters
The RTCRtpHeaderExtensionParameters dictionary

RtcRtpParameters
The RTCRtpParameters dictionary

RtcRtpReceiver
The RTCRtpReceiver object

RtcRtpSender
The RTCRtpSender object

RtcRtpSourceEntry
The RTCRtpSourceEntry dictionary

RtcRtpSynchronizationSource
The RTCRtpSynchronizationSource dictionary

RtcRtpTransceiver
The RTCRtpTransceiver object

RtcRtpTransceiverInit
The RTCRtpTransceiverInit dictionary

RtcRtxParameters
The RTCRtxParameters dictionary

RtcSessionDescription
The RTCSessionDescription object

RtcSessionDescriptionInit
The RTCSessionDescriptionInit dictionary

RtcStats
The RTCStats dictionary

RtcStatsReport
The RTCStatsReport object

RtcStatsReportInternal
The RTCStatsReportInternal dictionary

RtcTrackEvent
The RTCTrackEvent object

RtcTrackEventInit
The RTCTrackEventInit dictionary

RtcTransportStats
The RTCTransportStats dictionary

RtcdtmfSender
The RTCDTMFSender object

RtcdtmfToneChangeEvent
The RTCDTMFToneChangeEvent object

RtcdtmfToneChangeEventInit
The RTCDTMFToneChangeEventInit dictionary

RtcrtpContributingSourceStats
The RTCRTPContributingSourceStats dictionary

RtcrtpStreamStats
The RTCRTPStreamStats dictionary

Screen
The Screen object

ScreenLuminance
The ScreenLuminance object

ScreenOrientation
The ScreenOrientation object

ScriptProcessorNode
The ScriptProcessorNode object

ScrollAreaEvent
The ScrollAreaEvent object

ScrollBoxObject
The ScrollBoxObject object

ScrollIntoViewOptions
The ScrollIntoViewOptions dictionary

ScrollOptions
The ScrollOptions dictionary

ScrollToOptions
The ScrollToOptions dictionary

ScrollViewChangeEventInit
The ScrollViewChangeEventInit dictionary

SecurityPolicyViolationEvent
The SecurityPolicyViolationEvent object

SecurityPolicyViolationEventInit
The SecurityPolicyViolationEventInit dictionary

Selection
The Selection object

ServerSocketOptions
The ServerSocketOptions dictionary

ServiceWorker
The ServiceWorker object

ServiceWorkerContainer
The ServiceWorkerContainer object

ServiceWorkerGlobalScope
The ServiceWorkerGlobalScope object

ServiceWorkerRegistration
The ServiceWorkerRegistration object

ShadowRoot
The ShadowRoot object

ShadowRootInit
The ShadowRootInit dictionary

SharedWorker
The SharedWorker object

SharedWorkerGlobalScope
The SharedWorkerGlobalScope object

SignResponse
The SignResponse dictionary

SocketElement
The SocketElement dictionary

SocketOptions
The SocketOptions dictionary

SocketsDict
The SocketsDict dictionary

SourceBuffer
The SourceBuffer object

SourceBufferList
The SourceBufferList object

SpeechGrammar
The SpeechGrammar object

SpeechGrammarList
The SpeechGrammarList object

SpeechRecognition
The SpeechRecognition object

SpeechRecognitionAlternative
The SpeechRecognitionAlternative object

SpeechRecognitionError
The SpeechRecognitionError object

SpeechRecognitionErrorInit
The SpeechRecognitionErrorInit dictionary

SpeechRecognitionEvent
The SpeechRecognitionEvent object

SpeechRecognitionEventInit
The SpeechRecognitionEventInit dictionary

SpeechRecognitionResult
The SpeechRecognitionResult object

SpeechRecognitionResultList
The SpeechRecognitionResultList object

SpeechSynthesis
The SpeechSynthesis object

SpeechSynthesisErrorEvent
The SpeechSynthesisErrorEvent object

SpeechSynthesisErrorEventInit
The SpeechSynthesisErrorEventInit dictionary

SpeechSynthesisEvent
The SpeechSynthesisEvent object

SpeechSynthesisEventInit
The SpeechSynthesisEventInit dictionary

SpeechSynthesisUtterance
The SpeechSynthesisUtterance object

SpeechSynthesisVoice
The SpeechSynthesisVoice object

StereoPannerNode
The StereoPannerNode object

StereoPannerOptions
The StereoPannerOptions dictionary

Storage
The Storage object

StorageEstimate
The StorageEstimate dictionary

StorageEvent
The StorageEvent object

StorageEventInit
The StorageEventInit dictionary

StorageManager
The StorageManager object

StyleRuleChangeEventInit
The StyleRuleChangeEventInit dictionary

StyleSheet
The StyleSheet object

StyleSheetApplicableStateChangeEventInit
The StyleSheetApplicableStateChangeEventInit dictionary

StyleSheetChangeEventInit
The StyleSheetChangeEventInit dictionary

StyleSheetList
The StyleSheetList object

SubtleCrypto
The SubtleCrypto object

SvgAngle
The SVGAngle object

SvgAnimateElement
The SVGAnimateElement object

SvgAnimateMotionElement
The SVGAnimateMotionElement object

SvgAnimateTransformElement
The SVGAnimateTransformElement object

SvgAnimatedAngle
The SVGAnimatedAngle object

SvgAnimatedBoolean
The SVGAnimatedBoolean object

SvgAnimatedEnumeration
The SVGAnimatedEnumeration object

SvgAnimatedInteger
The SVGAnimatedInteger object

SvgAnimatedLength
The SVGAnimatedLength object

SvgAnimatedLengthList
The SVGAnimatedLengthList object

SvgAnimatedNumber
The SVGAnimatedNumber object

SvgAnimatedNumberList
The SVGAnimatedNumberList object

SvgAnimatedPreserveAspectRatio
The SVGAnimatedPreserveAspectRatio object

SvgAnimatedRect
The SVGAnimatedRect object

SvgAnimatedString
The SVGAnimatedString object

SvgAnimatedTransformList
The SVGAnimatedTransformList object

SvgAnimationElement
The SVGAnimationElement object

SvgBoundingBoxOptions
The SVGBoundingBoxOptions dictionary

SvgCircleElement
The SVGCircleElement object

SvgClipPathElement
The SVGClipPathElement object

SvgComponentTransferFunctionElement
The SVGComponentTransferFunctionElement object

SvgDefsElement
The SVGDefsElement object

SvgDescElement
The SVGDescElement object

SvgElement
The SVGElement object

SvgEllipseElement
The SVGEllipseElement object

SvgFilterElement
The SVGFilterElement object

SvgForeignObjectElement
The SVGForeignObjectElement object

SvgGeometryElement
The SVGGeometryElement object

SvgGradientElement
The SVGGradientElement object

SvgGraphicsElement
The SVGGraphicsElement object

SvgImageElement
The SVGImageElement object

SvgLength
The SVGLength object

SvgLengthList
The SVGLengthList object

SvgLineElement
The SVGLineElement object

SvgLinearGradientElement
The SVGLinearGradientElement object

SvgMarkerElement
The SVGMarkerElement object

SvgMaskElement
The SVGMaskElement object

SvgMatrix
The SVGMatrix object

SvgMetadataElement
The SVGMetadataElement object

SvgNumber
The SVGNumber object

SvgNumberList
The SVGNumberList object

SvgPathElement
The SVGPathElement object

SvgPathSeg
The SVGPathSeg object

SvgPathSegArcAbs
The SVGPathSegArcAbs object

SvgPathSegArcRel
The SVGPathSegArcRel object

SvgPathSegClosePath
The SVGPathSegClosePath object

SvgPathSegCurvetoCubicAbs
The SVGPathSegCurvetoCubicAbs object

SvgPathSegCurvetoCubicRel
The SVGPathSegCurvetoCubicRel object

SvgPathSegCurvetoCubicSmoothAbs
The SVGPathSegCurvetoCubicSmoothAbs object

SvgPathSegCurvetoCubicSmoothRel
The SVGPathSegCurvetoCubicSmoothRel object

SvgPathSegCurvetoQuadraticAbs
The SVGPathSegCurvetoQuadraticAbs object

SvgPathSegCurvetoQuadraticRel
The SVGPathSegCurvetoQuadraticRel object

SvgPathSegCurvetoQuadraticSmoothAbs
The SVGPathSegCurvetoQuadraticSmoothAbs object

SvgPathSegCurvetoQuadraticSmoothRel
The SVGPathSegCurvetoQuadraticSmoothRel object

SvgPathSegLinetoAbs
The SVGPathSegLinetoAbs object

SvgPathSegLinetoHorizontalAbs
The SVGPathSegLinetoHorizontalAbs object

SvgPathSegLinetoHorizontalRel
The SVGPathSegLinetoHorizontalRel object

SvgPathSegLinetoRel
The SVGPathSegLinetoRel object

SvgPathSegLinetoVerticalAbs
The SVGPathSegLinetoVerticalAbs object

SvgPathSegLinetoVerticalRel
The SVGPathSegLinetoVerticalRel object

SvgPathSegList
The SVGPathSegList object

SvgPathSegMovetoAbs
The SVGPathSegMovetoAbs object

SvgPathSegMovetoRel
The SVGPathSegMovetoRel object

SvgPatternElement
The SVGPatternElement object

SvgPoint
The SVGPoint object

SvgPointList
The SVGPointList object

SvgPolygonElement
The SVGPolygonElement object

SvgPolylineElement
The SVGPolylineElement object

SvgPreserveAspectRatio
The SVGPreserveAspectRatio object

SvgRadialGradientElement
The SVGRadialGradientElement object

SvgRect
The SVGRect object

SvgRectElement
The SVGRectElement object

SvgScriptElement
The SVGScriptElement object

SvgSetElement
The SVGSetElement object

SvgStopElement
The SVGStopElement object

SvgStringList
The SVGStringList object

SvgStyleElement
The SVGStyleElement object

SvgSwitchElement
The SVGSwitchElement object

SvgSymbolElement
The SVGSymbolElement object

SvgTextContentElement
The SVGTextContentElement object

SvgTextElement
The SVGTextElement object

SvgTextPathElement
The SVGTextPathElement object

SvgTextPositioningElement
The SVGTextPositioningElement object

SvgTitleElement
The SVGTitleElement object

SvgTransform
The SVGTransform object

SvgTransformList
The SVGTransformList object

SvgUnitTypes
The SVGUnitTypes object

SvgUseElement
The SVGUseElement object

SvgViewElement
The SVGViewElement object

SvgZoomAndPan
The SVGZoomAndPan object

SvgaElement
The SVGAElement object

SvgfeBlendElement
The SVGFEBlendElement object

SvgfeColorMatrixElement
The SVGFEColorMatrixElement object

SvgfeComponentTransferElement
The SVGFEComponentTransferElement object

SvgfeCompositeElement
The SVGFECompositeElement object

SvgfeConvolveMatrixElement
The SVGFEConvolveMatrixElement object

SvgfeDiffuseLightingElement
The SVGFEDiffuseLightingElement object

SvgfeDisplacementMapElement
The SVGFEDisplacementMapElement object

SvgfeDistantLightElement
The SVGFEDistantLightElement object

SvgfeDropShadowElement
The SVGFEDropShadowElement object

SvgfeFloodElement
The SVGFEFloodElement object

SvgfeFuncAElement
The SVGFEFuncAElement object

SvgfeFuncBElement
The SVGFEFuncBElement object

SvgfeFuncGElement
The SVGFEFuncGElement object

SvgfeFuncRElement
The SVGFEFuncRElement object

SvgfeGaussianBlurElement
The SVGFEGaussianBlurElement object

SvgfeImageElement
The SVGFEImageElement object

SvgfeMergeElement
The SVGFEMergeElement object

SvgfeMergeNodeElement
The SVGFEMergeNodeElement object

SvgfeMorphologyElement
The SVGFEMorphologyElement object

SvgfeOffsetElement
The SVGFEOffsetElement object

SvgfePointLightElement
The SVGFEPointLightElement object

SvgfeSpecularLightingElement
The SVGFESpecularLightingElement object

SvgfeSpotLightElement
The SVGFESpotLightElement object

SvgfeTileElement
The SVGFETileElement object

SvgfeTurbulenceElement
The SVGFETurbulenceElement object

SvggElement
The SVGGElement object

SvgmPathElement
The SVGMPathElement object

SvgsvgElement
The SVGSVGElement object

SvgtSpanElement
The SVGTSpanElement object

TcpServerSocket
The TCPServerSocket object

TcpServerSocketEvent
The TCPServerSocketEvent object

TcpServerSocketEventInit
The TCPServerSocketEventInit dictionary

TcpSocket
The TCPSocket object

TcpSocketErrorEvent
The TCPSocketErrorEvent object

TcpSocketErrorEventInit
The TCPSocketErrorEventInit dictionary

TcpSocketEvent
The TCPSocketEvent object

TcpSocketEventInit
The TCPSocketEventInit dictionary

Text
The Text object

TextDecodeOptions
The TextDecodeOptions dictionary

TextDecoder
The TextDecoder object

TextDecoderOptions
The TextDecoderOptions dictionary

TextEncoder
The TextEncoder object

TextMetrics
The TextMetrics object

TextTrack
The TextTrack object

TextTrackCue
The TextTrackCue object

TextTrackCueList
The TextTrackCueList object

TextTrackList
The TextTrackList object

TimeEvent
The TimeEvent object

TimeRanges
The TimeRanges object

Touch
The Touch object

TouchEvent
The TouchEvent object

TouchEventInit
The TouchEventInit dictionary

TouchInit
The TouchInit dictionary

TouchList
The TouchList object

TrackEvent
The TrackEvent object

TrackEventInit
The TrackEventInit dictionary

TransitionEvent
The TransitionEvent object

TransitionEventInit
The TransitionEventInit dictionary

TreeBoxObject
The TreeBoxObject object

TreeCellInfo
The TreeCellInfo dictionary

TreeView
The TreeView object

TreeWalker
The TreeWalker object

U2f
The U2F object

U2fClientData
The U2FClientData dictionary

UdpMessageEventInit
The UDPMessageEventInit dictionary

UdpOptions
The UDPOptions dictionary

UiEvent
The UIEvent object

UiEventInit
The UIEventInit dictionary

Url
The URL object

UrlSearchParams
The URLSearchParams object

UserProximityEvent
The UserProximityEvent object

UserProximityEventInit
The UserProximityEventInit dictionary

ValidityState
The ValidityState object

VideoConfiguration
The VideoConfiguration dictionary

VideoPlaybackQuality
The VideoPlaybackQuality object

VideoStreamTrack
The VideoStreamTrack object

VideoTrack
The VideoTrack object

VideoTrackList
The VideoTrackList object

VoidCallback

VrDisplay
The VRDisplay object

VrDisplayCapabilities
The VRDisplayCapabilities object

VrEyeParameters
The VREyeParameters object

VrFieldOfView
The VRFieldOfView object

VrFrameData
The VRFrameData object

VrLayer
The VRLayer dictionary

VrMockController
The VRMockController object

VrMockDisplay
The VRMockDisplay object

VrPose
The VRPose object

VrServiceTest
The VRServiceTest object

VrStageParameters
The VRStageParameters object

VrSubmitFrameResult
The VRSubmitFrameResult object

VttCue
The VTTCue object

VttRegion
The VTTRegion object

WaveShaperNode
The WaveShaperNode object

WaveShaperOptions
The WaveShaperOptions dictionary

WebGl2RenderingContext
The WebGL2RenderingContext object

WebGlActiveInfo
The WebGLActiveInfo object

WebGlBuffer
The WebGLBuffer object

WebGlContextAttributes
The WebGLContextAttributes dictionary

WebGlContextEvent
The WebGLContextEvent object

WebGlContextEventInit
The WebGLContextEventInit dictionary

WebGlFramebuffer
The WebGLFramebuffer object

WebGlProgram
The WebGLProgram object

WebGlQuery
The WebGLQuery object

WebGlRenderbuffer
The WebGLRenderbuffer object

WebGlRenderingContext
The WebGLRenderingContext object

WebGlSampler
The WebGLSampler object

WebGlShader
The WebGLShader object

WebGlShaderPrecisionFormat
The WebGLShaderPrecisionFormat object

WebGlSync
The WebGLSync object

WebGlTexture
The WebGLTexture object

WebGlTransformFeedback
The WebGLTransformFeedback object

WebGlUniformLocation
The WebGLUniformLocation object

WebGlVertexArrayObject
The WebGLVertexArrayObject object

WebGpu
The WebGPU object

WebGpuAdapter
The WebGPUAdapter object

WebGpuAdapterDescriptor
The WebGPUAdapterDescriptor dictionary

WebGpuAttachmentState
The WebGPUAttachmentState object

WebGpuAttachmentStateDescriptor
The WebGPUAttachmentStateDescriptor dictionary

WebGpuBindGroup
The WebGPUBindGroup object

WebGpuBindGroupBinding
The WebGPUBindGroupBinding dictionary

WebGpuBindGroupDescriptor
The WebGPUBindGroupDescriptor dictionary

WebGpuBindGroupLayout
The WebGPUBindGroupLayout object

WebGpuBindGroupLayoutDescriptor
The WebGPUBindGroupLayoutDescriptor dictionary

WebGpuBinding
The WebGPUBinding dictionary

WebGpuBindingType
The WebGPUBindingType object

WebGpuBlendDescriptor
The WebGPUBlendDescriptor dictionary

WebGpuBlendFactor
The WebGPUBlendFactor object

WebGpuBlendOperation
The WebGPUBlendOperation object

WebGpuBlendState
The WebGPUBlendState object

WebGpuBlendStateDescriptor
The WebGPUBlendStateDescriptor dictionary

WebGpuBuffer
The WebGPUBuffer object

WebGpuBufferBinding
The WebGPUBufferBinding dictionary

WebGpuBufferDescriptor
The WebGPUBufferDescriptor dictionary

WebGpuBufferUsage
The WebGPUBufferUsage object

WebGpuColorWriteBits
The WebGPUColorWriteBits object

WebGpuCommandBuffer
The WebGPUCommandBuffer object

WebGpuCommandEncoder
The WebGPUCommandEncoder object

WebGpuCommandEncoderDescriptor
The WebGPUCommandEncoderDescriptor dictionary

WebGpuCompareFunction
The WebGPUCompareFunction object

WebGpuComputePipeline
The WebGPUComputePipeline object

WebGpuComputePipelineDescriptor
The WebGPUComputePipelineDescriptor dictionary

WebGpuDepthStencilState
The WebGPUDepthStencilState object

WebGpuDepthStencilStateDescriptor
The WebGPUDepthStencilStateDescriptor dictionary

WebGpuDevice
The WebGPUDevice object

WebGpuDeviceDescriptor
The WebGPUDeviceDescriptor dictionary

WebGpuExtensions
The WebGPUExtensions dictionary

WebGpuFence
The WebGPUFence object

WebGpuFilterMode
The WebGPUFilterMode object

WebGpuIndexFormat
The WebGPUIndexFormat object

WebGpuInputState
The WebGPUInputState object

WebGpuInputStateDescriptor
The WebGPUInputStateDescriptor dictionary

WebGpuInputStepMode
The WebGPUInputStepMode object

WebGpuLimits
The WebGPULimits dictionary

WebGpuLoadOp
The WebGPULoadOp object

WebGpuLogEntry
The WebGPULogEntry object

WebGpuPipelineDescriptorBase
The WebGPUPipelineDescriptorBase dictionary

WebGpuPipelineLayout
The WebGPUPipelineLayout object

WebGpuPipelineLayoutDescriptor
The WebGPUPipelineLayoutDescriptor dictionary

WebGpuPipelineStageDescriptor
The WebGPUPipelineStageDescriptor dictionary

WebGpuPrimitiveTopology
The WebGPUPrimitiveTopology object

WebGpuQueue
The WebGPUQueue object

WebGpuRenderPassAttachmentDescriptor
The WebGPURenderPassAttachmentDescriptor dictionary

WebGpuRenderPassDescriptor
The WebGPURenderPassDescriptor dictionary

WebGpuRenderPipeline
The WebGPURenderPipeline object

WebGpuRenderPipelineDescriptor
The WebGPURenderPipelineDescriptor dictionary

WebGpuSampler
The WebGPUSampler object

WebGpuSamplerDescriptor
The WebGPUSamplerDescriptor dictionary

WebGpuShaderModule
The WebGPUShaderModule object

WebGpuShaderModuleDescriptor
The WebGPUShaderModuleDescriptor dictionary

WebGpuShaderStage
The WebGPUShaderStage object

WebGpuShaderStageBit
The WebGPUShaderStageBit object

WebGpuStencilOperation
The WebGPUStencilOperation object

WebGpuStencilStateFaceDescriptor
The WebGPUStencilStateFaceDescriptor dictionary

WebGpuStoreOp
The WebGPUStoreOp object

WebGpuSwapChain
The WebGPUSwapChain object

WebGpuSwapChainDescriptor
The WebGPUSwapChainDescriptor dictionary

WebGpuTexture
The WebGPUTexture object

WebGpuTextureDescriptor
The WebGPUTextureDescriptor dictionary

WebGpuTextureDimension
The WebGPUTextureDimension object

WebGpuTextureFormat
The WebGPUTextureFormat object

WebGpuTextureUsage
The WebGPUTextureUsage object

WebGpuTextureView
The WebGPUTextureView object

WebGpuTextureViewDescriptor
The WebGPUTextureViewDescriptor dictionary

WebGpuVertexAttributeDescriptor
The WebGPUVertexAttributeDescriptor dictionary

WebGpuVertexFormat
The WebGPUVertexFormat object

WebGpuVertexInputDescriptor
The WebGPUVertexInputDescriptor dictionary

WebKitCssMatrix
The WebKitCSSMatrix object

WebSocket
The WebSocket object

WebSocketDict
The WebSocketDict dictionary

WebSocketElement
The WebSocketElement dictionary

WebglColorBufferFloat
The WEBGL_color_buffer_float object

WebglCompressedTextureAstc
The WEBGL_compressed_texture_astc object

WebglCompressedTextureAtc
The WEBGL_compressed_texture_atc object

WebglCompressedTextureEtc
The WEBGL_compressed_texture_etc object

WebglCompressedTextureEtc1
The WEBGL_compressed_texture_etc1 object

WebglCompressedTexturePvrtc
The WEBGL_compressed_texture_pvrtc object

WebglCompressedTextureS3tc
The WEBGL_compressed_texture_s3tc object

WebglCompressedTextureS3tcSrgb
The WEBGL_compressed_texture_s3tc_srgb object

WebglDebugRendererInfo
The WEBGL_debug_renderer_info object

WebglDebugShaders
The WEBGL_debug_shaders object

WebglDepthTexture
The WEBGL_depth_texture object

WebglDrawBuffers
The WEBGL_draw_buffers object

WebglLoseContext
The WEBGL_lose_context object

WebrtcGlobalStatisticsReport
The WebrtcGlobalStatisticsReport dictionary

WheelEvent
The WheelEvent object

WheelEventInit
The WheelEventInit dictionary

WidevineCdmManifest
The WidevineCDMManifest dictionary

Window
The Window object

WindowClient
The WindowClient object

Worker
The Worker object

WorkerDebuggerGlobalScope
The WorkerDebuggerGlobalScope object

WorkerGlobalScope
The WorkerGlobalScope object

WorkerLocation
The WorkerLocation object

WorkerNavigator
The WorkerNavigator object

WorkerOptions
The WorkerOptions dictionary

Worklet
The Worklet object

WorkletGlobalScope
The WorkletGlobalScope object

XPathExpression
The XPathExpression object

XPathNsResolver

XPathResult
The XPathResult object

XmlDocument
The XMLDocument object

XmlHttpRequest
The XMLHttpRequest object

XmlHttpRequestEventTarget
The XMLHttpRequestEventTarget object

XmlHttpRequestUpload
The XMLHttpRequestUpload object

XmlSerializer
The XMLSerializer object

XsltProcessor
The XSLTProcessor object

Enums

AlignSetting

AnimationPlayState

AttestationConveyancePreference

AudioContextState

AuthenticatorAttachment

AuthenticatorTransport

AutoKeyword

BasicCardType

BinaryType

BiquadFilterType

BrowserFindCaseSensitivity

BrowserFindDirection

CacheStorageNamespace

CanvasWindingRule

CaretChangedReason

ChannelCountMode

ChannelInterpretation

ChannelPixelLayoutDataType

CheckerboardReason

ClientType

CompositeOperation

ConnectionType

ConsoleLevel

ConsoleLogLevel

CssBoxType

CssStyleSheetParsingMode

DecoderDoctorNotificationType

DirectionSetting

DistanceModelType

DomRequestReadyState

EndingTypes

FetchState

FillMode

FlashClassification

FlexLineGrowthState

FontFaceLoadStatus

FontFaceSetLoadStatus

FrameType

GamepadHand

GamepadHapticActuatorType

GamepadMappingType

GridDeclaration

GridTrackState

HeadersGuardEnum

IdbCursorDirection

IdbRequestReadyState

IdbTransactionMode

ImageBitmapFormat

IterationCompositeOperation

LineAlignSetting

MediaDecodingType

MediaDeviceKind

MediaEncodingType

MediaKeyMessageType

MediaKeySessionType

MediaKeyStatus

MediaKeySystemStatus

MediaKeysRequirement

MediaSourceEndOfStreamError

MediaSourceEnum

MediaSourceReadyState

MediaStreamTrackState

MidiPortConnectionState

MidiPortDeviceState

MidiPortType

NavigationType

NotificationDirection

NotificationPermission

OrientationLockType

OrientationType

OscillatorType

OverSampleType

PanningModelType

PaymentComplete

PcImplIceConnectionState

PcImplIceGatheringState

PcImplSignalingState

PcObserverStateType

PermissionName

PermissionState

PlaybackDirection

PositionAlignSetting

PresentationConnectionBinaryType

PresentationConnectionClosedReason

PresentationConnectionState

ProfileTimelineMessagePortOperationType

ProfileTimelineWorkerOperationType

PublicKeyCredentialType

PushEncryptionKeyName

PushPermissionState

RecordingState

ReferrerPolicy

RequestCache

RequestCredentials

RequestDestination

RequestMode

RequestRedirect

ResponseType

RtcBundlePolicy

RtcDataChannelState

RtcDataChannelType

RtcDegradationPreference

RtcIceConnectionState

RtcIceCredentialType

RtcIceGatheringState

RtcIceTransportPolicy

RtcLifecycleEvent

RtcPriorityType

RtcRtpSourceEntryType

RtcRtpTransceiverDirection

RtcSdpType

RtcSignalingState

RtcStatsIceCandidatePairState

RtcStatsIceCandidateType

RtcStatsType

ScreenColorGamut

ScrollBehavior

ScrollLogicalPosition

ScrollRestoration

ScrollSetting

ScrollState

SecurityPolicyViolationEventDisposition

ServiceWorkerState

ServiceWorkerUpdateViaCache

ShadowRootMode

SocketReadyState

SourceBufferAppendMode

SpeechRecognitionErrorCode

SpeechSynthesisErrorCode

StorageType

SupportedType

TcpReadyState

TcpSocketBinaryType

TextTrackKind

TextTrackMode

Transport

UserVerificationRequirement

VideoFacingModeEnum

VisibilityState

VrEye

WebGlPowerPreference

WebGpuLogEntryType

WebGpuObjectStatus

WebGpuPowerPreference

XmlHttpRequestResponseType

Functions

window
Getter for the Window object