web_sys
Enum GpuTextureFormat

#[non_exhaustive]

#[repr(u32)]
pub enum GpuTextureFormat {

    R8unorm = 0,
    R8snorm = 1,
    R8uint = 2,
    R8sint = 3,
    R16uint = 4,
    R16sint = 5,
    R16float = 6,
    Rg8unorm = 7,
    Rg8snorm = 8,
    Rg8uint = 9,
    Rg8sint = 10,
    R32uint = 11,
    R32sint = 12,
    R32float = 13,
    Rg16uint = 14,
    Rg16sint = 15,
    Rg16float = 16,
    Rgba8unorm = 17,
    Rgba8unormSrgb = 18,
    Rgba8snorm = 19,
    Rgba8uint = 20,
    Rgba8sint = 21,
    Bgra8unorm = 22,
    Bgra8unormSrgb = 23,
    Rgb9e5ufloat = 24,
    Rgb10a2uint = 25,
    Rgb10a2unorm = 26,
    Rg11b10ufloat = 27,
    Rg32uint = 28,
    Rg32sint = 29,
    Rg32float = 30,
    Rgba16uint = 31,
    Rgba16sint = 32,
    Rgba16float = 33,
    Rgba32uint = 34,
    Rgba32sint = 35,
    Rgba32float = 36,
    Stencil8 = 37,
    Depth16unorm = 38,
    Depth24plus = 39,
    Depth24plusStencil8 = 40,
    Depth32float = 41,
    Depth32floatStencil8 = 42,
    Bc1RgbaUnorm = 43,
    Bc1RgbaUnormSrgb = 44,
    Bc2RgbaUnorm = 45,
    Bc2RgbaUnormSrgb = 46,
    Bc3RgbaUnorm = 47,
    Bc3RgbaUnormSrgb = 48,
    Bc4RUnorm = 49,
    Bc4RSnorm = 50,
    Bc5RgUnorm = 51,
    Bc5RgSnorm = 52,
    Bc6hRgbUfloat = 53,
    Bc6hRgbFloat = 54,
    Bc7RgbaUnorm = 55,
    Bc7RgbaUnormSrgb = 56,
    Etc2Rgb8unorm = 57,
    Etc2Rgb8unormSrgb = 58,
    Etc2Rgb8a1unorm = 59,
    Etc2Rgb8a1unormSrgb = 60,
    Etc2Rgba8unorm = 61,
    Etc2Rgba8unormSrgb = 62,
    EacR11unorm = 63,
    EacR11snorm = 64,
    EacRg11unorm = 65,
    EacRg11snorm = 66,
    Astc4x4Unorm = 67,
    Astc4x4UnormSrgb = 68,
    Astc5x4Unorm = 69,
    Astc5x4UnormSrgb = 70,
    Astc5x5Unorm = 71,
    Astc5x5UnormSrgb = 72,
    Astc6x5Unorm = 73,
    Astc6x5UnormSrgb = 74,
    Astc6x6Unorm = 75,
    Astc6x6UnormSrgb = 76,
    Astc8x5Unorm = 77,
    Astc8x5UnormSrgb = 78,
    Astc8x6Unorm = 79,
    Astc8x6UnormSrgb = 80,
    Astc8x8Unorm = 81,
    Astc8x8UnormSrgb = 82,
    Astc10x5Unorm = 83,
    Astc10x5UnormSrgb = 84,
    Astc10x6Unorm = 85,
    Astc10x6UnormSrgb = 86,
    Astc10x8Unorm = 87,
    Astc10x8UnormSrgb = 88,
    Astc10x10Unorm = 89,
    Astc10x10UnormSrgb = 90,
    Astc12x10Unorm = 91,
    Astc12x10UnormSrgb = 92,
    Astc12x12Unorm = 93,
    Astc12x12UnormSrgb = 94,

}

The GpuTextureFormat enum.

This API requires the following crate features to be activated: GpuTextureFormat

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

R8unorm = 0

R8snorm = 1

R8uint = 2

R8sint = 3

R16uint = 4

R16sint = 5

R16float = 6

Rg8unorm = 7

Rg8snorm = 8

Rg8uint = 9

Rg8sint = 10

R32uint = 11

R32sint = 12

R32float = 13

Rg16uint = 14

Rg16sint = 15

Rg16float = 16

Rgba8unorm = 17

Rgba8unormSrgb = 18

Rgba8snorm = 19

Rgba8uint = 20

Rgba8sint = 21

Bgra8unorm = 22

Bgra8unormSrgb = 23

Rgb9e5ufloat = 24

Rgb10a2uint = 25

Rgb10a2unorm = 26

Rg11b10ufloat = 27

Rg32uint = 28

Rg32sint = 29

Rg32float = 30

Rgba16uint = 31

Rgba16sint = 32

Rgba16float = 33

Rgba32uint = 34

Rgba32sint = 35

Rgba32float = 36

Stencil8 = 37

Depth16unorm = 38

Depth24plus = 39

Depth24plusStencil8 = 40

Depth32float = 41

Depth32floatStencil8 = 42

Bc1RgbaUnorm = 43

Bc1RgbaUnormSrgb = 44

Bc2RgbaUnorm = 45

Bc2RgbaUnormSrgb = 46

Bc3RgbaUnorm = 47

Bc3RgbaUnormSrgb = 48

Bc4RUnorm = 49

Bc4RSnorm = 50

Bc5RgUnorm = 51

Bc5RgSnorm = 52

Bc6hRgbUfloat = 53

Bc6hRgbFloat = 54

Bc7RgbaUnorm = 55

Bc7RgbaUnormSrgb = 56

Etc2Rgb8unorm = 57

Etc2Rgb8unormSrgb = 58

Etc2Rgb8a1unorm = 59

Etc2Rgb8a1unormSrgb = 60

Etc2Rgba8unorm = 61

Etc2Rgba8unormSrgb = 62

EacR11unorm = 63

EacR11snorm = 64

EacRg11unorm = 65

EacRg11snorm = 66

Astc4x4Unorm = 67

Astc4x4UnormSrgb = 68

Astc5x4Unorm = 69

Astc5x4UnormSrgb = 70

Astc5x5Unorm = 71

Astc5x5UnormSrgb = 72

Astc6x5Unorm = 73

Astc6x5UnormSrgb = 74

Astc6x6Unorm = 75

Astc6x6UnormSrgb = 76

Astc8x5Unorm = 77

Astc8x5UnormSrgb = 78

Astc8x6Unorm = 79

Astc8x6UnormSrgb = 80

Astc8x8Unorm = 81

Astc8x8UnormSrgb = 82

Astc10x5Unorm = 83

Astc10x5UnormSrgb = 84

Astc10x6Unorm = 85

Astc10x6UnormSrgb = 86

Astc10x8Unorm = 87

Astc10x8UnormSrgb = 88

Astc10x10Unorm = 89

Astc10x10UnormSrgb = 90

Astc12x10Unorm = 91

Astc12x10UnormSrgb = 92

Astc12x12Unorm = 93

Astc12x12UnormSrgb = 94

Implementations

impl GpuTextureFormat

pub fn from_js_value(obj: &JsValue) -> Option<GpuTextureFormat>