Enum web_sys::GpuTextureFormat

pub enum GpuTextureFormat {

    R8unorm,
    R8snorm,
    R8uint,
    R8sint,
    R16uint,
    R16sint,
    R16float,
    Rg8unorm,
    Rg8snorm,
    Rg8uint,
    Rg8sint,
    R32uint,
    R32sint,
    R32float,
    Rg16uint,
    Rg16sint,
    Rg16float,
    Rgba8unorm,
    Rgba8unormSrgb,
    Rgba8snorm,
    Rgba8uint,
    Rgba8sint,
    Bgra8unorm,
    Bgra8unormSrgb,
    Rgb9e5ufloat,
    Rgb10a2unorm,
    Rg11b10ufloat,
    Rg32uint,
    Rg32sint,
    Rg32float,
    Rgba16uint,
    Rgba16sint,
    Rgba16float,
    Rgba32uint,
    Rgba32sint,
    Rgba32float,
    Stencil8,
    Depth16unorm,
    Depth24plus,
    Depth24plusStencil8,
    Depth32float,
    Depth32floatStencil8,
    Bc1RgbaUnorm,
    Bc1RgbaUnormSrgb,
    Bc2RgbaUnorm,
    Bc2RgbaUnormSrgb,
    Bc3RgbaUnorm,
    Bc3RgbaUnormSrgb,
    Bc4RUnorm,
    Bc4RSnorm,
    Bc5RgUnorm,
    Bc5RgSnorm,
    Bc6hRgbUfloat,
    Bc6hRgbFloat,
    Bc7RgbaUnorm,
    Bc7RgbaUnormSrgb,
    Etc2Rgb8unorm,
    Etc2Rgb8unormSrgb,
    Etc2Rgb8a1unorm,
    Etc2Rgb8a1unormSrgb,
    Etc2Rgba8unorm,
    Etc2Rgba8unormSrgb,
    EacR11unorm,
    EacR11snorm,
    EacRg11unorm,
    EacRg11snorm,
    Astc4x4Unorm,
    Astc4x4UnormSrgb,
    Astc5x4Unorm,
    Astc5x4UnormSrgb,
    Astc5x5Unorm,
    Astc5x5UnormSrgb,
    Astc6x5Unorm,
    Astc6x5UnormSrgb,
    Astc6x6Unorm,
    Astc6x6UnormSrgb,
    Astc8x5Unorm,
    Astc8x5UnormSrgb,
    Astc8x6Unorm,
    Astc8x6UnormSrgb,
    Astc8x8Unorm,
    Astc8x8UnormSrgb,
    Astc10x5Unorm,
    Astc10x5UnormSrgb,
    Astc10x6Unorm,
    Astc10x6UnormSrgb,
    Astc10x8Unorm,
    Astc10x8UnormSrgb,
    Astc10x10Unorm,
    Astc10x10UnormSrgb,
    Astc12x10Unorm,
    Astc12x10UnormSrgb,
    Astc12x12Unorm,
    Astc12x12UnormSrgb,
    // some variants omitted

}

The GpuTextureFormat enum.

This API requires the following crate features to be activated: GpuTextureFormat

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

Variants

R8unorm

R8snorm

R8uint

R8sint

R16uint

R16sint

R16float

Rg8unorm

Rg8snorm

Rg8uint

Rg8sint

R32uint

R32sint

R32float

Rg16uint

Rg16sint

Rg16float

Rgba8unorm

Rgba8unormSrgb

Rgba8snorm

Rgba8uint

Rgba8sint

Bgra8unorm

Bgra8unormSrgb

Rgb9e5ufloat

Rgb10a2unorm

Rg11b10ufloat

Rg32uint

Rg32sint

Rg32float

Rgba16uint

Rgba16sint

Rgba16float

Rgba32uint

Rgba32sint

Rgba32float

Stencil8

Depth16unorm

Depth24plus

Depth24plusStencil8

Depth32float

Depth32floatStencil8

Bc1RgbaUnorm

Bc1RgbaUnormSrgb

Bc2RgbaUnorm

Bc2RgbaUnormSrgb

Bc3RgbaUnorm

Bc3RgbaUnormSrgb

Bc4RUnorm

Bc4RSnorm

Bc5RgUnorm

Bc5RgSnorm

Bc6hRgbUfloat

Bc6hRgbFloat

Bc7RgbaUnorm

Bc7RgbaUnormSrgb

Etc2Rgb8unorm

Etc2Rgb8unormSrgb

Etc2Rgb8a1unorm

Etc2Rgb8a1unormSrgb

Etc2Rgba8unorm

Etc2Rgba8unormSrgb

EacR11unorm

EacR11snorm

EacRg11unorm

EacRg11snorm

Astc4x4Unorm

Astc4x4UnormSrgb

Astc5x4Unorm

Astc5x4UnormSrgb

Astc5x5Unorm

Astc5x5UnormSrgb

Astc6x5Unorm

Astc6x5UnormSrgb

Astc6x6Unorm

Astc6x6UnormSrgb

Astc8x5Unorm

Astc8x5UnormSrgb

Astc8x6Unorm

Astc8x6UnormSrgb

Astc8x8Unorm

Astc8x8UnormSrgb

Astc10x5Unorm

Astc10x5UnormSrgb

Astc10x6Unorm

Astc10x6UnormSrgb

Astc10x8Unorm

Astc10x8UnormSrgb

Astc10x10Unorm

Astc10x10UnormSrgb

Astc12x10Unorm

Astc12x10UnormSrgb

Astc12x12Unorm

Astc12x12UnormSrgb

Implementations

impl GpuTextureFormat

pub fn from_js_value(obj: &JsValue) -> Option<GpuTextureFormat>