Enum web_sys::GpuTextureFormat

pub enum GpuTextureFormat {
    R8unorm,
    R8snorm,
    R8uint,
    R8sint,
    R16uint,
    R16sint,
    R16float,
    Rg8unorm,
    Rg8snorm,
    Rg8uint,
    Rg8sint,
    R32uint,
    R32sint,
    R32float,
    Rg16uint,
    Rg16sint,
    Rg16float,
    Rgba8unorm,
    Rgba8unormSrgb,
    Rgba8snorm,
    Rgba8uint,
    Rgba8sint,
    Bgra8unorm,
    Bgra8unormSrgb,
    Rgb10a2unorm,
    Rg11b10float,
    Rg32uint,
    Rg32sint,
    Rg32float,
    Rgba16uint,
    Rgba16sint,
    Rgba16float,
    Rgba32uint,
    Rgba32sint,
    Rgba32float,
    Depth32float,
    Depth24plus,
    Depth24plusStencil8,
    Bc1RgbaUnorm,
    Bc1RgbaUnormSrgb,
    Bc2RgbaUnorm,
    Bc2RgbaUnormSrgb,
    Bc3RgbaUnorm,
    Bc3RgbaUnormSrgb,
    Bc4RUnorm,
    Bc4RSnorm,
    Bc5RgUnorm,
    Bc5RgSnorm,
    Bc6hRgbUfloat,
    Bc6hRgbSfloat,
    Bc7RgbaUnorm,
    Bc7RgbaUnormSrgb,
    // some variants omitted
}

The GpuTextureFormat enum.

This API requires the following crate features to be activated: GpuTextureFormat

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

Variants

R8unorm

R8snorm

R8uint

R8sint

R16uint

R16sint

R16float

Rg8unorm

Rg8snorm

Rg8uint

Rg8sint

R32uint

R32sint

R32float

Rg16uint

Rg16sint

Rg16float

Rgba8unorm

Rgba8unormSrgb

Rgba8snorm

Rgba8uint

Rgba8sint

Bgra8unorm

Bgra8unormSrgb

Rgb10a2unorm

Rg11b10float

Rg32uint

Rg32sint

Rg32float

Rgba16uint

Rgba16sint

Rgba16float

Rgba32uint

Rgba32sint

Rgba32float

Depth32float

Depth24plus

Depth24plusStencil8

Bc1RgbaUnorm

Bc1RgbaUnormSrgb

Bc2RgbaUnorm

Bc2RgbaUnormSrgb

Bc3RgbaUnorm

Bc3RgbaUnormSrgb

Bc4RUnorm

Bc4RSnorm

Bc5RgUnorm

Bc5RgSnorm

Bc6hRgbUfloat

Bc6hRgbSfloat

Bc7RgbaUnorm

Bc7RgbaUnormSrgb

Implementations

impl GpuTextureFormat

pub fn from_js_value(obj: &JsValue) -> Option<GpuTextureFormat>