Struct web_sys::Blob

#[repr(transparent)]pub struct Blob { /* fields omitted */ }

The Blob class.

MDN Documentation

This API requires the following crate features to be activated: Blob

Implementations

impl Blob

pub fn size(&self) -> f64

Getter for the size field of this object.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn type_(&self) -> String

Getter for the type field of this object.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn new() -> Result<Blob, JsValue>

The new Blob(..) constructor, creating a new instance of Blob.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn new_with_buffer_source_sequence(
blob_parts: &JsValue
) -> Result<Blob, JsValue>

The new Blob(..) constructor, creating a new instance of Blob.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn new_with_u8_array_sequence(blob_parts: &JsValue) -> Result<Blob, JsValue>

The new Blob(..) constructor, creating a new instance of Blob.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn new_with_blob_sequence(blob_parts: &JsValue) -> Result<Blob, JsValue>

The new Blob(..) constructor, creating a new instance of Blob.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn new_with_str_sequence(blob_parts: &JsValue) -> Result<Blob, JsValue>

The new Blob(..) constructor, creating a new instance of Blob.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn array_buffer(&self) -> Promise

The arrayBuffer() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice(&self) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_i32(&self, start: i32) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_f64(&self, start: f64) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_i32_and_i32(
&self,
start: i32,
end: i32
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_f64_and_i32(
&self,
start: f64,
end: i32
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_i32_and_f64(
&self,
start: i32,
end: f64
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_f64_and_f64(
&self,
start: f64,
end: f64
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_i32_and_i32_and_content_type(
&self,
start: i32,
end: i32,
content_type: &str
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_f64_and_i32_and_content_type(
&self,
start: f64,
end: i32,
content_type: &str
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_i32_and_f64_and_content_type(
&self,
start: i32,
end: f64,
content_type: &str
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn slice_with_f64_and_f64_and_content_type(
&self,
start: f64,
end: f64,
content_type: &str
) -> Result<Blob, JsValue>

The slice() method.

MDN Documentation

This API requires the following crate features to be activated: Blob

impl Blob

pub fn text(&self) -> Promise

The text() method.

MDN Documentation

This API requires the following crate features to be activated: Blob