web_sys
Struct CanvasRenderingContext2d

pub struct CanvasRenderingContext2d { /* private fields */ }

The CanvasRenderingContext2d class.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

Implementations

impl CanvasRenderingContext2d

pub fn canvas(&self) -> Option<HtmlCanvasElement>

Getter for the canvas field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn global_alpha(&self) -> f64

Getter for the globalAlpha field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_global_alpha(&self, value: f64)

Setter for the globalAlpha field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn global_composite_operation(&self) -> Result<String, JsValue>

Getter for the globalCompositeOperation field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_global_composite_operation(&self, value: &str) -> Result<(), JsValue>

Setter for the globalCompositeOperation field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_style(&self) -> JsValue

Getter for the strokeStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_stroke_style(&self, value: &JsValue)

๐Deprecated

Setter for the strokeStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_stroke_style_str(&self, value: &str)

Setter for the strokeStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_stroke_style_canvas_gradient(&self, value: &CanvasGradient)

Setter for the strokeStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasGradient, CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_stroke_style_canvas_pattern(&self, value: &CanvasPattern)

Setter for the strokeStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_style(&self) -> JsValue

Getter for the fillStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_fill_style(&self, value: &JsValue)

๐Deprecated

Setter for the fillStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_fill_style_str(&self, value: &str)

Setter for the fillStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_fill_style_canvas_gradient(&self, value: &CanvasGradient)

Setter for the fillStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasGradient, CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_fill_style_canvas_pattern(&self, value: &CanvasPattern)

Setter for the fillStyle field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn filter(&self) -> String

Getter for the filter field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_filter(&self, value: &str)

Setter for the filter field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn image_smoothing_enabled(&self) -> bool

Getter for the imageSmoothingEnabled field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_image_smoothing_enabled(&self, value: bool)

Setter for the imageSmoothingEnabled field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_width(&self) -> f64

Getter for the lineWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_width(&self, value: f64)

Setter for the lineWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_cap(&self) -> String

Getter for the lineCap field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_cap(&self, value: &str)

Setter for the lineCap field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_join(&self) -> String

Getter for the lineJoin field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_join(&self, value: &str)

Setter for the lineJoin field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn miter_limit(&self) -> f64

Getter for the miterLimit field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_miter_limit(&self, value: f64)

Setter for the miterLimit field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_dash_offset(&self) -> f64

Getter for the lineDashOffset field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_dash_offset(&self, value: f64)

Setter for the lineDashOffset field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_offset_x(&self) -> f64

Getter for the shadowOffsetX field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_offset_x(&self, value: f64)

Setter for the shadowOffsetX field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_offset_y(&self) -> f64

Getter for the shadowOffsetY field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_offset_y(&self, value: f64)

Setter for the shadowOffsetY field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_blur(&self) -> f64

Getter for the shadowBlur field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_blur(&self, value: f64)

Setter for the shadowBlur field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn shadow_color(&self) -> String

Getter for the shadowColor field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_shadow_color(&self, value: &str)

Setter for the shadowColor field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn font(&self) -> String

Getter for the font field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_font(&self, value: &str)

Setter for the font field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn text_align(&self) -> String

Getter for the textAlign field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_text_align(&self, value: &str)

Setter for the textAlign field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn text_baseline(&self) -> String

Getter for the textBaseline field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_text_baseline(&self, value: &str)

Setter for the textBaseline field of this object.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn draw_window( &self, window: &Window, x: f64, y: f64, w: f64, h: f64, bg_color: &str, ) -> Result<(), JsValue>

The drawWindow() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Window

impl CanvasRenderingContext2d

pub fn draw_window_with_flags( &self, window: &Window, x: f64, y: f64, w: f64, h: f64, bg_color: &str, flags: u32, ) -> Result<(), JsValue>

The drawWindow() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Window

impl CanvasRenderingContext2d

pub fn draw_image_with_html_image_element( &self, image: &HtmlImageElement, dx: f64, dy: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlImageElement

impl CanvasRenderingContext2d

pub fn draw_image_with_svg_image_element( &self, image: &SvgImageElement, dx: f64, dy: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, SvgImageElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_canvas_element( &self, image: &HtmlCanvasElement, dx: f64, dy: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_video_element( &self, image: &HtmlVideoElement, dx: f64, dy: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlVideoElement

impl CanvasRenderingContext2d

pub fn draw_image_with_image_bitmap( &self, image: &ImageBitmap, dx: f64, dy: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageBitmap

impl CanvasRenderingContext2d

pub fn draw_image_with_offscreen_canvas( &self, image: &OffscreenCanvas, dx: f64, dy: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, OffscreenCanvas

impl CanvasRenderingContext2d

pub fn draw_image_with_video_frame( &self, image: &VideoFrame, dx: f64, dy: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, VideoFrame

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl CanvasRenderingContext2d

pub fn draw_image_with_html_image_element_and_dw_and_dh( &self, image: &HtmlImageElement, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlImageElement

impl CanvasRenderingContext2d

pub fn draw_image_with_svg_image_element_and_dw_and_dh( &self, image: &SvgImageElement, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, SvgImageElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_canvas_element_and_dw_and_dh( &self, image: &HtmlCanvasElement, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_video_element_and_dw_and_dh( &self, image: &HtmlVideoElement, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlVideoElement

impl CanvasRenderingContext2d

pub fn draw_image_with_image_bitmap_and_dw_and_dh( &self, image: &ImageBitmap, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageBitmap

impl CanvasRenderingContext2d

pub fn draw_image_with_offscreen_canvas_and_dw_and_dh( &self, image: &OffscreenCanvas, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, OffscreenCanvas

impl CanvasRenderingContext2d

pub fn draw_image_with_video_frame_and_dw_and_dh( &self, image: &VideoFrame, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, VideoFrame

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl CanvasRenderingContext2d

pub fn draw_image_with_html_image_element_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &HtmlImageElement, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlImageElement

impl CanvasRenderingContext2d

pub fn draw_image_with_svg_image_element_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &SvgImageElement, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, SvgImageElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_canvas_element_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &HtmlCanvasElement, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn draw_image_with_html_video_element_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &HtmlVideoElement, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HtmlVideoElement

impl CanvasRenderingContext2d

pub fn draw_image_with_image_bitmap_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &ImageBitmap, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageBitmap

impl CanvasRenderingContext2d

pub fn draw_image_with_offscreen_canvas_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &OffscreenCanvas, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, OffscreenCanvas

impl CanvasRenderingContext2d

pub fn draw_image_with_video_frame_and_sw_and_sh_and_dx_and_dy_and_dw_and_dh( &self, image: &VideoFrame, sx: f64, sy: f64, sw: f64, sh: f64, dx: f64, dy: f64, dw: f64, dh: f64, ) -> Result<(), JsValue>

The drawImage() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, VideoFrame

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl CanvasRenderingContext2d

pub fn begin_path(&self)

The beginPath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn clip(&self)

The clip() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn clip_with_canvas_winding_rule(&self, winding: CanvasWindingRule)

The clip() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, CanvasWindingRule

impl CanvasRenderingContext2d

pub fn clip_with_path_2d(&self, path: &Path2d)

The clip() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Path2d

impl CanvasRenderingContext2d

pub fn clip_with_path_2d_and_winding( &self, path: &Path2d, winding: CanvasWindingRule, )

The clip() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, CanvasWindingRule, Path2d

impl CanvasRenderingContext2d

pub fn fill(&self)

The fill() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_with_canvas_winding_rule(&self, winding: CanvasWindingRule)

The fill() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, CanvasWindingRule

impl CanvasRenderingContext2d

pub fn fill_with_path_2d(&self, path: &Path2d)

The fill() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Path2d

impl CanvasRenderingContext2d

pub fn fill_with_path_2d_and_winding( &self, path: &Path2d, winding: CanvasWindingRule, )

The fill() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, CanvasWindingRule, Path2d

impl CanvasRenderingContext2d

pub fn is_point_in_path_with_f64(&self, x: f64, y: f64) -> bool

The isPointInPath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn is_point_in_path_with_f64_and_canvas_winding_rule( &self, x: f64, y: f64, winding: CanvasWindingRule, ) -> bool

The isPointInPath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, CanvasWindingRule

impl CanvasRenderingContext2d

pub fn is_point_in_path_with_path_2d_and_f64( &self, path: &Path2d, x: f64, y: f64, ) -> bool

The isPointInPath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Path2d

impl CanvasRenderingContext2d

pub fn is_point_in_path_with_path_2d_and_f64_and_winding( &self, path: &Path2d, x: f64, y: f64, winding: CanvasWindingRule, ) -> bool

The isPointInPath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, CanvasWindingRule, Path2d

impl CanvasRenderingContext2d

pub fn is_point_in_stroke_with_x_and_y(&self, x: f64, y: f64) -> bool

The isPointInStroke() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn is_point_in_stroke_with_path_and_x_and_y( &self, path: &Path2d, x: f64, y: f64, ) -> bool

The isPointInStroke() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Path2d

impl CanvasRenderingContext2d

pub fn stroke(&self)

The stroke() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_with_path(&self, path: &Path2d)

The stroke() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Path2d

impl CanvasRenderingContext2d

pub fn create_linear_gradient( &self, x0: f64, y0: f64, x1: f64, y1: f64, ) -> CanvasGradient

The createLinearGradient() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasGradient, CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn create_pattern_with_html_image_element( &self, image: &HtmlImageElement, repetition: &str, ) -> Result<Option<CanvasPattern>, JsValue>

The createPattern() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d, HtmlImageElement

impl CanvasRenderingContext2d

pub fn create_pattern_with_svg_image_element( &self, image: &SvgImageElement, repetition: &str, ) -> Result<Option<CanvasPattern>, JsValue>

The createPattern() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d, SvgImageElement

impl CanvasRenderingContext2d

pub fn create_pattern_with_html_canvas_element( &self, image: &HtmlCanvasElement, repetition: &str, ) -> Result<Option<CanvasPattern>, JsValue>

The createPattern() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d, HtmlCanvasElement

impl CanvasRenderingContext2d

pub fn create_pattern_with_html_video_element( &self, image: &HtmlVideoElement, repetition: &str, ) -> Result<Option<CanvasPattern>, JsValue>

The createPattern() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d, HtmlVideoElement

impl CanvasRenderingContext2d

pub fn create_pattern_with_image_bitmap( &self, image: &ImageBitmap, repetition: &str, ) -> Result<Option<CanvasPattern>, JsValue>

The createPattern() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d, ImageBitmap

impl CanvasRenderingContext2d

pub fn create_pattern_with_offscreen_canvas( &self, image: &OffscreenCanvas, repetition: &str, ) -> Result<Option<CanvasPattern>, JsValue>

The createPattern() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d, OffscreenCanvas

impl CanvasRenderingContext2d

pub fn create_pattern_with_video_frame( &self, image: &VideoFrame, repetition: &str, ) -> Result<Option<CanvasPattern>, JsValue>

The createPattern() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasPattern, CanvasRenderingContext2d, VideoFrame

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl CanvasRenderingContext2d

pub fn create_radial_gradient( &self, x0: f64, y0: f64, r0: f64, x1: f64, y1: f64, r1: f64, ) -> Result<CanvasGradient, JsValue>

The createRadialGradient() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasGradient, CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn add_hit_region(&self) -> Result<(), JsValue>

The addHitRegion() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn add_hit_region_with_options( &self, options: &HitRegionOptions, ) -> Result<(), JsValue>

The addHitRegion() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, HitRegionOptions

impl CanvasRenderingContext2d

pub fn clear_hit_regions(&self)

The clearHitRegions() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn remove_hit_region(&self, id: &str)

The removeHitRegion() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn create_image_data_with_sw_and_sh( &self, sw: f64, sh: f64, ) -> Result<ImageData, JsValue>

The createImageData() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageData

impl CanvasRenderingContext2d

pub fn create_image_data_with_imagedata( &self, imagedata: &ImageData, ) -> Result<ImageData, JsValue>

The createImageData() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageData

impl CanvasRenderingContext2d

pub fn get_image_data( &self, sx: f64, sy: f64, sw: f64, sh: f64, ) -> Result<ImageData, JsValue>

The getImageData() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageData

impl CanvasRenderingContext2d

pub fn put_image_data( &self, imagedata: &ImageData, dx: f64, dy: f64, ) -> Result<(), JsValue>

The putImageData() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageData

impl CanvasRenderingContext2d

pub fn put_image_data_with_dirty_x_and_dirty_y_and_dirty_width_and_dirty_height( &self, imagedata: &ImageData, dx: f64, dy: f64, dirty_x: f64, dirty_y: f64, dirty_width: f64, dirty_height: f64, ) -> Result<(), JsValue>

The putImageData() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, ImageData

impl CanvasRenderingContext2d

pub fn arc( &self, x: f64, y: f64, radius: f64, start_angle: f64, end_angle: f64, ) -> Result<(), JsValue>

The arc() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn arc_with_anticlockwise( &self, x: f64, y: f64, radius: f64, start_angle: f64, end_angle: f64, anticlockwise: bool, ) -> Result<(), JsValue>

The arc() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn arc_to( &self, x1: f64, y1: f64, x2: f64, y2: f64, radius: f64, ) -> Result<(), JsValue>

The arcTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn bezier_curve_to( &self, cp1x: f64, cp1y: f64, cp2x: f64, cp2y: f64, x: f64, y: f64, )

The bezierCurveTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn close_path(&self)

The closePath() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn ellipse( &self, x: f64, y: f64, radius_x: f64, radius_y: f64, rotation: f64, start_angle: f64, end_angle: f64, ) -> Result<(), JsValue>

The ellipse() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn ellipse_with_anticlockwise( &self, x: f64, y: f64, radius_x: f64, radius_y: f64, rotation: f64, start_angle: f64, end_angle: f64, anticlockwise: bool, ) -> Result<(), JsValue>

The ellipse() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn line_to(&self, x: f64, y: f64)

The lineTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn move_to(&self, x: f64, y: f64)

The moveTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn quadratic_curve_to(&self, cpx: f64, cpy: f64, x: f64, y: f64)

The quadraticCurveTo() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn rect(&self, x: f64, y: f64, w: f64, h: f64)

The rect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect(&self, x: f64, y: f64, w: f64, h: f64) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect_with_f64( &self, x: f64, y: f64, w: f64, h: f64, radii: f64, ) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect_with_dom_point_init( &self, x: f64, y: f64, w: f64, h: f64, radii: &DomPointInit, ) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, DomPointInit

impl CanvasRenderingContext2d

pub fn round_rect_with_f64_sequence( &self, x: f64, y: f64, w: f64, h: f64, radii: &JsValue, ) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn round_rect_with_dom_point_init_sequence( &self, x: f64, y: f64, w: f64, h: f64, radii: &JsValue, ) -> Result<(), JsValue>

The roundRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn get_line_dash(&self) -> Array

The getLineDash() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_line_dash(&self, segments: &JsValue) -> Result<(), JsValue>

The setLineDash() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn clear_rect(&self, x: f64, y: f64, w: f64, h: f64)

The clearRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_rect(&self, x: f64, y: f64, w: f64, h: f64)

The fillRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_rect(&self, x: f64, y: f64, w: f64, h: f64)

The strokeRect() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn reset(&self)

The reset() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn restore(&self)

The restore() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn save(&self)

The save() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_text(&self, text: &str, x: f64, y: f64) -> Result<(), JsValue>

The fillText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn fill_text_with_max_width( &self, text: &str, x: f64, y: f64, max_width: f64, ) -> Result<(), JsValue>

The fillText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn measure_text(&self, text: &str) -> Result<TextMetrics, JsValue>

The measureText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, TextMetrics

impl CanvasRenderingContext2d

pub fn stroke_text(&self, text: &str, x: f64, y: f64) -> Result<(), JsValue>

The strokeText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn stroke_text_with_max_width( &self, text: &str, x: f64, y: f64, max_width: f64, ) -> Result<(), JsValue>

The strokeText() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn get_transform(&self) -> Result<DomMatrix, JsValue>

The getTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, DomMatrix

impl CanvasRenderingContext2d

pub fn reset_transform(&self) -> Result<(), JsValue>

The resetTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn rotate(&self, angle: f64) -> Result<(), JsValue>

The rotate() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn scale(&self, x: f64, y: f64) -> Result<(), JsValue>

The scale() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_transform( &self, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64, ) -> Result<(), JsValue>

The setTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_transform_with_default_dom_matrix_2d_init( &self, ) -> Result<(), JsValue>

The setTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn set_transform_with_dom_matrix_2d_init( &self, transform: &DomMatrix2dInit, ) -> Result<(), JsValue>

The setTransform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, DomMatrix2dInit

impl CanvasRenderingContext2d

pub fn transform( &self, a: f64, b: f64, c: f64, d: f64, e: f64, f: f64, ) -> Result<(), JsValue>

The transform() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn translate(&self, x: f64, y: f64) -> Result<(), JsValue>

The translate() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d

impl CanvasRenderingContext2d

pub fn draw_custom_focus_ring(&self, element: &Element) -> bool

The drawCustomFocusRing() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Element

impl CanvasRenderingContext2d

pub fn draw_focus_if_needed(&self, element: &Element) -> Result<(), JsValue>

The drawFocusIfNeeded() method.

MDN Documentation

This API requires the following crate features to be activated: CanvasRenderingContext2d, Element

impl CanvasRenderingContext2d

pub const DRAWWINDOW_DRAW_CARET: u32 = 1u32

The CanvasRenderingContext2D.DRAWWINDOW_DRAW_CARET const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_DO_NOT_FLUSH: u32 = 2u32

The CanvasRenderingContext2D.DRAWWINDOW_DO_NOT_FLUSH const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_DRAW_VIEW: u32 = 4u32

The CanvasRenderingContext2D.DRAWWINDOW_DRAW_VIEW const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_USE_WIDGET_LAYERS: u32 = 8u32

The CanvasRenderingContext2D.DRAWWINDOW_USE_WIDGET_LAYERS const.

This API requires the following crate features to be activated: CanvasRenderingContext2d

pub const DRAWWINDOW_ASYNC_DECODE_IMAGES: u32 = 16u32

The CanvasRenderingContext2D.DRAWWINDOW_ASYNC_DECODE_IMAGES const.

This API requires the following crate features to be activated: CanvasRenderingContext2d