Struct web_sys::CssStyleDeclaration

#[repr(transparent)]

pub struct CssStyleDeclaration { /* fields omitted */ }

The CSSStyleDeclaration object

MDN Documentation

Methods

impl CssStyleDeclaration

pub fn get_property_priority(&self, property: &str) -> String

The getPropertyPriority() method

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn get_property_value(&self, property: &str) -> Result<String, JsValue>

The getPropertyValue() method

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn item(&self, index: u32) -> String

The item() method

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn remove_property(&self, property: &str) -> Result<String, JsValue>

The removeProperty() method

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn set_property(&self, property: &str, value: &str) -> Result<(), JsValue>

The setProperty() method

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn set_property_with_priority(
&self,
property: &str,
value: &str,
priority: &str
) -> Result<(), JsValue>

The setProperty() method

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn get(&self, index: u32) -> String

The indexing getter

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn css_text(&self) -> String

The cssText getter

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn set_css_text(&self, css_text: &str)

The cssText setter

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn length(&self) -> u32

The length getter

MDN Documentation

This function requires the following crate features to be activated: CssStyleDeclaration

impl CssStyleDeclaration

pub fn parent_rule(&self) -> Option<CssRule>

The parentRule getter

MDN Documentation

This function requires the following crate features to be activated: CssRule, CssStyleDeclaration