Struct web_sys::DocumentFragment

#[repr(transparent)]
pub struct DocumentFragment { /* fields omitted */ }

The DocumentFragment class.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

Implementations

impl DocumentFragment

pub fn children(&self) -> HtmlCollection

Getter for the children field of this object.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment, HtmlCollection

impl DocumentFragment

pub fn first_element_child(&self) -> Option<Element>

Getter for the firstElementChild field of this object.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment, Element

impl DocumentFragment

pub fn last_element_child(&self) -> Option<Element>

Getter for the lastElementChild field of this object.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment, Element

impl DocumentFragment

pub fn child_element_count(&self) -> u32

Getter for the childElementCount field of this object.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn new() -> Result<DocumentFragment, JsValue>

The new DocumentFragment(..) constructor, creating a new instance of DocumentFragment.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn get_element_by_id(&self, element_id: &str) -> Option<Element>

The getElementById() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment, Element

impl DocumentFragment

pub fn query_selector(
&self,
selectors: &str
) -> Result<Option<Element>, JsValue>

The querySelector() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment, Element

impl DocumentFragment

pub fn query_selector_all(&self, selectors: &str) -> Result<NodeList, JsValue>

The querySelectorAll() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment, NodeList

impl DocumentFragment

pub fn append_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_0(&self) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_0(&self) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn append_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_0(&self) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_2(
&self,
nodes_1: &Node,
nodes_2: &Node
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_3(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_4(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_5(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_6(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_node_7(
&self,
nodes_1: &Node,
nodes_2: &Node,
nodes_3: &Node,
nodes_4: &Node,
nodes_5: &Node,
nodes_6: &Node,
nodes_7: &Node
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_0(&self) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_2(
&self,
nodes_1: &str,
nodes_2: &str
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_3(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_4(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_5(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_6(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment

impl DocumentFragment

pub fn prepend_with_str_7(
&self,
nodes_1: &str,
nodes_2: &str,
nodes_3: &str,
nodes_4: &str,
nodes_5: &str,
nodes_6: &str,
nodes_7: &str
) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: DocumentFragment