web_sys
Struct Element

pub struct Element { /* private fields */ }

The Element class.

MDN Documentation

This API requires the following crate features to be activated: Element

Implementations

impl Element

pub fn namespace_uri(&self) -> Option<String>

Getter for the namespaceURI field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prefix(&self) -> Option<String>

Getter for the prefix field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn local_name(&self) -> String

Getter for the localName field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn tag_name(&self) -> String

Getter for the tagName field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn id(&self) -> String

Getter for the id field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_id(&self, value: &str)

Setter for the id field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn class_name(&self) -> String

Getter for the className field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_class_name(&self, value: &str)

Setter for the className field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn class_list(&self) -> DomTokenList

Getter for the classList field of this object.

MDN Documentation

This API requires the following crate features to be activated: DomTokenList, Element

impl Element

pub fn attributes(&self) -> NamedNodeMap

Getter for the attributes field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element, NamedNodeMap

impl Element

pub fn scroll_top(&self) -> i32

Getter for the scrollTop field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_scroll_top(&self, value: i32)

Setter for the scrollTop field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_left(&self) -> i32

Getter for the scrollLeft field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_scroll_left(&self, value: i32)

Setter for the scrollLeft field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_width(&self) -> i32

Getter for the scrollWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_height(&self) -> i32

Getter for the scrollHeight field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_top(&self) -> i32

Getter for the clientTop field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_left(&self) -> i32

Getter for the clientLeft field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_width(&self) -> i32

Getter for the clientWidth field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn client_height(&self) -> i32

Getter for the clientHeight field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn inner_html(&self) -> String

Getter for the innerHTML field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_inner_html(&self, value: &str)

Setter for the innerHTML field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn outer_html(&self) -> String

Getter for the outerHTML field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_outer_html(&self, value: &str)

Setter for the outerHTML field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn shadow_root(&self) -> Option<ShadowRoot>

Getter for the shadowRoot field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element, ShadowRoot

impl Element

pub fn assigned_slot(&self) -> Option<HtmlSlotElement>

Getter for the assignedSlot field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlSlotElement

impl Element

pub fn slot(&self) -> String

Getter for the slot field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_slot(&self, value: &str)

Setter for the slot field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn previous_element_sibling(&self) -> Option<Element>

Getter for the previousElementSibling field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn next_element_sibling(&self) -> Option<Element>

Getter for the nextElementSibling field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn children(&self) -> HtmlCollection

Getter for the children field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlCollection

impl Element

pub fn first_element_child(&self) -> Option<Element>

Getter for the firstElementChild field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn last_element_child(&self) -> Option<Element>

Getter for the lastElementChild field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn child_element_count(&self) -> u32

Getter for the childElementCount field of this object.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn attach_shadow( &self, shadow_root_init_dict: &ShadowRootInit, ) -> Result<ShadowRoot, JsValue>

The attachShadow() method.

MDN Documentation

This API requires the following crate features to be activated: Element, ShadowRoot, ShadowRootInit

impl Element

pub fn closest(&self, selector: &str) -> Result<Option<Element>, JsValue>

The closest() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_attribute(&self, name: &str) -> Option<String>

The getAttribute() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_attribute_ns( &self, namespace: Option<&str>, local_name: &str, ) -> Option<String>

The getAttributeNS() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_attribute_names(&self) -> Array

The getAttributeNames() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_attribute_node(&self, name: &str) -> Option<Attr>

The getAttributeNode() method.

MDN Documentation

This API requires the following crate features to be activated: Attr, Element

impl Element

pub fn get_attribute_node_ns( &self, namespace_uri: Option<&str>, local_name: &str, ) -> Option<Attr>

The getAttributeNodeNS() method.

MDN Documentation

This API requires the following crate features to be activated: Attr, Element

impl Element

pub fn get_bounding_client_rect(&self) -> DomRect

The getBoundingClientRect() method.

MDN Documentation

This API requires the following crate features to be activated: DomRect, Element

impl Element

pub fn get_client_rects(&self) -> DomRectList

The getClientRects() method.

MDN Documentation

This API requires the following crate features to be activated: DomRectList, Element

impl Element

pub fn get_elements_by_class_name(&self, class_names: &str) -> HtmlCollection

The getElementsByClassName() method.

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlCollection

impl Element

pub fn get_elements_by_tag_name(&self, local_name: &str) -> HtmlCollection

The getElementsByTagName() method.

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlCollection

impl Element

pub fn get_elements_by_tag_name_ns( &self, namespace: Option<&str>, local_name: &str, ) -> Result<HtmlCollection, JsValue>

The getElementsByTagNameNS() method.

MDN Documentation

This API requires the following crate features to be activated: Element, HtmlCollection

impl Element

pub fn has_attribute(&self, name: &str) -> bool

The hasAttribute() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn has_attribute_ns( &self, namespace: Option<&str>, local_name: &str, ) -> bool

The hasAttributeNS() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn has_attributes(&self) -> bool

The hasAttributes() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn has_pointer_capture(&self, pointer_id: i32) -> bool

The hasPointerCapture() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn insert_adjacent_element( &self, where_: &str, element: &Element, ) -> Result<Option<Element>, JsValue>

The insertAdjacentElement() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn insert_adjacent_html( &self, position: &str, text: &str, ) -> Result<(), JsValue>

The insertAdjacentHTML() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn insert_adjacent_text( &self, where_: &str, data: &str, ) -> Result<(), JsValue>

The insertAdjacentText() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn matches(&self, selector: &str) -> Result<bool, JsValue>

The matches() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn query_selector( &self, selectors: &str, ) -> Result<Option<Element>, JsValue>

The querySelector() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn query_selector_all(&self, selectors: &str) -> Result<NodeList, JsValue>

The querySelectorAll() method.

MDN Documentation

This API requires the following crate features to be activated: Element, NodeList

impl Element

pub fn release_capture(&self)

The releaseCapture() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn release_pointer_capture(&self, pointer_id: i32) -> Result<(), JsValue>

The releasePointerCapture() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn remove_attribute(&self, name: &str) -> Result<(), JsValue>

The removeAttribute() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn remove_attribute_ns( &self, namespace: Option<&str>, local_name: &str, ) -> Result<(), JsValue>

The removeAttributeNS() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn remove_attribute_node( &self, old_attr: &Attr, ) -> Result<Option<Attr>, JsValue>

The removeAttributeNode() method.

MDN Documentation

This API requires the following crate features to be activated: Attr, Element

impl Element

pub fn request_fullscreen(&self) -> Result<(), JsValue>

The requestFullscreen() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn request_pointer_lock(&self)

The requestPointerLock() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_with_x_and_y(&self, x: f64, y: f64)

The scroll() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll(&self)

The scroll() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_with_scroll_to_options(&self, options: &ScrollToOptions)

The scroll() method.

MDN Documentation

This API requires the following crate features to be activated: Element, ScrollToOptions

impl Element

pub fn scroll_by_with_x_and_y(&self, x: f64, y: f64)

The scrollBy() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_by(&self)

The scrollBy() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_by_with_scroll_to_options(&self, options: &ScrollToOptions)

The scrollBy() method.

MDN Documentation

This API requires the following crate features to be activated: Element, ScrollToOptions

impl Element

pub fn scroll_into_view(&self)

The scrollIntoView() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_into_view_with_bool(&self, arg: bool)

The scrollIntoView() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_into_view_with_scroll_into_view_options( &self, arg: &ScrollIntoViewOptions, )

The scrollIntoView() method.

MDN Documentation

This API requires the following crate features to be activated: Element, ScrollIntoViewOptions

impl Element

pub fn scroll_to_with_x_and_y(&self, x: f64, y: f64)

The scrollTo() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_to(&self)

The scrollTo() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn scroll_to_with_scroll_to_options(&self, options: &ScrollToOptions)

The scrollTo() method.

MDN Documentation

This API requires the following crate features to be activated: Element, ScrollToOptions

impl Element

pub fn set_attribute(&self, name: &str, value: &str) -> Result<(), JsValue>

The setAttribute() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_attribute_ns( &self, namespace: Option<&str>, name: &str, value: &str, ) -> Result<(), JsValue>

The setAttributeNS() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_attribute_node( &self, new_attr: &Attr, ) -> Result<Option<Attr>, JsValue>

The setAttributeNode() method.

MDN Documentation

This API requires the following crate features to be activated: Attr, Element

impl Element

pub fn set_attribute_node_ns( &self, new_attr: &Attr, ) -> Result<Option<Attr>, JsValue>

The setAttributeNodeNS() method.

MDN Documentation

This API requires the following crate features to be activated: Attr, Element

impl Element

pub fn set_capture(&self)

The setCapture() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_capture_with_retarget_to_element(&self, retarget_to_element: bool)

The setCapture() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn set_pointer_capture(&self, pointer_id: i32) -> Result<(), JsValue>

The setPointerCapture() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn toggle_attribute(&self, name: &str) -> Result<bool, JsValue>

The toggleAttribute() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn toggle_attribute_with_force( &self, name: &str, force: bool, ) -> Result<bool, JsValue>

The toggleAttribute() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn webkit_matches_selector(&self, selector: &str) -> Result<bool, JsValue>

The webkitMatchesSelector() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn animate(&self, keyframes: Option<&Object>) -> Animation

The animate() method.

MDN Documentation

This API requires the following crate features to be activated: Animation, Element

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Element

pub fn animate_with_f64( &self, keyframes: Option<&Object>, options: f64, ) -> Animation

The animate() method.

MDN Documentation

This API requires the following crate features to be activated: Animation, Element

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Element

pub fn animate_with_keyframe_animation_options( &self, keyframes: Option<&Object>, options: &KeyframeAnimationOptions, ) -> Animation

The animate() method.

MDN Documentation

This API requires the following crate features to be activated: Animation, Element, KeyframeAnimationOptions

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Element

pub fn get_animations(&self) -> Array

The getAnimations() method.

MDN Documentation

This API requires the following crate features to be activated: Element

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Element

pub fn get_animations_with_options( &self, options: &GetAnimationsOptions, ) -> Array

The getAnimations() method.

MDN Documentation

This API requires the following crate features to be activated: Element, GetAnimationsOptions

This API is unstable and requires --cfg=web_sys_unstable_apis to be activated, as described in the wasm-bindgen guide

impl Element

pub fn after_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_0(&self) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_2( &self, nodes_1: &Node, nodes_2: &Node, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_3( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_4( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_5( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_6( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_node_7( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, nodes_7: &Node, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_0(&self) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_2( &self, nodes_1: &str, nodes_2: &str, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_3( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_4( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_5( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_6( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn after_with_str_7( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, nodes_7: &str, ) -> Result<(), JsValue>

The after() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_0(&self) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_2( &self, nodes_1: &Node, nodes_2: &Node, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_3( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_4( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_5( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_6( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_node_7( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, nodes_7: &Node, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_0(&self) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_2( &self, nodes_1: &str, nodes_2: &str, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_3( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_4( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_5( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_6( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn before_with_str_7( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, nodes_7: &str, ) -> Result<(), JsValue>

The before() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn remove(&self)

The remove() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_0(&self) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_2( &self, nodes_1: &Node, nodes_2: &Node, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_3( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_4( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_5( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_6( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_node_7( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, nodes_7: &Node, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_0(&self) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_2( &self, nodes_1: &str, nodes_2: &str, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_3( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_4( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_5( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_6( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_with_with_str_7( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, nodes_7: &str, ) -> Result<(), JsValue>

The replaceWith() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn convert_point_from_node_with_text( &self, point: &DomPointInit, from: &Text, ) -> Result<DomPoint, JsValue>

The convertPointFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: DomPoint, DomPointInit, Element, Text

impl Element

pub fn convert_point_from_node_with_element( &self, point: &DomPointInit, from: &Element, ) -> Result<DomPoint, JsValue>

The convertPointFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: DomPoint, DomPointInit, Element

impl Element

pub fn convert_point_from_node_with_document( &self, point: &DomPointInit, from: &Document, ) -> Result<DomPoint, JsValue>

The convertPointFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: Document, DomPoint, DomPointInit, Element

impl Element

pub fn convert_point_from_node_with_text_and_options( &self, point: &DomPointInit, from: &Text, options: &ConvertCoordinateOptions, ) -> Result<DomPoint, JsValue>

The convertPointFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, DomPoint, DomPointInit, Element, Text

impl Element

pub fn convert_point_from_node_with_element_and_options( &self, point: &DomPointInit, from: &Element, options: &ConvertCoordinateOptions, ) -> Result<DomPoint, JsValue>

The convertPointFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, DomPoint, DomPointInit, Element

impl Element

pub fn convert_point_from_node_with_document_and_options( &self, point: &DomPointInit, from: &Document, options: &ConvertCoordinateOptions, ) -> Result<DomPoint, JsValue>

The convertPointFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, Document, DomPoint, DomPointInit, Element

impl Element

pub fn convert_quad_from_node_with_text( &self, quad: &DomQuad, from: &Text, ) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: DomQuad, Element, Text

impl Element

pub fn convert_quad_from_node_with_element( &self, quad: &DomQuad, from: &Element, ) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: DomQuad, Element

impl Element

pub fn convert_quad_from_node_with_document( &self, quad: &DomQuad, from: &Document, ) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: Document, DomQuad, Element

impl Element

pub fn convert_quad_from_node_with_text_and_options( &self, quad: &DomQuad, from: &Text, options: &ConvertCoordinateOptions, ) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, Element, Text

impl Element

pub fn convert_quad_from_node_with_element_and_options( &self, quad: &DomQuad, from: &Element, options: &ConvertCoordinateOptions, ) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, Element

impl Element

pub fn convert_quad_from_node_with_document_and_options( &self, quad: &DomQuad, from: &Document, options: &ConvertCoordinateOptions, ) -> Result<DomQuad, JsValue>

The convertQuadFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, Document, DomQuad, Element

impl Element

pub fn convert_rect_from_node_with_text( &self, rect: &DomRectReadOnly, from: &Text, ) -> Result<DomQuad, JsValue>

The convertRectFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: DomQuad, DomRectReadOnly, Element, Text

impl Element

pub fn convert_rect_from_node_with_element( &self, rect: &DomRectReadOnly, from: &Element, ) -> Result<DomQuad, JsValue>

The convertRectFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: DomQuad, DomRectReadOnly, Element

impl Element

pub fn convert_rect_from_node_with_document( &self, rect: &DomRectReadOnly, from: &Document, ) -> Result<DomQuad, JsValue>

The convertRectFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: Document, DomQuad, DomRectReadOnly, Element

impl Element

pub fn convert_rect_from_node_with_text_and_options( &self, rect: &DomRectReadOnly, from: &Text, options: &ConvertCoordinateOptions, ) -> Result<DomQuad, JsValue>

The convertRectFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, DomRectReadOnly, Element, Text

impl Element

pub fn convert_rect_from_node_with_element_and_options( &self, rect: &DomRectReadOnly, from: &Element, options: &ConvertCoordinateOptions, ) -> Result<DomQuad, JsValue>

The convertRectFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, DomQuad, DomRectReadOnly, Element

impl Element

pub fn convert_rect_from_node_with_document_and_options( &self, rect: &DomRectReadOnly, from: &Document, options: &ConvertCoordinateOptions, ) -> Result<DomQuad, JsValue>

The convertRectFromNode() method.

MDN Documentation

This API requires the following crate features to be activated: ConvertCoordinateOptions, Document, DomQuad, DomRectReadOnly, Element

impl Element

pub fn get_box_quads(&self) -> Result<Array, JsValue>

The getBoxQuads() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn get_box_quads_with_options( &self, options: &BoxQuadOptions, ) -> Result<Array, JsValue>

The getBoxQuads() method.

MDN Documentation

This API requires the following crate features to be activated: BoxQuadOptions, Element

impl Element

pub fn append_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_0(&self) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_2( &self, nodes_1: &Node, nodes_2: &Node, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_3( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_4( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_5( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_6( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_node_7( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, nodes_7: &Node, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_0(&self) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_2( &self, nodes_1: &str, nodes_2: &str, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_3( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_4( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_5( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_6( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn append_with_str_7( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, nodes_7: &str, ) -> Result<(), JsValue>

The append() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_0(&self) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_1(&self, nodes_1: &Node) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_2( &self, nodes_1: &Node, nodes_2: &Node, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_3( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_4( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_5( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_6( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_node_7( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, nodes_7: &Node, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str(&self, nodes: &Array) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_0(&self) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_1(&self, nodes_1: &str) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_2( &self, nodes_1: &str, nodes_2: &str, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_3( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_4( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_5( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_6( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn prepend_with_str_7( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, nodes_7: &str, ) -> Result<(), JsValue>

The prepend() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node(&self, nodes: &Array)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_0(&self)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_1(&self, nodes_1: &Node)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_2(&self, nodes_1: &Node, nodes_2: &Node)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_3( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_4( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_5( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_6( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_node_7( &self, nodes_1: &Node, nodes_2: &Node, nodes_3: &Node, nodes_4: &Node, nodes_5: &Node, nodes_6: &Node, nodes_7: &Node, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str(&self, nodes: &Array)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_0(&self)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_1(&self, nodes_1: &str)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_2(&self, nodes_1: &str, nodes_2: &str)

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_3( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_4( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_5( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_6( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element

impl Element

pub fn replace_children_with_str_7( &self, nodes_1: &str, nodes_2: &str, nodes_3: &str, nodes_4: &str, nodes_5: &str, nodes_6: &str, nodes_7: &str, )

The replaceChildren() method.

MDN Documentation

This API requires the following crate features to be activated: Element