Struct web_sys::Event

#[repr(transparent)]

pub struct Event { /* fields omitted */ }

The Event object

MDN Documentation

Methods

impl Event

pub fn new(type_: &str) -> Result<Event, JsValue>

This function requires the following crate features to be activated: Event

impl Event

pub fn new_with_event_init_dict(
type_: &str,
event_init_dict: &EventInit
) -> Result<Event, JsValue>

This function requires the following crate features to be activated: Event, EventInit

impl Event

pub fn init_event(&self, type_: &str)

The initEvent() method

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn init_event_with_bubbles(&self, type_: &str, bubbles: bool)

The initEvent() method

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn init_event_with_bubbles_and_cancelable(
&self,
type_: &str,
bubbles: bool,
cancelable: bool
)

The initEvent() method

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn prevent_default(&self)

The preventDefault() method

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn stop_immediate_propagation(&self)

The stopImmediatePropagation() method

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn stop_propagation(&self)

The stopPropagation() method

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn type_(&self) -> String

The type getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn target(&self) -> Option<EventTarget>

The target getter

MDN Documentation

This function requires the following crate features to be activated: Event, EventTarget

impl Event

pub fn current_target(&self) -> Option<EventTarget>

The currentTarget getter

MDN Documentation

This function requires the following crate features to be activated: Event, EventTarget

impl Event

pub fn event_phase(&self) -> u16

The eventPhase getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn bubbles(&self) -> bool

The bubbles getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn cancelable(&self) -> bool

The cancelable getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn default_prevented(&self) -> bool

The defaultPrevented getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn composed(&self) -> bool

The composed getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn is_trusted(&self) -> bool

The isTrusted getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn time_stamp(&self) -> f64

The timeStamp getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn cancel_bubble(&self) -> bool

The cancelBubble getter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub fn set_cancel_bubble(&self, cancel_bubble: bool)

The cancelBubble setter

MDN Documentation

This function requires the following crate features to be activated: Event

impl Event

pub const NONE: u16

impl Event

pub const CAPTURING_PHASE: u16

impl Event

pub const AT_TARGET: u16

impl Event

pub const BUBBLING_PHASE: u16