Struct web_sys::Event

#[repr(transparent)]
pub struct Event { /* fields omitted */ }

The Event class.

MDN Documentation

This API requires the following crate features to be activated: Event

Implementations

impl Event

pub fn type_(&self) -> String

Getter for the type field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn target(&self) -> Option<EventTarget>

Getter for the target field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event, EventTarget

impl Event

pub fn current_target(&self) -> Option<EventTarget>

Getter for the currentTarget field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event, EventTarget

impl Event

pub fn event_phase(&self) -> u16

Getter for the eventPhase field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn bubbles(&self) -> bool

Getter for the bubbles field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn cancelable(&self) -> bool

Getter for the cancelable field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn default_prevented(&self) -> bool

Getter for the defaultPrevented field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn composed(&self) -> bool

Getter for the composed field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn is_trusted(&self) -> bool

Getter for the isTrusted field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn time_stamp(&self) -> f64

Getter for the timeStamp field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn cancel_bubble(&self) -> bool

Getter for the cancelBubble field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn set_cancel_bubble(&self, value: bool)

Setter for the cancelBubble field of this object.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn new(type_: &str) -> Result<Event, JsValue>

The new Event(..) constructor, creating a new instance of Event.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn new_with_event_init_dict(
type_: &str,
event_init_dict: &EventInit
) -> Result<Event, JsValue>

The new Event(..) constructor, creating a new instance of Event.

MDN Documentation

This API requires the following crate features to be activated: Event, EventInit

impl Event

pub fn composed_path(&self) -> Array

The composedPath() method.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn init_event(&self, type_: &str)

The initEvent() method.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn init_event_with_bubbles(&self, type_: &str, bubbles: bool)

The initEvent() method.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn init_event_with_bubbles_and_cancelable(
&self,
type_: &str,
bubbles: bool,
cancelable: bool
)

The initEvent() method.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn prevent_default(&self)

The preventDefault() method.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn stop_immediate_propagation(&self)

The stopImmediatePropagation() method.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub fn stop_propagation(&self)

The stopPropagation() method.

MDN Documentation

This API requires the following crate features to be activated: Event

impl Event

pub const NONE: u16

The Event.NONE const.

This API requires the following crate features to be activated: Event

pub const CAPTURING_PHASE: u16

The Event.CAPTURING_PHASE const.

This API requires the following crate features to be activated: Event

pub const AT_TARGET: u16

The Event.AT_TARGET const.

This API requires the following crate features to be activated: Event

pub const BUBBLING_PHASE: u16

The Event.BUBBLING_PHASE const.

This API requires the following crate features to be activated: Event