Struct web_sys::HtmlDocument

#[repr(transparent)]

pub struct HtmlDocument { /* fields omitted */ }

The HTMLDocument object

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

Methods

impl HtmlDocument

pub fn capture_events(&self)

The captureEvents() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn clear(&self)

The clear() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn close(&self) -> Result<(), JsValue>

The close() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn exec_command(&self, command_id: &str) -> Result<bool, JsValue>

The execCommand() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn exec_command_with_show_ui(
&self,
command_id: &str,
show_ui: bool
) -> Result<bool, JsValue>

The execCommand() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn exec_command_with_show_ui_and_value(
&self,
command_id: &str,
show_ui: bool,
value: &str
) -> Result<bool, JsValue>

The execCommand() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn open(&self) -> Result<Document, JsValue>

The open() method

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlDocument

impl HtmlDocument

pub fn open_with_type(&self, type_: &str) -> Result<Document, JsValue>

The open() method

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlDocument

impl HtmlDocument

pub fn open_with_type_and_replace(
&self,
type_: &str,
replace: &str
) -> Result<Document, JsValue>

The open() method

MDN Documentation

This API requires the following crate features to be activated: Document, HtmlDocument

impl HtmlDocument

pub fn open_with_url_and_name_and_features(
&self,
url: &str,
name: &str,
features: &str
) -> Result<Option<Window>, JsValue>

The open() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument, Window

impl HtmlDocument

pub fn open_with_url_and_name_and_features_and_replace(
&self,
url: &str,
name: &str,
features: &str,
replace: bool
) -> Result<Option<Window>, JsValue>

The open() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument, Window

impl HtmlDocument

pub fn query_command_enabled(&self, command_id: &str) -> Result<bool, JsValue>

The queryCommandEnabled() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn query_command_indeterm(&self, command_id: &str) -> Result<bool, JsValue>

The queryCommandIndeterm() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn query_command_state(&self, command_id: &str) -> Result<bool, JsValue>

The queryCommandState() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn query_command_supported(&self, command_id: &str) -> bool

The queryCommandSupported() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn query_command_value(&self, command_id: &str) -> Result<String, JsValue>

The queryCommandValue() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn release_events(&self)

The releaseEvents() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write(&self, text: &Array) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_0(&self) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_1(&self, text_1: &str) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_2(&self, text_1: &str, text_2: &str) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_3(
&self,
text_1: &str,
text_2: &str,
text_3: &str
) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_4(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str
) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_5(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str,
text_5: &str
) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_6(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str,
text_5: &str,
text_6: &str
) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn write_7(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str,
text_5: &str,
text_6: &str,
text_7: &str
) -> Result<(), JsValue>

The write() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln(&self, text: &Array) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_0(&self) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_1(&self, text_1: &str) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_2(&self, text_1: &str, text_2: &str) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_3(
&self,
text_1: &str,
text_2: &str,
text_3: &str
) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_4(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str
) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_5(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str,
text_5: &str
) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_6(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str,
text_5: &str,
text_6: &str
) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn writeln_7(
&self,
text_1: &str,
text_2: &str,
text_3: &str,
text_4: &str,
text_5: &str,
text_6: &str,
text_7: &str
) -> Result<(), JsValue>

The writeln() method

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn get(&self, name: &str) -> Result<Object, JsValue>

The indexing getter

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn domain(&self) -> String

The domain getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_domain(&self, domain: &str)

The domain setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn cookie(&self) -> Result<String, JsValue>

The cookie getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_cookie(&self, cookie: &str) -> Result<(), JsValue>

The cookie setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn design_mode(&self) -> String

The designMode getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_design_mode(&self, design_mode: &str)

The designMode setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn fg_color(&self) -> String

The fgColor getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_fg_color(&self, fg_color: &str)

The fgColor setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn link_color(&self) -> String

The linkColor getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_link_color(&self, link_color: &str)

The linkColor setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn vlink_color(&self) -> String

The vlinkColor getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_vlink_color(&self, vlink_color: &str)

The vlinkColor setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn alink_color(&self) -> String

The alinkColor getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_alink_color(&self, alink_color: &str)

The alinkColor setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn bg_color(&self) -> String

The bgColor getter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn set_bg_color(&self, bg_color: &str)

The bgColor setter

MDN Documentation

This API requires the following crate features to be activated: HtmlDocument

impl HtmlDocument

pub fn all(&self) -> HtmlAllCollection

The all getter

MDN Documentation

This API requires the following crate features to be activated: HtmlAllCollection, HtmlDocument