web_sys
Struct KeyEvent

pub struct KeyEvent { /* private fields */ }

The KeyEvent class.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

Implementations

impl KeyEvent

pub fn init_key_event(&self, type_: &str)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble(&self, type_: &str, can_bubble: bool)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable( &self, type_: &str, can_bubble: bool, cancelable: bool, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view( &self, type_: &str, can_bubble: bool, cancelable: bool, view: Option<&Window>, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key( &self, type_: &str, can_bubble: bool, cancelable: bool, view: Option<&Window>, ctrl_key: bool, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key( &self, type_: &str, can_bubble: bool, cancelable: bool, view: Option<&Window>, ctrl_key: bool, alt_key: bool, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key( &self, type_: &str, can_bubble: bool, cancelable: bool, view: Option<&Window>, ctrl_key: bool, alt_key: bool, shift_key: bool, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key( &self, type_: &str, can_bubble: bool, cancelable: bool, view: Option<&Window>, ctrl_key: bool, alt_key: bool, shift_key: bool, meta_key: bool, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_key_code( &self, type_: &str, can_bubble: bool, cancelable: bool, view: Option<&Window>, ctrl_key: bool, alt_key: bool, shift_key: bool, meta_key: bool, key_code: u32, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_key_code_and_char_code( &self, type_: &str, can_bubble: bool, cancelable: bool, view: Option<&Window>, ctrl_key: bool, alt_key: bool, shift_key: bool, meta_key: bool, key_code: u32, char_code: u32, )

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub const DOM_VK_CANCEL: u32 = 3u32

The KeyEvent.DOM_VK_CANCEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HELP: u32 = 6u32

The KeyEvent.DOM_VK_HELP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_BACK_SPACE: u32 = 8u32

The KeyEvent.DOM_VK_BACK_SPACE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_TAB: u32 = 9u32

The KeyEvent.DOM_VK_TAB const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLEAR: u32 = 12u32

The KeyEvent.DOM_VK_CLEAR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_RETURN: u32 = 13u32

The KeyEvent.DOM_VK_RETURN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SHIFT: u32 = 16u32

The KeyEvent.DOM_VK_SHIFT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CONTROL: u32 = 17u32

The KeyEvent.DOM_VK_CONTROL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ALT: u32 = 18u32

The KeyEvent.DOM_VK_ALT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PAUSE: u32 = 19u32

The KeyEvent.DOM_VK_PAUSE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CAPS_LOCK: u32 = 20u32

The KeyEvent.DOM_VK_CAPS_LOCK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_KANA: u32 = 21u32

The KeyEvent.DOM_VK_KANA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HANGUL: u32 = 21u32

The KeyEvent.DOM_VK_HANGUL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EISU: u32 = 22u32

The KeyEvent.DOM_VK_EISU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_JUNJA: u32 = 23u32

The KeyEvent.DOM_VK_JUNJA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_FINAL: u32 = 24u32

The KeyEvent.DOM_VK_FINAL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HANJA: u32 = 25u32

The KeyEvent.DOM_VK_HANJA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_KANJI: u32 = 25u32

The KeyEvent.DOM_VK_KANJI const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ESCAPE: u32 = 27u32

The KeyEvent.DOM_VK_ESCAPE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CONVERT: u32 = 28u32

The KeyEvent.DOM_VK_CONVERT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NONCONVERT: u32 = 29u32

The KeyEvent.DOM_VK_NONCONVERT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ACCEPT: u32 = 30u32

The KeyEvent.DOM_VK_ACCEPT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_MODECHANGE: u32 = 31u32

The KeyEvent.DOM_VK_MODECHANGE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SPACE: u32 = 32u32

The KeyEvent.DOM_VK_SPACE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PAGE_UP: u32 = 33u32

The KeyEvent.DOM_VK_PAGE_UP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PAGE_DOWN: u32 = 34u32

The KeyEvent.DOM_VK_PAGE_DOWN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_END: u32 = 35u32

The KeyEvent.DOM_VK_END const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HOME: u32 = 36u32

The KeyEvent.DOM_VK_HOME const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_LEFT: u32 = 37u32

The KeyEvent.DOM_VK_LEFT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_UP: u32 = 38u32

The KeyEvent.DOM_VK_UP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_RIGHT: u32 = 39u32

The KeyEvent.DOM_VK_RIGHT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DOWN: u32 = 40u32

The KeyEvent.DOM_VK_DOWN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SELECT: u32 = 41u32

The KeyEvent.DOM_VK_SELECT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PRINT: u32 = 42u32

The KeyEvent.DOM_VK_PRINT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EXECUTE: u32 = 43u32

The KeyEvent.DOM_VK_EXECUTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PRINTSCREEN: u32 = 44u32

The KeyEvent.DOM_VK_PRINTSCREEN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_INSERT: u32 = 45u32

The KeyEvent.DOM_VK_INSERT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DELETE: u32 = 46u32

The KeyEvent.DOM_VK_DELETE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_0: u32 = 48u32

The KeyEvent.DOM_VK_0 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_1: u32 = 49u32

The KeyEvent.DOM_VK_1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_2: u32 = 50u32

The KeyEvent.DOM_VK_2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_3: u32 = 51u32

The KeyEvent.DOM_VK_3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_4: u32 = 52u32

The KeyEvent.DOM_VK_4 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_5: u32 = 53u32

The KeyEvent.DOM_VK_5 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_6: u32 = 54u32

The KeyEvent.DOM_VK_6 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_7: u32 = 55u32

The KeyEvent.DOM_VK_7 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_8: u32 = 56u32

The KeyEvent.DOM_VK_8 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_9: u32 = 57u32

The KeyEvent.DOM_VK_9 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_COLON: u32 = 58u32

The KeyEvent.DOM_VK_COLON const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SEMICOLON: u32 = 59u32

The KeyEvent.DOM_VK_SEMICOLON const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_LESS_THAN: u32 = 60u32

The KeyEvent.DOM_VK_LESS_THAN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EQUALS: u32 = 61u32

The KeyEvent.DOM_VK_EQUALS const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_GREATER_THAN: u32 = 62u32

The KeyEvent.DOM_VK_GREATER_THAN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_QUESTION_MARK: u32 = 63u32

The KeyEvent.DOM_VK_QUESTION_MARK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_AT: u32 = 64u32

The KeyEvent.DOM_VK_AT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_A: u32 = 65u32

The KeyEvent.DOM_VK_A const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_B: u32 = 66u32

The KeyEvent.DOM_VK_B const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_C: u32 = 67u32

The KeyEvent.DOM_VK_C const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_D: u32 = 68u32

The KeyEvent.DOM_VK_D const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_E: u32 = 69u32

The KeyEvent.DOM_VK_E const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F: u32 = 70u32

The KeyEvent.DOM_VK_F const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_G: u32 = 71u32

The KeyEvent.DOM_VK_G const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_H: u32 = 72u32

The KeyEvent.DOM_VK_H const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_I: u32 = 73u32

The KeyEvent.DOM_VK_I const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_J: u32 = 74u32

The KeyEvent.DOM_VK_J const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_K: u32 = 75u32

The KeyEvent.DOM_VK_K const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_L: u32 = 76u32

The KeyEvent.DOM_VK_L const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_M: u32 = 77u32

The KeyEvent.DOM_VK_M const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_N: u32 = 78u32

The KeyEvent.DOM_VK_N const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_O: u32 = 79u32

The KeyEvent.DOM_VK_O const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_P: u32 = 80u32

The KeyEvent.DOM_VK_P const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_Q: u32 = 81u32

The KeyEvent.DOM_VK_Q const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_R: u32 = 82u32

The KeyEvent.DOM_VK_R const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_S: u32 = 83u32

The KeyEvent.DOM_VK_S const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_T: u32 = 84u32

The KeyEvent.DOM_VK_T const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_U: u32 = 85u32

The KeyEvent.DOM_VK_U const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_V: u32 = 86u32

The KeyEvent.DOM_VK_V const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_W: u32 = 87u32

The KeyEvent.DOM_VK_W const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_X: u32 = 88u32

The KeyEvent.DOM_VK_X const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_Y: u32 = 89u32

The KeyEvent.DOM_VK_Y const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_Z: u32 = 90u32

The KeyEvent.DOM_VK_Z const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN: u32 = 91u32

The KeyEvent.DOM_VK_WIN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CONTEXT_MENU: u32 = 93u32

The KeyEvent.DOM_VK_CONTEXT_MENU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SLEEP: u32 = 95u32

The KeyEvent.DOM_VK_SLEEP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD0: u32 = 96u32

The KeyEvent.DOM_VK_NUMPAD0 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD1: u32 = 97u32

The KeyEvent.DOM_VK_NUMPAD1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD2: u32 = 98u32

The KeyEvent.DOM_VK_NUMPAD2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD3: u32 = 99u32

The KeyEvent.DOM_VK_NUMPAD3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD4: u32 = 100u32

The KeyEvent.DOM_VK_NUMPAD4 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD5: u32 = 101u32

The KeyEvent.DOM_VK_NUMPAD5 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD6: u32 = 102u32

The KeyEvent.DOM_VK_NUMPAD6 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD7: u32 = 103u32

The KeyEvent.DOM_VK_NUMPAD7 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD8: u32 = 104u32

The KeyEvent.DOM_VK_NUMPAD8 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD9: u32 = 105u32

The KeyEvent.DOM_VK_NUMPAD9 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_MULTIPLY: u32 = 106u32

The KeyEvent.DOM_VK_MULTIPLY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ADD: u32 = 107u32

The KeyEvent.DOM_VK_ADD const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SEPARATOR: u32 = 108u32

The KeyEvent.DOM_VK_SEPARATOR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SUBTRACT: u32 = 109u32

The KeyEvent.DOM_VK_SUBTRACT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DECIMAL: u32 = 110u32

The KeyEvent.DOM_VK_DECIMAL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DIVIDE: u32 = 111u32

The KeyEvent.DOM_VK_DIVIDE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F1: u32 = 112u32

The KeyEvent.DOM_VK_F1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F2: u32 = 113u32

The KeyEvent.DOM_VK_F2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F3: u32 = 114u32

The KeyEvent.DOM_VK_F3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F4: u32 = 115u32

The KeyEvent.DOM_VK_F4 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F5: u32 = 116u32

The KeyEvent.DOM_VK_F5 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F6: u32 = 117u32

The KeyEvent.DOM_VK_F6 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F7: u32 = 118u32

The KeyEvent.DOM_VK_F7 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F8: u32 = 119u32

The KeyEvent.DOM_VK_F8 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F9: u32 = 120u32

The KeyEvent.DOM_VK_F9 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F10: u32 = 121u32

The KeyEvent.DOM_VK_F10 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F11: u32 = 122u32

The KeyEvent.DOM_VK_F11 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F12: u32 = 123u32

The KeyEvent.DOM_VK_F12 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F13: u32 = 124u32

The KeyEvent.DOM_VK_F13 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F14: u32 = 125u32

The KeyEvent.DOM_VK_F14 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F15: u32 = 126u32

The KeyEvent.DOM_VK_F15 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F16: u32 = 127u32

The KeyEvent.DOM_VK_F16 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F17: u32 = 128u32

The KeyEvent.DOM_VK_F17 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F18: u32 = 129u32

The KeyEvent.DOM_VK_F18 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F19: u32 = 130u32

The KeyEvent.DOM_VK_F19 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F20: u32 = 131u32

The KeyEvent.DOM_VK_F20 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F21: u32 = 132u32

The KeyEvent.DOM_VK_F21 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F22: u32 = 133u32

The KeyEvent.DOM_VK_F22 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F23: u32 = 134u32

The KeyEvent.DOM_VK_F23 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F24: u32 = 135u32

The KeyEvent.DOM_VK_F24 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUM_LOCK: u32 = 144u32

The KeyEvent.DOM_VK_NUM_LOCK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SCROLL_LOCK: u32 = 145u32

The KeyEvent.DOM_VK_SCROLL_LOCK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_JISHO: u32 = 146u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_JISHO const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_MASSHOU: u32 = 147u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_MASSHOU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_TOUROKU: u32 = 148u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_TOUROKU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_LOYA: u32 = 149u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_LOYA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_ROYA: u32 = 150u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_ROYA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CIRCUMFLEX: u32 = 160u32

The KeyEvent.DOM_VK_CIRCUMFLEX const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EXCLAMATION: u32 = 161u32

The KeyEvent.DOM_VK_EXCLAMATION const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DOUBLE_QUOTE: u32 = 162u32

The KeyEvent.DOM_VK_DOUBLE_QUOTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HASH: u32 = 163u32

The KeyEvent.DOM_VK_HASH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DOLLAR: u32 = 164u32

The KeyEvent.DOM_VK_DOLLAR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PERCENT: u32 = 165u32

The KeyEvent.DOM_VK_PERCENT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_AMPERSAND: u32 = 166u32

The KeyEvent.DOM_VK_AMPERSAND const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_UNDERSCORE: u32 = 167u32

The KeyEvent.DOM_VK_UNDERSCORE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_OPEN_PAREN: u32 = 168u32

The KeyEvent.DOM_VK_OPEN_PAREN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLOSE_PAREN: u32 = 169u32

The KeyEvent.DOM_VK_CLOSE_PAREN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ASTERISK: u32 = 170u32

The KeyEvent.DOM_VK_ASTERISK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PLUS: u32 = 171u32

The KeyEvent.DOM_VK_PLUS const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PIPE: u32 = 172u32

The KeyEvent.DOM_VK_PIPE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HYPHEN_MINUS: u32 = 173u32

The KeyEvent.DOM_VK_HYPHEN_MINUS const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_OPEN_CURLY_BRACKET: u32 = 174u32

The KeyEvent.DOM_VK_OPEN_CURLY_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLOSE_CURLY_BRACKET: u32 = 175u32

The KeyEvent.DOM_VK_CLOSE_CURLY_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_TILDE: u32 = 176u32

The KeyEvent.DOM_VK_TILDE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_VOLUME_MUTE: u32 = 181u32

The KeyEvent.DOM_VK_VOLUME_MUTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_VOLUME_DOWN: u32 = 182u32

The KeyEvent.DOM_VK_VOLUME_DOWN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_VOLUME_UP: u32 = 183u32

The KeyEvent.DOM_VK_VOLUME_UP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_COMMA: u32 = 188u32

The KeyEvent.DOM_VK_COMMA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PERIOD: u32 = 190u32

The KeyEvent.DOM_VK_PERIOD const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SLASH: u32 = 191u32

The KeyEvent.DOM_VK_SLASH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_BACK_QUOTE: u32 = 192u32

The KeyEvent.DOM_VK_BACK_QUOTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_OPEN_BRACKET: u32 = 219u32

The KeyEvent.DOM_VK_OPEN_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_BACK_SLASH: u32 = 220u32

The KeyEvent.DOM_VK_BACK_SLASH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLOSE_BRACKET: u32 = 221u32

The KeyEvent.DOM_VK_CLOSE_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_QUOTE: u32 = 222u32

The KeyEvent.DOM_VK_QUOTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_META: u32 = 224u32

The KeyEvent.DOM_VK_META const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ALTGR: u32 = 225u32

The KeyEvent.DOM_VK_ALTGR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_ICO_HELP: u32 = 227u32

The KeyEvent.DOM_VK_WIN_ICO_HELP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_ICO_00: u32 = 228u32

The KeyEvent.DOM_VK_WIN_ICO_00 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PROCESSKEY: u32 = 229u32

The KeyEvent.DOM_VK_PROCESSKEY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_ICO_CLEAR: u32 = 230u32

The KeyEvent.DOM_VK_WIN_ICO_CLEAR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_RESET: u32 = 233u32

The KeyEvent.DOM_VK_WIN_OEM_RESET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_JUMP: u32 = 234u32

The KeyEvent.DOM_VK_WIN_OEM_JUMP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_PA1: u32 = 235u32

The KeyEvent.DOM_VK_WIN_OEM_PA1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_PA2: u32 = 236u32

The KeyEvent.DOM_VK_WIN_OEM_PA2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_PA3: u32 = 237u32

The KeyEvent.DOM_VK_WIN_OEM_PA3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_WSCTRL: u32 = 238u32

The KeyEvent.DOM_VK_WIN_OEM_WSCTRL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_CUSEL: u32 = 239u32

The KeyEvent.DOM_VK_WIN_OEM_CUSEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_ATTN: u32 = 240u32

The KeyEvent.DOM_VK_WIN_OEM_ATTN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FINISH: u32 = 241u32

The KeyEvent.DOM_VK_WIN_OEM_FINISH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_COPY: u32 = 242u32

The KeyEvent.DOM_VK_WIN_OEM_COPY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_AUTO: u32 = 243u32

The KeyEvent.DOM_VK_WIN_OEM_AUTO const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_ENLW: u32 = 244u32

The KeyEvent.DOM_VK_WIN_OEM_ENLW const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_BACKTAB: u32 = 245u32

The KeyEvent.DOM_VK_WIN_OEM_BACKTAB const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ATTN: u32 = 246u32

The KeyEvent.DOM_VK_ATTN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CRSEL: u32 = 247u32

The KeyEvent.DOM_VK_CRSEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EXSEL: u32 = 248u32

The KeyEvent.DOM_VK_EXSEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EREOF: u32 = 249u32

The KeyEvent.DOM_VK_EREOF const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PLAY: u32 = 250u32

The KeyEvent.DOM_VK_PLAY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ZOOM: u32 = 251u32

The KeyEvent.DOM_VK_ZOOM const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PA1: u32 = 253u32

The KeyEvent.DOM_VK_PA1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_CLEAR: u32 = 254u32

The KeyEvent.DOM_VK_WIN_OEM_CLEAR const.

This API requires the following crate features to be activated: KeyEvent