Struct web_sys::KeyEvent

#[repr(transparent)]pub struct KeyEvent { /* fields omitted */ }

The KeyEvent class.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

Implementations

impl KeyEvent

pub fn init_key_event(&self, type_: &str)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble(&self, type_: &str, can_bubble: bool)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_key_code(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
key_code: u32
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub fn init_key_event_with_can_bubble_and_cancelable_and_view_and_ctrl_key_and_alt_key_and_shift_key_and_meta_key_and_key_code_and_char_code(
&self,
type_: &str,
can_bubble: bool,
cancelable: bool,
view: Option<&Window>,
ctrl_key: bool,
alt_key: bool,
shift_key: bool,
meta_key: bool,
key_code: u32,
char_code: u32
)

The initKeyEvent() method.

MDN Documentation

This API requires the following crate features to be activated: KeyEvent, Window

impl KeyEvent

pub const DOM_VK_CANCEL: u32

The KeyEvent.DOM_VK_CANCEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HELP: u32

The KeyEvent.DOM_VK_HELP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_BACK_SPACE: u32

The KeyEvent.DOM_VK_BACK_SPACE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_TAB: u32

The KeyEvent.DOM_VK_TAB const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLEAR: u32

The KeyEvent.DOM_VK_CLEAR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_RETURN: u32

The KeyEvent.DOM_VK_RETURN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SHIFT: u32

The KeyEvent.DOM_VK_SHIFT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CONTROL: u32

The KeyEvent.DOM_VK_CONTROL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ALT: u32

The KeyEvent.DOM_VK_ALT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PAUSE: u32

The KeyEvent.DOM_VK_PAUSE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CAPS_LOCK: u32

The KeyEvent.DOM_VK_CAPS_LOCK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_KANA: u32

The KeyEvent.DOM_VK_KANA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HANGUL: u32

The KeyEvent.DOM_VK_HANGUL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EISU: u32

The KeyEvent.DOM_VK_EISU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_JUNJA: u32

The KeyEvent.DOM_VK_JUNJA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_FINAL: u32

The KeyEvent.DOM_VK_FINAL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HANJA: u32

The KeyEvent.DOM_VK_HANJA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_KANJI: u32

The KeyEvent.DOM_VK_KANJI const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ESCAPE: u32

The KeyEvent.DOM_VK_ESCAPE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CONVERT: u32

The KeyEvent.DOM_VK_CONVERT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NONCONVERT: u32

The KeyEvent.DOM_VK_NONCONVERT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ACCEPT: u32

The KeyEvent.DOM_VK_ACCEPT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_MODECHANGE: u32

The KeyEvent.DOM_VK_MODECHANGE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SPACE: u32

The KeyEvent.DOM_VK_SPACE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PAGE_UP: u32

The KeyEvent.DOM_VK_PAGE_UP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PAGE_DOWN: u32

The KeyEvent.DOM_VK_PAGE_DOWN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_END: u32

The KeyEvent.DOM_VK_END const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HOME: u32

The KeyEvent.DOM_VK_HOME const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_LEFT: u32

The KeyEvent.DOM_VK_LEFT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_UP: u32

The KeyEvent.DOM_VK_UP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_RIGHT: u32

The KeyEvent.DOM_VK_RIGHT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DOWN: u32

The KeyEvent.DOM_VK_DOWN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SELECT: u32

The KeyEvent.DOM_VK_SELECT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PRINT: u32

The KeyEvent.DOM_VK_PRINT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EXECUTE: u32

The KeyEvent.DOM_VK_EXECUTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PRINTSCREEN: u32

The KeyEvent.DOM_VK_PRINTSCREEN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_INSERT: u32

The KeyEvent.DOM_VK_INSERT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DELETE: u32

The KeyEvent.DOM_VK_DELETE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_0: u32

The KeyEvent.DOM_VK_0 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_1: u32

The KeyEvent.DOM_VK_1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_2: u32

The KeyEvent.DOM_VK_2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_3: u32

The KeyEvent.DOM_VK_3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_4: u32

The KeyEvent.DOM_VK_4 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_5: u32

The KeyEvent.DOM_VK_5 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_6: u32

The KeyEvent.DOM_VK_6 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_7: u32

The KeyEvent.DOM_VK_7 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_8: u32

The KeyEvent.DOM_VK_8 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_9: u32

The KeyEvent.DOM_VK_9 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_COLON: u32

The KeyEvent.DOM_VK_COLON const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SEMICOLON: u32

The KeyEvent.DOM_VK_SEMICOLON const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_LESS_THAN: u32

The KeyEvent.DOM_VK_LESS_THAN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EQUALS: u32

The KeyEvent.DOM_VK_EQUALS const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_GREATER_THAN: u32

The KeyEvent.DOM_VK_GREATER_THAN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_QUESTION_MARK: u32

The KeyEvent.DOM_VK_QUESTION_MARK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_AT: u32

The KeyEvent.DOM_VK_AT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_A: u32

The KeyEvent.DOM_VK_A const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_B: u32

The KeyEvent.DOM_VK_B const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_C: u32

The KeyEvent.DOM_VK_C const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_D: u32

The KeyEvent.DOM_VK_D const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_E: u32

The KeyEvent.DOM_VK_E const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F: u32

The KeyEvent.DOM_VK_F const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_G: u32

The KeyEvent.DOM_VK_G const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_H: u32

The KeyEvent.DOM_VK_H const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_I: u32

The KeyEvent.DOM_VK_I const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_J: u32

The KeyEvent.DOM_VK_J const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_K: u32

The KeyEvent.DOM_VK_K const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_L: u32

The KeyEvent.DOM_VK_L const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_M: u32

The KeyEvent.DOM_VK_M const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_N: u32

The KeyEvent.DOM_VK_N const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_O: u32

The KeyEvent.DOM_VK_O const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_P: u32

The KeyEvent.DOM_VK_P const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_Q: u32

The KeyEvent.DOM_VK_Q const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_R: u32

The KeyEvent.DOM_VK_R const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_S: u32

The KeyEvent.DOM_VK_S const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_T: u32

The KeyEvent.DOM_VK_T const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_U: u32

The KeyEvent.DOM_VK_U const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_V: u32

The KeyEvent.DOM_VK_V const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_W: u32

The KeyEvent.DOM_VK_W const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_X: u32

The KeyEvent.DOM_VK_X const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_Y: u32

The KeyEvent.DOM_VK_Y const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_Z: u32

The KeyEvent.DOM_VK_Z const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN: u32

The KeyEvent.DOM_VK_WIN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CONTEXT_MENU: u32

The KeyEvent.DOM_VK_CONTEXT_MENU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SLEEP: u32

The KeyEvent.DOM_VK_SLEEP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD0: u32

The KeyEvent.DOM_VK_NUMPAD0 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD1: u32

The KeyEvent.DOM_VK_NUMPAD1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD2: u32

The KeyEvent.DOM_VK_NUMPAD2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD3: u32

The KeyEvent.DOM_VK_NUMPAD3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD4: u32

The KeyEvent.DOM_VK_NUMPAD4 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD5: u32

The KeyEvent.DOM_VK_NUMPAD5 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD6: u32

The KeyEvent.DOM_VK_NUMPAD6 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD7: u32

The KeyEvent.DOM_VK_NUMPAD7 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD8: u32

The KeyEvent.DOM_VK_NUMPAD8 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUMPAD9: u32

The KeyEvent.DOM_VK_NUMPAD9 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_MULTIPLY: u32

The KeyEvent.DOM_VK_MULTIPLY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ADD: u32

The KeyEvent.DOM_VK_ADD const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SEPARATOR: u32

The KeyEvent.DOM_VK_SEPARATOR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SUBTRACT: u32

The KeyEvent.DOM_VK_SUBTRACT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DECIMAL: u32

The KeyEvent.DOM_VK_DECIMAL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DIVIDE: u32

The KeyEvent.DOM_VK_DIVIDE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F1: u32

The KeyEvent.DOM_VK_F1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F2: u32

The KeyEvent.DOM_VK_F2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F3: u32

The KeyEvent.DOM_VK_F3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F4: u32

The KeyEvent.DOM_VK_F4 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F5: u32

The KeyEvent.DOM_VK_F5 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F6: u32

The KeyEvent.DOM_VK_F6 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F7: u32

The KeyEvent.DOM_VK_F7 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F8: u32

The KeyEvent.DOM_VK_F8 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F9: u32

The KeyEvent.DOM_VK_F9 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F10: u32

The KeyEvent.DOM_VK_F10 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F11: u32

The KeyEvent.DOM_VK_F11 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F12: u32

The KeyEvent.DOM_VK_F12 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F13: u32

The KeyEvent.DOM_VK_F13 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F14: u32

The KeyEvent.DOM_VK_F14 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F15: u32

The KeyEvent.DOM_VK_F15 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F16: u32

The KeyEvent.DOM_VK_F16 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F17: u32

The KeyEvent.DOM_VK_F17 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F18: u32

The KeyEvent.DOM_VK_F18 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F19: u32

The KeyEvent.DOM_VK_F19 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F20: u32

The KeyEvent.DOM_VK_F20 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F21: u32

The KeyEvent.DOM_VK_F21 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F22: u32

The KeyEvent.DOM_VK_F22 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F23: u32

The KeyEvent.DOM_VK_F23 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_F24: u32

The KeyEvent.DOM_VK_F24 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_NUM_LOCK: u32

The KeyEvent.DOM_VK_NUM_LOCK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SCROLL_LOCK: u32

The KeyEvent.DOM_VK_SCROLL_LOCK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_JISHO: u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_JISHO const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_MASSHOU: u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_MASSHOU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_TOUROKU: u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_TOUROKU const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_LOYA: u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_LOYA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FJ_ROYA: u32

The KeyEvent.DOM_VK_WIN_OEM_FJ_ROYA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CIRCUMFLEX: u32

The KeyEvent.DOM_VK_CIRCUMFLEX const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EXCLAMATION: u32

The KeyEvent.DOM_VK_EXCLAMATION const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DOUBLE_QUOTE: u32

The KeyEvent.DOM_VK_DOUBLE_QUOTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HASH: u32

The KeyEvent.DOM_VK_HASH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_DOLLAR: u32

The KeyEvent.DOM_VK_DOLLAR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PERCENT: u32

The KeyEvent.DOM_VK_PERCENT const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_AMPERSAND: u32

The KeyEvent.DOM_VK_AMPERSAND const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_UNDERSCORE: u32

The KeyEvent.DOM_VK_UNDERSCORE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_OPEN_PAREN: u32

The KeyEvent.DOM_VK_OPEN_PAREN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLOSE_PAREN: u32

The KeyEvent.DOM_VK_CLOSE_PAREN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ASTERISK: u32

The KeyEvent.DOM_VK_ASTERISK const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PLUS: u32

The KeyEvent.DOM_VK_PLUS const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PIPE: u32

The KeyEvent.DOM_VK_PIPE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_HYPHEN_MINUS: u32

The KeyEvent.DOM_VK_HYPHEN_MINUS const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_OPEN_CURLY_BRACKET: u32

The KeyEvent.DOM_VK_OPEN_CURLY_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLOSE_CURLY_BRACKET: u32

The KeyEvent.DOM_VK_CLOSE_CURLY_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_TILDE: u32

The KeyEvent.DOM_VK_TILDE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_VOLUME_MUTE: u32

The KeyEvent.DOM_VK_VOLUME_MUTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_VOLUME_DOWN: u32

The KeyEvent.DOM_VK_VOLUME_DOWN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_VOLUME_UP: u32

The KeyEvent.DOM_VK_VOLUME_UP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_COMMA: u32

The KeyEvent.DOM_VK_COMMA const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PERIOD: u32

The KeyEvent.DOM_VK_PERIOD const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_SLASH: u32

The KeyEvent.DOM_VK_SLASH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_BACK_QUOTE: u32

The KeyEvent.DOM_VK_BACK_QUOTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_OPEN_BRACKET: u32

The KeyEvent.DOM_VK_OPEN_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_BACK_SLASH: u32

The KeyEvent.DOM_VK_BACK_SLASH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CLOSE_BRACKET: u32

The KeyEvent.DOM_VK_CLOSE_BRACKET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_QUOTE: u32

The KeyEvent.DOM_VK_QUOTE const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_META: u32

The KeyEvent.DOM_VK_META const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ALTGR: u32

The KeyEvent.DOM_VK_ALTGR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_ICO_HELP: u32

The KeyEvent.DOM_VK_WIN_ICO_HELP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_ICO_00: u32

The KeyEvent.DOM_VK_WIN_ICO_00 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PROCESSKEY: u32

The KeyEvent.DOM_VK_PROCESSKEY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_ICO_CLEAR: u32

The KeyEvent.DOM_VK_WIN_ICO_CLEAR const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_RESET: u32

The KeyEvent.DOM_VK_WIN_OEM_RESET const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_JUMP: u32

The KeyEvent.DOM_VK_WIN_OEM_JUMP const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_PA1: u32

The KeyEvent.DOM_VK_WIN_OEM_PA1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_PA2: u32

The KeyEvent.DOM_VK_WIN_OEM_PA2 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_PA3: u32

The KeyEvent.DOM_VK_WIN_OEM_PA3 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_WSCTRL: u32

The KeyEvent.DOM_VK_WIN_OEM_WSCTRL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_CUSEL: u32

The KeyEvent.DOM_VK_WIN_OEM_CUSEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_ATTN: u32

The KeyEvent.DOM_VK_WIN_OEM_ATTN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_FINISH: u32

The KeyEvent.DOM_VK_WIN_OEM_FINISH const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_COPY: u32

The KeyEvent.DOM_VK_WIN_OEM_COPY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_AUTO: u32

The KeyEvent.DOM_VK_WIN_OEM_AUTO const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_ENLW: u32

The KeyEvent.DOM_VK_WIN_OEM_ENLW const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_BACKTAB: u32

The KeyEvent.DOM_VK_WIN_OEM_BACKTAB const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ATTN: u32

The KeyEvent.DOM_VK_ATTN const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_CRSEL: u32

The KeyEvent.DOM_VK_CRSEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EXSEL: u32

The KeyEvent.DOM_VK_EXSEL const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_EREOF: u32

The KeyEvent.DOM_VK_EREOF const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PLAY: u32

The KeyEvent.DOM_VK_PLAY const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_ZOOM: u32

The KeyEvent.DOM_VK_ZOOM const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_PA1: u32

The KeyEvent.DOM_VK_PA1 const.

This API requires the following crate features to be activated: KeyEvent

pub const DOM_VK_WIN_OEM_CLEAR: u32

The KeyEvent.DOM_VK_WIN_OEM_CLEAR const.

This API requires the following crate features to be activated: KeyEvent