web_sys
Struct NetworkCommandOptions

pub struct NetworkCommandOptions { /* private fields */ }

The NetworkCommandOptions dictionary.

This API requires the following crate features to be activated: NetworkCommandOptions

Implementations

impl NetworkCommandOptions

pub fn get_cmd(&self) -> Option<String>

Get the cmd field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_cmd(&self, val: &str)

Change the cmd field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_cur_external_ifname(&self) -> Option<String>

Get the curExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_cur_external_ifname(&self, val: &str)

Change the curExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_cur_internal_ifname(&self) -> Option<String>

Get the curInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_cur_internal_ifname(&self, val: &str)

Change the curInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_dns1(&self) -> Option<String>

Get the dns1 field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_dns1(&self, val: &str)

Change the dns1 field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_dns1_long(&self) -> Option<i32>

Get the dns1_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_dns1_long(&self, val: i32)

Change the dns1_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_dns2(&self) -> Option<String>

Get the dns2 field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_dns2(&self, val: &str)

Change the dns2 field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_dns2_long(&self) -> Option<i32>

Get the dns2_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_dns2_long(&self, val: i32)

Change the dns2_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_dnses(&self) -> Option<Array>

Get the dnses field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_dnses(&self, val: &JsValue)

Change the dnses field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_domain(&self) -> Option<String>

Get the domain field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_domain(&self, val: &str)

Change the domain field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_enable(&self) -> Option<bool>

Get the enable field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_enable(&self, val: bool)

Change the enable field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_enabled(&self) -> Option<bool>

Get the enabled field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_enabled(&self, val: bool)

Change the enabled field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_end_ip(&self) -> Option<String>

Get the endIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_end_ip(&self, val: &str)

Change the endIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_external_ifname(&self) -> Option<String>

Get the externalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_external_ifname(&self, val: &str)

Change the externalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_gateway(&self) -> Option<String>

Get the gateway field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_gateway(&self, val: &str)

Change the gateway field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_gateway_long(&self) -> Option<i32>

Get the gateway_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_gateway_long(&self, val: i32)

Change the gateway_long field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_gateways(&self) -> Option<Array>

Get the gateways field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_gateways(&self, val: &JsValue)

Change the gateways field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_id(&self) -> Option<i32>

Get the id field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_id(&self, val: i32)

Change the id field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_ifname(&self) -> Option<String>

Get the ifname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_ifname(&self, val: &str)

Change the ifname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_interface_list(&self) -> Option<Array>

Get the interfaceList field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_interface_list(&self, val: &JsValue)

Change the interfaceList field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_internal_ifname(&self) -> Option<String>

Get the internalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_internal_ifname(&self, val: &str)

Change the internalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_ip(&self) -> Option<String>

Get the ip field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_ip(&self, val: &str)

Change the ip field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_ipaddr(&self) -> Option<i32>

Get the ipaddr field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_ipaddr(&self, val: i32)

Change the ipaddr field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_key(&self) -> Option<String>

Get the key field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_key(&self, val: &str)

Change the key field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_link(&self) -> Option<String>

Get the link field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_link(&self, val: &str)

Change the link field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_mask(&self) -> Option<i32>

Get the mask field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_mask(&self, val: i32)

Change the mask field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_mask_length(&self) -> Option<String>

Get the maskLength field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_mask_length(&self, val: &str)

Change the maskLength field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_mode(&self) -> Option<String>

Get the mode field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_mode(&self, val: &str)

Change the mode field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_mtu(&self) -> Option<i32>

Get the mtu field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_mtu(&self, val: i32)

Change the mtu field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_pre_external_ifname(&self) -> Option<String>

Get the preExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_pre_external_ifname(&self, val: &str)

Change the preExternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_pre_internal_ifname(&self) -> Option<String>

Get the preInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_pre_internal_ifname(&self, val: &str)

Change the preInternalIfname field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_prefix(&self) -> Option<String>

Get the prefix field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_prefix(&self, val: &str)

Change the prefix field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_prefix_length(&self) -> Option<u32>

Get the prefixLength field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_prefix_length(&self, val: u32)

Change the prefixLength field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_report(&self) -> Option<bool>

Get the report field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_report(&self, val: bool)

Change the report field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_security(&self) -> Option<String>

Get the security field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_security(&self, val: &str)

Change the security field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_server_ip(&self) -> Option<String>

Get the serverIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_server_ip(&self, val: &str)

Change the serverIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_ssid(&self) -> Option<String>

Get the ssid field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_ssid(&self, val: &str)

Change the ssid field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_start_ip(&self) -> Option<String>

Get the startIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_start_ip(&self, val: &str)

Change the startIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_threshold(&self) -> Option<f64>

Get the threshold field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_threshold(&self, val: f64)

Change the threshold field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_usb_end_ip(&self) -> Option<String>

Get the usbEndIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_usb_end_ip(&self, val: &str)

Change the usbEndIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_usb_start_ip(&self) -> Option<String>

Get the usbStartIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_usb_start_ip(&self, val: &str)

Change the usbStartIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_wifi_end_ip(&self) -> Option<String>

Get the wifiEndIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_wifi_end_ip(&self, val: &str)

Change the wifiEndIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_wifi_start_ip(&self) -> Option<String>

Get the wifiStartIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_wifi_start_ip(&self, val: &str)

Change the wifiStartIp field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn get_wifictrlinterfacename(&self) -> Option<String>

Get the wifictrlinterfacename field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn set_wifictrlinterfacename(&self, val: &str)

Change the wifictrlinterfacename field of this object.

This API requires the following crate features to be activated: NetworkCommandOptions

impl NetworkCommandOptions

pub fn new() -> Self

Construct a new NetworkCommandOptions.

This API requires the following crate features to be activated: NetworkCommandOptions

pub fn cmd(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_cmd() instead.

pub fn cur_external_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_cur_external_ifname() instead.

pub fn cur_internal_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_cur_internal_ifname() instead.

pub fn dns1(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_dns1() instead.

pub fn dns1_long(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_dns1_long() instead.

pub fn dns2(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_dns2() instead.

pub fn dns2_long(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_dns2_long() instead.

pub fn dnses(&mut self, val: &JsValue) -> &mut Self

👎Deprecated: Use set_dnses() instead.

pub fn domain(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_domain() instead.

pub fn enable(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_enable() instead.

pub fn enabled(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_enabled() instead.

pub fn end_ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_end_ip() instead.

pub fn external_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_external_ifname() instead.

pub fn gateway(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_gateway() instead.

pub fn gateway_long(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_gateway_long() instead.

pub fn gateways(&mut self, val: &JsValue) -> &mut Self

👎Deprecated: Use set_gateways() instead.

pub fn id(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_id() instead.

pub fn ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_ifname() instead.

pub fn interface_list(&mut self, val: &JsValue) -> &mut Self

👎Deprecated: Use set_interface_list() instead.

pub fn internal_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_internal_ifname() instead.

pub fn ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_ip() instead.

pub fn ipaddr(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_ipaddr() instead.

pub fn key(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_key() instead.

pub fn link(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_link() instead.

pub fn mask(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_mask() instead.

pub fn mask_length(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_mask_length() instead.

pub fn mode(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_mode() instead.

pub fn mtu(&mut self, val: i32) -> &mut Self

👎Deprecated: Use set_mtu() instead.

pub fn pre_external_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_pre_external_ifname() instead.

pub fn pre_internal_ifname(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_pre_internal_ifname() instead.

pub fn prefix(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_prefix() instead.

pub fn prefix_length(&mut self, val: u32) -> &mut Self

👎Deprecated: Use set_prefix_length() instead.

pub fn report(&mut self, val: bool) -> &mut Self

👎Deprecated: Use set_report() instead.

pub fn security(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_security() instead.

pub fn server_ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_server_ip() instead.

pub fn ssid(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_ssid() instead.

pub fn start_ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_start_ip() instead.

pub fn threshold(&mut self, val: f64) -> &mut Self

👎Deprecated: Use set_threshold() instead.

pub fn usb_end_ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_usb_end_ip() instead.

pub fn usb_start_ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_usb_start_ip() instead.

pub fn wifi_end_ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_wifi_end_ip() instead.

pub fn wifi_start_ip(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_wifi_start_ip() instead.

pub fn wifictrlinterfacename(&mut self, val: &str) -> &mut Self

👎Deprecated: Use set_wifictrlinterfacename() instead.