Struct web_sys::Notification

#[repr(transparent)]

pub struct Notification { /* fields omitted */ }

The Notification object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl Notification

pub fn new(title: &str) -> Result<Notification, JsValue>

This function requires the following crate features to be activated: Notification

impl Notification

pub fn new_with_options(
title: &str,
options: &NotificationOptions
) -> Result<Notification, JsValue>

This function requires the following crate features to be activated: Notification, NotificationOptions

impl Notification

pub fn close(&self)

The close() method

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn get() -> Result<Promise, JsValue>

The get() method

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn get_with_filter(
filter: &GetNotificationOptions
) -> Result<Promise, JsValue>

The get() method

MDN Documentation

This function requires the following crate features to be activated: GetNotificationOptions, Notification

impl Notification

pub fn request_permission() -> Result<Promise, JsValue>

The requestPermission() method

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn request_permission_with_permission_callback(
permission_callback: &Function
) -> Result<Promise, JsValue>

The requestPermission() method

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn permission() -> NotificationPermission

The permission getter

MDN Documentation

This function requires the following crate features to be activated: Notification, NotificationPermission

impl Notification

pub fn onclick(&self) -> Option<Function>

The onclick getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn set_onclick(&self, onclick: Option<&Function>)

The onclick setter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn onshow(&self) -> Option<Function>

The onshow getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn set_onshow(&self, onshow: Option<&Function>)

The onshow setter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn onclose(&self) -> Option<Function>

The onclose getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn set_onclose(&self, onclose: Option<&Function>)

The onclose setter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn title(&self) -> String

The title getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn dir(&self) -> NotificationDirection

The dir getter

MDN Documentation

This function requires the following crate features to be activated: Notification, NotificationDirection

impl Notification

pub fn lang(&self) -> Option<String>

The lang getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn body(&self) -> Option<String>

The body getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn tag(&self) -> Option<String>

The tag getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn icon(&self) -> Option<String>

The icon getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn require_interaction(&self) -> bool

The requireInteraction getter

MDN Documentation

This function requires the following crate features to be activated: Notification

impl Notification

pub fn data(&self) -> JsValue

The data getter

MDN Documentation

This function requires the following crate features to be activated: Notification