Struct web_sys::Range

#[repr(transparent)]

pub struct Range { /* fields omitted */ }

The Range object

MDN Documentation

Methods

impl Range

pub fn new() -> Result<Range, JsValue>

This function requires the following crate features to be activated: Range

impl Range

pub fn clone_contents(&self) -> Result<DocumentFragment, JsValue>

The cloneContents() method

MDN Documentation

This function requires the following crate features to be activated: DocumentFragment, Range

impl Range

pub fn clone_range(&self) -> Range

The cloneRange() method

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn collapse(&self)

The collapse() method

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn collapse_with_to_start(&self, to_start: bool)

The collapse() method

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn compare_boundary_points(
&self,
how: u16,
source_range: &Range
) -> Result<i16, JsValue>

The compareBoundaryPoints() method

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn compare_point(&self, node: &Node, offset: u32) -> Result<i16, JsValue>

The comparePoint() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn create_contextual_fragment(
&self,
fragment: &str
) -> Result<DocumentFragment, JsValue>

The createContextualFragment() method

MDN Documentation

This function requires the following crate features to be activated: DocumentFragment, Range

impl Range

pub fn delete_contents(&self) -> Result<(), JsValue>

The deleteContents() method

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn detach(&self)

The detach() method

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn extract_contents(&self) -> Result<DocumentFragment, JsValue>

The extractContents() method

MDN Documentation

This function requires the following crate features to be activated: DocumentFragment, Range

impl Range

pub fn get_bounding_client_rect(&self) -> DomRect

The getBoundingClientRect() method

MDN Documentation

This function requires the following crate features to be activated: DomRect, Range

impl Range

pub fn get_client_rects(&self) -> Option<DomRectList>

The getClientRects() method

MDN Documentation

This function requires the following crate features to be activated: DomRectList, Range

impl Range

pub fn insert_node(&self, node: &Node) -> Result<(), JsValue>

The insertNode() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn intersects_node(&self, node: &Node) -> Result<bool, JsValue>

The intersectsNode() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn is_point_in_range(
&self,
node: &Node,
offset: u32
) -> Result<bool, JsValue>

The isPointInRange() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn select_node(&self, ref_node: &Node) -> Result<(), JsValue>

The selectNode() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn select_node_contents(&self, ref_node: &Node) -> Result<(), JsValue>

The selectNodeContents() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn set_end(&self, ref_node: &Node, offset: u32) -> Result<(), JsValue>

The setEnd() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn set_end_after(&self, ref_node: &Node) -> Result<(), JsValue>

The setEndAfter() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn set_end_before(&self, ref_node: &Node) -> Result<(), JsValue>

The setEndBefore() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn set_start(&self, ref_node: &Node, offset: u32) -> Result<(), JsValue>

The setStart() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn set_start_after(&self, ref_node: &Node) -> Result<(), JsValue>

The setStartAfter() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn set_start_before(&self, ref_node: &Node) -> Result<(), JsValue>

The setStartBefore() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn surround_contents(&self, new_parent: &Node) -> Result<(), JsValue>

The surroundContents() method

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn start_container(&self) -> Result<Node, JsValue>

The startContainer getter

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn start_offset(&self) -> Result<u32, JsValue>

The startOffset getter

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn end_container(&self) -> Result<Node, JsValue>

The endContainer getter

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub fn end_offset(&self) -> Result<u32, JsValue>

The endOffset getter

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn collapsed(&self) -> bool

The collapsed getter

MDN Documentation

This function requires the following crate features to be activated: Range

impl Range

pub fn common_ancestor_container(&self) -> Result<Node, JsValue>

The commonAncestorContainer getter

MDN Documentation

This function requires the following crate features to be activated: Node, Range

impl Range

pub const START_TO_START: u16

impl Range

pub const START_TO_END: u16

impl Range

pub const END_TO_END: u16

impl Range

pub const END_TO_START: u16