web_sys
Struct RtcPeerConnection

pub struct RtcPeerConnection { /* private fields */ }

The RtcPeerConnection class.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

Implementations

impl RtcPeerConnection

pub fn local_description(&self) -> Option<RtcSessionDescription>

Getter for the localDescription field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescription

impl RtcPeerConnection

pub fn current_local_description(&self) -> Option<RtcSessionDescription>

Getter for the currentLocalDescription field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescription

impl RtcPeerConnection

pub fn pending_local_description(&self) -> Option<RtcSessionDescription>

Getter for the pendingLocalDescription field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescription

impl RtcPeerConnection

pub fn remote_description(&self) -> Option<RtcSessionDescription>

Getter for the remoteDescription field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescription

impl RtcPeerConnection

pub fn current_remote_description(&self) -> Option<RtcSessionDescription>

Getter for the currentRemoteDescription field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescription

impl RtcPeerConnection

pub fn pending_remote_description(&self) -> Option<RtcSessionDescription>

Getter for the pendingRemoteDescription field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescription

impl RtcPeerConnection

pub fn signaling_state(&self) -> RtcSignalingState

Getter for the signalingState field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSignalingState

impl RtcPeerConnection

pub fn can_trickle_ice_candidates(&self) -> Option<bool>

Getter for the canTrickleIceCandidates field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn ice_gathering_state(&self) -> RtcIceGatheringState

Getter for the iceGatheringState field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcIceGatheringState, RtcPeerConnection

impl RtcPeerConnection

pub fn ice_connection_state(&self) -> RtcIceConnectionState

Getter for the iceConnectionState field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcIceConnectionState, RtcPeerConnection

impl RtcPeerConnection

pub fn connection_state(&self) -> RtcPeerConnectionState

Getter for the connectionState field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcPeerConnectionState

impl RtcPeerConnection

pub fn peer_identity(&self) -> Promise

Getter for the peerIdentity field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn idp_login_url(&self) -> Option<String>

Getter for the idpLoginUrl field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onnegotiationneeded(&self) -> Option<Function>

Getter for the onnegotiationneeded field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onnegotiationneeded(&self, value: Option<&Function>)

Setter for the onnegotiationneeded field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onicecandidate(&self) -> Option<Function>

Getter for the onicecandidate field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onicecandidate(&self, value: Option<&Function>)

Setter for the onicecandidate field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onsignalingstatechange(&self) -> Option<Function>

Getter for the onsignalingstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onsignalingstatechange(&self, value: Option<&Function>)

Setter for the onsignalingstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onaddstream(&self) -> Option<Function>

Getter for the onaddstream field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onaddstream(&self, value: Option<&Function>)

Setter for the onaddstream field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onaddtrack(&self) -> Option<Function>

Getter for the onaddtrack field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onaddtrack(&self, value: Option<&Function>)

Setter for the onaddtrack field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn ontrack(&self) -> Option<Function>

Getter for the ontrack field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_ontrack(&self, value: Option<&Function>)

Setter for the ontrack field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onremovestream(&self) -> Option<Function>

Getter for the onremovestream field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onremovestream(&self, value: Option<&Function>)

Setter for the onremovestream field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn oniceconnectionstatechange(&self) -> Option<Function>

Getter for the oniceconnectionstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_oniceconnectionstatechange(&self, value: Option<&Function>)

Setter for the oniceconnectionstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onicegatheringstatechange(&self) -> Option<Function>

Getter for the onicegatheringstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onicegatheringstatechange(&self, value: Option<&Function>)

Setter for the onicegatheringstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn onconnectionstatechange(&self) -> Option<Function>

Getter for the onconnectionstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_onconnectionstatechange(&self, value: Option<&Function>)

Setter for the onconnectionstatechange field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn ondatachannel(&self) -> Option<Function>

Getter for the ondatachannel field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_ondatachannel(&self, value: Option<&Function>)

Setter for the ondatachannel field of this object.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn new() -> Result<RtcPeerConnection, JsValue>

The new RtcPeerConnection(..) constructor, creating a new instance of RtcPeerConnection.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn new_with_configuration( configuration: &RtcConfiguration, ) -> Result<RtcPeerConnection, JsValue>

The new RtcPeerConnection(..) constructor, creating a new instance of RtcPeerConnection.

MDN Documentation

This API requires the following crate features to be activated: RtcConfiguration, RtcPeerConnection

impl RtcPeerConnection

pub fn new_with_configuration_and_constraints( configuration: &RtcConfiguration, constraints: Option<&Object>, ) -> Result<RtcPeerConnection, JsValue>

The new RtcPeerConnection(..) constructor, creating a new instance of RtcPeerConnection.

MDN Documentation

This API requires the following crate features to be activated: RtcConfiguration, RtcPeerConnection

impl RtcPeerConnection

pub fn add_ice_candidate_with_opt_rtc_ice_candidate_init( &self, candidate: Option<&RtcIceCandidateInit>, ) -> Promise

The addIceCandidate() method.

MDN Documentation

This API requires the following crate features to be activated: RtcIceCandidateInit, RtcPeerConnection

impl RtcPeerConnection

pub fn add_ice_candidate_with_opt_rtc_ice_candidate( &self, candidate: Option<&RtcIceCandidate>, ) -> Promise

The addIceCandidate() method.

MDN Documentation

This API requires the following crate features to be activated: RtcIceCandidate, RtcPeerConnection

impl RtcPeerConnection

pub fn add_ice_candidate_with_rtc_ice_candidate_and_success_callback_and_failure_callback( &self, candidate: &RtcIceCandidate, success_callback: &Function, failure_callback: &Function, ) -> Promise

The addIceCandidate() method.

MDN Documentation

This API requires the following crate features to be activated: RtcIceCandidate, RtcPeerConnection

impl RtcPeerConnection

pub fn add_stream(&self, stream: &MediaStream)

The addStream() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, RtcPeerConnection

impl RtcPeerConnection

pub fn add_track( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams: &Array, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_0( &self, track: &MediaStreamTrack, stream: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_1( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams_1: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_2( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams_1: &MediaStream, more_streams_2: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_3( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams_1: &MediaStream, more_streams_2: &MediaStream, more_streams_3: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_4( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams_1: &MediaStream, more_streams_2: &MediaStream, more_streams_3: &MediaStream, more_streams_4: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_5( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams_1: &MediaStream, more_streams_2: &MediaStream, more_streams_3: &MediaStream, more_streams_4: &MediaStream, more_streams_5: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_6( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams_1: &MediaStream, more_streams_2: &MediaStream, more_streams_3: &MediaStream, more_streams_4: &MediaStream, more_streams_5: &MediaStream, more_streams_6: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_track_7( &self, track: &MediaStreamTrack, stream: &MediaStream, more_streams_1: &MediaStream, more_streams_2: &MediaStream, more_streams_3: &MediaStream, more_streams_4: &MediaStream, more_streams_5: &MediaStream, more_streams_6: &MediaStream, more_streams_7: &MediaStream, ) -> RtcRtpSender

The addTrack() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStream, MediaStreamTrack, RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn add_transceiver_with_media_stream_track( &self, track_or_kind: &MediaStreamTrack, ) -> RtcRtpTransceiver

The addTransceiver() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStreamTrack, RtcPeerConnection, RtcRtpTransceiver

impl RtcPeerConnection

pub fn add_transceiver_with_str(&self, track_or_kind: &str) -> RtcRtpTransceiver

The addTransceiver() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcRtpTransceiver

impl RtcPeerConnection

pub fn add_transceiver_with_media_stream_track_and_init( &self, track_or_kind: &MediaStreamTrack, init: &RtcRtpTransceiverInit, ) -> RtcRtpTransceiver

The addTransceiver() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStreamTrack, RtcPeerConnection, RtcRtpTransceiver, RtcRtpTransceiverInit

impl RtcPeerConnection

pub fn add_transceiver_with_str_and_init( &self, track_or_kind: &str, init: &RtcRtpTransceiverInit, ) -> RtcRtpTransceiver

The addTransceiver() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcRtpTransceiver, RtcRtpTransceiverInit

impl RtcPeerConnection

pub fn close(&self)

The close() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn create_answer(&self) -> Promise

The createAnswer() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn create_answer_with_rtc_answer_options( &self, options: &RtcAnswerOptions, ) -> Promise

The createAnswer() method.

MDN Documentation

This API requires the following crate features to be activated: RtcAnswerOptions, RtcPeerConnection

impl RtcPeerConnection

pub fn create_answer_with_success_callback_and_failure_callback( &self, success_callback: &Function, failure_callback: &Function, ) -> Promise

The createAnswer() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn create_data_channel(&self, label: &str) -> RtcDataChannel

The createDataChannel() method.

MDN Documentation

This API requires the following crate features to be activated: RtcDataChannel, RtcPeerConnection

impl RtcPeerConnection

pub fn create_data_channel_with_data_channel_dict( &self, label: &str, data_channel_dict: &RtcDataChannelInit, ) -> RtcDataChannel

The createDataChannel() method.

MDN Documentation

This API requires the following crate features to be activated: RtcDataChannel, RtcDataChannelInit, RtcPeerConnection

impl RtcPeerConnection

pub fn create_offer(&self) -> Promise

The createOffer() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn create_offer_with_rtc_offer_options( &self, options: &RtcOfferOptions, ) -> Promise

The createOffer() method.

MDN Documentation

This API requires the following crate features to be activated: RtcOfferOptions, RtcPeerConnection

impl RtcPeerConnection

pub fn create_offer_with_callback_and_failure_callback( &self, success_callback: &Function, failure_callback: &Function, ) -> Promise

The createOffer() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn create_offer_with_callback_and_failure_callback_and_options( &self, success_callback: &Function, failure_callback: &Function, options: &RtcOfferOptions, ) -> Promise

The createOffer() method.

MDN Documentation

This API requires the following crate features to be activated: RtcOfferOptions, RtcPeerConnection

impl RtcPeerConnection

pub fn generate_certificate_with_object( keygen_algorithm: &Object, ) -> Result<Promise, JsValue>

The generateCertificate() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn generate_certificate_with_str( keygen_algorithm: &str, ) -> Result<Promise, JsValue>

The generateCertificate() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn get_configuration(&self) -> RtcConfiguration

The getConfiguration() method.

MDN Documentation

This API requires the following crate features to be activated: RtcConfiguration, RtcPeerConnection

impl RtcPeerConnection

pub fn get_identity_assertion(&self) -> Promise

The getIdentityAssertion() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn get_local_streams(&self) -> Array

The getLocalStreams() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn get_receivers(&self) -> Array

The getReceivers() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn get_remote_streams(&self) -> Array

The getRemoteStreams() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn get_senders(&self) -> Array

The getSenders() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn get_stats(&self) -> Promise

The getStats() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn get_stats_with_selector( &self, selector: Option<&MediaStreamTrack>, ) -> Promise

The getStats() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStreamTrack, RtcPeerConnection

impl RtcPeerConnection

pub fn get_stats_with_selector_and_success_callback_and_failure_callback( &self, selector: Option<&MediaStreamTrack>, success_callback: &Function, failure_callback: &Function, ) -> Promise

The getStats() method.

MDN Documentation

This API requires the following crate features to be activated: MediaStreamTrack, RtcPeerConnection

impl RtcPeerConnection

pub fn get_transceivers(&self) -> Array

The getTransceivers() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn remove_track(&self, sender: &RtcRtpSender)

The removeTrack() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcRtpSender

impl RtcPeerConnection

pub fn set_configuration(&self) -> Result<(), JsValue>

The setConfiguration() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_configuration_with_configuration( &self, configuration: &RtcConfiguration, ) -> Result<(), JsValue>

The setConfiguration() method.

MDN Documentation

This API requires the following crate features to be activated: RtcConfiguration, RtcPeerConnection

impl RtcPeerConnection

pub fn set_identity_provider(&self, provider: &str)

The setIdentityProvider() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection

impl RtcPeerConnection

pub fn set_identity_provider_with_options( &self, provider: &str, options: &RtcIdentityProviderOptions, )

The setIdentityProvider() method.

MDN Documentation

This API requires the following crate features to be activated: RtcIdentityProviderOptions, RtcPeerConnection

impl RtcPeerConnection

pub fn set_local_description( &self, description: &RtcSessionDescriptionInit, ) -> Promise

The setLocalDescription() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescriptionInit

impl RtcPeerConnection

pub fn set_local_description_with_success_callback_and_failure_callback( &self, description: &RtcSessionDescriptionInit, success_callback: &Function, failure_callback: &Function, ) -> Promise

The setLocalDescription() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescriptionInit

impl RtcPeerConnection

pub fn set_remote_description( &self, description: &RtcSessionDescriptionInit, ) -> Promise

The setRemoteDescription() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescriptionInit

impl RtcPeerConnection

pub fn set_remote_description_with_success_callback_and_failure_callback( &self, description: &RtcSessionDescriptionInit, success_callback: &Function, failure_callback: &Function, ) -> Promise

The setRemoteDescription() method.

MDN Documentation

This API requires the following crate features to be activated: RtcPeerConnection, RtcSessionDescriptionInit