Struct web_sys::SharedWorkerGlobalScope

#[repr(transparent)]pub struct SharedWorkerGlobalScope { /* fields omitted */ }

The SharedWorkerGlobalScope class.

MDN Documentation

This API requires the following crate features to be activated: SharedWorkerGlobalScope

Implementations

impl SharedWorkerGlobalScope

pub fn name(&self) -> String

Getter for the name field of this object.

MDN Documentation

This API requires the following crate features to be activated: SharedWorkerGlobalScope

impl SharedWorkerGlobalScope

pub fn onconnect(&self) -> Option<Function>

Getter for the onconnect field of this object.

MDN Documentation

This API requires the following crate features to be activated: SharedWorkerGlobalScope

impl SharedWorkerGlobalScope

pub fn set_onconnect(&self, value: Option<&Function>)

Setter for the onconnect field of this object.

MDN Documentation

This API requires the following crate features to be activated: SharedWorkerGlobalScope

impl SharedWorkerGlobalScope

pub fn close(&self)

The close() method.

MDN Documentation

This API requires the following crate features to be activated: SharedWorkerGlobalScope