Struct web_sys::SvgsvgElement

#[repr(transparent)]pub struct SvgsvgElement { /* fields omitted */ }

The SvgsvgElement class.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

Implementations

impl SvgsvgElement

pub fn x(&self) -> SvgAnimatedLength

Getter for the x field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn y(&self) -> SvgAnimatedLength

Getter for the y field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn width(&self) -> SvgAnimatedLength

Getter for the width field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn height(&self) -> SvgAnimatedLength

Getter for the height field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedLength, SvgsvgElement

impl SvgsvgElement

pub fn use_current_view(&self) -> bool

Getter for the useCurrentView field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn current_scale(&self) -> f32

Getter for the currentScale field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn set_current_scale(&self, value: f32)

Setter for the currentScale field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn current_translate(&self) -> SvgPoint

Getter for the currentTranslate field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgPoint, SvgsvgElement

impl SvgsvgElement

pub fn view_box(&self) -> SvgAnimatedRect

Getter for the viewBox field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedRect, SvgsvgElement

impl SvgsvgElement

pub fn preserve_aspect_ratio(&self) -> SvgAnimatedPreserveAspectRatio

Getter for the preserveAspectRatio field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgAnimatedPreserveAspectRatio, SvgsvgElement

impl SvgsvgElement

pub fn zoom_and_pan(&self) -> u16

Getter for the zoomAndPan field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn set_zoom_and_pan(&self, value: u16)

Setter for the zoomAndPan field of this object.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn animations_paused(&self) -> bool

The animationsPaused() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn create_svg_angle(&self) -> SvgAngle

The createSVGAngle() method.

MDN Documentation

This API requires the following crate features to be activated: SvgAngle, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_length(&self) -> SvgLength

The createSVGLength() method.

MDN Documentation

This API requires the following crate features to be activated: SvgLength, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_matrix(&self) -> SvgMatrix

The createSVGMatrix() method.

MDN Documentation

This API requires the following crate features to be activated: SvgMatrix, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_number(&self) -> SvgNumber

The createSVGNumber() method.

MDN Documentation

This API requires the following crate features to be activated: SvgNumber, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_point(&self) -> SvgPoint

The createSVGPoint() method.

MDN Documentation

This API requires the following crate features to be activated: SvgPoint, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_rect(&self) -> SvgRect

The createSVGRect() method.

MDN Documentation

This API requires the following crate features to be activated: SvgRect, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_transform(&self) -> SvgTransform

The createSVGTransform() method.

MDN Documentation

This API requires the following crate features to be activated: SvgTransform, SvgsvgElement

impl SvgsvgElement

pub fn create_svg_transform_from_matrix(
&self,
matrix: &SvgMatrix
) -> SvgTransform

The createSVGTransformFromMatrix() method.

MDN Documentation

This API requires the following crate features to be activated: SvgMatrix, SvgTransform, SvgsvgElement

impl SvgsvgElement

pub fn deselect_all(&self)

The deselectAll() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn force_redraw(&self)

The forceRedraw() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn get_current_time(&self) -> f32

The getCurrentTime() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn get_element_by_id(&self, element_id: &str) -> Option<Element>

The getElementById() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn pause_animations(&self)

The pauseAnimations() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn set_current_time(&self, seconds: f32)

The setCurrentTime() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn suspend_redraw(&self, max_wait_milliseconds: u32) -> u32

The suspendRedraw() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn unpause_animations(&self)

The unpauseAnimations() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn unsuspend_redraw(&self, suspend_handle_id: u32)

The unsuspendRedraw() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub fn unsuspend_redraw_all(&self)

The unsuspendRedrawAll() method.

MDN Documentation

This API requires the following crate features to be activated: SvgsvgElement

impl SvgsvgElement

pub const SVG_ZOOMANDPAN_UNKNOWN: u16

The SVGSVGElement.SVG_ZOOMANDPAN_UNKNOWN const.

This API requires the following crate features to be activated: SvgsvgElement

pub const SVG_ZOOMANDPAN_DISABLE: u16

The SVGSVGElement.SVG_ZOOMANDPAN_DISABLE const.

This API requires the following crate features to be activated: SvgsvgElement

pub const SVG_ZOOMANDPAN_MAGNIFY: u16

The SVGSVGElement.SVG_ZOOMANDPAN_MAGNIFY const.

This API requires the following crate features to be activated: SvgsvgElement