Struct web_sys::UiEvent

#[repr(transparent)]

pub struct UiEvent { /* fields omitted */ }

The UIEvent object

MDN Documentation

This function requires the following crate features to be activated: Event

Methods

impl UiEvent

pub fn new(type_: &str) -> Result<UiEvent, JsValue>

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn new_with_event_init_dict(
type_: &str,
event_init_dict: &UiEventInit
) -> Result<UiEvent, JsValue>

This function requires the following crate features to be activated: UiEvent, UiEventInit

impl UiEvent

pub fn init_ui_event(&self, a_type: &str)

The initUIEvent() method

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn init_ui_event_with_a_can_bubble(&self, a_type: &str, a_can_bubble: bool)

The initUIEvent() method

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn init_ui_event_with_a_can_bubble_and_a_cancelable(
&self,
a_type: &str,
a_can_bubble: bool,
a_cancelable: bool
)

The initUIEvent() method

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn init_ui_event_with_a_can_bubble_and_a_cancelable_and_a_view(
&self,
a_type: &str,
a_can_bubble: bool,
a_cancelable: bool,
a_view: Option<&Window>
)

The initUIEvent() method

MDN Documentation

This function requires the following crate features to be activated: UiEvent, Window

impl UiEvent

pub fn init_ui_event_with_a_can_bubble_and_a_cancelable_and_a_view_and_a_detail(
&self,
a_type: &str,
a_can_bubble: bool,
a_cancelable: bool,
a_view: Option<&Window>,
a_detail: i32
)

The initUIEvent() method

MDN Documentation

This function requires the following crate features to be activated: UiEvent, Window

impl UiEvent

pub fn detail(&self) -> i32

The detail getter

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn layer_x(&self) -> i32

The layerX getter

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn layer_y(&self) -> i32

The layerY getter

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn page_x(&self) -> i32

The pageX getter

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn page_y(&self) -> i32

The pageY getter

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn which(&self) -> u32

The which getter

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub fn range_parent(&self) -> Option<Node>

The rangeParent getter

MDN Documentation

This function requires the following crate features to be activated: Node, UiEvent

impl UiEvent

pub fn range_offset(&self) -> i32

The rangeOffset getter

MDN Documentation

This function requires the following crate features to be activated: UiEvent

impl UiEvent

pub const SCROLL_PAGE_UP: i32

impl UiEvent

pub const SCROLL_PAGE_DOWN: i32