Struct web_sys::Url

#[repr(transparent)]

pub struct Url { /* fields omitted */ }

The URL object

MDN Documentation

This API requires the following crate features to be activated: Url

Methods

impl Url

pub fn new(url: &str) -> Result<Url, JsValue>

The new URL(..) constructor, creating a new instance of URL

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn new_with_base(url: &str, base: &str) -> Result<Url, JsValue>

The new URL(..) constructor, creating a new instance of URL

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn create_object_url_with_blob(blob: &Blob) -> Result<String, JsValue>

The createObjectURL() method

MDN Documentation

This API requires the following crate features to be activated: Blob, Url

impl Url

pub fn create_object_url_with_source(
source: &MediaSource
) -> Result<String, JsValue>

The createObjectURL() method

MDN Documentation

This API requires the following crate features to be activated: MediaSource, Url

impl Url

pub fn revoke_object_url(url: &str) -> Result<(), JsValue>

The revokeObjectURL() method

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn to_json(&self) -> String

The toJSON() method

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn href(&self) -> String

The href getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_href(&self, href: &str)

The href setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn origin(&self) -> String

The origin getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn protocol(&self) -> String

The protocol getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_protocol(&self, protocol: &str)

The protocol setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn username(&self) -> String

The username getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_username(&self, username: &str)

The username setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn password(&self) -> String

The password getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_password(&self, password: &str)

The password setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn host(&self) -> String

The host getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_host(&self, host: &str)

The host setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn hostname(&self) -> String

The hostname getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_hostname(&self, hostname: &str)

The hostname setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn port(&self) -> String

The port getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_port(&self, port: &str)

The port setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn pathname(&self) -> String

The pathname getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_pathname(&self, pathname: &str)

The pathname setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn search(&self) -> String

The search getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_search(&self, search: &str)

The search setter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn search_params(&self) -> UrlSearchParams

The searchParams getter

MDN Documentation

This API requires the following crate features to be activated: Url, UrlSearchParams

impl Url

pub fn hash(&self) -> String

The hash getter

MDN Documentation

This API requires the following crate features to be activated: Url

impl Url

pub fn set_hash(&self, hash: &str)

The hash setter

MDN Documentation

This API requires the following crate features to be activated: Url