Struct web_sys::WebGpuCommandEncoder

#[repr(transparent)]
pub struct WebGpuCommandEncoder { /* fields omitted */ }

The WebGPUCommandEncoder object

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

Methods

impl WebGpuCommandEncoder

pub fn begin_compute_pass(&self)

The beginComputePass() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn begin_render_pass(&self)

The beginRenderPass() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn begin_render_pass_with_descriptor(
&self,
descriptor: &WebGpuRenderPassDescriptor
)

The beginRenderPass() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder, WebGpuRenderPassDescriptor

impl WebGpuCommandEncoder

pub fn blit(&self)

The blit() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn copy_buffer_to_buffer(
&self,
src: &WebGpuBuffer,
src_offset: u32,
dst: &WebGpuBuffer,
dst_offset: u32,
size: u32
)

The copyBufferToBuffer() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuBuffer, WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn copy_buffer_to_texture(&self)

The copyBufferToTexture() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn copy_texture_to_buffer(&self)

The copyTextureToBuffer() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn copy_texture_to_texture(&self)

The copyTextureToTexture() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn dispatch(&self, x: u32, y: u32, z: u32)

The dispatch() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn draw(
&self,
vertex_count: u32,
instance_count: u32,
first_vertex: u32,
first_instance: u32
)

The draw() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn draw_indexed(
&self,
index_count: u32,
instance_count: u32,
first_index: u32,
first_instance: u32,
first_vertex: u32
)

The drawIndexed() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn end_compute_pass(&self)

The endComputePass() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn end_render_pass(&self)

The endRenderPass() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn finish_encoding(&self) -> WebGpuCommandBuffer

The finishEncoding() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandBuffer, WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn set_bind_group(&self, index: u32, bind_group: &WebGpuBindGroup)

The setBindGroup() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuBindGroup, WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn set_blend_color(&self, r: f32, g: f32, b: f32, a: f32)

The setBlendColor() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn set_index_buffer(&self, buffer: &WebGpuBuffer, offset: u32)

The setIndexBuffer() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuBuffer, WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn set_pipeline_with_web_gpu_compute_pipeline(
&self,
pipeline: &WebGpuComputePipeline
)

The setPipeline() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder, WebGpuComputePipeline

impl WebGpuCommandEncoder

pub fn set_pipeline_with_web_gpu_render_pipeline(
&self,
pipeline: &WebGpuRenderPipeline
)

The setPipeline() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder, WebGpuRenderPipeline

impl WebGpuCommandEncoder

pub fn set_push_constants(
&self,
stage: u32,
offset: u32,
count: u32,
data: &ArrayBuffer
)

The setPushConstants() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn set_vertex_buffers(
&self,
start_slot: u32,
buffers: &JsValue,
offsets: &JsValue
)

The setVertexBuffers() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuCommandEncoder

impl WebGpuCommandEncoder

pub fn transition_buffer(&self, b: &WebGpuBuffer, f: u32)

The transitionBuffer() method

MDN Documentation

This API requires the following crate features to be activated: WebGpuBuffer, WebGpuCommandEncoder