Struct web_sys::WebSocket

#[repr(transparent)]pub struct WebSocket { /* fields omitted */ }

The WebSocket class.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

Implementations

impl WebSocket

pub fn url(&self) -> String

Getter for the url field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn ready_state(&self) -> u16

Getter for the readyState field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn buffered_amount(&self) -> u32

Getter for the bufferedAmount field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onopen(&self) -> Option<Function>

Getter for the onopen field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onopen(&self, value: Option<&Function>)

Setter for the onopen field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onerror(&self) -> Option<Function>

Getter for the onerror field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onerror(&self, value: Option<&Function>)

Setter for the onerror field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onclose(&self) -> Option<Function>

Getter for the onclose field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onclose(&self, value: Option<&Function>)

Setter for the onclose field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn extensions(&self) -> String

Getter for the extensions field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn protocol(&self) -> String

Getter for the protocol field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn onmessage(&self) -> Option<Function>

Getter for the onmessage field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn set_onmessage(&self, value: Option<&Function>)

Setter for the onmessage field of this object.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn binary_type(&self) -> BinaryType

Getter for the binaryType field of this object.

MDN Documentation

This API requires the following crate features to be activated: BinaryType, WebSocket

impl WebSocket

pub fn set_binary_type(&self, value: BinaryType)

Setter for the binaryType field of this object.

MDN Documentation

This API requires the following crate features to be activated: BinaryType, WebSocket

impl WebSocket

pub fn new(url: &str) -> Result<WebSocket, JsValue>

The new WebSocket(..) constructor, creating a new instance of WebSocket.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn new_with_str(url: &str, protocols: &str) -> Result<WebSocket, JsValue>

The new WebSocket(..) constructor, creating a new instance of WebSocket.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn new_with_str_sequence(
url: &str,
protocols: &JsValue
) -> Result<WebSocket, JsValue>

The new WebSocket(..) constructor, creating a new instance of WebSocket.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn close(&self) -> Result<(), JsValue>

The close() method.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn close_with_code(&self, code: u16) -> Result<(), JsValue>

The close() method.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn close_with_code_and_reason(
&self,
code: u16,
reason: &str
) -> Result<(), JsValue>

The close() method.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_str(&self, data: &str) -> Result<(), JsValue>

The send() method.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_blob(&self, data: &Blob) -> Result<(), JsValue>

The send() method.

MDN Documentation

This API requires the following crate features to be activated: Blob, WebSocket

impl WebSocket

pub fn send_with_array_buffer(&self, data: &ArrayBuffer) -> Result<(), JsValue>

The send() method.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_array_buffer_view(&self, data: &Object) -> Result<(), JsValue>

The send() method.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub fn send_with_u8_array(&self, data: &[u8]) -> Result<(), JsValue>

The send() method.

MDN Documentation

This API requires the following crate features to be activated: WebSocket

impl WebSocket

pub const CONNECTING: u16

The WebSocket.CONNECTING const.

This API requires the following crate features to be activated: WebSocket

pub const OPEN: u16

The WebSocket.OPEN const.

This API requires the following crate features to be activated: WebSocket

pub const CLOSING: u16

The WebSocket.CLOSING const.

This API requires the following crate features to be activated: WebSocket

pub const CLOSED: u16

The WebSocket.CLOSED const.

This API requires the following crate features to be activated: WebSocket