Struct web_sys::WorkerGlobalScope

#[repr(transparent)]

pub struct WorkerGlobalScope { /* fields omitted */ }

The WorkerGlobalScope object

MDN Documentation

This function requires the following crate features to be activated: EventTarget

Methods

impl WorkerGlobalScope

pub fn import_scripts(&self, urls: &Array) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_0(&self) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_1(&self, urls_1: &str) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_2(
&self,
urls_1: &str,
urls_2: &str
) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_3(
&self,
urls_1: &str,
urls_2: &str,
urls_3: &str
) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_4(
&self,
urls_1: &str,
urls_2: &str,
urls_3: &str,
urls_4: &str
) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_5(
&self,
urls_1: &str,
urls_2: &str,
urls_3: &str,
urls_4: &str,
urls_5: &str
) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_6(
&self,
urls_1: &str,
urls_2: &str,
urls_3: &str,
urls_4: &str,
urls_5: &str,
urls_6: &str
) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn import_scripts_7(
&self,
urls_1: &str,
urls_2: &str,
urls_3: &str,
urls_4: &str,
urls_5: &str,
urls_6: &str,
urls_7: &str
) -> Result<(), JsValue>

The importScripts() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn self_(&self) -> WorkerGlobalScope

The self getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn location(&self) -> WorkerLocation

The location getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope, WorkerLocation

impl WorkerGlobalScope

pub fn navigator(&self) -> WorkerNavigator

The navigator getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope, WorkerNavigator

impl WorkerGlobalScope

pub fn onerror(&self) -> Option<Function>

The onerror getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_onerror(&self, onerror: Option<&Function>)

The onerror setter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn onoffline(&self) -> Option<Function>

The onoffline getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_onoffline(&self, onoffline: Option<&Function>)

The onoffline setter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn ononline(&self) -> Option<Function>

The ononline getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_ononline(&self, ononline: Option<&Function>)

The ononline setter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn crypto(&self) -> Result<Crypto, JsValue>

The crypto getter

MDN Documentation

This function requires the following crate features to be activated: Crypto, WorkerGlobalScope

impl WorkerGlobalScope

pub fn atob(&self, atob: &str) -> Result<String, JsValue>

The atob() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn btoa(&self, btoa: &str) -> Result<String, JsValue>

The btoa() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn clear_interval(&self)

The clearInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn clear_interval_with_handle(&self, handle: i32)

The clearInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn clear_timeout(&self)

The clearTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn clear_timeout_with_handle(&self, handle: i32)

The clearTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_html_image_element(
&self,
a_image: &HtmlImageElement
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: HtmlImageElement, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_html_video_element(
&self,
a_image: &HtmlVideoElement
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: HtmlVideoElement, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_html_canvas_element(
&self,
a_image: &HtmlCanvasElement
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: HtmlCanvasElement, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_blob(
&self,
a_image: &Blob
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: Blob, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_image_data(
&self,
a_image: &ImageData
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: ImageData, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_canvas_rendering_context_2d(
&self,
a_image: &CanvasRenderingContext2d
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: CanvasRenderingContext2d, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_image_bitmap(
&self,
a_image: &ImageBitmap
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: ImageBitmap, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_buffer_source(
&self,
a_image: &Object
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_u8_array(
&self,
a_image: &mut [u8]
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_html_image_element_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &HtmlImageElement,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: HtmlImageElement, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_html_video_element_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &HtmlVideoElement,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: HtmlVideoElement, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_html_canvas_element_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &HtmlCanvasElement,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: HtmlCanvasElement, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_blob_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &Blob,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: Blob, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_image_data_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &ImageData,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: ImageData, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_canvas_rendering_context_2d_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &CanvasRenderingContext2d,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: CanvasRenderingContext2d, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_image_bitmap_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &ImageBitmap,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: ImageBitmap, WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_buffer_source_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &Object,
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn create_image_bitmap_with_u8_array_and_a_sx_and_a_sy_and_a_sw_and_a_sh(
&self,
a_image: &mut [u8],
a_sx: i32,
a_sy: i32,
a_sw: i32,
a_sh: i32
) -> Result<Promise, JsValue>

The createImageBitmap() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn fetch_with_request(&self, input: &Request) -> Promise

The fetch() method

MDN Documentation

This function requires the following crate features to be activated: Request, WorkerGlobalScope

impl WorkerGlobalScope

pub fn fetch_with_str(&self, input: &str) -> Promise

The fetch() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn fetch_with_request_and_init(
&self,
input: &Request,
init: &RequestInit
) -> Promise

The fetch() method

MDN Documentation

This function requires the following crate features to be activated: Request, RequestInit, WorkerGlobalScope

impl WorkerGlobalScope

pub fn fetch_with_str_and_init(
&self,
input: &str,
init: &RequestInit
) -> Promise

The fetch() method

MDN Documentation

This function requires the following crate features to be activated: RequestInit, WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback(
&self,
handler: &Function
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments(
&self,
handler: &Function,
timeout: i32,
arguments: &Array
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_0(
&self,
handler: &Function,
timeout: i32
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_1(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_2(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_3(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_4(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_5(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue,
arguments_5: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_6(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue,
arguments_5: &JsValue,
arguments_6: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_callback_and_timeout_and_arguments_7(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue,
arguments_5: &JsValue,
arguments_6: &JsValue,
arguments_7: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str(&self, handler: &str) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused(
&self,
handler: &str,
timeout: i32,
unused: &Array
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_0(
&self,
handler: &str,
timeout: i32
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_1(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_2(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_3(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_4(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_5(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue,
unused_5: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_6(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue,
unused_5: &JsValue,
unused_6: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_interval_with_str_and_timeout_and_unused_7(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue,
unused_5: &JsValue,
unused_6: &JsValue,
unused_7: &JsValue
) -> Result<i32, JsValue>

The setInterval() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback(
&self,
handler: &Function
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments(
&self,
handler: &Function,
timeout: i32,
arguments: &Array
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_0(
&self,
handler: &Function,
timeout: i32
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_1(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_2(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_3(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_4(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_5(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue,
arguments_5: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_6(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue,
arguments_5: &JsValue,
arguments_6: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_callback_and_timeout_and_arguments_7(
&self,
handler: &Function,
timeout: i32,
arguments_1: &JsValue,
arguments_2: &JsValue,
arguments_3: &JsValue,
arguments_4: &JsValue,
arguments_5: &JsValue,
arguments_6: &JsValue,
arguments_7: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str(&self, handler: &str) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused(
&self,
handler: &str,
timeout: i32,
unused: &Array
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_0(
&self,
handler: &str,
timeout: i32
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_1(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_2(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_3(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_4(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_5(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue,
unused_5: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_6(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue,
unused_5: &JsValue,
unused_6: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn set_timeout_with_str_and_timeout_and_unused_7(
&self,
handler: &str,
timeout: i32,
unused_1: &JsValue,
unused_2: &JsValue,
unused_3: &JsValue,
unused_4: &JsValue,
unused_5: &JsValue,
unused_6: &JsValue,
unused_7: &JsValue
) -> Result<i32, JsValue>

The setTimeout() method

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn origin(&self) -> String

The origin getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn is_secure_context(&self) -> bool

The isSecureContext getter

MDN Documentation

This function requires the following crate features to be activated: WorkerGlobalScope

impl WorkerGlobalScope

pub fn indexed_db(&self) -> Result<Option<IdbFactory>, JsValue>

The indexedDB getter

MDN Documentation

This function requires the following crate features to be activated: IdbFactory, WorkerGlobalScope

impl WorkerGlobalScope

pub fn caches(&self) -> Result<CacheStorage, JsValue>

The caches getter

MDN Documentation

This function requires the following crate features to be activated: CacheStorage, WorkerGlobalScope