Struct web_sys::XmlHttpRequest

#[repr(transparent)]

pub struct XmlHttpRequest { /* fields omitted */ }

The XMLHttpRequest object

MDN Documentation

This function requires the following crate features to be activated: EventTarget, XmlHttpRequestEventTarget

Methods

impl XmlHttpRequest

pub fn new() -> Result<XmlHttpRequest, JsValue>

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn new_with_moz_xml_http_request_parameters(
params: &MozXmlHttpRequestParameters
) -> Result<XmlHttpRequest, JsValue>

This function requires the following crate features to be activated: MozXmlHttpRequestParameters, XmlHttpRequest

impl XmlHttpRequest

pub fn new_with_ignored(ignored: &str) -> Result<XmlHttpRequest, JsValue>

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn abort(&self) -> Result<(), JsValue>

The abort() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn get_all_response_headers(&self) -> Result<String, JsValue>

The getAllResponseHeaders() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn get_response_header(
&self,
header: &str
) -> Result<Option<String>, JsValue>

The getResponseHeader() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn open(&self, method: &str, url: &str) -> Result<(), JsValue>

The open() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn open_with_async(
&self,
method: &str,
url: &str,
async: bool
) -> Result<(), JsValue>

The open() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn open_with_async_and_user(
&self,
method: &str,
url: &str,
async: bool,
user: Option<&str>
) -> Result<(), JsValue>

The open() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn open_with_async_and_user_and_password(
&self,
method: &str,
url: &str,
async: bool,
user: Option<&str>,
password: Option<&str>
) -> Result<(), JsValue>

The open() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn override_mime_type(&self, mime: &str) -> Result<(), JsValue>

The overrideMimeType() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn send(&self) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn send_with_opt_document(
&self,
body: Option<&Document>
) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: Document, XmlHttpRequest

impl XmlHttpRequest

pub fn send_with_opt_blob(&self, body: Option<&Blob>) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: Blob, XmlHttpRequest

impl XmlHttpRequest

pub fn send_with_opt_buffer_source(
&self,
body: Option<&Object>
) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn send_with_opt_u8_array(
&self,
body: Option<&mut [u8]>
) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn send_with_opt_form_data(
&self,
body: Option<&FormData>
) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: FormData, XmlHttpRequest

impl XmlHttpRequest

pub fn send_with_opt_url_search_params(
&self,
body: Option<&UrlSearchParams>
) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: UrlSearchParams, XmlHttpRequest

impl XmlHttpRequest

pub fn send_with_opt_str(&self, body: Option<&str>) -> Result<(), JsValue>

The send() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn set_request_header(
&self,
header: &str,
value: &str
) -> Result<(), JsValue>

The setRequestHeader() method

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn onreadystatechange(&self) -> Option<Function>

The onreadystatechange getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn set_onreadystatechange(&self, onreadystatechange: Option<&Function>)

The onreadystatechange setter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn ready_state(&self) -> u16

The readyState getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn timeout(&self) -> u32

The timeout getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn set_timeout(&self, timeout: u32)

The timeout setter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn with_credentials(&self) -> bool

The withCredentials getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn set_with_credentials(&self, with_credentials: bool)

The withCredentials setter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn upload(&self) -> Result<XmlHttpRequestUpload, JsValue>

The upload getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest, XmlHttpRequestUpload

impl XmlHttpRequest

pub fn response_url(&self) -> String

The responseURL getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn status(&self) -> Result<u16, JsValue>

The status getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn status_text(&self) -> Result<String, JsValue>

The statusText getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn response_type(&self) -> XmlHttpRequestResponseType

The responseType getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest, XmlHttpRequestResponseType

impl XmlHttpRequest

pub fn set_response_type(&self, response_type: XmlHttpRequestResponseType)

The responseType setter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest, XmlHttpRequestResponseType

impl XmlHttpRequest

pub fn response(&self) -> Result<JsValue, JsValue>

The response getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn response_text(&self) -> Result<Option<String>, JsValue>

The responseText getter

MDN Documentation

This function requires the following crate features to be activated: XmlHttpRequest

impl XmlHttpRequest

pub fn response_xml(&self) -> Result<Option<Document>, JsValue>

The responseXML getter

MDN Documentation

This function requires the following crate features to be activated: Document, XmlHttpRequest

impl XmlHttpRequest

pub const UNSENT: u16

impl XmlHttpRequest

pub const OPENED: u16

impl XmlHttpRequest

pub const HEADERS_RECEIVED: u16

impl XmlHttpRequest

pub const LOADING: u16

impl XmlHttpRequest

pub const DONE: u16